% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{all_equal}
\alias{all_equal}
\title{Wrapper for all.equal}
\usage{
all_equal(
  target,
  current,
  tolerance = 1e-10,
  check.names = FALSE,
  check.attributes = FALSE
)
}
\arguments{
\item{target}{\R object.}

\item{current}{other \R object, to be compared with \code{target}.}

\item{tolerance}{numeric >= 0. Differences smaller than tolerance are not reported.
The default value is close to 1e-10.}

\item{check.names}{logical indicating if the \code{\link[base]{names}(.)}
    of \code{target} and \code{current} should be compared.}

\item{check.attributes}{logical indicating if the
    \code{\link[base]{attributes}} of \code{target} and \code{current}
    (other than the names) should be compared.}
}
\value{
\code{TRUE} is \code{target} and \code{current} do not differ
up to desired level of accuracy, \code{FALSE} otherwise.
}
\description{
It ensures a default tolerance level equal to 1e-10,
and ignores names and the other attributes (if any) by default.
In contrast to \code{\link[base:all.equal]{base::all.equal}}, it always returns a logical type object.
}
