% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Backfill-methods.R
\name{and,Opening,Opening-method}
\alias{and,Opening,Opening-method}
\alias{&,Opening,Opening-method}
\title{Logical AND Operator for Opening Objects}
\usage{
\S4method{&}{Opening,Opening}(e1, e2)
}
\arguments{
\item{e1}{(\code{Opening})
the first opening object.}

\item{e2}{(\code{Opening})
the second opening object.}
}
\value{
An \code{\link{OpeningAll}} object combining \code{e1} and \code{e2}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Combines two \code{\link{Opening}} objects with AND logic using the \code{&} operator.
This creates an \code{\link{OpeningAll}} object.
}
\seealso{
\code{\link{OpeningAll}} for more details.
}
