% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchicalDesign-class.R
\docType{class}
\name{ArmCondition-class}
\alias{ArmCondition-class}
\alias{.ArmCondition}
\alias{ArmCondition}
\alias{.DefaultArmCondition}
\title{\code{ArmCondition}}
\usage{
.DefaultArmCondition()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{\link{ArmCondition}} is a virtual class for criteria controlling when a
\code{\link{DesignArm}} opens for enrollment in a \code{\link{HierarchicalDesign}}.
}
\note{
Typically, end users will not use the \code{.DefaultArmCondition()}
function.
}
\seealso{
\code{\link{NoArmCondition}}, \code{\link{ArmFinishedCondition}},
\code{\link{ArmMinDoseCondition}}, \code{\link{ArmConditionAll}}, \code{\link{ArmConditionAny}}.
}
