% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchicalDesign-class.R
\docType{class}
\name{ArmConditionAll-class}
\alias{ArmConditionAll-class}
\alias{.ArmConditionAll}
\alias{ArmConditionAll}
\alias{.DefaultArmConditionAll}
\title{\code{ArmConditionAll}}
\usage{
ArmConditionAll(...)

.DefaultArmConditionAll()
}
\arguments{
\item{...}{(\code{ArmCondition})\cr arm condition objects to combine with AND
logic.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{\link{ArmConditionAll}} combines multiple \code{\link{ArmCondition}} objects using AND
logic. An arm opens only if ALL conditions in the list are satisfied.
}
\section{Slots}{

\describe{
\item{\code{condition_list}}{(\code{list})\cr a list of \code{\link{ArmCondition}} objects to be
combined with AND logic.}
}}

\note{
Typically, end users will not use the \code{.DefaultArmConditionAll()}
function.
}
\seealso{
\code{\link{ArmCondition}}, \code{\link{ArmConditionAny}}, \code{\link{ArmConditionList}}.
}
