% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchicalDesign-class.R
\docType{class}
\name{ArmConditionAny-class}
\alias{ArmConditionAny-class}
\alias{.ArmConditionAny}
\alias{ArmConditionAny}
\alias{.DefaultArmConditionAny}
\title{\code{ArmConditionAny}}
\usage{
ArmConditionAny(...)

.DefaultArmConditionAny()
}
\arguments{
\item{...}{(\code{ArmCondition})\cr arm condition objects to combine with OR
logic.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{\link{ArmConditionAny}} combines multiple \code{\link{ArmCondition}} objects using OR
logic. An arm opens if ANY condition in the list is satisfied.
}
\section{Slots}{

\describe{
\item{\code{condition_list}}{(\code{list})\cr a list of \code{\link{ArmCondition}} objects to be
combined with OR logic.}
}}

\note{
Typically, end users will not use the \code{.DefaultArmConditionAny()}
function.
}
\seealso{
\code{\link{ArmCondition}}, \code{\link{ArmConditionAll}}, \code{\link{ArmConditionList}}.
}
