% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchicalDesign-class.R
\docType{class}
\name{ArmConditionList-class}
\alias{ArmConditionList-class}
\alias{.ArmConditionList}
\alias{ArmConditionList}
\alias{.DefaultArmConditionList}
\title{\code{ArmConditionList}}
\usage{
ArmConditionList(...)

.DefaultArmConditionList()
}
\arguments{
\item{...}{(\code{ArmCondition})\cr arm condition objects to combine.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{\link{ArmConditionList}} is a virtual class for combining multiple
\code{\link{ArmCondition}} objects using logical operators. It is used as a base class
for \code{\link{ArmConditionAll}} and \code{\link{ArmConditionAny}}.
}
\section{Slots}{

\describe{
\item{\code{condition_list}}{(\code{list})\cr a list of \code{\link{ArmCondition}} objects to be
combined.}
}}

\note{
Typically, end users will not use the \code{.DefaultArmConditionList()}
function.
}
\seealso{
\code{\link{ArmCondition}}, \code{\link{ArmConditionAll}}, \code{\link{ArmConditionAny}}.
}
