% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchicalDesign-class.R
\docType{class}
\name{ArmFinishedCondition-class}
\alias{ArmFinishedCondition-class}
\alias{.ArmFinishedCondition}
\alias{ArmFinishedCondition}
\alias{.DefaultArmFinishedCondition}
\title{\code{ArmFinishedCondition}}
\usage{
ArmFinishedCondition(arm_name)

.DefaultArmFinishedCondition()
}
\arguments{
\item{arm_name}{(\code{string})\cr see slot definition.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{\link{ArmFinishedCondition}} opens an arm when the named arm has finished dose
escalation.
}
\section{Slots}{

\describe{
\item{\code{arm_name}}{(\code{string})\cr the name of the arm that must have finished.}
}}

\note{
Typically, end users will not use the
\code{.DefaultArmFinishedCondition()} function.
}
