% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchicalDesign-class.R
\docType{class}
\name{ArmMinDoseCondition-class}
\alias{ArmMinDoseCondition-class}
\alias{.ArmMinDoseCondition}
\alias{ArmMinDoseCondition}
\alias{.DefaultArmMinDoseCondition}
\title{\code{ArmMinDoseCondition}}
\usage{
ArmMinDoseCondition(arm_name, min_dose)

.DefaultArmMinDoseCondition()
}
\arguments{
\item{arm_name}{(\code{string})\cr see slot definition.}

\item{min_dose}{(\code{numeric})\cr see slot definition.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{\link{ArmMinDoseCondition}} opens an arm when the named arm has enrolled at
least one patient at the specified minimum dose or higher.
}
\section{Slots}{

\describe{
\item{\code{arm_name}}{(\code{string})\cr the name of the arm whose data are checked.}

\item{\code{min_dose}}{(\code{numeric})\cr the minimum dose that must have been reached.
For combination arms, this can be a vector of minimum doses for each agent.}
}}

\note{
Typically, end users will not use the \code{.DefaultArmMinDoseCondition()}
function.
}
