% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-methods.R
\name{as.list,GeneralData-method}
\alias{as.list,GeneralData-method}
\title{Coerce to List Method for \code{GeneralData} Objects}
\usage{
\S4method{as.list}{GeneralData}(x, ...)
}
\arguments{
\item{x}{(\code{GeneralData})\cr the object we want to convert into list.}

\item{\dots}{unused.}
}
\value{
A \code{list} with all slots in object \code{x}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

A method that coerces \code{\link{GeneralData}} object into list.
}
\examples{
# Create some data of class 'Data'
my_data <- Data(
  x = c(0.1, 0.5, 1.5, 3, 6, 10, 10, 10),
  y = c(0, 0, 0, 0, 0, 0, 1, 0),
  doseGrid = c(0.1, 0.5, 1.5, 3, 6, seq(from = 10, to = 80, by = 2))
)

# Convert Data object into list.
as.list(my_data)
}
