% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate.R
\name{assert_equal}
\alias{assert_equal}
\title{Assert That All Arguments Are Equal}
\usage{
assert_equal(
  ...,
  tol = sqrt(.Machine$double.eps),
  .var.name = vname(x),
  add = NULL
)
}
\arguments{
\item{...}{(\code{numeric})\cr vectors to be compared}

\item{tol}{(\code{numeric})\cr the maximum difference to be tolerated when
judging equality}

\item{.var.name}{[\code{character(1)}]\cr
Name of the checked object to print in assertions. Defaults to
the heuristic implemented in \code{\link[checkmate]{vname}}.}

\item{add}{[\code{AssertCollection}]\cr
Collection to store assertion messages. See \code{\link[checkmate]{AssertCollection}}.}
}
\value{
\code{list(...)}, invisibly.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Elements of \code{...} must be numeric vectors or scalars.

This function performs an element-by-element comparison of the first object
provided in \code{...} with every other object in \code{...} and throws an error if they
are not.
}
\note{
If there are any missing or infinite values in \code{...}, this function
throws an error, regardless of the values of other elements in \code{...}.

If elements in \code{...} are not all of the same length, an error is thrown.
}
\examples{
assert_equal(1:2, 1:2) # no error
assert_equal(0.01, 0.02, tol = 0.05) # no error
}
\seealso{
\code{\link{assertions}} for more details.
}
