% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate.R
\name{assertions}
\alias{assertions}
\title{Additional Assertions for \code{checkmate}}
\value{
Depending on the function prefix.
\itemize{
\item \code{assert_} functions return the object invisibly if successful, and otherwise
throw an error message.
\item \code{check_} functions return \code{TRUE} if successful, otherwise a string with the
error message.
\item \code{test_} functions just return \code{TRUE} or \code{FALSE}.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

We provide additional assertion functions that can be used together with
the \code{checkmate} functions. These are described in individual help pages
linked below.
}
\seealso{
\code{\link[=assert_probabilities]{assert_probabilities()}}, \code{\link[=assert_probability]{assert_probability()}},
\code{\link[=assert_probability_range]{assert_probability_range()}}, \code{\link[=assert_common_length]{assert_common_length()}}.
}
