% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_simulations.R
\name{barplot_percentages}
\alias{barplot_percentages}
\title{Convenience function to make barplots of percentages}
\usage{
barplot_percentages(x, description, xaxisround = 0)
}
\arguments{
\item{x}{vector of samples}

\item{description}{xlab string}

\item{xaxisround}{rounding for xaxis labels (default: 0, i.e. integers will
be used)}
}
\value{
the ggplot2 object
}
\description{
Convenience function to make barplots of percentages
}
\examples{
crmPack:::barplot_percentages(
  rpois(n=100, lambda=2),
  "test"
)
}
\keyword{internal}
