% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Model-methods.R
\docType{methods}
\name{biomLevel}
\alias{biomLevel}
\alias{biomLevel,numeric,DualEndpoint,Samples-method}
\title{Compute the biomarker level for a given dose, given model and samples}
\usage{
biomLevel(dose, model, samples, ...)

\S4method{biomLevel}{numeric,DualEndpoint,Samples}(dose, model, samples, xLevel,
  ...)
}
\arguments{
\item{dose}{the dose}

\item{model}{the \code{\linkS4class{DualEndpoint}} object}

\item{samples}{the \code{\linkS4class{Samples}} object}

\item{xLevel}{the grid index of \code{dose}}

\item{\dots}{unused}
}
\description{
Compute the biomarker level for a given dose, given model and samples
}
\section{Methods (by class)}{
\itemize{
\item \code{dose = numeric,model = DualEndpoint,samples = Samples}: Here it is very easy, we just return the corresponding
column (index \code{xLevel}) of the biomarker samples matrix, since we save
that in the samples
}}
\keyword{methods}

