% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate.R
\name{check_common_length}
\alias{check_common_length}
\alias{assert_common_length}
\alias{test_common_length}
\alias{expect_common_length}
\title{Check if vectors are of compatible lengths}
\usage{
check_common_length(x, y, y_len)

assert_common_length(x, y, y_len, .var.name = checkmate::vname(x), add = NULL)

test_common_length(x, y, y_len)

expect_common_length(x, y, y_len, info = NULL, label = vname(x))
}
\arguments{
\item{x}{(\code{any})\cr the first vector, any object for which \code{\link[=length]{length()}}
function is defined.}

\item{y}{(\code{any})\cr the second vector, any object for which \code{\link[=length]{length()}}
function is defined.}

\item{y_len}{(\code{count})\cr length of the second vector. It must be specified
if and only if \code{y} is missing.}

\item{.var.name}{[\code{character(1)}]\cr
Name of the checked object to print in assertions. Defaults to
the heuristic implemented in \code{\link[checkmate]{vname}}.}

\item{add}{[\code{AssertCollection}]\cr
Collection to store assertion messages. See \code{\link[checkmate]{AssertCollection}}.}

\item{info}{[\code{character(1)}]\cr
Extra information to be included in the message for the testthat reporter.
See \code{\link[testthat]{expect_that}}.}

\item{label}{[\code{character(1)}]\cr
Name of the checked object to print in messages. Defaults to
the heuristic implemented in \code{\link[checkmate]{vname}}.}
}
\value{
\code{TRUE} if successful, otherwise a string with the error message.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Two vectors are of compatible size if and only if: \cr
\enumerate{
\item At least one vector has size 1 \cr
\item or both vectors are of the same size. \cr
}
}
\examples{
check_common_length(1:5, 1)
check_common_length(1:5, 5:6)
check_common_length(1:5, 11:15)
check_common_length(10, 1)
check_common_length(10, 1:6)
check_common_length(10, y_len = 14)
}
\seealso{
\code{\link{assertions}} for more details.
}
