% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{.DefaultCohortSize}
\alias{.DefaultCohortSize}
\alias{CohortSize-class}
\alias{CohortSize}
\title{\code{CohortSize}}
\usage{
.DefaultCohortSize()

.DefaultCohortSize()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{\link{CohortSize}} is a class for cohort sizes.
}
\note{
Typically, end users will not use the \code{DefaultCohortSize()} function.

Typically, end users will not use the \code{DefaultCohortSize()} function.
}
\seealso{
\code{\link{CohortSizeRange}}, \code{\link{CohortSizeDLT}}, \code{\link{CohortSizeConst}},
\code{\link{CohortSizeParts}}, \code{\link{CohortSizeMin}}, \code{\link{CohortSizeMin}}.
}
