% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{CohortSizeConst-class}
\alias{.CohortSizeConst}
\alias{CohortSizeConst-class}
\title{Constant cohort size}
\description{
This class is used when the cohort size should be kept constant.
}
\section{Slots}{

\describe{
\item{\code{size}}{the constant integer size}
}}
\keyword{classes}

