% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\name{CohortSizeConst}
\alias{CohortSizeConst}
\title{Initialization function for "CohortSizeConst"}
\usage{
CohortSizeConst(size)
}
\arguments{
\item{size}{see \code{\linkS4class{CohortSizeConst}}}
}
\value{
the \code{\linkS4class{CohortSizeConst}} object
}
\description{
Initialization function for "CohortSizeConst"
}
\keyword{methods}
