\docType{class}
\name{CohortSizeDLT-class}
\alias{CohortSizeDLT-class}
\title{Cohort size based on number of DLTs}
\description{
Cohort size based on number of DLTs
}
\section{Slots}{

\describe{
\item{\code{DLTintervals}}{a vector with the bounds of the relevant DLT intervals of
length \code{n}}

\item{\code{cohortSize}}{an integer vector of length \code{n-1} with the cohort
sizes in the \code{DLTintervals}}
}}
\keyword{classes}

