% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\name{CohortSizeDLT}
\alias{CohortSizeDLT}
\title{Initialization function for "CohortSizeDLT"}
\usage{
CohortSizeDLT(DLTintervals, cohortSize)
}
\arguments{
\item{DLTintervals}{see \code{\linkS4class{CohortSizeDLT}}}

\item{cohortSize}{see \code{\linkS4class{CohortSizeDLT}}}
}
\value{
the \code{\linkS4class{CohortSizeDLT}} object
}
\description{
Initialization function for "CohortSizeDLT"
}
\keyword{methods}
