% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\name{CohortSizeMax}
\alias{CohortSizeMax}
\title{Initialization function for "CohortSizeMax"}
\usage{
CohortSizeMax(cohortSizeList)
}
\arguments{
\item{cohortSizeList}{see \code{\linkS4class{CohortSizeMax}}}
}
\value{
the \code{\linkS4class{CohortSizeMax}} object
}
\description{
Initialization function for "CohortSizeMax"
}
\keyword{methods}
