% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\name{CohortSizeMin}
\alias{CohortSizeMin}
\title{Initialization function for "CohortSizeMin"}
\usage{
CohortSizeMin(cohortSizeList)
}
\arguments{
\item{cohortSizeList}{see \code{\linkS4class{CohortSizeMin}}}
}
\value{
the \code{\linkS4class{CohortSizeMin}} object
}
\description{
Initialization function for "CohortSizeMin"
}
\keyword{methods}
