% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{CohortSizeOrdinal-class}
\alias{CohortSizeOrdinal-class}
\alias{.CohortSizeOrdinal}
\alias{CohortSizeOrdinal}
\alias{.DefaultCohortSizeOrdinal}
\title{\code{CohortSizeOrdinal}}
\usage{
CohortSizeOrdinal(grade, rule)

.DefaultCohortSizeOrdinal()
}
\arguments{
\item{grade}{(\code{integer})\cr see slot definition.}

\item{rule}{(\code{CohortSize})\cr see slot definition.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{\link{CohortSizeOrdinal}} is the class for cohort size for an ordinal CRM trial.
}
\section{Slots}{

\describe{
\item{\code{grade}}{(\code{integer})\cr the grade at which the rule should be applied}

\item{\code{rule}}{(\code{CohortSize})\cr the \code{CohortSize} rule to apply.}
}}

\note{
Typically, end users will not use the \code{.DefaultCohortSizeOrdinal()} function.
}
\examples{
CohortSizeOrdinal(
  grade = 1L,
  rule = CohortSizeRange(intervals = c(0, 30), cohort_size = c(1L, 3L))
)
}
