% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\name{CohortSizeParts}
\alias{CohortSizeParts}
\title{Initialization function for "CohortSizeParts"}
\usage{
CohortSizeParts(sizes)
}
\arguments{
\item{sizes}{see \code{\linkS4class{CohortSizeParts}}}
}
\value{
the \code{\linkS4class{CohortSizeParts}} object
}
\description{
Initialization function for "CohortSizeParts"
}
\keyword{methods}
