% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{CohortSizeRandom-class}
\alias{CohortSizeRandom-class}
\alias{.CohortSizeRandom}
\alias{CohortSizeRandom}
\alias{.DefaultCohortSizeRandom}
\title{\code{CohortSizeRandom}}
\usage{
CohortSizeRandom(min_size, max_size)

.DefaultCohortSizeRandom()
}
\arguments{
\item{min_size}{(\code{integer})\cr see slot definition.}

\item{max_size}{(\code{integer})\cr see slot definition.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{\link{CohortSizeRandom}} is the class for random cohort sizes drawn from a
uniform distribution between \code{min_size} and \code{max_size} (inclusive).
}
\section{Slots}{

\describe{
\item{\code{min_size}}{(\code{integer})\cr minimum cohort size.}

\item{\code{max_size}}{(\code{integer})\cr maximum cohort size.}
}}

\note{
Typically, end users will not use the \code{.DefaultCohortSizeRandom()} function.
}
\examples{
# Random cohort size between 1 and 5.
my_size <- CohortSizeRandom(min_size = 1, max_size = 5)
}
