\docType{class}
\name{CohortSizeRange-class}
\alias{CohortSizeRange-class}
\title{Cohort size based on dose range}
\description{
Cohort size based on dose range
}
\section{Slots}{

\describe{
\item{\code{intervals}}{a vector with the bounds of the relevant dose intervals of
length \code{n}}

\item{\code{cohortSize}}{an integer vector of length \code{n-1} with the cohort
sizes in the \code{intervals}}
}}
\keyword{classes}

