% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\name{CohortSizeRange}
\alias{CohortSizeRange}
\title{Initialization function for "CohortSizeRange"}
\usage{
CohortSizeRange(intervals, cohortSize)
}
\arguments{
\item{intervals}{see \code{\linkS4class{CohortSizeRange}}}

\item{cohortSize}{see \code{\linkS4class{CohortSizeRange}}}
}
\value{
the \code{\linkS4class{CohortSizeRange}} object
}
\description{
Initialization function for "CohortSizeRange"
}
\keyword{methods}
