% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Model-class.R
\docType{class}
\name{ComboLogistic-class}
\alias{.ComboLogistic}
\alias{ComboLogistic-class}
\title{Combo model with logistic regression}
\description{
Currently, this model is for double combination dose escalation trials.
}
\details{
todo: Later, it will be extended to higher-order combinations. The model and
code is building on the work by Simon Wandel et al (Novartis).

The regression model is defined as follows. Let \eqn{odds(p)=p/(1-p)} be the
odds transformation of the probability \eqn{p}, such that \eqn{logit(p) =
log(odds(p))}. Let \eqn{x_i} be the dose of drug i=1,2, and \eqn{p(x_1, x_2)}
be the probability of DLT with doses \eqn{x_1} and \eqn{x_2}. The reference
doses for the two compounds are again denoted by stars. Then the model
assumes:
\deqn{odds(p(x_1, x_2)) = odds(p_0(x_1, x_2)) \cdot \exp(\eta x_1/x_1^{*}
x_2/x_2^{*}),}
where \eqn{\eta} is the interaction coefficient (positive values correspond
to synergistic toxicity, zero corresponds to additive effect without
interaction, and negative values correspond to antagonistic toxicity). A
normal prior
\deqn{\eta \sim Normal(\gamma, \tau^-1)}
is used.
Under no interaction with \eqn{\eta=0}, this reduces the probability
\eqn{\eta{odds(p(x_1, x_2))}} to
\deqn{p_0(x_1, x_2) = p(x_1) + p(x_2) - p(x_1)p(x_2) = 1 - (1 - p(x_1))(1 -
p(x_2)).}
Now for the single-agent DLT probabilities \eqn{p(x_1)} and \eqn{p(x_2)} we
assume the logistic log-normal models (compare
\code{\linkS4class{LogisticLogNormal}}):
\deqn{logit[p(x_i)] = \alpha_i + \beta_i \cdot \log(x_i/x_i^{*}),}
with prior
\deqn{(\alpha_i, \log(\beta_i)) \sim Normal(\mu_i, \Sigma_i)}
for \eqn{i=1,2}. todo: Note that in principle any model could be used for the
individual agents. For now we stick to this simple, fixed implementation.

The \code{prob} function has as first argument \code{doses}, which is a dose
combination: a vector with names specifying the drug names. Additional
arguments are the model parameters \code{alpha0} (intercepts, nSamples x
nDrugs matrix), \code{alpha1} (slopes, nSamples x nDrugs matrix) and
\code{eta} (vector of length nSamples), containing the nSamples MCMC
samples. Then \code{prob} computes the resulting samples of the probability
of toxicity at that dose combination. Again here, the function must
vectorize over the model parameters.
}
\section{Slots}{

\describe{
\item{\code{singlePriors}}{a list with one \code{\linkS4class{LogisticLogNormal}}
model per drug specifying the bivariate log normal prior described above.
The names of this list are the drug names for this model - they have to be
specified correctly in any interactions with the model object (e.g. when
calling the \code{prob} function).}

\item{\code{gamma}}{the mean for the interaction parameter}

\item{\code{tau}}{the precision for the interaction parameter}

\item{\code{prob}}{function calculating the probability of toxicity for a specific
dose combination, based on the model parameters (see the details above)}
}}
\keyword{classes}

