% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Model-class.R
\name{ComboLogistic}
\alias{ComboLogistic}
\title{Initialization function for the "ComboLogistic" class}
\usage{
ComboLogistic(singlePriors, gamma, tau)
}
\arguments{
\item{singlePriors}{a named list where each element is a
\code{\linkS4class{LogisticLogNormal}} object, specifying the prior for this
drug. The list names are the drug names.}

\item{gamma}{see \code{\linkS4class{ComboLogistic}}}

\item{tau}{see \code{\linkS4class{ComboLogistic}}}
}
\value{
the \code{\linkS4class{ComboLogistic}} object
}
\description{
Initialization function for the "ComboLogistic" class
}
\keyword{methods}

