% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations-class.R
\docType{class}
\name{ComboSimulationsSummary-class}
\alias{ComboSimulationsSummary-class}
\alias{.ComboSimulationsSummary}
\alias{ComboSimulationsSummary}
\alias{.DefaultComboSimulationsSummary}
\title{\code{ComboSimulationsSummary}}
\usage{
.DefaultComboSimulationsSummary()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This class captures summary output from \code{\link{ComboSimulations}} objects.
}
\section{Slots}{

\describe{
\item{\code{target}}{(\code{numeric})
target toxicity interval. Empty if no truth function was supplied.}

\item{\code{nsim}}{(\code{integer})
number of simulations.}

\item{\code{n_obs}}{(\code{integer})
number of patients in each simulation.}

\item{\code{prop_dlts}}{(\code{numeric})
observed proportion of DLTs in each simulation.}

\item{\code{mean_tox_risk}}{(\code{numeric})
average fitted toxicity risk in each simulation.}

\item{\code{dose_selected}}{(\code{matrix})
selected dose combinations; one row per simulation.}

\item{\code{tox_at_doses_selected}}{(\code{numeric})
true toxicity at selected dose combinations (if truth supplied).}

\item{\code{prop_at_target}}{(\code{numeric})
proportion of selected combinations within target interval (if truth supplied).}

\item{\code{dose_most_selected}}{(\code{numeric})
most frequently selected dose combination.}

\item{\code{obs_tox_rate_at_dose_most_selected}}{(\code{numeric})
observed toxicity rate at the most frequently selected combination.}

\item{\code{dose_grid}}{(\code{list})
dose grid for each drug.}

\item{\code{stop_report}}{(\code{matrix})
matrix of stopping rule outcomes.}

\item{\code{stop_reasons}}{(\code{list})
stopping reasons by simulation.}

\item{\code{additional_stats}}{(\code{list})
additional statistics.}
}}

\note{
Typically, end users will not use the \code{.DefaultComboSimulationsSummary()} function.
}
