% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crmPack-package.R
\docType{package}
\name{crmPack}
\alias{crmPack-package}
\alias{crmPack}
\title{Object-oriented implementation of CRM designs}
\description{
Object-oriented implementation of CRM designs
}
\references{
\insertRef{SabanesBoveYeungPalermoJaki2019}{crmPack}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/openpharma/crmPack}
  \item \url{https://openpharma.github.io/crmPack/}
  \item Report bugs at \url{https://github.com/openpharma/crmPack/issues}
}

}
\author{
\strong{Maintainer}: Daniel Sabanes Bove \email{daniel.sabanes_bove@rconis.com}

Authors:
\itemize{
  \item Wai Yin Yeung \email{winnie.yeung@roche.com}
  \item Burak Kuersad Guenhan \email{burakgunhan@gmail.com}
  \item Giuseppe Palermo \email{giuseppe.palermo@roche.com}
  \item Thomas Jaki \email{jaki.thomas@gmail.com}
  \item Jiawen Zhu \email{zhu.jiawen@gene.com}
  \item Ziwei Liao \email{ziwei.liao.fdu@gmail.com}
  \item Dimitris Kontos \email{dimitris.kontos@bayer.com}
  \item Marlene Schulte-Goebel \email{marlene.schulte-goebel@merckgroup.com}
  \item Doug Kelkhoff \email{doug.kelkhoff@gmail.com} (\href{https://orcid.org/0009-0003-7845-4061}{ORCID})
  \item Oliver Boix \email{oliver.boix@bayer.com}
  \item Robert Adams \email{robert.adams@bayer.com}
  \item Clara Beck \email{clara.beck@bayer.com}
  \item John Kirkpatrick \email{john@puzzledface.net}
}

Other contributors:
\itemize{
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
  \item Merck Healthcare KGaA [copyright holder, funder]
  \item Bayer AG [copyright holder, funder]
}

}
\keyword{package}
