% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrmPackClass-class.R
\docType{class}
\name{CrmPackClass-class}
\alias{CrmPackClass-class}
\alias{.CrmPackClass}
\alias{CrmPackClass}
\title{\code{CrmPackClass}}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{\link{CrmPackClass}} is a virtual class, from which all other \code{crmPack} classes
inherit.
}
