\name{crmPackExample}
\alias{crmPackExample}
\title{Open the example pdf for crmPack}
\usage{
crmPackExample()
}
\value{
nothing
}
\description{
Calling this helper function should open the example.pdf
document, residing in the doc subfolder of the package
installation directory.
}
\author{
Daniel Sabanes Bove \email{sabanesd@roche.com}
}
\keyword{documentation}

