% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/crmUpgrade.R
\name{crmPackUpgrade}
\alias{crmPackUpgrade}
\title{Upgrade your crmPack installation with the latest version}
\usage{
crmPackUpgrade(lib = NULL, devel = FALSE, force = FALSE,
  repos = structure(c(CRAN = "http://stat.ethz.ch/CRAN/")))
}
\arguments{
\item{lib}{library where to install the new version (and other required
packages) into. Default: same location as the last crmPack version.}

\item{devel}{Should the development version from Stash be installed?
(default FALSE) Otherwise and by default the latest version on the internal
package repository GRAN is installed.}

\item{force}{Should the installation be forced, i.e., even if you have
already the latest version, should the package be re-installed?}

\item{repos}{which repository to use for installing required packages
(default: ETHZ in Switzerland)}
}
\value{
nothing
}
\description{
Executing this function upgrades your crmPack installation with the newest
version available on the server. You will need connection to the Roche
network (i.e. RANGE connection when you are working off-site) in order to
successfully run it.
}
\details{
After installation, the new features in this version will be shown by
printing the relevant parts of the NEWS file.
}
\author{
Daniel Sabanes Bove \email{sabanesd@roche.com}
}
\keyword{programming}

