% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design-class.R
\name{DADesign}
\alias{DADesign}
\title{Initialization function for \code{DADesign}}
\usage{
DADesign(model, data, safetyWindow, ...)
}
\arguments{
\item{model}{see \code{\linkS4class{DADesign}}}

\item{data}{see \code{\linkS4class{DADesign}}}

\item{safetyWindow}{see \code{\linkS4class{DADesign}}}

\item{\dots}{additional arguments for \code{\link{Design}}}
}
\value{
the \code{\linkS4class{DADesign}} object
}
\description{
Initialization function for \code{DADesign}
}
\keyword{methods}
