% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-class.R
\docType{class}
\name{DALogisticLogNormal-class}
\alias{DALogisticLogNormal-class}
\alias{.DALogisticLogNormal}
\title{Logistic model with bivariate (log) normal prior and data augmentation}
\description{
This is a modified data augmented CRM with logistic regression model using
a bivariate normal prior on the intercept and log slope parameters.
This class inherits from the normal logistic model class.
}
\details{
todo: need to include here formula for the lambda prior.
}
\section{Slots}{

\describe{
\item{\code{npiece}}{the number of pieces in the PEM}

\item{\code{l}}{a vector used in the lambda prior}

\item{\code{C_par}}{a parameter used in the lambda prior,
according to Liu's paper, C_par=2 is recommended
todo: cite "Liu's paper" here (should be dedicated Roxygen tag for that)}

\item{\code{conditionalPEM}}{is a conditional piecewise-exponential model used?
(default) Otherwise an unconditional model is used.}
}}

\examples{
npiece <- 10
Tmax <- 60

lambda_prior <- function(k)
{
  npiece / (Tmax * (npiece - k + 0.5))
}

## todo_with_reply: need to explain why "l" is defined like that.
## reply: using the same assumption in Liu, Yin, Yuan 2013, it is assumed that a priori toxicity occurs 
## uniformly throughout the assessment period (0, Tmax), which represents a neutral prior opinion
## between early-onset and late-onset toxicity. Under this assumption, the hazard at the middle of 
## the kth partition is caluclated as the lambda_prior function;

model <- DALogisticLogNormal(mean=c(-0.85,1),
                           cov=matrix(c(1, -0.5, -0.5, 1),
                                      nrow=2),
                           refDose=56,
                           npiece=npiece,
                           l=as.numeric(t(apply(as.matrix(c(1:npiece), 1, npiece),
                                                2,
                                                lambda_prior))),
                           C_par=2)
}
\keyword{classes}
