% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-class.R
\name{DALogisticLogNormal}
\alias{DALogisticLogNormal}
\title{Initialization function for the "DALogisticLogNormal" class}
\usage{
DALogisticLogNormal(npiece = 3, l, C_par = 2, conditionalPEM = TRUE, ...)
}
\arguments{
\item{npiece}{see \code{\linkS4class{DALogisticLogNormal}}}

\item{l}{see \code{\linkS4class{DALogisticLogNormal}}}

\item{C_par}{see \code{\linkS4class{DALogisticLogNormal}}}

\item{conditionalPEM}{see \code{\linkS4class{DALogisticLogNormal}}}

\item{\dots}{additional parameters from \code{\link{LogisticLogNormal}}}
}
\value{
the \code{\linkS4class{DALogisticLogNormal}} object
}
\description{
Initialization function for the "DALogisticLogNormal" class
}
\keyword{programming}
