% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations-class.R
\name{DASimulations}
\alias{DASimulations}
\title{Initialization function for \code{DASimulations}}
\usage{
DASimulations(trialduration, ...)
}
\arguments{
\item{\dots}{additional parameters from \code{\link{Simulations}}}

\item{trialdurationt}{see \code{\linkS4class{DASimulations}}}
}
\value{
the \code{\linkS4class{DASimulations}} object
}
\description{
Initialization function for \code{DASimulations}
}
\keyword{methods}
