% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-class.R
\docType{class}
\name{Data-class}
\alias{Data-class}
\alias{.Data}
\alias{Data}
\title{\code{Data}}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{\link{Data}} is a class for the data input.
It inherits from \code{\link{GeneralData}}.
}
\section{Slots}{

\describe{
\item{\code{x}}{(\code{numeric})\cr the doses for the patients.}

\item{\code{y}}{(\code{integer})\cr the vector of toxicity events (0 or 1 integers).}

\item{\code{doseGrid}}{(\code{numeric})\cr the vector of all possible doses (sorted),
i.e. the dose grid.}

\item{\code{nGrid}}{(\code{integer})\cr number of gridpoints.}

\item{\code{xLevel}}{(\code{integer})\cr the levels for the doses the patients have been given,
w.r.t \code{doseGrid}.}

\item{\code{placebo}}{(\code{logical})\cr if \code{TRUE} the first dose level
in the \code{doseGrid}is considered as PLACEBO.}
}}

