% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-class.R
\name{Data}
\alias{Data}
\title{\code{Data}}
\usage{
Data(
  x = numeric(),
  y = integer(),
  ID = integer(),
  cohort = integer(),
  doseGrid = numeric(),
  placebo = FALSE,
  ...
)
}
\arguments{
\item{x}{(\code{numeric})\cr the doses for the patients.}

\item{y}{(\code{integer})\cr the vector of toxicity events (0 or 1 integers).
You can also supply numeric vectors, but these will then be converted to integers.}

\item{ID}{(\code{integer})\cr unique patient IDs (integer vector).}

\item{cohort}{(\code{integer})\cr the cohort indices (sorted values from \{0, 1, 2, ...\}).}

\item{doseGrid}{(\code{numeric})\cr the vector of all possible doses.}

\item{placebo}{(\code{logical})\cr if \code{TRUE} the first dose level
in the \code{doseGrid} is considered as PLACEBO.}

\item{\dots}{not used.}
}
\value{
the initialized \code{\linkS4class{Data}} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This is the function for initializing a \code{\linkS4class{Data}} object.

Note that \code{ID} and \code{cohort} can be missing, then a warning
will be issued and the variables will be filled with default
IDs and best guesses, respectively.
}
