% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-class.R
\docType{class}
\name{DataCombo-class}
\alias{DataCombo-class}
\alias{.DataCombo}
\alias{DataCombo}
\alias{.DefaultDataCombo}
\title{\code{DataCombo}}
\usage{
DataCombo(
  x = matrix(numeric(), ncol = 2L),
  y = integer(),
  ID = integer(),
  cohort = integer(),
  doseGrid = list(drug1 = numeric(), drug2 = numeric()),
  drugNames = names(doseGrid),
  backfilled = rep(FALSE, nrow(x)),
  response = rep(NA_integer_, nrow(x)),
  ...
)

.DefaultDataCombo()
}
\arguments{
\item{x}{(\code{matrix})
numeric matrix with 2 columns giving the doses of the two drugs.}

\item{y}{(\code{integer})
the vector of toxicity events (0 or 1).}

\item{ID}{(\code{integer})
unique patient IDs.}

\item{cohort}{(\code{integer})
the cohort (non-negative sorted) indices.}

\item{doseGrid}{(\code{list})
named list of length 2 with all possible doses for each drug.}

\item{drugNames}{(\code{character})
names of the two drugs.}

\item{backfilled}{(\code{logical})
whether each patient was in a backfill cohort.}

\item{response}{(\code{integer})
whether each patient had a positive efficacy response (1 = yes, 0 = no).
May contain \code{NA}.}

\item{...}{not used.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{\link{DataCombo}} is a class for two-drug combination toxicity data.
It inherits from \code{\link{GeneralData}} and stores patient-level dose combinations.
}
\section{Slots}{

\describe{
\item{\code{x}}{(\code{matrix})
numeric matrix with 2 columns giving the doses of the two drugs.}

\item{\code{y}}{(\code{integer})
the vector of toxicity events (0 or 1 integers).}

\item{\code{doseGrid}}{(\code{list})
named list of length 2 with the possible doses for each drug.}

\item{\code{nGrid}}{(\code{integer})
number of grid points for each drug.}

\item{\code{xLevel}}{(\code{matrix})
integer matrix with 2 columns giving the dose levels of \code{x} with respect to
\code{doseGrid}.}

\item{\code{drugNames}}{(\code{character})
names of the two drugs.}

\item{\code{backfilled}}{(\code{logical})
whether this patient was in a backfill cohort.}

\item{\code{response}}{(\code{integer})
whether this patient had a positive efficacy response (0/1/\code{NA}).}
}}

\note{
\code{ID} and \code{cohort} can be missing. Then a message will be issued
and the variables will be filled with default IDs and best guess cohort
indices.

Typically, end users will not use the \code{.DefaultDataCombo()} function.
}
\examples{
## Example for DataCombo ----

DataCombo(
  x = cbind(
    drug1 = c(10, 10, 10, 20, 20, 20),
    drug2 = c(20, 20, 20, 20, 20, 20)
  ),
  y = c(0L, 0L, 1L, 0L, 0L, 0L),
  ID = 1L:6L,
  cohort = c(1L, 1L, 1L, 2L, 2L, 2L),
  doseGrid = list(
    drug1 = c(10, 20, 30),
    drug2 = c(20, 40)
  )
)
}
