% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Data-class.R
\name{DataCombo}
\alias{DataCombo}
\title{Initialization function for the "DataCombo" class}
\usage{
DataCombo(x, y, ID, cohort, doseGrid)
}
\arguments{
\item{x}{the matrix with the doses for the patients. Recommendation: Use
\code{cbind(drugA=..., drugB=...)} to create this matrix.}

\item{y}{the vector of toxicity events (0 or 1 integers). You can also
normal numeric vectors, but these will then be converted to integers.}

\item{ID}{unique patient IDs (integer vector)}

\item{cohort}{the cohort indices (sorted values from 0, 1, 2, ...)}

\item{doseGrid}{the list with vectors of all possible doses for each of the
drugs}
}
\value{
the initialized \code{\linkS4class{DataCombo}} object
}
\description{
This is the function for initializing a "DataCombo" class object.
}
\details{
Note that \code{ID} and \code{cohort} can be missing, then a warning
will be issued and the variables will be filled with default
IDs and best guesses, respectively.
}
\keyword{programming}

