% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-class.R
\name{DataDA}
\alias{DataDA}
\title{Initialization function for the "DataDA" class}
\usage{
DataDA(u = numeric(), Tmax = numeric(), t0 = numeric(), ...)
}
\arguments{
\item{u}{the continuous vector of DLT free survival values}

\item{Tmax}{the DLT observation period}

\item{t0}{time of initial dosing for each patient (default: 0)}

\item{\dots}{additional parameters from \code{\link{Data}}}
}
\value{
the initialized \code{\linkS4class{DataDA}} object
}
\description{
This is the function for initializing a "DataDA" class object.
}
\keyword{programming}
