% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-class.R
\name{DataDual}
\alias{DataDual}
\title{Initialization function for the "DataDual" class}
\usage{
DataDual(w = numeric(), ...)
}
\arguments{
\item{w}{the continuous vector of biomarker values}

\item{\dots}{additional parameters from \code{\link{Data}}}
}
\value{
the initialized \code{\linkS4class{DataDual}} object
}
\description{
This is the function for initializing a "DataDual" class object.
}
\keyword{programming}
