% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-class.R
\name{DataMixture}
\alias{DataMixture}
\title{Initialization function for the "DataMixture" class}
\usage{
DataMixture(xshare = numeric(), yshare = integer(), ...)
}
\arguments{
\item{xshare}{see \code{\linkS4class{DataMixture}}}

\item{yshare}{see \code{\linkS4class{DataMixture}}}

\item{\dots}{additional arguments for the underlying Data slots}
}
\value{
the initialized \code{\linkS4class{DataMixture}} object
}
\description{
This is the function for initializing a "DataMixture" class object.
}
\keyword{programming}
