% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-class.R
\name{DataParts}
\alias{DataParts}
\title{Initialization function for the "DataParts" class}
\usage{
DataParts(part = integer(), nextPart = 1L, part1Ladder = numeric(),
  ...)
}
\arguments{
\item{part}{which part does each of the patients belong to?}

\item{nextPart}{what is the part for the next cohort? (1 or 2)}

\item{part1Ladder}{what is the escalation ladder for
part 1?}

\item{\dots}{additional parameters from \code{\link{Data}}}
}
\value{
the initialized \code{\linkS4class{DataParts}} object
}
\description{
This is the function for initializing a \code{\linkS4class{DataParts}}
object.
}
\keyword{programming}
