% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Design-class.R
\docType{class}
\name{Design-class}
\alias{.Design}
\alias{Design-class}
\title{Class for the CRM design}
\description{
In addition to the slots in the more simple \code{\linkS4class{RuleDesign}},
objects of this class contain:
}
\section{Slots}{

\describe{
\item{\code{model}}{the model to be used, an object of class
\code{\linkS4class{Model}}}

\item{\code{stopping}}{stopping rule(s) for the trial, an object of class
\code{\linkS4class{Stopping}}}

\item{\code{increments}}{how to control increments between dose levels,
an object of class \code{\linkS4class{Increments}}}
}}
\keyword{classes}

