% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchicalDesign-class.R
\docType{class}
\name{DesignArm-class}
\alias{DesignArm-class}
\alias{.DesignArm}
\alias{DesignArm}
\alias{HistoricalArm}
\alias{.DefaultDesignArm}
\title{\code{DesignArm}}
\usage{
DesignArm(
  name,
  design,
  borrow = TRUE,
  open_when = NoArmCondition()
)

HistoricalArm(
  name,
  data,
  model,
  borrow = TRUE
)

.DefaultDesignArm()
}
\arguments{
\item{name}{(\code{string})\cr see slot definition.}

\item{design}{(\code{Design})\cr see slot definition.}

\item{borrow}{(\code{flag})\cr see slot definition.}

\item{open_when}{(\code{ArmCondition})\cr see slot definition.}

\item{data}{(\code{Data} or \code{DataCombo})\cr arm data for historical arms.}

\item{model}{(\code{GeneralModel} or \code{TwoDrugsCombo})\cr arm model for
historical arms.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{\link{DesignArm}} is a light-weight wrapper around \code{Design} object
for use in a \code{\link{HierarchicalDesign}}. Use \code{\link{DesignArm}()} for active arms
and \code{\link{HistoricalArm}()} for non-enrolling historical arms.
}
\section{Slots}{

\describe{
\item{\code{name}}{(\code{string})\cr the name of the arm.}

\item{\code{active}}{(\code{flag})\cr whether the arm is enrolling or not (historical).}

\item{\code{borrow}}{(\code{flag})\cr whether this arm may borrow information from other
arms when making dose escalation decisions. Only relevant if \code{active = TRUE}.}

\item{\code{open_when}}{(\code{ArmCondition})\cr the condition that must be satisfied
before this arm opens for enrollment.}

\item{\code{design}}{(\code{Design})\cr the design object for this arm.}
}}

\note{
Typically, end users will not use the \code{.DefaultDesignArm()} function.
}
\examples{
design <- .DefaultDesign()
design_arm <- DesignArm(
  name = "Arm A",
  design = design
)
}
