% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design-class.R
\docType{class}
\name{DesignCombo-class}
\alias{DesignCombo-class}
\alias{.DesignCombo}
\alias{DesignCombo}
\alias{.DefaultDesignCombo}
\title{\code{DesignCombo}}
\usage{
DesignCombo(
  model,
  nextBest,
  stopping,
  increments,
  cohort_size,
  data,
  startingDose,
  backfill = Backfill(opening = OpeningNone())
)

.DefaultDesignCombo()
}
\arguments{
\item{model}{(\code{TwoDrugsCombo})
see slot definition.}

\item{nextBest}{(\code{NextBestNCRM})
see slot definition.}

\item{stopping}{(\code{Stopping})
see slot definition.}

\item{increments}{(\code{Increments})
see slot definition.}

\item{cohort_size}{(\code{CohortSize})
see slot definition.}

\item{data}{(\code{DataCombo})
see slot definition.}

\item{startingDose}{(\code{numeric})
see slot definition.}

\item{backfill}{(\code{Backfill})
see slot definition.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{\link{DesignCombo}} is the class for two-drug combination dose-escalation
designs.
}
\details{
This class stores the same core design components as \code{\link{Design}}, but uses
\code{\link{DataCombo}} and \code{\link{TwoDrugsCombo}} to represent two-drug
combinations.
}
\section{Slots}{

\describe{
\item{\code{model}}{(\code{TwoDrugsCombo})
the model to be used.}

\item{\code{nextBest}}{(\code{NextBestNCRM})
the rule to select the next dose combination.}

\item{\code{stopping}}{(\code{Stopping})
stopping rule(s) for the trial.}

\item{\code{increments}}{(\code{Increments})
how to control increments between dose combinations.}

\item{\code{cohort_size}}{(\code{CohortSize})
rules for the cohort sizes.}

\item{\code{data}}{(\code{DataCombo})
specifies dose grids and any previously observed data.}

\item{\code{startingDose}}{(\code{numeric})
starting dose combination as a numeric vector of length 2.}

\item{\code{backfill}}{(\code{Backfill})
rules for backfilling patients in the trial.}
}}

\note{
Typically, end users will not use the \code{.DefaultDesignCombo()} function.
}
\examples{
empty_data <- DataCombo(
  doseGrid = list(
    drug1 = c(10, 20, 30),
    drug2 = c(20, 40, 60)
  )
)

my_model <- TwoDrugsCombo(
  single_models = list(
    drug1 = LogisticLogNormal(
      mean = c(-0.85, 1),
      cov = matrix(c(1, -0.5, -0.5, 1), nrow = 2),
      ref_dose = 10
    ),
    drug2 = LogisticLogNormal(
      mean = c(-0.7, 0.8),
      cov = matrix(c(1.1, -0.3, -0.3, 0.9), nrow = 2),
      ref_dose = 20
    )
  ),
  gamma = 0,
  tau = 1
)

my_next_best <- NextBestNCRM(
  target = c(0.2, 0.35),
  overdose = c(0.35, 1),
  max_overdose_prob = 0.25
)

my_increments <- IncrementsMin(
  increments_list = list(
    IncrementsComboOneDrugOnly(),
    IncrementsComboCartesian(
      drug1 = IncrementsRelative(intervals = c(0), increments = c(1)),
      drug2 = IncrementsRelative(intervals = c(0), increments = c(1))
    )
  )
)

my_stopping <- StoppingMinPatients(nPatients = 20)

design_combo <- DesignCombo(
  model = my_model,
  nextBest = my_next_best,
  stopping = my_stopping,
  increments = my_increments,
  cohort_size = CohortSizeConst(3L),
  data = empty_data,
  startingDose = c(drug1 = 10, drug2 = 20)
)
}
