% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design-class.R
\docType{class}
\name{DesignGrouped-class}
\alias{DesignGrouped-class}
\alias{.DesignGrouped}
\alias{DesignGrouped}
\alias{.DefaultDesignGrouped}
\title{\code{DesignGrouped}}
\usage{
DesignGrouped(
  model,
  mono,
  combo = mono,
  first_cohort_mono_only = TRUE,
  same_dose = TRUE,
  ...
)
}
\arguments{
\item{model}{(\code{LogisticLogNormalGrouped})\cr see slot definition.}

\item{mono}{(\code{Design})\cr see slot definition.}

\item{combo}{(\code{Design})\cr see slot definition.}

\item{first_cohort_mono_only}{(\code{flag})\cr see slot definition.}

\item{same_dose}{(\code{flag})\cr see slot definition.}

\item{...}{not used.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{\link{DesignGrouped}} combines two \code{\link{Design}} objects: one for the mono and one
for the combo arm of a joint dose escalation design.
}
\section{Slots}{

\describe{
\item{\code{model}}{(\code{LogisticLogNormalGrouped})\cr the model to be used, currently only one
class is allowed.}

\item{\code{mono}}{(\code{Design})\cr including the rules to be followed for the mono arm, the
\code{model} slot will be ignored since the joint model is used instead.}

\item{\code{combo}}{(\code{Design})\cr including the rules to be followed for the combo arm, the
\code{model} slot will be ignored since the joint model is used instead.}

\item{\code{first_cohort_mono_only}}{(\code{flag})\cr whether first test one mono agent cohort, and then
once its DLT data has been collected, we proceed from the second cohort onwards with
concurrent mono and combo cohorts.}

\item{\code{same_dose}}{(\code{flag})\cr whether the lower dose of the separately determined mono and combo
doses should be used as the next dose for both mono and combo.}
}}

\note{
Typically, end-users will not use the \code{.DefaultDesignGrouped()} function.
}
\examples{
empty_data <- Data(doseGrid = c(1, 3, 5, 10, 15, 20, 25, 40, 50, 80, 100))

# Initialize the joint model.
my_model <- LogisticLogNormalGrouped(
  mean = c(-0.85, 0, 1, 0),
  cov = diag(1, 4),
  ref_dose = 56
)

# Choose the rule for selecting the next dose.
my_next_best <- NextBestNCRM(
  target = c(0.2, 0.35),
  overdose = c(0.35, 1),
  max_overdose_prob = 0.25
)

# Choose the rule for the cohort-size.
my_size1 <- CohortSizeRange(
  intervals = c(0, 30),
  cohort_size = c(1, 3)
)
my_size2 <- CohortSizeDLT(
  intervals = c(0, 1),
  cohort_size = c(1, 3)
)
my_size <- maxSize(my_size1, my_size2)

# Choose the rule for stopping.
my_stopping1 <- StoppingMinCohorts(nCohorts = 3)
my_stopping2 <- StoppingTargetProb(
  target = c(0.2, 0.35),
  prob = 0.5
)
my_stopping3 <- StoppingMinPatients(nPatients = 20)
my_stopping <- (my_stopping1 & my_stopping2) | my_stopping3

# Choose the rule for dose increments.
my_increments <- IncrementsRelative(
  intervals = c(0, 20),
  increments = c(1, 0.33)
)

# Initialize the design.
design <- DesignGrouped(
  model = my_model,
  mono = Design(
    model = .DefaultModelLogNormal(), # Ignored.
    nextBest = my_next_best,
    stopping = my_stopping,
    increments = my_increments,
    cohort_size = my_size,
    data = empty_data,
    startingDose = 3
  )
)
}
