% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{dinvGamma}
\alias{dinvGamma}
\title{Compute the Density of Inverse Gamma Distribution}
\usage{
dinvGamma(x, a, b, log = FALSE, normalize = TRUE)
}
\arguments{
\item{x}{(\code{numeric})\cr vector of quantiles.}

\item{a}{(\code{number})\cr the shape parameter of the inverse gamma distribution.}

\item{b}{(\code{number})\cr the scale parameter of the inverse gamma distribution.}

\item{log}{(\code{flag})\cr if \code{TRUE}, probabilities \code{p} are given as \code{log(p)}.}

\item{normalize}{(\code{flag})\cr if \code{TRUE}, the output will be normalized.}
}
\value{
The density value(s).
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\keyword{internal}
