\docType{methods}
\name{dose,numeric,Model,Samples-method}
\alias{dose,numeric,Model,Samples-method}
\title{Compute the doses for a given probability, given model and samples}
\usage{
\S4method{dose}{numeric,Model,Samples}(prob, model, samples, ...)
}
\arguments{
  \item{prob}{the probability}

  \item{model}{the \code{\linkS4class{Model}}}

  \item{samples}{the \code{\linkS4class{Samples}}}

  \item{...}{unused}
}
\description{
Compute the doses for a given probability, given model and
samples
}

