% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Model-methods.R
\docType{methods}
\name{dose,numeric,ModelTox,ANY-method}
\alias{dose,numeric,ModelTox,ANY-method}
\title{Compute the dose for a given probability and a given Pseudo DLE model without samples}
\usage{
\S4method{dose}{numeric,ModelTox,ANY}(prob, model, samples, ...)
}
\arguments{
\item{prob}{the probability}

\item{model}{the \code{\linkS4class{ModelTox}}}

\item{\dots}{unused}
}
\description{
Compute the dose for a given probability and a given Pseudo DLE model without samples
}
\keyword{methods}

