% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Model-methods.R
\docType{methods}
\name{dose,numeric,ModelTox,Samples-method}
\alias{dose,numeric,ModelTox,Samples-method}
\title{Compute the doses for a given probability, given Pseudo DLE model with samples}
\usage{
\S4method{dose}{numeric,ModelTox,Samples}(prob, model, samples, ...)
}
\arguments{
\item{prob}{the probability}

\item{model}{the \code{\linkS4class{ModelTox}}}

\item{samples}{the \code{\linkS4class{Samples}}}

\item{\dots}{unused}
}
\description{
Compute the doses for a given probability, given Pseudo DLE model with samples
}
\keyword{methods}

