\name{dose}
\alias{dose}
\title{Compute the doses for a given probability, given model and samples}
\usage{
dose(prob, model, samples, ...)
}
\arguments{
  \item{prob}{the probability}

  \item{model}{the \code{\linkS4class{Model}}}

  \item{samples}{the \code{\linkS4class{Samples}}}

  \item{\dots}{unused}
}
\description{
Compute the doses for a given probability, given model and
samples
}
\keyword{methods}

