% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Model-methods.R
\docType{methods}
\name{doseforEff}
\alias{doseforEff}
\alias{doseforEff,numeric,ModelEff-method}
\title{Compute the dose for a given Expected Efficacy and a given Pseudo Efficacy model}
\usage{
doseforEff(ExpEff, model, ...)

\S4method{doseforEff}{numeric,ModelEff}(ExpEff, model, ...)
}
\arguments{
\item{ExpEff}{the Expected Efficacy value}

\item{model}{the \code{\linkS4class{ModelEff}} class object}

\item{sample}{the \code{\linkS4class{Samples}} class object}
}
\description{
Compute the dose for a given Expected Efficacy and a given Pseudo Efficacy model
}
\section{Methods (by class)}{
\itemize{
\item \code{ExpEff = numeric,model = ModelEff}: 
}}
\keyword{methods}

