% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design-class.R
\docType{class}
\name{DualDesign-class}
\alias{DualDesign-class}
\alias{.DualDesign}
\alias{DualDesign}
\title{\code{DualDesign}}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{\link{DualDesign}} is the class for the dual-endpoint CRM design. This class has
special requirements for the \code{model} and \code{data} slots in comparison to the
parent class \code{\link{Design}}.
}
\section{Slots}{

\describe{
\item{\code{model}}{(\code{DualEndpoint})\cr the model to be used.}

\item{\code{data}}{(\code{DataDual})\cr specifies dose grid, any previous data, etc.}
}}

\note{
the \code{nextBest} slot can be of any class, this allows for easy comparison
with recommendation methods that don't use the biomarker information.
}
