% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design-class.R
\name{DualDesign}
\alias{DualDesign}
\title{Initialization function for "DualDesign"}
\usage{
DualDesign(model, data, ...)
}
\arguments{
\item{model}{see \code{\linkS4class{DualDesign}}}

\item{data}{see \code{\linkS4class{DualDesign}}}

\item{\dots}{additional arguments for \code{\link{Design}}}
}
\value{
the \code{\linkS4class{DualDesign}} object
}
\description{
Initialization function for "DualDesign"
}
\keyword{methods}
