% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Model-class.R
\docType{class}
\name{DualEndpoint-class}
\alias{.DualEndpoint}
\alias{DualEndpoint-class}
\title{General class for the dual endpoint model}
\description{
Note that only the subclasses of this can be used. This class here
shall contain all the common features to reduce duplicate code.
(However, this class must not be virtual, because we need to create
objects of it during the construction of subclass objects.)
}
\section{Slots}{

\describe{
\item{\code{mu}}{For the probit toxicity model, \code{mu} contains the prior mean
vector}

\item{\code{Sigma}}{For the probit toxicity model, contains the prior covariance
matrix}

\item{\code{sigma2W}}{Either a fixed value for the biomarker variance, or a vector
with elements \code{a} and \code{b} for the inverse-gamma prior parameters.}

\item{\code{rho}}{Either a fixed value for the correlation (between -1 and 1), or a
vector with elements \code{a} and \code{b} for the Beta prior on the
transformation kappa = (rho + 1) / 2, which is in (0, 1). For example,
\code{a=1,b=1} leads to a uniform prior on rho.}

\item{\code{useFixed}}{a list with logical value for each of the parameters
indicating whether a fixed value is used or not.}
}}
\keyword{classes}

