% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Model-class.R
\name{DualEndpoint}
\alias{DualEndpoint}
\title{Initialization function for the "DualEndpoint" class}
\usage{
DualEndpoint(mu, Sigma, sigma2W, rho)
}
\arguments{
\item{mu}{see \code{\linkS4class{DualEndpoint}}}

\item{Sigma}{see \code{\linkS4class{DualEndpoint}}}

\item{sigma2W}{see \code{\linkS4class{DualEndpoint}}}

\item{rho}{see \code{\linkS4class{DualEndpoint}}}
}
\value{
the \code{\linkS4class{DualEndpoint}} object
}
\description{
Initialization function for the "DualEndpoint" class
}
\keyword{methods}

