% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Model-class.R
\docType{class}
\name{DualEndpointBeta-class}
\alias{.DualEndpointBeta}
\alias{DualEndpointBeta-class}
\title{Dual endpoint model with beta prior for biomarker}
\description{
This class extends the \code{\linkS4class{DualEndpoint}} class.
}
\details{
Here the biomarker mean is parametrized in the form of a rescaled
beta density function:

mu = E0 + (Emax - E0) * Beta(delta1, delta2) * (x/x*)^delta1 * (1 - x/x*)^delta2

where x* is a reference dose, delta1 and delta2 are the two beta
parameters, and E0 and Emax are the minimum and maximum levels,
respectively.
For ease of interpretation, we parametrize with delta1 and the mode
of the curve instead, where mode = delta1 / (delta1 + delta2).
}
\section{Slots}{

\describe{
\item{\code{E0}}{either a fixed number or the two uniform distribution parameters}

\item{\code{Emax}}{either a fixed number or the two uniform distribution parameters}

\item{\code{delta1}}{either a fixed number or the two uniform distribution parameters}

\item{\code{mode}}{either a fixed number or the two uniform distribution parameters}

\item{\code{refDose}}{the reference dose \eqn{x^{*}}}
}}
\keyword{classes}

