% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-class.R
\docType{class}
\name{DualEndpointOld-class}
\alias{DualEndpointOld-class}
\title{Dual endpoint model}
\description{
todo: describe the model
}
\section{Slots}{

\describe{
\item{\code{mu}}{For the probit toxicity model, \code{mu} contains the prior mean
vector}

\item{\code{Sigma}}{For the probit toxicity model, contains the prior covariance
matrix}

\item{\code{sigma2betaW}}{For the biomarker model, contains the prior variance
factor of the random walk prior. If it is not a single number, it can also
contain a vector with elements \code{a} and {b} for the inverse-gamma prior
on \code{sigma2betaW}.}

\item{\code{sigma2W}}{Either a fixed value for the biomarker variance, or a vector
with elements \code{a} and \code{b} for the inverse-gamma prior parameters.}

\item{\code{rho}}{Either a fixed value for the correlation (between -1 and 1), or a
vector with elements \code{a} and \code{b} for the Beta prior on the
transformation kappa = (rho + 1) / 2, which is in (0, 1). For example,
\code{a=1,b=1} leads to a uniform prior on rho.}

\item{\code{useRW1}}{for specifying the random walk prior on the biomarker level: if
\code{TRUE}, RW1 is used, otherwise RW2.}

\item{\code{useFixed}}{a list with logical value for each of the three parameters
\code{sigma2betaW}, \code{sigma2W} and \code{rho} indicating whether
a fixed value is used or not.}
}}

\keyword{classes}
\keyword{internal}
