% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Model-class.R
\docType{class}
\name{DualEndpointRW-class}
\alias{.DualEndpointRW}
\alias{DualEndpointRW-class}
\title{Dual endpoint model with RW prior for biomarker}
\description{
This class extends the \code{\linkS4class{DualEndpoint}} class.
}
\details{
todo: describe RW priors
}
\section{Slots}{

\describe{
\item{\code{sigma2betaW}}{Contains the prior variance factor of the random walk
prior for the biomarker model. If it is not a single number, it can also
contain a vector with elements \code{a} and {b} for the inverse-gamma prior
on \code{sigma2betaW}.}

\item{\code{useRW1}}{for specifying the random walk prior on the biomarker level: if
\code{TRUE}, RW1 is used, otherwise RW2.}
}}
\keyword{classes}

