% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-class.R
\name{DualEndpointRW}
\alias{DualEndpointRW}
\title{Initialization function for the "DualEndpointRW" class}
\usage{
DualEndpointRW(sigma2betaW, smooth = c("RW1", "RW2"), ...)
}
\arguments{
\item{sigma2betaW}{see \code{\linkS4class{DualEndpointRW}}}

\item{smooth}{either \dQuote{RW1} (default) or \dQuote{RW2}, for
specifying the random walk prior on the biomarker level.}

\item{\dots}{additional parameters, see \code{\linkS4class{DualEndpoint}}}
}
\value{
the \code{\linkS4class{DualEndpointRW}} object
}
\description{
Initialization function for the "DualEndpointRW" class
}
\keyword{methods}
