% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design-class.R
\docType{class}
\name{DualResponsesDesign-class}
\alias{DualResponsesDesign-class}
\alias{.DualResponsesDesign}
\alias{DualResponsesDesign}
\title{\code{DualResponsesDesign}}
\usage{
DualResponsesDesign(eff_model, data, ...)
}
\arguments{
\item{eff_model}{please refer to \code{\link{DualResponsesDesign}} class object}

\item{data}{please refer to \code{\link{DualResponsesDesign}} class object}

\item{\dots}{additional arguments for \code{\link{TDDesign}}}

\item{...}{
  Arguments passed on to \code{\link[=RuleDesign]{RuleDesign}}
  \describe{
    \item{\code{nextBest}}{(\code{NextBest})\cr see slot definition.}
    \item{\code{cohortSize}}{(\code{CohortSize})\cr see slot definition.}
    \item{\code{startingDose}}{(\code{number})\cr see slot definition.}
  }}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This is a class of design based on DLE responses using the \code{\link{LogisticIndepBeta}} model
model and efficacy responses using \code{\link{ModelEff}}  model class
without DLE and efficacy samples. It contain all slots in
\code{\link{RuleDesign}} and \code{\link{TDDesign}} class object
}
\section{Slots}{

\describe{
\item{\code{data}}{the data set of \code{\link{DataDual}} class object}

\item{\code{eff_model}}{the pseudo efficacy model to be used, an object class of
\code{\link{ModelEff}}}
}}

\examples{
# nolint start
##Construct the DualResponsesDesign for simulations
##The design comprises the DLE and efficacy models, the escalation rule, starting data,
##a cohort size and a starting dose
##Define your data set first using an empty data set
## with dose levels from 25 to 300 with increments 25
data <- DataDual(doseGrid=seq(25,300,25),placebo=FALSE)
##First for the DLE model
##The DLE model must be of 'ModelTox' (e.g 'LogisticIndepBeta') class
DLEmodel <- LogisticIndepBeta(binDLE=c(1.05,1.8),
                              DLEweights=c(3,3),
                              DLEdose=c(25,300),
                              data=data)

##The efficacy model of 'ModelEff' (e.g 'Effloglog') class
Effmodel<-Effloglog(eff=c(1.223,2.513), eff_dose=c(25,300), nu=c(a=1,b=0.025),data=data)

##The escalation rule using the 'NextBestMaxGain' class
mynextbest<-NextBestMaxGain(prob_target_drt=0.35,
                            prob_target_eot=0.3)


##The increments (see Increments class examples)
## 200\% allowable increase for dose below 300 and 200\% increase for dose above 300
myIncrements<-IncrementsRelative(intervals=c(25,300),
                                 increments=c(2,2))
##cohort size of 3
mySize<-CohortSizeConst(size=3)
##Stop only when 36 subjects are treated
myStopping <- StoppingMinPatients(nPatients=36)
##Now specified the design with all the above information and starting with a dose of 25

design <- DualResponsesDesign(nextBest=mynextbest,
                              model=DLEmodel,
                              Effmodel=Effmodel,
                              stopping=myStopping,
                              increments=myIncrements,
                              cohortSize=mySize,
                              data=data,startingDose=25)
# nolint end
}
