% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Design-class.R
\name{DualResponsesDesign}
\alias{DualResponsesDesign}
\title{Init function}
\usage{
DualResponsesDesign(nextBest, cohortSize, startingDose, DLEmodel, Effmodel,
  data, stopping, increments)
}
\description{
Init function
}

