% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design-class.R
\name{DualResponsesSamplesDesign}
\alias{DualResponsesSamplesDesign}
\title{Initialization function for 'DualResponsesSamplesDesign"}
\usage{
DualResponsesSamplesDesign(Effmodel, data, ...)
}
\arguments{
\item{Effmodel}{please refer to \code{\linkS4class{DualResponsesSamplesDesign}} class object}

\item{data}{please refer to \code{\linkS4class{DualResponsesSamplesDesign}} class object}

\item{\dots}{additional arguments for \code{\link{TDsamplesDesign}}}
}
\value{
the \code{\linkS4class{DualResponsesSamplesDesign}} class object
}
\description{
Initialization function for 'DualResponsesSamplesDesign"
}
\keyword{methods}
