% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations-class.R
\name{DualSimulations}
\alias{DualSimulations}
\title{Initialization function for "DualSimulations"}
\usage{
DualSimulations(rhoEst, sigma2West, fitBiomarker, ...)
}
\arguments{
\item{rhoEst}{see \code{\linkS4class{DualSimulations}}}

\item{sigma2West}{see \code{\linkS4class{DualSimulations}}}

\item{fitBiomarker}{see \code{\linkS4class{DualSimulations}}}

\item{\dots}{additional parameters from \code{\link{Simulations}}}
}
\value{
the \code{\linkS4class{DualSimulations}} object
}
\description{
Initialization function for "DualSimulations"
}
\keyword{methods}
