% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Model-class.R
\name{EffFlexi}
\alias{EffFlexi}
\title{Initialization function for the "EffFlexi" class}
\usage{
EffFlexi(Eff, Effdose, sigma2, sigma2betaW, smooth = c("RW1", "RW2"), data)
}
\arguments{
\item{Eff}{the pseudo efficacy responses}

\item{Effdose}{the corresponding dose levels for the pseudo efficacy responses}

\item{sigma2}{the variance of the efficacy responses which can be a fixed value or from the
inverse gamma distribution, that is either specific a fixed value or specified the values
of the two paramters of the inverse gamma distribtuion}

\item{sigma2betaW}{the variance of the random walk model used for smoothing data which can be
a fixed value or using the inverse gamm distribution}

\item{smooth}{used for smoothing data for this efficacy model, either the "RW1", the
first-order random walk model or "RW2", the second-order random walk model can be used.}

\item{data}{the input data to update estimates of model parameters and
follow the \code{\linkS4class{DataDual}} object class specification}
}
\description{
Initialization function for the "EffFlexi" class
}
\keyword{methods}

