% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Model-class.R
\docType{class}
\name{Effloglog-class}
\alias{.Effloglog}
\alias{Effloglog-class}
\title{Class for the linear log-log efficacy model using pseudo data prior
This is the efficay model which describe the realtionship of the normal efficacy responses
and their corresponding log-log dose levels}
\description{
Class for the linear log-log efficacy model using pseudo data prior
This is the efficay model which describe the realtionship of the normal efficacy responses
and their corresponding log-log dose levels
}
\keyword{methods}

