% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Model-class.R
\name{Effloglog}
\alias{Effloglog}
\title{Initialization function for the "Effloglog" class}
\usage{
Effloglog(Eff, Effdose, nu, data)
}
\arguments{
\item{Eff}{the pseudo efficacy responses}

\item{Effdose}{the corresponding dose levels for the pseudo efficacy responses}

\item{nu}{the precision (inverse of the variance) of the efficacy responses}

\item{data}{the input data of \code{\linkS4class{DataDual}} class to update model estimates}
}
\value{
the \code{\linkS4class{Effloglog}} object
}
\description{
Initialization function for the "Effloglog" class
}
\keyword{methods}

