% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Model-methods.R
\docType{methods}
\name{ExpEff,numeric,EffFlexi,Samples-method}
\alias{ExpEff,numeric,EffFlexi,Samples-method}
\title{Compute the Expected Efficacy based a given dose, Efficacy Flexible  model with
samples}
\usage{
\S4method{ExpEff}{numeric,EffFlexi,Samples}(dose, model, samples, ...)
}
\arguments{
\item{dose}{the dose}

\item{model}{the \code{\linkS4class{ModelEff}} class object}

\item{samples}{the \code{\linkS4class{Samples}} class object}
}
\description{
Compute the Expected Efficacy based a given dose, Efficacy Flexible  model with
samples
}
\keyword{methods}

