% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Model-methods.R
\docType{methods}
\name{ExpEff,numeric,ModelEff,ANY-method}
\alias{ExpEff,numeric,ModelEff,ANY-method}
\title{Compute the Expected Efficacy based a given dose and a given Pseudo Efficacy model without
samples}
\usage{
\S4method{ExpEff}{numeric,ModelEff,ANY}(dose, model, samples, ...)
}
\arguments{
\item{dose}{the dose}

\item{model}{the \code{\linkS4class{ModelEff}} class object}
}
\description{
Compute the Expected Efficacy based a given dose and a given Pseudo Efficacy model without
samples
}
\keyword{methods}

