% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Model-methods.R
\docType{methods}
\name{ExpEff}
\alias{ExpEff}
\alias{ExpEff,numeric,ModelEff,Samples-method}
\title{Compute the expected efficacy based on a given dose, a given pseudo Efficacy model and a given sample}
\usage{
ExpEff(dose, model, samples, ...)

\S4method{ExpEff}{numeric,ModelEff,Samples}(dose, model, samples, ...)
}
\arguments{
\item{dose}{the dose}

\item{model}{the \code{\linkS4class{ModelEff}} class object}

\item{samples}{the \code{\linkS4class{Samples}} class object}
}
\description{
Compute the expected efficacy based on a given dose, a given pseudo Efficacy model and a given sample
}
\section{Methods (by class)}{
\itemize{
\item \code{dose = numeric,model = ModelEff,samples = Samples}: 
}}
\keyword{methods}

