\docType{methods}
\name{extract,Samples-method}
\alias{extract,Samples-method}
\title{Extract certain parameter from Samples object}
\usage{
\S4method{extract}{Samples}(object, parameter, ...)
}
\arguments{
  \item{object}{the \code{\linkS4class{Samples}} object}

  \item{parameter}{the name of the parameter}

  \item{...}{unused}
}
\value{
the data frame suitable for use with
\code{\link[ggmcmc]{ggmcmc}}
}
\description{
Extract certain parameter from Samples object
}
\keyword{methods}

