\name{extract}
\alias{extract}
\title{Extract something from an object and produce a data.frame}
\usage{
extract(object, ...)
}
\arguments{
  \item{object}{the object}

  \item{\dots}{unused}
}
\value{
the data frame
}
\description{
Extract something from an object and produce a data.frame
}
\keyword{methods}

