% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Samples-methods.R
\docType{methods}
\name{fit}
\alias{fit}
\alias{fit,Samples,DualEndpoint,DataDual-method}
\alias{fit,Samples,Model,Data-method}
\title{Fit method for the Samples class}
\usage{
fit(object, model, data, ...)

\S4method{fit}{Samples,Model,Data}(object, model, data,
  points = data@doseGrid, quantiles = c(0.025, 0.975), middle = mean, ...)

\S4method{fit}{Samples,DualEndpoint,DataDual}(object, model, data,
  quantiles = c(0.025, 0.975), middle = mean, ...)
}
\arguments{
\item{object}{the \code{\linkS4class{Samples}} object}

\item{model}{the \code{\linkS4class{Model}} object}

\item{data}{the \code{\linkS4class{Data}} object}

\item{points}{at which dose levels is the fit requested? default is the dose
grid}

\item{quantiles}{the quantiles to be calculated (default: 0.025 and
0.975)}

\item{middle}{the function for computing the middle point. Default:
\code{\link{mean}}}

\item{\dots}{unused}
}
\value{
the data frame with required information (see method details)
}
\description{
Note this new generic function is necessary because the \code{\link{fitted}}
function only allows the first argument \code{object} to appear in the
signature. But we need also other arguments in the signature.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = Samples,model = Model,data = Data}: This method returns a data frame with dose, middle, lower
and upper quantiles for the dose-toxicity curve

\item \code{object = Samples,model = DualEndpoint,data = DataDual}: This method returns a data frame with dose, and middle,
lower and upper quantiles, for both the dose-tox and dose-biomarker (suffix
"Biomarker") curves, for all grid points (Note that currently only the grid
points can be used, because the DualEndpointRW models only allow that)
}}
\keyword{methods}

