% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Samples-methods.R
\docType{methods}
\name{fitGain,ModelTox,Samples,ModelEff,Samples,DataDual-method}
\alias{fitGain,ModelTox,Samples,ModelEff,Samples,DataDual-method}
\title{init' function}
\usage{
\S4method{fitGain}{ModelTox,Samples,ModelEff,Samples,DataDual}(DLEmodel,
  DLEsamples, Effmodel, Effsamples, data, points = data@doseGrid,
  quantiles = c(0.025, 0.975), middle = mean, ...)
}
\description{
init' function
}

