% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Samples-methods.R
\name{fitGain}
\alias{fitGain}
\title{Get the fiited values for the gain values at all dose levels based on
a given DLE model, DLE sample, an efficacy model, Efficacy sample and data}
\usage{
fitGain(DLEmodel, DLEsamples, Effmodel, Effsamples, data, ...)
}
\arguments{
\item{DLEmodel}{the DLE pseudo model of \code{\linkS4class{ModelTox}} class object}

\item{DLEsamples}{the DLE samples of \code{\linkS4class{Samples}} class object}

\item{Effmodel}{the efficacy pseudo model of \code{\linkS4class{ModelEff}} class object}

\item{Effsamples}{the efficacy samples of \code{\linkS4class{Samples}} class object}

\item{the}{data input of \code{\linkS4class{DataDual}} class object}
}
\description{
Get the fiited values for the gain values at all dose levels based on
a given DLE model, DLE sample, an efficacy model, Efficacy sample and data
}
\keyword{methods}

