% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-class.R
\docType{class}
\name{FractionalCRM-class}
\alias{FractionalCRM-class}
\alias{.FractionalCRM}
\alias{FractionalCRM}
\title{Fractional CRM following paper and code by Guosheng Yin et al}
\usage{
FractionalCRM(...)
}
\description{
This is a fractional CRM model based on a one parameter CRM (with normal
prior on the log-power parameter) as well as Kaplan-Meier based estimation
of the conditional probability to experience a DLT for non-complete
observations.
}
\section{Functions}{
\itemize{
\item \code{FractionalCRM}: Initialization function for the fractional CRM.
Takes the same arguments as \code{\link{OneParExpNormalPrior}}
}}

\keyword{classes}
