% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Model-methods.R
\docType{methods}
\name{gain,numeric,ModelTox,missing,ModelEff,missing-method}
\alias{gain,numeric,ModelTox,missing,ModelEff,missing-method}
\title{Compute the gain given a dose level, a given DLE model and a given Efficacy model}
\usage{
\S4method{gain}{numeric,ModelTox,missing,ModelEff,missing}(dose, DLEmodel,
  DLEsamples, Effmodel, Effsamples, ...)
}
\description{
Compute the gain given a dose level, a given DLE model and a given Efficacy model
}
\keyword{methods}

