% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Model-methods.R
\docType{methods}
\name{gain,numeric,ModelTox,Samples,EffFlexi,Samples-method}
\alias{gain,numeric,ModelTox,Samples,EffFlexi,Samples-method}
\title{Compute the gain given a dose level, a given DLE model, a given DLE sample,
the EffFlexi model and a given Efficacy sample}
\usage{
\S4method{gain}{numeric,ModelTox,Samples,EffFlexi,Samples}(dose, DLEmodel,
  DLEsamples, Effmodel, Effsamples, ...)
}
\description{
Compute the gain given a dose level, a given DLE model, a given DLE sample,
the EffFlexi model and a given Efficacy sample
}
\keyword{methods}

