% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Model-methods.R
\docType{methods}
\name{gain,numeric,ModelTox,Samples,ModelEff,Samples-method}
\alias{gain,numeric,ModelTox,Samples,ModelEff,Samples-method}
\title{Compute the gain value samples given a dose level, a DLE model, a DLE sample, a Efficacy model
and a Efficacy sample}
\usage{
\S4method{gain}{numeric,ModelTox,Samples,ModelEff,Samples}(dose, DLEmodel,
  DLEsamples, Effmodel, Effsamples, ...)
}
\description{
Compute the gain value samples given a dose level, a DLE model, a DLE sample, a Efficacy model
and a Efficacy sample
}
\keyword{methods}

