% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Model-methods.R
\name{gain}
\alias{gain}
\title{Compute the gain value with a given dose level, given DLE model, given DLE sample, given Efficacy
model and a given Efficacy sample}
\usage{
gain(dose, DLEmodel, DLEsamples, Effmodel, Effsamples, ...)
}
\description{
Compute the gain value with a given dose level, given DLE model, given DLE sample, given Efficacy
model and a given Efficacy sample
}
\keyword{methods}

