% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-class.R
\docType{class}
\name{GeneralModel-class}
\alias{GeneralModel-class}
\alias{.GeneralModel}
\alias{GeneralModel}
\title{\code{GeneralModel}}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{\link{GeneralModel}} is a general model class, from which all other specific
model-like classes inherit.
}
\section{Slots}{

\describe{
\item{\code{datamodel}}{(\code{function})\cr a function representing the \code{JAGS} data model
specification.}

\item{\code{priormodel}}{(\code{function})\cr a function representing the \code{JAGS} prior
specification.}

\item{\code{modelspecs}}{(\code{function})\cr a function computing the list of the data
model and prior model specifications that are required to be specified
completely (e.g. prior parameters, reference dose, etc.), based on the data
slots that are required as arguments of this function.}

\item{\code{init}}{(\code{function})\cr a function computing the list of starting values
for parameters required to be initialized in the MCMC sampler, based on the
data slots that are required as arguments of this function.}

\item{\code{datanames}}{(\code{character})\cr the names of all data slots that are used
by \code{datamodel} JAGS function.}

\item{\code{datanames_prior}}{(\code{character})\cr the names of all data slots that are
used by \code{priormodel} JAGS function.}

\item{\code{sample}}{(\code{character})\cr names of all parameters from which you would
like to save the MCMC samples.}
}}

\note{
The \code{datamodel} must obey the convention that the data input is
called exactly in the same way as in the corresponding data class.
All prior distributions for parameters should be contained in the
model function \code{priormodel}. The background is that this can
be used to simulate from the prior distribution, before obtaining any data.
}
\seealso{
\code{\link{ModelPseudo}}.
}
