% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations-class.R
\name{GeneralSimulations}
\alias{GeneralSimulations}
\title{Initialization function for "GeneralSimulations"}
\usage{
GeneralSimulations(data, doses, seed)
}
\arguments{
\item{data}{see \code{\linkS4class{GeneralSimulations}}}

\item{doses}{see \code{\linkS4class{GeneralSimulations}}}

\item{seed}{see \code{\linkS4class{GeneralSimulations}}}
}
\value{
the \code{\linkS4class{GeneralSimulations}} object
}
\description{
Initialization function for "GeneralSimulations"
}
\keyword{methods}
