% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Samples-methods.R
\docType{methods}
\name{get,Samples,character-method}
\alias{get,Samples,character-method}
\title{Get specific parameter samples and produce a data.frame}
\usage{
\S4method{get}{Samples,character}(x, pos = -1L, envir = NULL, mode = NULL,
  inherits = NULL)
}
\arguments{
\item{x}{the \code{\linkS4class{Samples}} object}

\item{pos}{the name of the parameter}

\item{envir}{for vectorial parameters, you can give the indices of the
elements you would like to extract. If \code{NULL}, the whole vector samples
will be returned}

\item{mode}{not used}

\item{inherits}{not used}
}
\value{
the data frame suitable for use with \code{\link[ggmcmc]{ggmcmc}}
}
\description{
Here you have to specify with \code{pos} which
parameter you would like to extract from the \code{\linkS4class{Samples}}
object
}
\keyword{methods}

