% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Model-methods.R
\docType{methods}
\name{getEff}
\alias{getEff}
\alias{getEff,DataDual-method}
\title{Extracting efficay responses for subjects without an DLE}
\usage{
getEff(object, ...)

\S4method{getEff}{DataDual}(object, x, y, w, ...)
}
\arguments{
\item{object}{for data input from \code{\linkS4class{Data}} object}
}
\description{
Extracting efficay responses for subjects without an DLE
}
\section{Methods (by class)}{
\itemize{
\item \code{DataDual}: 
}}
\keyword{methods}

