% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromQuantiles.R
\name{getMinInfBeta}
\alias{getMinInfBeta}
\title{Get the minimal informative unimodal beta distribution}
\usage{
getMinInfBeta(p, q)
}
\arguments{
\item{p}{the probability (> 0 and < 1)}

\item{q}{the quantile (> 0 and < 1)}
}
\value{
the two resulting beta parameters a and b in a list
}
\description{
As defined in Neuenschwander et al (2008), this function computes the
parameters of the minimal informative unimodal beta distribution, given the
request that the p-quantile should be q, i.e. X ~ Be(a, b) with Pr(X <= q) =
p.
}
\keyword{programming}
