\name{getThreePlusThreeDesign}
\alias{getThreePlusThreeDesign}
\title{Creates a new 3+3 design object from a dose grid}
\usage{
getThreePlusThreeDesign(doseGrid)
}
\arguments{
  \item{doseGrid}{the dose grid to be used}
}
\value{
the object of class \code{\linkS4class{RuleDesign}} with
the 3+3 design
}
\description{
Creates a new 3+3 design object from a dose grid
}
\author{
Daniel Sabanes Bove \email{sabanesd@roche.com}
}
\keyword{programming}

