% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_design.R
\name{h_add_dlts}
\alias{h_add_dlts}
\title{Helper Function to Add Randomly Generated DLTs During Simulations}
\usage{
h_add_dlts(data, dose, truth, cohort_size, first_separate)
}
\arguments{
\item{data}{(\code{Data})\cr what data to start from.}

\item{dose}{(\code{number})\cr current dose.}

\item{truth}{(\code{function})\cr defines the true probability for a DLT at a dose.}

\item{cohort_size}{(\code{CohortSize})\cr the cohort size rule to use.}

\item{first_separate}{(\code{flag})\cr whether the first patient is enrolled separately.}
}
\value{
The updated \code{data}.
}
\description{
Helper Function to Add Randomly Generated DLTs During Simulations
}
\keyword{internal}
