% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_rules.R
\name{h_arrange_plots_aligned}
\alias{h_arrange_plots_aligned}
\title{Arrange Plots with Aligned Panels}
\usage{
h_arrange_plots_aligned(..., nrow = length(list(...)))
}
\arguments{
\item{...}{(\code{ggplot}) plots to arrange.}

\item{nrow}{(\code{count}) number of rows in the arrangement.}
}
\value{
A \code{gtable} object containing the aligned plots.
}
\description{
Aligns the widths of all plot-table columns before arranging
plots vertically. This keeps corresponding x-axes aligned when plots have
legends of different widths.
}
\keyword{internal}
