% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{h_assert_dose}
\alias{h_assert_dose}
\title{Assert that the value of \code{dose} argument of \code{\link{prob}} method is valid}
\usage{
h_assert_dose(dose, is_sample_scalar)
}
\arguments{
\item{dose}{(\code{number} or \code{numeric})\cr the dose to be validated.}

\item{is_sample_scalar}{(\code{flag})\cr is model parametr sample scalar?}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

A simple helper function that asserts that \code{dose} argument of \code{\link{prob}}
methods is valid. In particular, if the model parameter sample is a vector,
then the \code{dose} must be a scalar value.
}
\examples{
h_assert_dose(c(34, 12), TRUE)
\dontrun{
h_assert_dose(c(34, 12), FALSE) # Error.
}
}
