% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{h_assert_prob}
\alias{h_assert_prob}
\title{Assert that the value of \code{prob} argument of \code{\link{dose}} method is valid}
\usage{
h_assert_prob(prob, is_sample_scalar)
}
\arguments{
\item{prob}{(\code{number} or \code{numeric})\cr the prob to be validated.}

\item{is_sample_scalar}{(\code{flag})\cr is model parametr sample scalar?}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

A simple helper function that asserts that \code{prob} argument of \code{\link{dose}}
methods is valid. In particular, if the model parameter sample is a vector,
then the \code{prob} must be a scalar value.
}
\examples{
h_assert_prob(c(0.5, 0.9), TRUE)
\dontrun{
h_assert_prob(c(0.5, 0.9), FALSE) # Error.
}

}
