% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_simulations.R
\name{h_barplot_percentages}
\alias{h_barplot_percentages}
\title{Convenience function to make barplots of percentages}
\usage{
h_barplot_percentages(
  x,
  description,
  xaxisround = 0,
  x_is_discrete = FALSE,
  axis_text_angle = 0,
  discrete_levels = NULL
)
}
\arguments{
\item{x}{vector of samples}

\item{description}{xlab string}

\item{xaxisround}{rounding for xaxis labels (default: 0, i.e. integers will
be used)}

\item{x_is_discrete}{whether the values on the x-axis should be treated as
discrete categories}

\item{axis_text_angle}{(\code{number}) rotation angle for x-axis tick labels.}

\item{discrete_levels}{(\code{numeric} or \code{NULL}) complete ordered set of discrete
x-axis values. Only used when \code{x_is_discrete} is \code{TRUE}.}
}
\value{
the ggplot2 object
}
\description{
Convenience function to make barplots of percentages
}
\keyword{internal}
