% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_data.R
\name{h_blind_plot_data}
\alias{h_blind_plot_data}
\title{Helper Function to Blind Plot Data}
\usage{
h_blind_plot_data(df, blind, has_placebo, pbo_dose)
}
\arguments{
\item{df}{(\code{GeneralData})\cr The data to be blinded}

\item{blind}{(\code{flag})\cr Should the data be blinded?}

\item{has_placebo}{(\code{flag})\cr Does the data contain a placebo dose?}

\item{placebo}{(\code{positive_number})\cr The dose to be taken as placebo.
Ignored if \code{has_placebo} is \code{FALSE}}
}
\value{
The blinded data
}
\description{
Helper Function to Blind Plot Data
}
