% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{h_combine_validation_results}
\alias{h_combine_validation_results}
\title{Combining validation results}
\usage{
h_combine_validation_results(...)
}
\arguments{
\item{...}{outputs from S4 object validation functions.}
}
\value{
\code{character} vector if there was at least object of type \code{character}
in the input. Otherwise, \code{TRUE}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

A simple helper function that combines outputs from S4 object validation
functions into one consistent object. Namely, it combines the validation
\code{character} messages supplied to the input of this function into one \code{character}
vector, or simply returns \code{TRUE} if there were no such messages.
}
