% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_data.R
\name{h_convert_ordinal_data}
\alias{h_convert_ordinal_data}
\title{Convert a Ordinal Data to the Equivalent Binary Data for a Specific
Grade}
\usage{
h_convert_ordinal_data(data_ord, grade)
}
\arguments{
\item{data_ord}{(\code{DataOrdinal})\cr the \code{DataOrdinal} object to covert}

\item{grade}{(\code{integer})\cr the toxicity grade for which the equivalent data
is required.}
}
\value{
A \code{\link{Data}} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

A simple helper function that takes a \code{\link{DataOrdinal}} object and an
integer grade and converts them to the equivalent \code{Data} object.
}
