% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_model.R
\name{h_convert_ordinal_model}
\alias{h_convert_ordinal_model}
\title{Convert an ordinal CRM model to the Equivalent Binary CRM Model for a Specific
Grade}
\usage{
h_convert_ordinal_model(x, grade)
}
\arguments{
\item{x}{(\code{LogisticLogNormalOrdinal})\cr the \code{LogisticLogNormalOrdinal}
model to covert}

\item{grade}{(\code{integer})\cr the toxicity grade for which the equivalent model
is required.}
}
\value{
A \code{\link{LogisticLogNormal}} model.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

A simple helper function that takes a \code{\link{LogisticLogNormalOrdinal}} and an
integer grade and converts them to the equivalent \code{LogisticLogNormal} model.
}
