% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_samples.R
\name{h_convert_ordinal_samples}
\alias{h_convert_ordinal_samples}
\title{Convert a Samples Object from an ordinal Model to the Equivalent Samples Object
from a Binary Model}
\usage{
h_convert_ordinal_samples(obj, grade)
}
\arguments{
\item{obj}{(\code{DataOrdinal})\cr the \code{DataOrdinal} object to covert}

\item{grade}{(\code{integer})\cr the toxicity grade for which the equivalent data
is required.}
}
\value{
A \code{\link{Samples}} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

A simple helper function that converts a \code{\link{Samples}} object from the fit of an
ordinal CRM model to that which would have been obtained from fitting a binary
CRM model for toxicities of a specified grade to the same observed data.
}
