% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_covr.R
\name{h_covr_detrace}
\alias{h_covr_detrace}
\title{Remove \code{covr} traces from an expression}
\usage{
h_covr_detrace(expr)
}
\arguments{
\item{expr}{(\code{language})\cr an R expression or call to strip of \code{covr} trace
counters.}
}
\value{
A transformed expression, with calls to \code{covr:::count} removed.
}
\description{
'\code{covr} interferes with expressions (often function bodies) as a primary
interface of this package. To work with \code{covr}, we need to strip out the
code that \code{covr} injects into the package before using these expressions
for internal logic such as writing model files.
}
\details{
This method is non-exhaustive, covering only a subset of \code{covr}'s tracing
behaviors necessary for this package.
}
