% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h_data_blind}
\alias{h_data_blind}
\title{This is to blind the data.
For each cohort, the placebo is set to the active dose level for that cohort.
In addition, all DLTs are assigned to the first subjects in the cohort.}
\usage{
h_data_blind(x)
}
\description{
This is to blind the data.
For each cohort, the placebo is set to the active dose level for that cohort.
In addition, all DLTs are assigned to the first subjects in the cohort.
}
