% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_rules.R
\name{h_delta_g_yeung}
\alias{h_delta_g_yeung}
\title{Find a Matrix for Variance According to Yeung et. al (2015)}
\usage{
h_delta_g_yeung(log_dose_mg, model, model_eff)
}
\arguments{
\item{log_dose_mg}{(\code{number})\cr the log of dose corresponding to the maximum gain.}

\item{model}{(\code{ModelTox})\cr the DLT model.}

\item{model_eff}{(\code{Effloglog})\cr the efficacy model.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Helper function which computes the matrix required for calculation of the
variance of the logarithm of the maximum gain dose, based on Yeung et. al (2015)
paper. This helper function is used by \code{\link[=nextBest-NextBestMaxGain]{nextBest-NextBestMaxGain()}} method.
}
\references{
Yeung, W.Y., Whitehead, J., Reigner, B., Beyer, U., Diack, Ch., Jaki, T. (2015),
Bayesian adaptive dose-escalation procedures for binary and continuous responses utilizing a gain function,
\emph{Pharmaceutical Statistics},
\doi{10.1002/pst.1706} \cr
}
