% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_design.R
\name{h_enroll_backfill_patients}
\alias{h_enroll_backfill_patients}
\title{Helper Function to Enroll Backfill Patients}
\usage{
h_enroll_backfill_patients(
  backfill_cohorts,
  data,
  backfill,
  cohort_size,
  backfill_patients,
  current_args,
  truth,
  trueResponse
)
}
\arguments{
\item{backfill_cohorts}{(\code{list})\cr current queue of backfill cohorts.}

\item{data}{(\code{Data})\cr current trial data.}

\item{backfill}{(\code{Backfill})\cr with priority and recruitment rules.}

\item{cohort_size}{(\code{count})\cr size of the active cohort.}

\item{backfill_patients}{(\code{count})\cr number of enrolled backfill patients.}

\item{current_args}{(\code{data.frame})\cr arguments for the truth function.}

\item{truth}{(\code{function})\cr defining true DLT probability.}

\item{trueResponse}{(\code{function})\cr defining true response probability.}
}
\value{
List with updated \code{data}, \code{backfill_cohorts}, and
\code{backfill_patients}.
}
\description{
This function executes backfill enrollment according to the priority rule,
updating the data object, backfill cohort queue, and patient counter.
}
\keyword{internal}
