% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{h_extract_jags_samples}
\alias{h_extract_jags_samples}
\title{Extracting Samples from \code{JAGS} Model}
\usage{
h_extract_jags_samples(x)
}
\arguments{
\item{x}{an object returned from \code{\link[rjags:jags.samples]{rjags::jags.samples()}}, i.e. a list of
\code{mcarray} objects.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

A simple helper function that extracts parameters samples from the output of
\code{\link[rjags:jags.samples]{rjags::jags.samples()}} function.
}
