% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{h_find_interval}
\alias{h_find_interval}
\title{Find Interval Numbers or Indices and Return Custom Number For 0.}
\usage{
h_find_interval(..., replacement = -Inf)
}
\arguments{
\item{...}{further arguments passed to \code{\link[base:findInterval]{base::findInterval()}} function.}

\item{replacement}{(\code{number})\cr a custom number to be used as a replacement
for \eqn{0}. Default to \code{-Inf}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

A simple wrapper of \code{\link[base:findInterval]{base::findInterval()}} function that invokes
\code{\link[base:findInterval]{base::findInterval()}}, takes its output and replaces all the
elements with \eqn{0} value to a custom number as specified in \code{replacement}
argument.
}
\examples{
h_find_interval(1, c(2, 4, 6))
h_find_interval(3, c(2, 4, 6))
h_find_interval(1, c(2, 4, 6), replacement = -1)
}
