% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromQuantiles.R
\name{h_get_min_inf_beta}
\alias{h_get_min_inf_beta}
\title{Helper for Minimal Informative Unimodal Beta Distribution}
\usage{
h_get_min_inf_beta(p, q)
}
\arguments{
\item{p}{(\code{number})\cr the probability.}

\item{q}{(\code{number})\cr the quantile.}
}
\value{
A list with the two resulting beta parameters \code{a} and \code{b}.
}
\description{
As defined in Neuenschwander et al (2008), this function computes the
parameters of the minimal informative unimodal beta distribution, given the
request that the p-quantile should be q, i.e. \code{X ~ Be(a, b)} with
\code{Pr(X <= q) = p}.
}
\keyword{internal}
