% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromQuantiles.R
\name{h_get_quantiles_start_values}
\alias{h_get_quantiles_start_values}
\title{Get Starting Values for Quantiles Optimization}
\usage{
h_get_quantiles_start_values(parstart, median, dosegrid, refDose, logNormal)
}
\arguments{
\item{parstart}{(\code{numeric} or \code{NULL})\cr starting parameter values.}

\item{median}{(\code{numeric})\cr median values.}

\item{dosegrid}{(\code{numeric})\cr dose grid.}

\item{refDose}{(\code{number})\cr reference dose.}

\item{logNormal}{(\code{flag})\cr use log-normal prior?}
}
\value{
Numeric vector of starting values.
}
\description{
Get Starting Values for Quantiles Optimization
}
\keyword{internal}
