% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{h_group_data}
\alias{h_group_data}
\title{Group Together Mono and Combo Data}
\usage{
h_group_data(mono_data, combo_data)
}
\arguments{
\item{mono_data}{(\code{Data})\cr mono data.}

\item{combo_data}{(\code{Data})\cr combo data.}
}
\value{
A \code{\link{DataGrouped}} object containing both \code{mono_data} and \code{combo_data},
but with arbitrary ID and cohort slots.
}
\description{
This is only used in the simulation method for \code{DesignGrouped} to combine
the separately generated data sets from mono and combo arms and to fit the
combined logistic regression model.
Hence the ID and cohort information is not relevant and will be
arbitrarily assigned to avoid problems with the \code{\link{DataGrouped}} validation.
}
\keyword{internal}
