% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-class.R
\name{h_hierarchical_compile_datamodel}
\alias{h_hierarchical_compile_datamodel}
\title{Compile the Hierarchical Data Model}
\usage{
h_hierarchical_compile_datamodel(models_to_arms)
}
\arguments{
\item{models_to_arms}{(\code{list})\cr named arm-specific models.}
}
\value{
A function representing the compiled JAGS data model.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Concatenates the arm-specific likelihood code into a single JAGS model
function. Each arm receives a unique variable-name prefix derived from its
user-facing name.
}
\keyword{internal}
