% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchicalModel-class.R
\name{h_hierarchical_expr_key}
\alias{h_hierarchical_expr_key}
\title{Normalize an Expression Key}
\usage{
h_hierarchical_expr_key(expr)
}
\arguments{
\item{expr}{(\code{language})\cr expression to identify.}
}
\value{
A character scalar.
}
\description{
Normalize an Expression Key
}
\keyword{internal}
