% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchicalDesign-methods.R
\name{h_hierarchical_get_decision_samples}
\alias{h_hierarchical_get_decision_samples}
\title{Helper function to get the samples used for decision rules in a hierarchical design}
\usage{
h_hierarchical_get_decision_samples(
  samples,
  arm_name,
  arm,
  arm_data,
  mcmcOptions
)
}
\arguments{
\item{samples}{the overall samples from the hierarchical model.}

\item{arm_name}{the name of the arm for which we want to get the samples for decision rules.}

\item{arm}{the \code{DesignArm} object for this arm.}

\item{arm_data}{the data for this arm.}

\item{mcmcOptions}{the MCMC options to use if we need to fit an arm-specific model.}
}
\value{
the samples to be used for decision rules for this arm.
}
\description{
In a hierarchical design, the samples used for decision rules in each arm may either be the overall
samples from the hierarchical model (if borrowing is allowed)
or the arm-specific samples (if no borrowing).
}
\keyword{internal}
