% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchicalModel-class.R
\name{h_hierarchical_is_single_model}
\alias{h_hierarchical_is_single_model}
\title{Is a Compatible Hierarchical Single-Agent Model}
\usage{
h_hierarchical_is_single_model(model)
}
\arguments{
\item{model}{(\code{GeneralModel})\cr model object.}
}
\value{
A flag.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Uses the same structural compatibility contract as \code{\link[=TwoDrugsCombo]{TwoDrugsCombo()}} for
single-agent binary outcome models.
}
\keyword{internal}
