% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchicalModel-class.R
\name{h_hierarchical_model_body_lines}
\alias{h_hierarchical_model_body_lines}
\title{Deparse a Model Body into JAGS Lines}
\usage{
h_hierarchical_model_body_lines(model_fun)
}
\value{
Character vector containing the body expressions without braces.
}
\description{
Deparse a Model Body into JAGS Lines
}
\keyword{internal}
