% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchicalModel-class.R
\name{h_hierarchical_model_specs}
\alias{h_hierarchical_model_specs}
\title{Suffix Single-Arm Model Specifications}
\usage{
h_hierarchical_model_specs(model, arm_name, from_prior)
}
\arguments{
\item{model}{(\code{GeneralModel})\cr arm-specific model object.}

\item{arm_name}{(\code{string})\cr hierarchical arm name.}

\item{from_prior}{(\code{flag})\cr whether to request prior-only specifications.}
}
\value{
A named list of suffixed model specifications.
}
\description{
Suffix Single-Arm Model Specifications
}
\keyword{internal}
