% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchicalModel-class.R
\name{h_hierarchical_namespace_model}
\alias{h_hierarchical_namespace_model}
\title{Add an Arm Suffix to a Hierarchical Model Fragment}
\usage{
h_hierarchical_namespace_model(model, arm_name, slot_name)
}
\arguments{
\item{model}{(\code{GeneralModel})\cr arm-specific model object.}

\item{arm_name}{(\code{string})\cr hierarchical arm name.}

\item{slot_name}{(\code{string})\cr model-function slot to namespace.}
}
\value{
A function containing the namespaced JAGS model fragment.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Namespaces one arm-specific model fragment using code supplied by the arm
model class itself.
}
\keyword{internal}
