% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchicalModel-class.R
\name{h_hierarchical_pooled_nodes}
\alias{h_hierarchical_pooled_nodes}
\title{Find Pooled Nodes for an Arm}
\usage{
h_hierarchical_pooled_nodes(model, arm_name, refs, pool_names)
}
\arguments{
\item{model}{(\code{GeneralModel})\cr arm-specific model object.}

\item{arm_name}{(\code{string})\cr hierarchical arm name.}

\item{refs}{(\code{character})\cr hierarchical references.}

\item{pool_names}{(\code{character})\cr pool names matching \code{refs}.}
}
\value{
Character vector of namespaced nodes with exchangeable priors.
}
\description{
Find Pooled Nodes for an Arm
}
\keyword{internal}
