% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchicalModel-class.R
\name{h_hierarchical_prior_relations}
\alias{h_hierarchical_prior_relations}
\title{Collect Prior Relations}
\usage{
h_hierarchical_prior_relations(expr, operator)
}
\arguments{
\item{expr}{(\code{language})\cr prior model body.}

\item{operator}{(\code{string})\cr relation operator, either \code{"~"} or \code{"<-"}.}
}
\value{
A list of relation calls.
}
\description{
Collect Prior Relations
}
\keyword{internal}
