% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchicalModel-class.R
\name{h_hierarchical_reference_expr}
\alias{h_hierarchical_reference_expr}
\title{Create the Namespaced Expression for a Parameter Reference}
\usage{
h_hierarchical_reference_expr(model, arm_name, ref)
}
\arguments{
\item{model}{(\code{GeneralModel})\cr arm-specific model object.}

\item{arm_name}{(\code{string})\cr the user-facing arm name.}

\item{ref}{(\code{string})\cr parameter reference.}
}
\value{
A symbol or indexed call for the public sample node.
}
\description{
Create the Namespaced Expression for a Parameter Reference
}
\keyword{internal}
