% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchicalModel-class.R
\name{h_hierarchical_reference_stochastic_node}
\alias{h_hierarchical_reference_stochastic_node}
\title{Infer the Stochastic Node for a Parameter Reference}
\usage{
h_hierarchical_reference_stochastic_node(model, arm_name, ref)
}
\arguments{
\item{model}{(\code{GeneralModel})\cr arm-specific model object.}

\item{arm_name}{(\code{string})\cr the user-facing arm name.}

\item{ref}{(\code{string})\cr parameter reference.}
}
\value{
Character scalar naming the stochastic node to pool.
}
\description{
Infer the Stochastic Node for a Parameter Reference
}
\keyword{internal}
