% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchicalModel-class.R
\name{h_hierarchical_remove_pooled_prior_lines}
\alias{h_hierarchical_remove_pooled_prior_lines}
\title{Remove Fixed Priors for Pooled Nodes}
\usage{
h_hierarchical_remove_pooled_prior_lines(model_fun, pooled_nodes)
}
\arguments{
\item{model_fun}{(\code{function})\cr namespaced prior model function.}

\item{pooled_nodes}{(\code{character})\cr namespaced nodes with exchangeable priors.}
}
\value{
A function with fixed prior distribution statements removed.
}
\description{
Remove Fixed Priors for Pooled Nodes
}
\keyword{internal}
