% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchicalModel-class.R
\name{h_hierarchical_root_symbols}
\alias{h_hierarchical_root_symbols}
\title{Find the Root Symbol of an Expression}
\usage{
h_hierarchical_root_symbols(expr)
}
\arguments{
\item{expr}{(\code{language})\cr expression to inspect.}
}
\value{
Character vector of root symbols.
}
\description{
Find the Root Symbol of an Expression
}
\keyword{internal}
