% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchicalModel-class.R
\name{h_hierarchical_stochastic_subexpressions}
\alias{h_hierarchical_stochastic_subexpressions}
\title{Find Stochastic Subexpressions in an Expression}
\usage{
h_hierarchical_stochastic_subexpressions(expr, stochastic_roots)
}
\arguments{
\item{expr}{(\code{language})\cr expression to inspect.}

\item{stochastic_roots}{(\code{character})\cr stochastic node roots.}
}
\value{
Named list of stochastic subexpressions keyed by expression text.
}
\description{
Find Stochastic Subexpressions in an Expression
}
\keyword{internal}
