% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_simulations.R
\name{h_histogram_percentages}
\alias{h_histogram_percentages}
\title{Convenience Function to Make Histograms of Percentages}
\usage{
h_histogram_percentages(x, description, bins = 30L)
}
\arguments{
\item{x}{(\code{numeric})
vector of samples.}

\item{description}{(\code{string})
x-axis label.}

\item{bins}{(\code{count})
number of histogram bins.}
}
\value{
A \code{ggplot2} object.
}
\description{
Creates a histogram where the height of each bin is the percentage of all
observations in that bin.
}
\keyword{internal}
