% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchicalDesign-class.R
\name{h_historical_arm_design}
\alias{h_historical_arm_design}
\title{Construct a simplified DesignArm with hardcoded rule objects}
\usage{
h_historical_arm_design(data, model)
}
\arguments{
\item{data}{(\code{Data} or \code{DataCombo})\cr arm data for historical arms.}

\item{model}{(\code{GeneralModel} or \code{TwoDrugsCombo})\cr arm model for historical arms.}
}
\value{
A \code{Design} object with hardcoded rules for historical arms. The rules
don't matter because they won't be used, but they are required to construct a \code{Design} object.
}
\description{
Construct a simplified DesignArm with hardcoded rule objects
}
\keyword{internal}
