% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_jags.R
\name{h_jags_extract_samples}
\alias{h_jags_extract_samples}
\title{Extracting Samples from \code{JAGS} \code{mcarray} Object}
\usage{
h_jags_extract_samples(x)
}
\arguments{
\item{x}{an \code{\link[rjags:mcarray.object]{rjags::mcarray.object}} object.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

A simple helper function that extracts a sample from
\code{\link[rjags:mcarray.object]{rjags::mcarray.object}} S3 class object. The \code{\link[rjags:mcarray.object]{rjags::mcarray.object}}
object is used by the \code{\link[rjags:jags.samples]{rjags::jags.samples()}} function to represent MCMC
output from a \code{JAGS} model.
}
