% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_jags.R
\name{h_jags_get_data}
\alias{h_jags_get_data}
\title{Getting Data for \code{JAGS}}
\usage{
h_jags_get_data(model, data, from_prior)
}
\arguments{
\item{model}{(\code{GeneralModel})\cr an input model.}

\item{data}{(\code{GeneralData})\cr an input data.}

\item{from_prior}{(\code{flag})\cr sample from the prior only? In this case
data will not be appended to the output.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

A simple helper function that prepares an object for \code{data} argument of
\code{\link[rjags:jags.model]{rjags::jags.model()}}, which is invoked by \code{\link[=mcmc]{mcmc()}} method.
}
