% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_jags.R
\name{h_jags_join_models}
\alias{h_jags_join_models}
\title{Joining \code{JAGS} Models}
\usage{
h_jags_join_models(model1, model2)
}
\arguments{
\item{model1}{(\code{function})\cr the first model to join.}

\item{model2}{(\code{function})\cr the second model to join.}
}
\value{
joined models.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This helper function joins two JAGS models in the way that the body of the
second model is appended to the body of the first model (in this order).
After that, the first, body-extended model is returned.
}
\note{
\code{model1} and \code{model2} functions must have a multi-expression
body, i.e. braced expression(s). Environments or any attributes of the
function bodies are not preserved in any way after joining.
}
