% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{h_join_models}
\alias{h_join_models}
\title{Joining JAGS Models}
\usage{
h_join_models(model1, model2)
}
\arguments{
\item{model1}{(\code{function})\cr the first model to join.}

\item{model2}{(\code{function})\cr the second model to join.}
}
\value{
joined models.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This helper function joins two JAGS models in the way that the body of the
second model is appended into the body of the first model (in this order).
After that, the first, body-extended model is returned.
}
\note{
The functions behind \code{model1} and \code{model2} must have a multi-expression
body, i.e. braced expression(s). Environments or any attributes of the
function bodies are not preserved in any way after joining.
}
