% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_knitr.R
\name{h_kable_param_default}
\alias{h_kable_param_default}
\title{Set Default Values for kable Parameters}
\usage{
h_kable_param_default(param, col.names = NULL, caption = NULL)
}
\arguments{
\item{param}{(\code{list})\cr The list of \code{...} parameters passed to \code{knit_print}}

\item{col.names}{(\code{character})\cr Default column names for the table}

\item{caption}{(\code{character})\cr Default caption for the table}
}
\value{
The updated parameter list with defaults applied
}
\description{
Sets default values for \code{col.names} and \code{caption} if not already provided
by the user in the parameter list.
}
\keyword{internal}
