% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_knitr_GeneralModel.R
\name{h_knit_print_render_model_params}
\alias{h_knit_print_render_model_params}
\title{Render the Parameters of a Normal Distribution}
\usage{
h_knit_print_render_model_params(x, use_values = TRUE, fmt = "\%5.2f", ...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\keyword{internal}
