% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{h_mcmc_get_hierarchical_arm_samples}
\alias{h_mcmc_get_hierarchical_arm_samples}
\title{Build Arm-Specific Sample Name Mappings for a \code{HierarchicalModel}}
\usage{
h_mcmc_get_hierarchical_arm_samples(model)
}
\arguments{
\item{model}{(\code{HierarchicalModel})\cr hierarchical model object.}
}
\value{
Named list mapping arm-level sample names to hierarchical sample
names.
}
\description{
Build Arm-Specific Sample Name Mappings for a \code{HierarchicalModel}
}
\keyword{internal}
