% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_model.R
\name{h_model_dual_endpoint_rho}
\alias{h_model_dual_endpoint_rho}
\title{Prepare \code{DualEndpoint} class model components with regard to DLT and biomarker
correlation.}
\usage{
h_model_dual_endpoint_rho(use_fixed, rho, priormodel, modelspecs, init, sample)
}
\arguments{
\item{use_fixed}{(\code{flag})\cr TODO}

\item{rho}{(\code{numeric})\cr TODO}

\item{priormodel}{(\code{function})\cr TODO}

\item{modelspecs}{(\code{function})\cr TODO}

\item{init}{(\code{function})\cr TODO}

\item{sample}{(\code{character})\cr TODO}
}
\value{
\code{list} with updated model components.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

A simple helper function that takes \code{\link{DualEndpoint}} model existing components
(\code{priormodel}, \code{modelspecs}, \code{init}, \code{sample}), and updates them with regard to
DLT and biomarker correlation \code{rho}.
}
