% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_model.R
\name{h_model_dual_endpoint_sigma2W}
\alias{h_model_dual_endpoint_sigma2W}
\title{Prepare \code{DualEndpoint} class model components with regard to biomarker
regression variance.}
\usage{
h_model_dual_endpoint_sigma2W(
  use_fixed,
  sigma2W,
  priormodel,
  modelspecs,
  init,
  sample
)
}
\arguments{
\item{use_fixed}{(\code{flag})\cr TODO}

\item{sigma2W}{(\code{numeric})\cr TODO}

\item{priormodel}{(\code{function})\cr TODO}

\item{modelspecs}{(\code{function})\cr TODO}

\item{init}{(\code{function})\cr TODO}

\item{sample}{(\code{character})\cr TODO}
}
\value{
\code{list} with updated model components.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

A simple helper function that takes \code{\link{DualEndpoint}} model existing components
(\code{priormodel}, \code{modelspecs}, \code{init}, \code{sample}), and updates them with regard to
to biomarker regression variance \code{sigma2W}.
}
