% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_rules.R
\name{h_next_best_marker}
\alias{h_next_best_marker}
\title{Add the Next-Best Dose Marker}
\usage{
h_next_best_marker(
  plot,
  dose,
  y,
  dose_grid,
  dose_scale,
  label = "Dose Recommendation",
  colour = "blue"
)
}
\arguments{
\item{plot}{(\code{ggplot}) plot to decorate.}

\item{dose}{(\code{number}) dose to mark.}

\item{y}{(\code{number}) vertical marker position.}

\item{dose_grid}{(\code{numeric}) ordered dose grid.}

\item{dose_scale}{(\code{string}) dose-axis scale.}

\item{label}{(\code{string}) marker label in the legend.}

\item{colour}{(\code{string}) marker colour.}
}
\value{
The decorated \code{ggplot2} object.
}
\description{
Add the Next-Best Dose Marker
}
\keyword{internal}
