% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_rules.R
\name{h_next_best_mg_plot}
\alias{h_next_best_mg_plot}
\title{Building the Plot for nextBest-NextBestMaxGain Method.}
\usage{
h_next_best_mg_plot(
  prob_target_drt,
  dose_target_drt,
  prob_target_eot,
  dose_target_eot,
  dose_mg,
  max_gain,
  next_dose,
  doselimit,
  data,
  model,
  model_eff
)
}
\arguments{
\item{prob_target_drt}{(\code{proportion})\cr target DLT probability during the trial.}

\item{dose_target_drt}{(\code{number})\cr target dose estimate during the trial.}

\item{prob_target_eot}{(\code{proportion})\cr target DLT probability at the end of the trial.}

\item{dose_target_eot}{(\code{number})\cr target dose estimate at the end of the trial.}

\item{dose_mg}{(\code{number})\cr the dose corresponding to the maximum gain.}

\item{max_gain}{(\code{number})\cr the dose corresponding to the maximum gain.}

\item{next_dose}{(\code{number})\cr next best dose.}

\item{doselimit}{(\code{number})\cr the maximum allowed next dose.}

\item{data}{(\code{DataDual})\cr the data object from which the dose grid will be fetched.}

\item{model}{(\code{ModelTox})\cr the DLT model.}

\item{model_eff}{(\code{Effloglog})\cr the efficacy model.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Helper function which creates the plot for \code{\link[=nextBest-NextBestMaxGain]{nextBest-NextBestMaxGain()}} method.
}
