% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_rules.R
\name{h_next_best_mgsamples_plot}
\alias{h_next_best_mgsamples_plot}
\title{Building the Plot for nextBest-NextBestMaxGainSamples Method.}
\usage{
h_next_best_mgsamples_plot(
  prob_target_drt,
  dose_target_drt,
  prob_target_eot,
  dose_target_eot,
  dose_mg,
  dose_mg_samples,
  next_dose,
  doselimit,
  dose_grid_range
)
}
\arguments{
\item{prob_target_drt}{(\code{proportion})\cr target DLT probability during the trial.}

\item{dose_target_drt}{(\code{number})\cr target dose estimate during the trial.}

\item{prob_target_eot}{(\code{proportion})\cr target DLT probability at the end of the trial.}

\item{dose_target_eot}{(\code{number})\cr target dose estimate at the end of the trial.}

\item{dose_mg}{(\code{number})\cr the dose corresponding to the maximum gain.}

\item{dose_mg_samples}{(\code{numeric})\cr for every sample, the dose (from the dose grid)
that gives the maximum gain value.}

\item{next_dose}{(\code{number})\cr next best dose.}

\item{doselimit}{(\code{number})\cr the maximum allowed next dose.}

\item{dose_grid_range}{(\code{numeric})\cr dose grid range.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Helper function which creates the plot for \code{\link[=nextBest-NextBestMaxGainSamples]{nextBest-NextBestMaxGainSamples()}} method.
}
