% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_rules.R
\name{h_next_best_td_plot}
\alias{h_next_best_td_plot}
\title{Building the Plot for nextBest-NextBestTD Method.}
\usage{
h_next_best_td_plot(
  prob_target_drt,
  dose_target_drt,
  prob_target_eot,
  dose_target_eot,
  data,
  prob_dlt,
  doselimit,
  next_dose
)
}
\arguments{
\item{prob_target_drt}{(\code{proportion})\cr target DLT probability during the trial.}

\item{dose_target_drt}{(\code{number})\cr target dose estimate during the trial.}

\item{prob_target_eot}{(\code{proportion})\cr target DLT probability at the end of the trial.}

\item{dose_target_eot}{(\code{number})\cr target dose estimate at the end of the trial.}

\item{data}{(\code{Data})\cr the data object from which the dose grid will be fetched.}

\item{prob_dlt}{(\code{numeric})\cr}

\item{doselimit}{(\code{number})\cr the maximum allowed next dose.}

\item{next_dose}{(\code{number})\cr next best dose.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Helper function which creates the plot for nextBest-NextBestTD Method.
}
