% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{h_null_if_scalar}
\alias{h_null_if_scalar}
\title{Getting \code{NULL} for scalar object}
\usage{
h_null_if_scalar(x)
}
\arguments{
\item{x}{any object for which length function is defined.}
}
\value{
\code{NULL} if \code{x} is of length 1, otherwise, 1L.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

A simple helper function that returns \code{NULL} if \code{x} is a scalar object, i.e.
an object of length equals 1. Otherwise it returns 1L.
}
\examples{
h_null_if_scalar(c(1, 3))
h_null_if_scalar(2)
}
