% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_data.R
\name{h_obtain_dose_grid_range}
\alias{h_obtain_dose_grid_range}
\title{Helper Function Containing Common Functionality}
\usage{
h_obtain_dose_grid_range(object, ignore_placebo)
}
\arguments{
\item{object}{(\code{Data} or \code{DataOrdinal})\cr the object for which the dose grid
range is required}

\item{ignore_placebo}{(\code{flag})\cr should placebo dose (if any) not be counted?}
}
\description{
Used by \code{dose_grid_range-Data} and \code{dose_grid_range-DataOrdinal}
}
