% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design-methods.R
\name{h_overdose_threshold}
\alias{h_overdose_threshold}
\alias{h_overdose_threshold,NextBest-method}
\alias{h_overdose_threshold,NextBestDualEndpoint-method}
\title{Obtain the Toxicity Threshold Used for Simulation Overdose Reporting}
\usage{
h_overdose_threshold(next_best)

\S4method{h_overdose_threshold}{NextBest}(next_best)

\S4method{h_overdose_threshold}{NextBestDualEndpoint}(next_best)
}
\arguments{
\item{next_best}{(\code{\link{NextBest}})\cr next-best rule defining the threshold.}
}
\value{
A single toxicity probability threshold.
}
\description{
Obtain the Toxicity Threshold Used for Simulation Overdose Reporting
}
\section{Functions}{
\itemize{
\item \code{h_overdose_threshold(NextBest)}: use the upper toxicity target boundary.

\item \code{h_overdose_threshold(NextBestDualEndpoint)}: use the lower boundary of the toxicity
overdose interval instead of the biomarker target.

}}
\keyword{internal}
