% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations-methods.R
\name{h_plot_combo_evolution}
\alias{h_plot_combo_evolution}
\title{Helper Function to Plot 2D Combination Evolution}
\usage{
h_plot_combo_evolution(sim_data)
}
\arguments{
\item{sim_data}{(\code{list})
list of \code{\link{DataCombo}} objects.}
}
\value{
A \code{ggplot} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Creates a 2D combination-plane plot with arrows showing step-to-step
dose transitions aggregated across simulations.
}
\keyword{internal}
