% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_knitr.R
\name{h_prepare_labels}
\alias{h_prepare_labels}
\title{Check That Labels Are Valid and Useful}
\usage{
h_prepare_labels(x)
}
\arguments{
\item{x}{(\code{character})\cr The vector to be checked}
}
\value{
a character vector of length 2 whose values are distinct
}
\description{
A vector of labels is valid and useful if it is of length 2, of type character
and its values are distinct.
}
\details{
If \code{x} is a scalar, a second element is added, whose value is the value of the
scalar with "s" appended.  If \code{x} is \code{"toxicity"}, the plural is handled appropriately.
}
\keyword{internal}
