% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_knitr.R
\name{h_prepare_units}
\alias{h_prepare_units}
\title{Append Units to a Numeric Dose}
\usage{
h_prepare_units(units = NA)
}
\arguments{
\item{units}{(\code{character})\cr the units to be displayed}
}
\value{
if \code{units} is \code{NA}, then \code{NA}.  Otherwise, \code{units}, ensuring that exactly
one space precedes the first non-whitespace character
}
\description{
\code{d} may be a vector
}
\keyword{internal}
