% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations-methods.R
\name{h_pseudo_sim_fit_summary}
\alias{h_pseudo_sim_fit_summary}
\title{Helper Function to Calculate Fit Summary}
\usage{
h_pseudo_sim_fit_summary(fit_list, x_most_selected, dose_grid, truth)
}
\arguments{
\item{fit_list}{(\code{list})\cr list of fit objects from simulations.}

\item{x_most_selected}{(\code{integer})\cr index of dose most often selected.}

\item{dose_grid}{(\code{numeric})\cr dose grid.}

\item{truth}{(\code{function})\cr truth function.}
}
\value{
A list with \code{fit_at_dose_most_selected} and \code{mean_fit} components.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Calculates fit summary statistics for pseudo simulations.
}
\keyword{internal}
