% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations-methods.R
\name{h_pseudo_sim_inverse_dose}
\alias{h_pseudo_sim_inverse_dose}
\title{Helper Function to Calculate Inverse Dose}
\usage{
h_pseudo_sim_inverse_dose(f, lower = -100, upper = 100)
}
\arguments{
\item{f}{(\code{function})\cr the truth function mapping dose to probability.}

\item{lower}{(\code{number})\cr lower bound for root finding.}

\item{upper}{(\code{number})\cr upper bound for root finding.}
}
\value{
A function that takes a probability and returns the corresponding
dose.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Creates an inverse function to find the dose corresponding to a target
probability.
}
\keyword{internal}
