% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromQuantiles.R
\name{h_quantiles_target_function}
\alias{h_quantiles_target_function}
\title{Target Function for Quantiles Optimization}
\usage{
h_quantiles_target_function(
  dosegrid,
  refDose,
  lower,
  median,
  upper,
  level,
  logNormal,
  seed
)
}
\arguments{
\item{dosegrid}{(\code{numeric})\cr dose grid.}

\item{refDose}{(\code{number})\cr reference dose.}

\item{lower}{(\code{numeric})\cr lower quantiles.}

\item{median}{(\code{numeric})\cr median quantiles.}

\item{upper}{(\code{numeric})\cr upper quantiles.}

\item{level}{(\code{number})\cr credible level.}

\item{logNormal}{(\code{flag})\cr use log-normal prior?}

\item{seed}{(\code{count})\cr random seed.}
}
\value{
Function that computes target value for optimization.
}
\description{
Target Function for Quantiles Optimization
}
\keyword{internal}
