% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{h_set_env}
\alias{h_set_env}
\title{Setting Environment for a Function or Formula}
\usage{
h_set_env(fun, env)
}
\arguments{
\item{fun}{(\code{function} or \code{formula})\cr an object for which an environment is set.}

\item{env}{(\code{environment})\cr an environment to be set for \code{fun}.}
}
\value{
\code{fun} with environment \code{env} set.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

A simpler helper function that sets an enclosing environment for a given
function or formula.
}
