% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_design.R
\name{h_simulations_output_format}
\alias{h_simulations_output_format}
\title{Helper Function to create return list for Simulations output}
\usage{
h_simulations_output_format(resultList)
}
\arguments{
\item{resultList}{(\code{list})\cr raw iteration output.}
}
\value{
aggregated output for simulation object \code{list}.
}
\description{
Helper Function to create return list for Simulations output
}
\keyword{internal}
