% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_design.R
\name{h_this_truth}
\alias{h_this_truth}
\title{Helper Function to call truth calculation}
\usage{
h_this_truth(dose, this_args, truth)
}
\arguments{
\item{dose}{(\code{number})\cr current dose.}

\item{this_args}{(\code{data.frame})\cr list of arguments for the truth.}

\item{truth}{(\code{function})\cr defines the true probability for a DLT at a dose.}
}
\value{
The updated \code{this_truth}.
}
\description{
Helper Function to call truth calculation
}
