% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-class.R
\name{h_tite_logistic_datamodel}
\alias{h_tite_logistic_datamodel}
\title{Build the common JAGS likelihood for TITE logistic CRM models.}
\usage{
h_tite_logistic_datamodel(dose_term)
}
\arguments{
\item{dose_term}{(\code{name} or \code{call})\cr expression defining the dose
transformation used in the linear predictor.}
}
\value{
A function implementing the shared zero-trick JAGS likelihood.
}
\description{
Build the common JAGS likelihood for TITE logistic CRM models.
}
\keyword{internal}
