% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-class.R
\name{h_tite_logistic_weights}
\alias{h_tite_logistic_weights}
\title{Calculate weights for TITE logistic CRM models.}
\usage{
h_tite_logistic_weights(u, Tmax, y, weight_method)
}
\arguments{
\item{u}{(\code{numeric})\cr follow-up times.}

\item{Tmax}{(\code{number})\cr DLT assessment window.}

\item{y}{(\code{integer})\cr DLT indicators.}

\item{weight_method}{(\code{string})\cr either \code{"linear"} or \code{"adaptive"}.}
}
\value{
A numeric vector of observation weights.
}
\description{
Calculate weights for TITE logistic CRM models.
}
\keyword{internal}
