% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{h_unpack_stopit}
\alias{h_unpack_stopit}
\title{unpack stopping rules and return list}
\usage{
h_unpack_stopit(stopit_tree)
}
\arguments{
\item{stopit_tree}{object from simulate method}
}
\description{
recursively unpack nested stopping rules logical value and label given
}
