% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_design.R
\name{h_update_backfill_queue}
\alias{h_update_backfill_queue}
\title{Helper Function to Update Backfill Queue}
\usage{
h_update_backfill_queue(backfill_cohorts, data, dose, backfill)
}
\arguments{
\item{backfill_cohorts}{(\code{list})\cr current queue of backfill cohorts.}

\item{data}{(\code{Data})\cr current trial data.}

\item{dose}{(\code{number})\cr current dose being evaluated.}

\item{backfill}{(\code{Backfill})\cr with opening rules and cohort size.}
}
\value{
Updated \code{backfill_cohorts} list.
}
\description{
This function manages the backfill cohort queue by determining which
previous cohorts are open or closed for backfill enrollment.
}
\keyword{internal}
