% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{h_validate_combine_results}
\alias{h_validate_combine_results}
\title{Combining S4 Class Validation Results}
\usage{
h_validate_combine_results(v1, v2)
}
\arguments{
\item{v1}{(\code{logical} or \code{character})\cr an output from \code{result()} function from
\code{\link[=Validate]{Validate()}} reference class, to be combined with \code{v2}.}

\item{v2}{(\code{logical} or \code{character})\cr an output from \code{result()} function from
\code{\link[=Validate]{Validate()}} reference class, to be combined with \code{v1}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

A simple helper function that combines two outputs from calls to \code{result()}
function which is placed in a slot of \code{\link[=Validate]{Validate()}} reference class.
}
\examples{
h_validate_combine_results(TRUE, "some_message")
}
