% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-validity.R
\name{h_validate_common_data_slots}
\alias{h_validate_common_data_slots}
\title{Helper Function performing validation Common to Data and DataOrdinal}
\usage{
h_validate_common_data_slots(object)
}
\arguments{
\item{object}{(\code{Data} or \code{DataOrdinal})\cr the object to be validated}
}
\value{
a \code{Validate} object containing the result of the validation
}
\description{
Helper Function performing validation Common to Data and DataOrdinal
}
