% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-class.R
\docType{class}
\name{HierarchicalData-class}
\alias{HierarchicalData-class}
\alias{.HierarchicalData}
\alias{HierarchicalData}
\alias{.DefaultHierarchicalData}
\title{\code{HierarchicalData}}
\usage{
HierarchicalData(...)

.DefaultHierarchicalData()
}
\arguments{
\item{...}{named \code{Data} objects, one for each arm in the trial.
The argument names are used as the arm names.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{\link{HierarchicalData}} is a class for hierarchical designs,
storing the data for all arms together.
}
\section{Slots}{

\describe{
\item{\code{arms}}{(\code{list})\cr a named list of \code{Data} objects, one
for each arm in the trial. The names of the list are the
arm names.}
}}

\note{
Typically, end users will not use the \code{.DefaultHierarchicalData()} function
directly.
}
\examples{
dat_hierarchical <- HierarchicalData(
  mono = .DefaultData(),
  combo = .DefaultDataCombo()
)
}
