% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design-class.R
\docType{class}
\name{HierarchicalDesign-class}
\alias{HierarchicalDesign-class}
\alias{.HierarchicalDesign}
\alias{HierarchicalDesign}
\alias{.DefaultHierarchicalDesign}
\title{\code{HierarchicalDesign}}
\usage{
HierarchicalDesign(..., exchangeable_parameters = list())

.DefaultHierarchicalDesign()
}
\arguments{
\item{...}{\code{\link{DesignArm}} objects describing the trial arms.}

\item{exchangeable_parameters}{(\code{list})\cr see
\code{\link[=HierarchicalModel]{HierarchicalModel()}}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{\link{HierarchicalDesign}} stores a collection of named \code{\link{DesignArm}} objects
and derives the corresponding \code{\link{HierarchicalData}} and \code{\link{HierarchicalModel}}
objects from them.
}
\section{Slots}{

\describe{
\item{\code{arms}}{(\code{list})\cr a named list of \code{\link{DesignArm}} objects.}

\item{\code{data}}{(\code{HierarchicalData})\cr the hierarchical data object assembled
from the arm-specific design data.}

\item{\code{model}}{(\code{HierarchicalModel})\cr the hierarchical model object assembled
from the arm-specific design models.}
}}

\note{
Typically, end users will not use the \code{.DefaultHierarchicalDesign()}
function directly.
}
