% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations-class.R
\docType{class}
\name{HierarchicalSimulations-class}
\alias{HierarchicalSimulations-class}
\alias{.HierarchicalSimulations}
\alias{HierarchicalSimulations}
\title{\code{HierarchicalSimulations}}
\usage{
HierarchicalSimulations(
  data,
  doses,
  samples,
  fit,
  stop_reasons,
  stop_report,
  additional_stats,
  seed
)
}
\arguments{
\item{data}{(\code{list})\cr see slot definition.}

\item{doses}{(\code{list})\cr see slot definition.}

\item{samples}{(\code{list})\cr see slot definition.}

\item{fit}{(\code{list})\cr see slot definition.}

\item{stop_reasons}{(\code{list})\cr see slot definition.}

\item{stop_report}{(\code{list})\cr see slot definition.}

\item{additional_stats}{(\code{list})\cr see slot definition.}

\item{seed}{(\code{integer})\cr see slot definition.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This class captures trial simulations from hierarchical designs
(\code{\link{HierarchicalDesign}}).
}
\section{Slots}{

\describe{
\item{\code{data}}{(\code{list})\cr produced \code{\link{HierarchicalData}} objects.}

\item{\code{doses}}{(\code{list})\cr final dose recommendations, one named list per
simulation run with entries for each hierarchical arm.}

\item{\code{samples}}{(\code{list})\cr final \code{\link{HierarchicalSamples}} objects.}

\item{\code{fit}}{(\code{list})\cr final arm-specific fits for each simulation run.}

\item{\code{stop_reasons}}{(\code{list})\cr arm-specific stopping reasons for each
simulation run.}

\item{\code{stop_report}}{(\code{list})\cr arm-specific stopping rule outcomes for each
simulation run.}

\item{\code{additional_stats}}{(\code{list})\cr additional arm-specific statistics.}

\item{\code{seed}}{(\code{integer})\cr random generator state before starting the
simulation.}
}}

