% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations-class.R
\docType{class}
\name{HierarchicalSimulationsSummary-class}
\alias{HierarchicalSimulationsSummary-class}
\alias{.HierarchicalSimulationsSummary}
\alias{HierarchicalSimulationsSummary}
\alias{.DefaultHierarchicalSimulationsSummary}
\title{\code{HierarchicalSimulationsSummary}}
\usage{
.DefaultHierarchicalSimulationsSummary()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This class captures arm-specific summaries from \code{\link{HierarchicalSimulations}}
objects.
}
\section{Slots}{

\describe{
\item{\code{arms}}{(\code{list})\cr named list of arm-level simulation summary objects.}

\item{\code{nsim}}{(\code{integer})\cr number of simulations.}
}}

\note{
Typically, end users will not use the \code{.DefaultHierarchicalSimulationsSummary()} function.
}
