% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\name{IncrementMin}
\alias{IncrementMin}
\title{Initialization function for "IncrementMin"}
\usage{
IncrementMin(IncrementsList)
}
\arguments{
\item{IncrementsList}{see \code{\linkS4class{IncrementMin}}}
}
\value{
the \code{\linkS4class{IncrementMin}} object
}
\description{
Initialization function for "IncrementMin"
}
\keyword{methods}
