\docType{class}
\name{Increments-class}
\alias{Increments-class}
\title{The virtual class for controlling increments}
\description{
The virtual class for controlling increments
}
\seealso{
\code{\linkS4class{IncrementsRelative}},
\code{\linkS4class{IncrementsRelativeDLT}},
\code{\linkS4class{IncrementsRelativeParts}}
}
\keyword{classes}

