% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{IncrementsAbsoluteDLT-class}
\alias{IncrementsAbsoluteDLT-class}
\alias{.IncrementsAbsoluteDLT}
\alias{IncrementsAbsoluteDLT}
\title{\code{IncrementsAbsoluteDLT}}
\usage{
IncrementsAbsoluteDLT(intervals = 1:2, increments = 2:1)
}
\arguments{
\item{intervals}{(\code{numeric}) \cr a vector of right hand boundaries of the DLT intervals}

\item{increments}{(\code{numeric}) \cr a vector of the maximum increment allowed in the corresponding DLT interval}
}
\value{
the \link{IncrementsAbsoluteDLT} object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Increments control based on absolute differences in intervals determined by
the number of DLTS reported to date

#' Note that \code{intervals} is to be read as follows. If for example,
we want to specify three intervals: First 0 DLTs, second exactly one DLT, and
third two or more DLTs, then we specify \code{intervals} to be \code{1:3}.
That means, the right bound of the intervals are exclusive to the interval,
and the last interval goes from the last value to infinity.
}
\section{Slots}{

\describe{
\item{\code{intervals}}{a vector with the left bounds of the relevant DLT counts}

\item{\code{increments}}{a vector of the same length with the maximum allowable
absolute increments in the \code{intervals}}
}}

