% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{IncrementsNumDoseLevels-class}
\alias{IncrementsNumDoseLevels-class}
\alias{.IncrementsNumDoseLevels}
\alias{IncrementsNumDoseLevels}
\title{Increments Control Based on Number of Dose Levels}
\usage{
IncrementsNumDoseLevels(maxLevels = 1, basisLevel = "lastGiven")
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\section{Slots}{

\describe{
\item{\code{maxLevels}}{(\verb{scalar positive integer})\cr corresponding to the number of maximum
dose levels to increment for the next dose. It defaults to 1,
which means that no dose skipping is allowed - the next dose
can be maximum one level higher than the current dose.}

\item{\code{basisLevel}}{(\verb{character scalar})\cr corresponding to the dose level used to increment from.
It can take two possible values 'lastGiven' or 'maxGiven'. If lastGiven (default)
is specified the increments is applied to the last given dose and if
maxGiven is specified the increment is applied from the max given dose
level.}
}}

\examples{

# As example, here is the rule for: 
# maximum skip one dose level, that is 2 dose levels higher is maximum
# increment
myIncrements <- IncrementsNumDoseLevels(maxLevels=2)

}
\keyword{methods}
