% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\name{IncrementsNumDoseLevels}
\alias{IncrementsNumDoseLevels}
\title{Initialization function for "IncrementsNumDoseLevels"}
\usage{
IncrementsNumDoseLevels(maxLevels = 1)
}
\arguments{
\item{maxLevels}{see \code{\linkS4class{IncrementsNumDoseLevels}}}
}
\value{
the \code{\linkS4class{IncrementsNumDoseLevels}} object
}
\description{
Initialization function for "IncrementsNumDoseLevels"
}
\keyword{methods}
