% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\name{IncrementsRelative}
\alias{IncrementsRelative}
\title{Initialization function for "IncrementsRelative"}
\usage{
IncrementsRelative(intervals, increments)
}
\arguments{
\item{intervals}{see \code{\linkS4class{IncrementsRelative}}}

\item{increments}{see \code{\linkS4class{IncrementsRelative}}}
}
\value{
the \code{\linkS4class{IncrementsRelative}} object
}
\description{
Initialization function for "IncrementsRelative"
}
\keyword{methods}
