\docType{class}
\name{IncrementsRelativeDLT-class}
\alias{IncrementsRelativeDLT-class}
\title{Increments control based on relative differences in terms of DLTs}
\description{
Increments control based on relative differences in terms
of DLTs
}
\section{Slots}{

\describe{
\item{\code{DLTintervals}}{a vector with the bounds of the relevant DLT intervals of
length \code{n}}

\item{\code{increments}}{a vector of length \code{n-1} with the maximum allowable
relative increments in the \code{DLTintervals}}
}}
\keyword{classes}

