% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\name{IncrementsRelativeDLT}
\alias{IncrementsRelativeDLT}
\title{Initialization function for "IncrementsRelativeDLT"}
\usage{
IncrementsRelativeDLT(DLTintervals, increments)
}
\arguments{
\item{DLTintervals}{see \code{\linkS4class{IncrementsRelativeDLT}}}

\item{increments}{see \code{\linkS4class{IncrementsRelativeDLT}}}
}
\value{
the \code{\linkS4class{IncrementsRelativeDLT}} object
}
\description{
Initialization function for "IncrementsRelativeDLT"
}
\keyword{methods}
