% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{IncrementsRelativeDLTCumulative-class}
\alias{IncrementsRelativeDLTCumulative-class}
\alias{.IncrementsRelativeDLTCumulative}
\alias{IncrementsRelativeDLTCumulative}
\title{Increment Control based on Relative Differences and Cumulative DLTs}
\usage{
IncrementsRelativeDLTCumulative(DLTintervals, increments)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The \code{DLTintervals} is to be read as follows. If for example,
we want to specify three intervals: First 0 DLTs, second 1 or 2 DLTs, and
third at least 3 DLTs, then we specify \code{DLTintervals} to be \code{c(0, 1, 3)}.
That means, the right bound of the intervals are exclusive to the interval
-- the vector only gives the left bounds of the intervals.
The last interval goes from 3 to infinity.
}
\section{Slots}{

\describe{
\item{\code{DLTintervals}}{(\code{integer})\cr left bounds of the relevant
DLT intervals.}

\item{\code{increments}}{(\code{numeric})\cr corresponding maximum allowable
relative increments in the \code{DLTintervals}.}
}}

\seealso{
\code{\link[=IncrementsRelativeDLT]{IncrementsRelativeDLT()}}.
}
