% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\name{IncrementsRelativeParts}
\alias{IncrementsRelativeParts}
\title{Initialization function for "IncrementsRelativeParts"}
\usage{
IncrementsRelativeParts(dltStart, cleanStart, ...)
}
\arguments{
\item{dltStart}{see \code{\linkS4class{IncrementsRelativeParts}}}

\item{cleanStart}{see \code{\linkS4class{IncrementsRelativeParts}}}

\item{\dots}{additional slots from \code{\linkS4class{IncrementsRelative}}}
}
\value{
the \code{\linkS4class{IncrementsRelativeParts}} object
}
\description{
Initialization function for "IncrementsRelativeParts"
}
\keyword{methods}
