% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-class.R
\docType{methods}
\name{initialize,DualEndpointOld-method}
\alias{initialize,DualEndpointOld-method}
\title{Initialization method for the "DualEndpointOld" class}
\usage{
\S4method{initialize}{DualEndpointOld}(.Object, mu, Sigma, sigma2betaW,
  sigma2W, rho, smooth = c("RW1", "RW2"), ...)
}
\arguments{
\item{.Object}{the \code{\linkS4class{DualEndpointOld}} we want to
initialize}

\item{mu}{see \code{\linkS4class{DualEndpointOld}}}

\item{Sigma}{see \code{\linkS4class{DualEndpointOld}}}

\item{sigma2betaW}{see \code{\linkS4class{DualEndpointOld}}}

\item{sigma2W}{see \code{\linkS4class{DualEndpointOld}}}

\item{rho}{see \code{\linkS4class{DualEndpointOld}}}

\item{smooth}{either \dQuote{RW1} (default) or \dQuote{RW2}, for
specifying the random walk prior on the biomarker level.}

\item{\dots}{not used}
}
\description{
Initialization method for the "DualEndpointOld" class
}
\keyword{methods}
