% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is.bool}
\alias{is.bool}
\title{Predicate checking for a boolean option}
\usage{
is.bool(x)
}
\arguments{
\item{x}{the object being checked}
}
\value{
Returns \code{TRUE} if \code{x} is a length one logical vector (i.e., a
scalar)
}
\description{
Predicate checking for a boolean option
}
\keyword{internal}
