% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is.range}
\alias{is.range}
\title{Predicate checking for a numeric range}
\usage{
is.range(x)
}
\arguments{
\item{x}{the object being checked}
}
\value{
Returns \code{TRUE} if \code{x} is a numeric range
}
\description{
Predicate checking for a numeric range
}
\keyword{internal}
