% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is.scalar}
\alias{is.scalar}
\title{Checking for scalar}
\usage{
is.scalar(x)
}
\arguments{
\item{x}{the input}
}
\value{
Returns \code{TRUE} if \code{x} is a length one vector
(i.e., a scalar)
}
\description{
Checking for scalar
}
\keyword{internal}
