% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{checks for whole numbers (integers)}
\usage{
is.wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{the numeric vector}

\item{tol}{the tolerance}
}
\value{
TRUE or FALSE for each element of x
}
\description{
checks for whole numbers (integers)
}
\keyword{internal}
