% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_knitr_GeneralModel.R
\name{knit_print.ModelParamsNormal}
\alias{knit_print.ModelParamsNormal}
\title{Render a Normal Prior}
\usage{
\method{knit_print}{ModelParamsNormal}(
  x,
  use_values = TRUE,
  fmt = "\%5.2f",
  params = c("alpha", "beta"),
  preamble = "The prior for &theta; is given by\\\\n",
  asis = TRUE,
  theta = "\\\\theta",
  ...
)
}
\arguments{
\item{x}{(\code{ModelParamsNormal})\cr the object to be rendered}

\item{use_values}{(\code{flag})\cr print the values associated with hyperparameters,
or the symbols used to define the hyper-parameters.  That is, for example, mu or 1.}

\item{fmt}{(\code{character})\cr the \code{sprintf} format string used to render
numerical values.  Ignored if \code{use_values} is \code{FALSE}.}

\item{params}{(\code{character})\cr The names of the model parameters.  See Usage
Notes below.}

\item{preamble}{(\code{character})\cr The text used to introduce the LaTex representation
of the model}

\item{asis}{(\code{flag})\cr wrap the return value in a call to \code{knitr::asis_output}?}

\item{theta}{(\code{character})\cr the LaTex representation of the theta vector}

\item{...\cr}{Not used at present}
}
\value{
A character string containing a LaTeX rendition of the object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\section{Usage Notes}{

\code{params} must be a character vector of length equal to that of \code{x@mean} (and
\code{x@cov}).  Its values represent the parameters of the model as entries in the
vector \code{theta}, on the left-hand side of "~" in the definition of the prior.
If named, names should be valid LaTex, escaped as usual for R character variables.
For example, \code{"\\\\alpha"} or \code{"\\\\beta_0"}.  If unnamed, names are constructed by
pre-pending an escaped backslash to each value provided.
}

\keyword{internal}
