% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Model-class.R
\docType{class}
\name{LogisticIndepBeta-class}
\alias{.LogisticIndepBeta}
\alias{LogisticIndepBeta-class}
\title{class for the two-parameter logistic regression DLE model
This is the two-parameter logistic regression model using pseudo data prior
which describe the relationship of the probabilities of an occurence of DLE
with their corresponding log dose levels. These probabilities of occurence of a DLE
have independent Beta distributions}
\description{
class for the two-parameter logistic regression DLE model
This is the two-parameter logistic regression model using pseudo data prior
which describe the relationship of the probabilities of an occurence of DLE
with their corresponding log dose levels. These probabilities of occurence of a DLE
have independent Beta distributions
}
\keyword{classes}

