% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Model-class.R
\name{LogisticIndepBeta}
\alias{LogisticIndepBeta}
\title{Intialization function for "LogisticIndepBeta" class}
\usage{
LogisticIndepBeta(binDLE, DLEdose, DLEweights, data)
}
\arguments{
\item{binDLE}{the pseudo DLE responses}

\item{DLEdose}{the corresponding levels for the pseudo DLE responses}

\item{DLEweights}{number of subjects treated for this binDLE response}

\item{data}{the input data to update estimates of model parameters and
follow the \code{\linkS4class{Data}} object class specification}
}
\description{
Intialization function for "LogisticIndepBeta" class
}
\keyword{methods}

