% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-class.R
\name{LogisticKadane}
\alias{LogisticKadane}
\title{Initialization function for the "LogisticKadane" class}
\usage{
LogisticKadane(theta, xmin, xmax)
}
\arguments{
\item{theta}{the target toxicity probability}

\item{xmin}{the minimum of the dose range}

\item{xmax}{the maximum of the dose range}
}
\value{
the \code{\linkS4class{LogisticKadane}}
}
\description{
Initialization function for the "LogisticKadane" class
}
\keyword{methods}
