% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-class.R
\name{LogisticLogNormal}
\alias{LogisticLogNormal}
\title{Initialization function for the "LogisticLogNormal" class}
\usage{
LogisticLogNormal(mean, cov, refDose)
}
\arguments{
\item{mean}{the prior mean vector}

\item{cov}{the prior covariance matrix}

\item{refDose}{the reference dose}
}
\value{
the \code{\linkS4class{LogisticLogNormal}} object
}
\description{
Initialization function for the "LogisticLogNormal" class
}
\keyword{methods}
