% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-class.R
\name{LogisticLogNormalMixture}
\alias{LogisticLogNormalMixture}
\title{Initialization function for the "LogisticLogNormalMixture" class}
\usage{
LogisticLogNormalMixture(mean, cov, refDose, shareWeight)
}
\arguments{
\item{mean}{the prior mean vector}

\item{cov}{the prior covariance matrix}

\item{refDose}{the reference dose}

\item{shareWeight}{the prior weight for the share component}
}
\value{
the \code{\linkS4class{LogisticLogNormalMixture}} object
}
\description{
Initialization function for the "LogisticLogNormalMixture" class
}
\keyword{methods}
