% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-class.R
\docType{class}
\name{LogisticLogNormalOrdinal-class}
\alias{LogisticLogNormalOrdinal-class}
\alias{.LogisticLogNormalOrdinal}
\alias{OrdinalLogisticLogNormal}
\alias{LogisticLogNormalOrdinal}
\alias{.DefaultLogisticLogNormalOrdinal}
\title{\code{OrdinalLogisticLogNormal}}
\usage{
LogisticLogNormalOrdinal(mean, cov, ref_dose)

.DefaultLogisticLogNormalOrdinal()
}
\arguments{
\item{mean}{(\code{numeric})\cr the prior mean vector.}

\item{cov}{(\code{matrix})\cr the prior covariance matrix. The precision matrix
\code{prec} is internally calculated as an inverse of \code{cov}.}

\item{ref_dose}{(\code{number})\cr the reference dose \eqn{x*} (strictly positive
number).}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{\link{OrdinalLogisticLogNormal}} is the class for a logistic lognormal CRM model
using an ordinal toxicity scale.
}
\note{
Typically, end users will not use the \code{.DefaultLogisticLogNormalOrdinal()} function.
}
\examples{
LogisticLogNormalOrdinal(
  mean=c(-3, -4, 1),
  cov=diag(c(3, 4, 1)),
  ref_dose = 50
)
}
