% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-class.R
\docType{class}
\name{LogisticLogNormalSub-class}
\alias{LogisticLogNormalSub-class}
\alias{.LogisticLogNormalSub}
\alias{LogisticLogNormalSub}
\title{\code{LogisticLogNormalSub}}
\usage{
LogisticLogNormalSub(mean, cov, refDose)
}
\arguments{
\item{mean}{(\code{numeric})\cr the prior mean vector.}

\item{cov}{(\code{matrix})\cr the prior covariance matrix.}

\item{refDose}{(\code{number})\cr the reference dose.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{\link{LogisticLogNormalSub}} is the class for a standard logistic model with
bivariate (log) normal prior with subtractive dose standardization.
}
\details{
This is the usual logistic regression model with a bivariate normal
prior on the intercept and log slope. The covariate is the dose \eqn{x}
minus the reference dose \eqn{x^{*}}:
\deqn{logit[p(x)] = \alpha + \beta \cdot (x - x^{*})}, where \eqn{p(x)}
is the probability of observing a DLT for a given dose \eqn{x}.
The prior is \deqn{(\alpha, \log(\beta)) \sim Normal(\mu, \Sigma)}.
}
\section{Slots}{

\describe{
\item{\code{mean}}{(\code{numeric})\cr the prior mean vector.}

\item{\code{cov}}{(\code{matrix})\cr the prior covariance matrix.}

\item{\code{refDose}}{(\code{number})\cr the reference dose.}
}}

\examples{
my_model <- LogisticLogNormalSub(
  mean = c(-0.85, 1),
  cov = matrix(c(1, -0.5, -0.5, 1), nrow = 2),
  refDose = 50
)
}
\seealso{
\code{\link{LogisticLogNormal}}.
}
