% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-class.R
\name{LogisticNormal}
\alias{LogisticNormal}
\title{Initialization function for the "LogisticNormal" class}
\usage{
LogisticNormal(mean, cov, refDose)
}
\arguments{
\item{mean}{the prior mean vector}

\item{cov}{the prior covariance matrix}

\item{refDose}{the reference dose}
}
\value{
the \code{\linkS4class{LogisticNormal}} object
}
\description{
Initialization function for the "LogisticNormal" class
}
\keyword{methods}
