% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-class.R
\name{LogisticNormalFixedMixture}
\alias{LogisticNormalFixedMixture}
\title{Initialization function for the "LogisticNormalFixedMixture" class}
\usage{
LogisticNormalFixedMixture(components, weights, refDose, logNormal = FALSE)
}
\arguments{
\item{components}{the specifications of the mixture components: a list with
one list of \code{mean} and \code{cov} for each bivariate (log) normal prior}

\item{weights}{the weights of the components, these must be positive and
will be normalized to sum to 1}

\item{refDose}{the reference dose}

\item{logNormal}{should a log normal prior be specified, such that the mean
vectors and covariance matrices are valid for the intercept and log slope?
(not default)}
}
\value{
the \code{\linkS4class{LogisticNormalFixedMixture}} object
}
\description{
Initialization function for the "LogisticNormalFixedMixture" class
}
\keyword{methods}
