% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-class.R
\name{LogisticNormalMixture}
\alias{LogisticNormalMixture}
\title{Initialization function for the "LogisticNormalMixture" class}
\usage{
LogisticNormalMixture(comp1, comp2, weightpar, refDose)
}
\arguments{
\item{comp1}{the specifications of the first component: a list with
\code{mean} and \code{cov} for the first bivariate normal prior}

\item{comp2}{the specifications of the second component}

\item{weightpar}{the beta parameters for the weight of the first component}

\item{refDose}{the reference dose}
}
\value{
the \code{\linkS4class{LogisticNormalMixture}} object
}
\description{
Initialization function for the "LogisticNormalMixture" class
}
\keyword{methods}
