% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{logit}
\alias{logit}
\title{Shorthand for Logit Function}
\usage{
logit(x)
}
\arguments{
\item{x}{(\code{numeric})\cr the function argument.}
}
\value{
The logit of \code{x}, computed as \code{log(x / (1 - x))}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Computes the logit transformation.
}
\examples{
logit(0.5)
logit(c(0.1, 0.5, 0.9))
}
