\docType{methods}
\name{maxDose,IncrementsRelative,Data-method}
\alias{maxDose,IncrementsRelative,Data-method}
\title{Determine the maximum possible next dose based on relative increments}
\usage{
\S4method{maxDose}{IncrementsRelative,Data}(increments, data, ...)
}
\arguments{
  \item{increments}{The rule, an object of class
  \code{\linkS4class{Increments}}}

  \item{data}{The data input, an object of class
  \code{\linkS4class{Data}}}

  \item{...}{further arguments}
}
\description{
Determine the maximum possible next dose based on relative
increments
}

