\docType{methods}
\name{maxDose,IncrementsRelativeDLT,Data-method}
\alias{maxDose,IncrementsRelativeDLT,Data-method}
\title{Determine the maximum possible next dose based on relative increments
determined by DLTs so far}
\usage{
\S4method{maxDose}{IncrementsRelativeDLT,Data}(increments, data, ...)
}
\arguments{
  \item{increments}{The rule, an object of class
  \code{\linkS4class{Increments}}}

  \item{data}{The data input, an object of class
  \code{\linkS4class{Data}}}

  \item{...}{further arguments}
}
\description{
Determine the maximum possible next dose based on relative
increments determined by DLTs so far
}

