\docType{methods}
\name{maxDose,IncrementsRelativeParts,DataParts-method}
\alias{maxDose,IncrementsRelativeParts,DataParts-method}
\title{Determine the maximum possible next dose based on relative increments
and part 1 and 2}
\usage{
\S4method{maxDose}{IncrementsRelativeParts,DataParts}(increments, data, ...)
}
\arguments{
  \item{increments}{The rule, an object of class
  \code{\linkS4class{Increments}}}

  \item{data}{The data input, an object of class
  \code{\linkS4class{Data}}}

  \item{...}{further arguments}
}
\description{
Determine the maximum possible next dose based on relative
increments and part 1 and 2
}

