\docType{methods}
\name{maxSize,CohortSize-method}
\alias{maxSize,CohortSize-method}
\title{The method combining cohort size rules by taking maximum}
\usage{
\S4method{maxSize}{CohortSize}(...)
}
\arguments{
  \item{...}{Objects of class
  \code{\linkS4class{CohortSize}}}
}
\description{
The method combining cohort size rules by taking maximum
}

