\name{maxSize}
\alias{maxSize}
\title{"MAX" combination of cohort size rules}
\usage{
maxSize(...)
}
\arguments{
  \item{\dots}{Objects of class
  \code{\linkS4class{CohortSize}}}
}
\value{
the combination as an object of class
\code{\linkS4class{CohortSizeMax}}
}
\description{
This function combines cohort size rules by taking the
maximum of all sizes.
}
\seealso{
\code{\link{minSize}}
}
\keyword{methods}

