% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mcmc.R
\docType{methods}
\name{mcmc,Data,LogisticIndepBeta,McmcOptions-method}
\alias{mcmc,Data,LogisticIndepBeta,McmcOptions-method}
\title{Obtain postrior samples for the two-parameter logistic pseudo DLE model}
\usage{
\S4method{mcmc}{Data,LogisticIndepBeta,McmcOptions}(data, model, options,
  verbose = FALSE, ...)
}
\arguments{
\item{data}{The data input, an object of class \code{\linkS4class{Data}}}

\item{model}{The model input, an object of class \code{\linkS4class{LogisticIndepBeta}}}

\item{options}{MCMC options, an object of class
\code{\linkS4class{McmcOptions}}}

\item{\dots}{unused}
}
\value{
The posterior samples, an object of class
\code{\linkS4class{Samples}}.
}
\description{
Obtain postrior samples for the two-parameter logistic pseudo DLE model
}
\keyword{methods}

