\docType{methods}
\name{mcmc,Data,LogisticLogNormal,McmcOptions-method}
\alias{mcmc,Data,LogisticLogNormal,McmcOptions-method}
\title{The fast method for the LogisticLogNormal class}
\usage{
\S4method{mcmc}{Data,LogisticLogNormal,McmcOptions}(data, model, options,
  verbose = FALSE, ...)
}
\arguments{
  \item{verbose}{shall messages be printed? (not default)}

  \item{data}{The data input, an object of class
  \code{\linkS4class{Data}}}

  \item{model}{The model input, an object of class
  \code{\linkS4class{Model}}}

  \item{options}{MCMC options, an object of class
  \code{\linkS4class{McmcOptions}}}

  \item{...}{unused}
}
\description{
The fast method for the LogisticLogNormal class
}

