\docType{methods}
\name{mcmc,Data,LogisticNormalMixture,McmcOptions-method}
\alias{mcmc,Data,LogisticNormalMixture,McmcOptions-method}
\title{JAGS for the LogisticNormalMixture method}
\usage{
\S4method{mcmc}{Data,LogisticNormalMixture,McmcOptions}(data, model, options,
  verbose = FALSE, ...)
}
\arguments{
  \item{verbose}{shall messages be printed? (not default)}

  \item{data}{The data input, an object of class
  \code{\linkS4class{Data}}}

  \item{model}{The model input, an object of class
  \code{\linkS4class{Model}}}

  \item{options}{MCMC options, an object of class
  \code{\linkS4class{McmcOptions}}}

  \item{...}{unused}
}
\description{
JAGS for the LogisticNormalMixture method
}

