% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mcmc.R
\docType{methods}
\name{mcmc,DataDual,Effloglog,McmcOptions-method}
\alias{mcmc,DataDual,Effloglog,McmcOptions-method}
\title{Obtain the posterior samples for the efficacy linear log log model using pseudo
data as prior}
\usage{
\S4method{mcmc}{DataDual,Effloglog,McmcOptions}(data, model, options,
  verbose = FALSE, ...)
}
\description{
Obtain the posterior samples for the efficacy linear log log model using pseudo
data as prior
}
\keyword{methods}

