\docType{class}
\name{McmcOptions-class}
\alias{McmcOptions-class}
\title{Class for the three canonical MCMC options}
\description{
Class for the three canonical MCMC options
}
\section{Slots}{

\describe{
\item{\code{iterations}}{number of MCMC iterations}

\item{\code{burnin}}{number of burn-in iterations which are not saved}

\item{\code{step}}{only every step-th iteration is saved after the burn-in}
}}
\keyword{classes}

