% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McmcOptions-class.R
\name{McmcOptions}
\alias{McmcOptions}
\title{Initialization function for the "McmcOptions" class}
\usage{
McmcOptions(burnin = 10000L, step = 2L, samples = 10000L)
}
\arguments{
\item{burnin}{number of burn-in iterations which are not saved (default:
\code{10,000})}

\item{step}{only every step-th iteration is saved after the burn-in
(default: \code{2})}

\item{samples}{number of resulting samples (by default \code{10,000} will
result)}
}
\value{
the \code{\linkS4class{McmcOptions}} object
}
\description{
Initialization function for the "McmcOptions" class
}
\keyword{methods}
