\docType{methods}
\name{minSize,CohortSize-method}
\alias{minSize,CohortSize-method}
\title{The method combining cohort size rules by taking minimum}
\usage{
\S4method{minSize}{CohortSize}(...)
}
\arguments{
  \item{...}{Objects of class
  \code{\linkS4class{CohortSize}}}
}
\description{
The method combining cohort size rules by taking minimum
}

