\name{minSize}
\alias{minSize}
\title{"MIN" combination of cohort size rules}
\usage{
minSize(...)
}
\arguments{
  \item{\dots}{Objects of class
  \code{\linkS4class{CohortSize}}}
}
\value{
the combination as an object of class
\code{\linkS4class{CohortSizeMin}}
}
\description{
This function combines cohort size rules by taking the
minimum of all sizes.
}
\seealso{
\code{\link{maxSize}}
}
\keyword{methods}

