% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Model-class.R
\docType{class}
\name{ModelEff-class}
\alias{.ModelEff}
\alias{ModelEff-class}
\title{class for Efficacy models using pseudo data prior}
\description{
class for Efficacy models using pseudo data prior
}
\keyword{classes}

