% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-class.R
\docType{class}
\name{ModelNormal-class}
\alias{ModelNormal-class}
\alias{.ModelNormal}
\alias{ModelNormal}
\title{\code{ModelNormal}}
\usage{
ModelNormal(mean, cov, ref_dose)
}
\arguments{
\item{mean}{(\code{numeric})\cr the prior mean vector.}

\item{cov}{(\code{matrix})\cr the prior covariance matrix.}

\item{ref_dose}{(\code{number})\cr the reference dose.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{\link{ModelNormal}} is the class for a model with a bivariate normal prior on the
model parameters \code{alpha0} and \code{alpha1}, i.e.:
\deqn{(alpha0, alpha1) ~ Normal(mean, cov)} All other specific models
inherit from this class.
The slots of this class contain the mean vector, the covariance and
precision matrices of the bivariate normal distribution, as well as the
reference dose.
}
\section{Slots}{

\describe{
\item{\code{mean}}{(\code{numeric})\cr the prior mean vector.}

\item{\code{cov}}{(\code{matrix})\cr the prior covariance matrix.}

\item{\code{prec}}{the prior precision matrix, which is an inverse matrix of the \code{cov}.}

\item{\code{ref_dose}}{(\code{number})\cr the reference dose.}
}}

\seealso{
\code{\link{LogisticNormal}}, \code{\link{LogisticLogNormal}}, \code{\link{LogisticLogNormalSub}},
\code{\link{ProbitLogNormal}}, \code{\link{ProbitLogNormalLogDose}},
}
