% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Model-class.R
\docType{class}
\name{ModelTox-class}
\alias{.ModelTox}
\alias{ModelTox-class}
\title{Class for DLE models using pesudo data prior}
\description{
Class for DLE models using pesudo data prior
}
\keyword{classes}

