% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{myBayesLogit}
\alias{myBayesLogit}
\title{MCMC Sampling for Bayesian Logistic Regression Model}
\usage{
myBayesLogit(y, X, m0, P0, options)
}
\arguments{
\item{y}{(\code{integer})\cr 0/1 vector of responses.}

\item{X}{(\code{matrix})\cr design matrix.}

\item{m0}{(\code{numeric})\cr prior mean vector.}

\item{P0}{(\code{matrix})\cr precision matrix.}

\item{options}{(\code{McmcOptions})\cr MCMC options.}
}
\value{
The matrix of samples (samples x parameters).
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Replacement for \code{BayesLogit::logit}. Performs MCMC sampling for Bayesian
logistic regression using JAGS.
}
\keyword{internal}
