% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Samples-methods.R
\name{names,Samples-method}
\alias{names,Samples-method}
\alias{names-Samples}
\title{The Names of the Sampled Parameters}
\usage{
\S4method{names}{Samples}(x)
}
\arguments{
\item{x}{(\code{Samples})\cr object with samples.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

A method that returns names of the parameters that are sampled.
}
\examples{
my_samples <- Samples(
  data = list(alpha = 1:5, beta = 15:19),
  options = McmcOptions(burnin = 2, step = 2, samples = 5)
)

names(my_samples)
}
