% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{NextBest-class}
\alias{NextBest-class}
\alias{NextBest}
\title{\code{NextBest}}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{\link{NextBest}} is a virtual class for finding next best dose, from which all
other specific next best dose classes inherit.
}
\seealso{
\code{\link{NextBestMTD}}, \code{\link{NextBestNCRM}}, \code{\link{NextBestDualEndpoint}},
\code{\link{NextBestThreePlusThree}}, \code{\link{NextBestDualEndpoint}}, \code{\link{NextBestMinDist}},
\code{\link{NextBestInfTheory}}, \code{\link{NextBestTD}}, \code{\link{NextBestTDsamples}},
\code{\link{NextBestMaxGain}}, \code{\link{NextBestMaxGainSamples}}.
}
