% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rules-methods.R
\docType{methods}
\name{nextBest,NextBestMaxGain,numeric,ANY,ModelTox,DataDual-method}
\alias{nextBest,NextBestMaxGain,numeric,ANY,ModelTox,DataDual-method}
\title{This is the nextBest method to choose the next dose with the maximum gain value using
a given model (DLE model) and a given Effmodel}
\usage{
\S4method{nextBest}{NextBestMaxGain,numeric,ANY,ModelTox,DataDual}(nextBest,
  doselimit, model, data, Effmodel, ...)
}
\description{
This is the nextBest method to choose the next dose with the maximum gain value using
a given model (DLE model) and a given Effmodel
}
\keyword{methods}

