% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rules-methods.R
\docType{methods}
\name{nextBest,NextBestTD,numeric,missing,LogisticIndepBeta,Data-method}
\alias{nextBest,NextBestTD,numeric,missing,LogisticIndepBeta,Data-method}
\title{nextBest method using pseudo DLE model
This is the nextBest method using only DLE responses and based on pseudo DLE model without
using any samples}
\usage{

  \S4method{nextBest}{NextBestTD,numeric,missing,LogisticIndepBeta,Data}(nextBest,
  doselimit, samples, model, data, ...)
}
\description{
nextBest method using pseudo DLE model
This is the nextBest method using only DLE responses and based on pseudo DLE model without
using any samples
}
\keyword{methods}

