% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\name{NextBestDualEndpoint}
\alias{NextBestDualEndpoint}
\title{Initialization function for "NextBestDualEndpoint"}
\usage{
NextBestDualEndpoint(target, scale = c("relative", "absolute"), overdose,
  maxOverdoseProb, targetThresh = 0.01)
}
\arguments{
\item{target}{see \code{\linkS4class{NextBestDualEndpoint}}}

\item{scale}{see \code{\linkS4class{NextBestDualEndpoint}}}

\item{overdose}{see \code{\linkS4class{NextBestDualEndpoint}}}

\item{maxOverdoseProb}{see \code{\linkS4class{NextBestDualEndpoint}}}

\item{targetThresh}{see \code{\linkS4class{NextBestDualEndpoint}}}
}
\value{
the \code{\linkS4class{NextBestDualEndpoint}} object
}
\description{
Initialization function for "NextBestDualEndpoint"
}
\keyword{methods}
