% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{NextBestInfTheory-class}
\alias{NextBestInfTheory-class}
\alias{.NextBestInfTheory}
\alias{NextBestInfTheory}
\title{Class for Next Best Dose based on Information Theory.}
\usage{
NextBestInfTheory(target, asymmetry)
}
\arguments{
\item{target}{target DLT probability.}

\item{asymmetry}{value of the asymmetry exponent}
}
\description{
Class for Next Best Dose based on Information Theory.
}
\section{Functions}{
\itemize{
\item \code{NextBestInfTheory}: Initialization function for \code{NextBestInfTheory}.
}}

\section{Slots}{

\describe{
\item{\code{target}}{target DLT probability.}

\item{\code{asymmetry}}{value of the asymmetry exponent}
}}

