% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rules-class.R
\name{NextBestMaxGain}
\alias{NextBestMaxGain}
\title{Init function for NextBestMaxGain}
\usage{
NextBestMaxGain(DLEDuringTrialtarget, DLEEndOfTrialtarget)
}
\description{
Init function for NextBestMaxGain
}

