% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\name{NextBestMaxGainSamples}
\alias{NextBestMaxGainSamples}
\title{Initialization function for class "NextBestMaxGainSamples"}
\usage{
NextBestMaxGainSamples(
  DLEDuringTrialtarget,
  DLEEndOfTrialtarget,
  TDderive,
  Gstarderive
)
}
\arguments{
\item{DLEDuringTrialtarget}{please refer to \code{\linkS4class{NextBestMaxGainSamples}} class object}

\item{DLEEndOfTrialtarget}{please refer to \code{\linkS4class{NextBestMaxGainSamples}} class object}

\item{TDderive}{please refer to \code{\linkS4class{NextBestMaxGainSamples}} class object}

\item{Gstarderive}{please refer to \code{\linkS4class{NextBestMaxGainSamples}} class object}
}
\value{
the \code{\linkS4class{NextBestMaxGainSamples}} class object
}
\description{
Initialization function for class "NextBestMaxGainSamples"
}
\keyword{methods}
