% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{NextBestMinDist-class}
\alias{NextBestMinDist-class}
\alias{.NextBestMinDist}
\alias{NextBestMinDist}
\title{\code{NextBestMinDist}}
\usage{
NextBestMinDist(target)
}
\arguments{
\item{target}{(\code{proportion})\cr target toxicity probability.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{\link{NextBestMinDist}} is the class for next best dose that is based on minimum
distance to target probability.
}
\section{Slots}{

\describe{
\item{\code{target}}{(\code{proportion})\cr target toxicity probability.}
}}

