\docType{class}
\name{NextBestMTD-class}
\alias{NextBestMTD-class}
\title{The class with the input for finding the next best MTD estimate}
\description{
The class with the input for finding the next best MTD
estimate
}
\section{Slots}{

\describe{
\item{\code{target}}{the target toxicity probability}

\item{\code{derive}}{the function which derives from the input, a vector of
posterior MTD samples called \code{mtdSamples}, the final next best MTD
estimate.}
}}
\keyword{classes}

