% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\name{NextBestMTD}
\alias{NextBestMTD}
\title{Initialization function for class "NextBestMTD"}
\usage{
NextBestMTD(target, derive)
}
\arguments{
\item{target}{see \code{\linkS4class{NextBestMTD}}}

\item{derive}{see \code{\linkS4class{NextBestMTD}}}
}
\value{
the \code{\linkS4class{NextBestMTD}} object
}
\description{
Initialization function for class "NextBestMTD"
}
\keyword{methods}
