\docType{class}
\name{NextBestNCRM-class}
\alias{NextBestNCRM-class}
\title{The class with the input for finding the next dose in target interval}
\description{
The class with the input for finding the next dose in
target interval
}
\section{Slots}{

\describe{
\item{\code{target}}{the target toxicity interval}

\item{\code{overdose}}{the overdose toxicity interval}

\item{\code{maxOverdoseProb}}{maximum overdose probability that is allowed}
}}
\keyword{classes}

