% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\name{NextBestNCRM}
\alias{NextBestNCRM}
\title{Initialization function for "NextBestNCRM"}
\usage{
NextBestNCRM(target, overdose, maxOverdoseProb)
}
\arguments{
\item{target}{see \code{\linkS4class{NextBestNCRM}}}

\item{overdose}{see \code{\linkS4class{NextBestNCRM}}}

\item{maxOverdoseProb}{see \code{\linkS4class{NextBestNCRM}}}
}
\value{
the \code{\linkS4class{NextBestNCRM}} object
}
\description{
Initialization function for "NextBestNCRM"
}
\keyword{methods}
