% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{NextBestNCRMLoss-class}
\alias{NextBestNCRMLoss-class}
\alias{.NextBestNCRMLoss}
\title{Next best dose based on NCRM rule and loss function}
\description{
This class is similar to NCRM class, the only difference is the addition of
loss function. As in NCRM rule, first admissible doses are found, which are those
with probability to fall in \code{overdose} category being below
\code{maxOverdoseProb}. Next, within the admissible doses, the loss function is
calculated, i.e. \code{losses} \%*\% \code{target}. Finally, the corresponding
dose with lowest loss function (Bayes risk) is recommended for the next dose.
}
\section{Slots}{

\describe{
\item{\code{target}}{the target toxicity interval (limits included)}

\item{\code{overdose}}{the overdose toxicity interval (lower limit excluded, upper limit included)}

\item{\code{maxOverdoseProb}}{maximum overdose probability that is allowed}

\item{\code{losses}}{loss function}
}}

\examples{

# In the example below, the target toxicity interval [0.2, 0.35] while the
# overdose interval is (0.35,1]. We would like to constrain the probability
# of overdosing below 25\%. The loss function is c(1, 0, 1, 2)
my_next_best <- NextBestNCRMLoss(
  target = c(0.2, 0.35),
  overdose = c(0.35, 0.6, 1),
  maxOverdoseProb = 0.25,
  losses <- c(1, 0, 1, 2)
)
}
\keyword{classes}
