% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\name{NextBestNCRMLoss}
\alias{NextBestNCRMLoss}
\title{Initialization function for "NextBestNCRMLoss"}
\usage{
NextBestNCRMLoss(target, overdose, maxOverdoseProb, losses)
}
\arguments{
\item{target}{see \code{\linkS4class{NextBestNCRMLoss}}}

\item{overdose}{see \code{\linkS4class{NextBestNCRMLoss}}}

\item{maxOverdoseProb}{see \code{\linkS4class{NextBestNCRMLoss}}}

\item{losses}{see \code{\linkS4class{NextBestNCRMLoss}}}
}
\value{
the \code{\linkS4class{NextBestNCRMLoss}} object
}
\description{
Initialization function for "NextBestNCRMLoss"
}
