% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{NextBestTD-class}
\alias{.NextBestTD}
\alias{NextBestTD-class}
\title{Class for next best based only on the DLE responses without using any samples}
\description{
Class for next best based only on the DLE responses without using any samples
}
\keyword{methods}

