% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{NextBestTDsamples-class}
\alias{.NextBestTDsamples}
\alias{NextBestTDsamples-class}
\title{Class for next best based on DLE responses also using DLE samples}
\description{
Class for next best based on DLE responses also using DLE samples
}
\keyword{class}

