% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rules-class.R
\name{NextBestTDsamples}
\alias{NextBestTDsamples}
\title{Init function for NextBestTDsamples}
\usage{
NextBestTDsamples(targetDuringTrial, targetEndOfTrial, derive)
}
\description{
Init function for NextBestTDsamples
}

