\docType{class}
\name{NextBestThreePlusThree-class}
\alias{NextBestThreePlusThree-class}
\title{The class with the input for finding the next dose in target interval}
\description{
Implements the classical 3+3 dose recommendation. No input
is required, hence this class has no slots.
}
\keyword{classes}

