% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\name{NextBestThreePlusThree}
\alias{NextBestThreePlusThree}
\title{Initialization function for "NextBestThreePlusThree"}
\usage{
NextBestThreePlusThree()
}
\value{
the \code{\linkS4class{NextBestThreePlusThree}} object
}
\description{
Initialization function for "NextBestThreePlusThree"
}
\keyword{methods}
