% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchicalDesign-class.R
\docType{class}
\name{NoArmCondition-class}
\alias{NoArmCondition-class}
\alias{.NoArmCondition}
\alias{NoArmCondition}
\alias{.DefaultNoArmCondition}
\title{\code{NoArmCondition}}
\usage{
NoArmCondition()

.DefaultNoArmCondition()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{\link{NoArmCondition}} always opens an active arm immediately. It is the default
opening condition for \code{\link{DesignArm}} objects.
}
\note{
Typically, end users will not use the \code{.DefaultNoArmCondition()}
function.
}
