% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{noOverlap}
\alias{noOverlap}
\title{Check overlap of two character vectors}
\usage{
noOverlap(a, b)
}
\arguments{
\item{a}{first character vector}

\item{b}{second character vector}
}
\value{
returns TRUE if there is no overlap between the two character
vectors, otherwise FALSE
}
\description{
Check overlap of two character vectors
}
\keyword{internal}
