% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-class.R
\docType{class}
\name{OneParExpExpPrior-class}
\alias{OneParExpExpPrior-class}
\alias{.OneParExpExpPrior}
\alias{OneParExpExpPrior}
\title{\code{OneParExpExpPrior}}
\usage{
OneParExpExpPrior(skel_probs, dose_grid, lambda)
}
\arguments{
\item{skel_probs}{(\code{numeric})\cr skeleton prior probabilities. This is a vector
of unique and sorted probability values between 0 and 1.}

\item{dose_grid}{(\code{numeric})\cr dose grid. It must be must be a sorted vector
of the same length as \code{skel_probs}.}

\item{lambda}{(\code{number})\cr rate parameter of prior exponential distribution for theta.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{\link{OneParExpExpPrior}} is the class for a standard CRM with an exponential prior on
the log power parameter for the skeleton prior probabilities. It is an
implementation of a version of the one-parameter CRM (O’Quigley et al. 1990).
}
\section{Slots}{

\describe{
\item{\code{skel_fun}}{(\code{function})\cr function to calculate the prior DLT probabilities.}

\item{\code{skel_fun_inv}}{(\code{function})\cr inverse function of \code{skel_fun}.}

\item{\code{skel_probs}}{(\code{numeric})\cr skeleton prior probabilities. This is a vector
of unique and sorted probability values between 0 and 1.}

\item{\code{lambda}}{(\code{number})\cr rate parameter of prior exponential distribution for theta.}
}}

\examples{
my_model <- OneParExpExpPrior(
  skel_probs = seq(from = 0.1, to = 0.9, length = 5),
  dose_grid = 1:5,
  lambda = 2
)
}
