% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-class.R
\docType{class}
\name{OneParExpNormalPrior-class}
\alias{OneParExpNormalPrior-class}
\alias{.OneParExpNormalPrior}
\alias{OneParExpNormalPrior}
\title{CRM Model with Skeleton Prior Probabilities}
\usage{
OneParExpNormalPrior(skeletonProbs, doseGrid, sigma2)
}
\description{
This is a standard CRM with a normal prior on the log power parameter
for the skeleton prior probabilities.
}
\section{Functions}{
\itemize{
\item \code{OneParExpNormalPrior}: Initialization function for the
"OneParExpNormalPrior" class.
}}

\section{Slots}{

\describe{
\item{\code{skeletonFun}}{function to calculate the prior DLT probabilities.}

\item{\code{skeletonProbs}}{skeleton prior probabilities.}

\item{\code{sigma2}}{prior variance of log power parameter alpha.}
}}

\keyword{classes}
\keyword{methods}
