% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchicalDesign-methods.R
\name{openArm}
\alias{openArm}
\alias{openArm,NoArmCondition-method}
\alias{openArm-NoArmCondition}
\alias{openArm,ArmFinishedCondition-method}
\alias{openArm-ArmFinishedCondition}
\alias{openArm,ArmMinDoseCondition-method}
\alias{openArm-ArmMinDoseCondition}
\alias{openArm,ArmConditionList-method}
\alias{openArm-ArmConditionList}
\alias{openArm,ArmConditionAll-method}
\alias{openArm-ArmConditionAll}
\alias{openArm,ArmConditionAny-method}
\alias{openArm-ArmConditionAny}
\title{Open a hierarchical design arm?}
\usage{
openArm(condition, data, ...)

\S4method{openArm}{NoArmCondition}(condition, data, ...)

\S4method{openArm}{ArmFinishedCondition}(condition, data, finished_arms, ...)

\S4method{openArm}{ArmMinDoseCondition}(condition, data, ...)

\S4method{openArm}{ArmConditionList}(condition, data, summary_fun, ...)

\S4method{openArm}{ArmConditionAll}(condition, data, ...)

\S4method{openArm}{ArmConditionAny}(condition, data, ...)
}
\arguments{
\item{condition}{(\code{ArmCondition})\cr opening condition to be applied.}

\item{data}{(\code{HierarchicalData})\cr current hierarchical trial data.}

\item{...}{further arguments passed to condition-specific methods.}

\item{finished_arms}{(\code{logical})\cr named vector indicating which arms have
finished dose escalation.}

\item{summary_fun}{(\code{function})\cr to apply to the list of results
(e.g. \code{all} or \code{any}). Only used for \code{ArmConditionList} and its
subclasses.}
}
\value{
\code{TRUE} if this arm can be opened, \code{FALSE} otherwise.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\section{Functions}{
\itemize{
\item \code{openArm(NoArmCondition)}: method for \code{NoArmCondition} class, which always opens
the arm.

\item \code{openArm(ArmFinishedCondition)}: method for \code{ArmFinishedCondition} class.

\item \code{openArm(ArmMinDoseCondition)}: method for \code{ArmMinDoseCondition} class.

\item \code{openArm(ArmConditionList)}: method for \code{ArmConditionList} class.

\item \code{openArm(ArmConditionAll)}: method for \code{ArmConditionAll} class. Returns \code{TRUE} if
ALL arm opening criteria are satisfied.

\item \code{openArm(ArmConditionAny)}: method for \code{ArmConditionAny} class. Returns \code{TRUE} if
ANY arm opening criterion is satisfied.

}}
