% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Backfill-class.R
\docType{class}
\name{Opening-class}
\alias{Opening-class}
\alias{.Opening}
\alias{Opening}
\alias{.DefaultOpening}
\title{\code{Opening}}
\usage{
.DefaultOpening()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{\link{Opening}} is a virtual class for opening criteria, from which all
other specific opening criteria classes inherit.
The subclasses are used for backfill cohort designs.
}
\note{
Typically, end users will not use the \code{.DefaultOpening()} function.
}
\seealso{
\code{\link{OpeningMinDose}}, \code{\link{OpeningMinCohorts}}, \code{\link{OpeningNone}},
\code{\link{OpeningMinResponses}}, \code{\link{OpeningAll}}, \code{\link{OpeningAny}}.
}
