% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Backfill-class.R
\docType{class}
\name{OpeningMinCohorts-class}
\alias{OpeningMinCohorts-class}
\alias{.OpeningMinCohorts}
\alias{OpeningMinCohorts}
\alias{.DefaultOpeningMinCohorts}
\title{\code{OpeningMinCohorts}}
\usage{
OpeningMinCohorts(min_cohorts = 2L)

.DefaultOpeningMinCohorts()
}
\arguments{
\item{min_cohorts}{(\code{integer})\cr see slot definition.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{\link{OpeningMinCohorts}} opens backfill cohorts when the overall number of
cohorts treated so far in the trial reaches or exceeds a minimum threshold.
This can be used to implement a "delayed backfill cohort opening" rule.
}
\section{Slots}{

\describe{
\item{\code{min_cohorts}}{(\code{integer})\cr the minimum number of cohorts that must
have been treated before backfilling can be opened.}
}}

\note{
Typically, end users will not use the \code{.DefaultOpeningMinCohorts()} function.
}
\seealso{
\code{\link{Opening}} and the other subclasses listed in there.
}
