% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Backfill-class.R
\docType{class}
\name{OpeningNone-class}
\alias{OpeningNone-class}
\alias{.OpeningNone}
\alias{OpeningNone}
\alias{.DefaultOpeningNone}
\title{\code{OpeningNone}}
\usage{
OpeningNone()

.DefaultOpeningNone()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{\link{OpeningNone}} never opens any backfill cohorts. This can be used when
no backfill cohorts should be available in a trial design.
}
\note{
Typically, end users will not use the \code{.DefaultOpeningNone()} function.
}
