% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchicalDesign-methods.R
\name{or,ArmCondition,ArmCondition-method}
\alias{or,ArmCondition,ArmCondition-method}
\alias{|,ArmCondition,ArmCondition-method}
\title{Logical OR Operator for ArmCondition Objects}
\usage{
\S4method{|}{ArmCondition,ArmCondition}(e1, e2)
}
\arguments{
\item{e1}{(\code{ArmCondition})\cr the first arm condition object.}

\item{e2}{(\code{ArmCondition})\cr the second arm condition object.}
}
\value{
An \code{\link{ArmConditionAny}} object combining \code{e1} and \code{e2}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Combines two \code{\link{ArmCondition}} objects with OR logic using the \code{|} operator.
This creates an \code{\link{ArmConditionAny}} object.
}
