% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rules-methods.R
\docType{methods}
\name{or-Stopping-Stopping}
\alias{or-Stopping-Stopping}
\alias{|,Stopping,Stopping-method}
\title{The method combining two atomic stopping rules}
\usage{
\S4method{|}{Stopping,Stopping}(e1, e2)
}
\arguments{
\item{e1}{First \code{\linkS4class{Stopping}} object}

\item{e2}{Second \code{\linkS4class{Stopping}} object}
}
\value{
The \code{\linkS4class{StoppingAny}} object
}
\description{
The method combining two atomic stopping rules
}
\keyword{methods}

