% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pinvGamma}
\alias{pinvGamma}
\title{Compute the distribution function of Inverse gamma distribution}
\usage{
pinvGamma(q, a, b, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{q}{vector of quantiles}

\item{a}{the shape parameter of the inverse gamma distribution}

\item{b}{the scale parameter of the inverse gamma distribution}

\item{lower.tail}{logical; if TRUE (default), probabilities are \code{P(X  > x)},
otherwise, \code{P(X <= x)}.}

\item{log.p}{if TRUE, probabilities/densities p are returned as \code{log(p)}}
}
\description{
Compute the distribution function of Inverse gamma distribution
}
\keyword{internal}
