% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations-methods.R
\name{plot,ComboSimulations,missing-method}
\alias{plot,ComboSimulations,missing-method}
\alias{plot-ComboSimulations-missing}
\title{Plot \code{ComboSimulations}}
\usage{
\S4method{plot}{ComboSimulations,missing}(x, y, type = c("trajectory", "dosesTried", "trajectory2D"), ...)
}
\arguments{
\item{x}{(\code{ComboSimulations})\cr the object we want to plot from.}

\item{y}{(\code{missing})\cr not used.}

\item{type}{(\code{character})\cr the type of plots you want to obtain.}

\item{...}{not used.}
}
\value{
A single \code{ggplot} object if a single panel is produced, otherwise
a \code{gtable} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Summarize two-drug combination simulations with plots.

This plot method can be applied to \code{\link{ComboSimulations}} objects. Possible
\code{type}s of plots are:
\describe{
\item{trajectory}{Summary of simulated dose trajectories for each drug}
\item{dosesTried}{Average proportions of tested doses for each drug}
\item{trajectory2D}{2D dose-plane evolution with arrows for dose transitions,
and labels showing DLTs/patients plus a step index hint}
}
}
