% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations-methods.R
\name{plot,ComboSimulationsSummary,missing-method}
\alias{plot,ComboSimulationsSummary,missing-method}
\alias{plot-ComboSimulationsSummary-missing}
\title{Plot \code{ComboSimulationsSummary}}
\usage{
\S4method{plot}{ComboSimulationsSummary,missing}(
  x,
  y,
  type = c("nObs", "doseSelectedDrug1", "doseSelectedDrug2", "propDLTs"),
  ...
)
}
\arguments{
\item{x}{(\code{ComboSimulationsSummary})
the object we want to plot from.}

\item{y}{(\code{missing})
not used.}

\item{type}{(\code{character})
the types of plots you want to obtain.}

\item{...}{not used.}
}
\value{
A single \code{ggplot} object if a single plot is asked for,
otherwise a \code{gtable} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Graphical display of combination simulation summaries.
}
