% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-methods.R
\name{plot,Data,missing-method}
\alias{plot,Data,missing-method}
\alias{plot-Data}
\title{Plot Method for the \code{\link{Data}} Class}
\usage{
\S4method{plot}{Data,missing}(x, y, blind = FALSE, legend = TRUE, ...)
}
\arguments{
\item{x}{(\code{Data})\cr object we want to plot.}

\item{y}{(\code{missing})\cr missing object, for compatibility with the generic
function.}

\item{blind}{(\code{flag})\cr indicates whether to blind the data.
If \code{TRUE}, then placebo subjects are reported at the same level
as the active dose level in the corresponding cohort,
and DLTs are always assigned to the first subjects in a cohort.}

\item{legend}{(\code{flag})\cr whether the legend should be added.}

\item{...}{not used.}
}
\value{
The \code{\link{ggplot2}} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

A method that creates a plot for \code{\link{Data}} object.
}
\examples{
# Create some data of class 'Data'.
my_data <- Data(
  x = c(0.001, 0.1, 0.1, 0.5, 0.001, 3, 3, 0.001, 10, 10, 10),
  y = c(0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0),
  cohort = c(1, 1, 1, 2, 3, 3, 3, 4, 4, 4, 4),
  doseGrid = c(0.001, 0.1, 0.5, 1.5, 3, 6, seq(from = 10, to = 80, by = 2)),
  placeb = TRUE
)

# Plot the data.
plot(my_data)
}
