% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Data-methods.R
\docType{methods}
\name{plot,Data,ModelTox-method}
\alias{plot,Data,ModelTox-method}
\title{Plot of the fitted dose-tox based with a given pseudo DLE model and data witout samples}
\usage{
\S4method{plot}{Data,ModelTox}(x, y, ..., xlab = "Dose level",
  ylab = "Probability of DLE", showLegend = TRUE)
}
\arguments{
\item{x}{the data input from the \code{\linkS4class{Data}} class object}

\item{y}{the model from the \code{\linkS4class{ModelTox}} class object}
}
\description{
Plot of the fitted dose-tox based with a given pseudo DLE model and data witout samples
}
\keyword{methods}

