% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-methods.R
\name{h_plot_data_dataordinal}
\alias{h_plot_data_dataordinal}
\alias{plot,Data,missing-method}
\alias{plot,DataOrdinal,missing-method}
\title{Helper Function for the Plot Method of the Data and DataOrdinal Classes}
\usage{
h_plot_data_dataordinal(
  x,
  blind = FALSE,
  legend = TRUE,
  tox_labels = c(Yes = "red", No = "black"),
  tox_shapes = c(Yes = 17L, No = 16L),
  ...
)

\S4method{plot}{Data,missing}(x, y, blind = FALSE, legend = TRUE, ...)

\S4method{plot}{DataOrdinal,missing}(
  x,
  y,
  blind = FALSE,
  legend = TRUE,
  tox_labels = NULL,
  tox_shapes = NULL,
  ...
)
}
\arguments{
\item{x}{(\code{DataOrdinal})\cr object we want to plot.}

\item{blind}{(\code{flag})\cr indicates whether to blind the data.
If \code{TRUE}, then placebo subjects are reported at the same level
as the active dose level in the corresponding cohort,
and DLTs are always assigned to the first subjects in a cohort.}

\item{legend}{(\code{flag})\cr whether the legend should be added.}

\item{tox_labels}{(\verb{named list of character})\cr The labels of the toxicity
categories}

\item{tox_shapes}{(\verb{names list of integers})\cr The symbols used to identify
the toxicity categories}

\item{...}{not used.}

\item{y}{(\code{missing})\cr missing object, for compatibility with the generic
function.}
}
\value{
The \code{\link{ggplot2}} object.

The \code{\link{ggplot2}} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

A method that creates a plot for \code{\link{Data}}  and \code{\link{DataOrdinal}}objects.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

A method that creates a plot for \code{\link{Data}} object.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

A method that creates a plot for \code{\link{DataOrdinal}} object.
}
\note{
The default values of \code{tox_shapes} and \code{tox_labels} result in DLTs
being displayed as red triangles and other responses as black circles.

With more than 9 toxicity categories, toxicity symbols must be
specified manually.\cr With more than 5 toxicity categories, toxicity labels
must be specified manually.
}
\examples{
# Create some data of class 'Data'.
my_data <- Data(
  x = c(0.001, 0.1, 0.1, 0.5, 0.001, 3, 3, 0.001, 10, 10, 10),
  y = c(0, 0, 1, 0, 0, 0, 1, 0, 0, 1, 0),
  cohort = c(1, 1, 1, 2, 3, 3, 3, 4, 4, 4, 4),
  doseGrid = c(0.001, 0.1, 0.5, 1.5, 3, 6, seq(from = 10, to = 80, by = 2)),
  placeb = TRUE
)

# Plot the data.
plot(my_data)
d <- DataOrdinal(
  x = c(25, 25, 25, 25, 50, 50, 50, 50, 100, 100, 100, 100),
  y = c(0L, 0L, 0L, 0L, 0L, 0L, 0L, 0L, 0L, 0L, 1L, 2L),
  doseGrid = seq(25, 300, 25),
  placebo = FALSE,
  ID = 1:12,
  cohort = c(1L, 1L, 1L, 1L, 2L, 2L, 2L, 2L, 3L, 3L, 3L, 3L),
  yCategories = c("No tox" = 0L, "Sub-tox AE" = 1L, "DLT" = 2L)
)

plot(d, legend = TRUE)
}
