% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-methods.R
\name{plot,DataCombo,missing-method}
\alias{plot,DataCombo,missing-method}
\alias{plot-DataCombo}
\title{Plot Method for the \code{\link{DataCombo}} Class}
\usage{
\S4method{plot}{DataCombo,missing}(x, y, legend = TRUE, ...)
}
\arguments{
\item{x}{(\code{DataCombo})}

\item{y}{(\code{missing})}

\item{legend}{(\code{flag})
whether legends should be displayed.}

\item{...}{not used.}
}
\value{
A \code{gtable} object combining two \code{\link[ggplot2:ggplot2]{ggplot2}} plots.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

A method that creates a plot for \code{\link{DataCombo}} object.
}
\examples{
data <- .DefaultDataCombo()

plot(data)
}
