% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Data-methods.R
\docType{methods}
\name{plot,DataDual,missing-method}
\alias{plot,DataDual,missing-method}
\title{Plot method for the "DataDual" class}
\usage{
\S4method{plot}{DataDual,missing}(x, y, ...)
}
\arguments{
\item{x}{the \code{\linkS4class{DataDual}} object we want to plot}

\item{y}{missing}

\item{\dots}{not used}
}
\value{
the \code{\link[ggplot2]{ggplot}} object
}
\description{
Plot method for the "DataDual" class
}
\keyword{methods}

