% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-methods.R
\name{plot,DataOrdinal,missing-method}
\alias{plot,DataOrdinal,missing-method}
\alias{plot-Data}
\title{Plot Method for the \code{\link{DataOrdinal}} Class}
\usage{
\S4method{plot}{DataOrdinal,missing}(
  x,
  y,
  blind = FALSE,
  legend = TRUE,
  tox_labels = NULL,
  tox_shapes = NULL,
  ...
)
}
\arguments{
\item{x}{(\code{DataOrdinal})\cr object we want to plot.}

\item{y}{(\code{missing})\cr missing object, for compatibility with the generic
function.}

\item{blind}{(\code{flag})\cr indicates whether to blind the data.
If \code{TRUE}, then placebo subjects are reported at the same level
as the active dose level in the corresponding cohort,
and DLTs are always assigned to the first subjects in a cohort.}

\item{legend}{(\code{flag})\cr whether the legend should be added.}

\item{...}{not used.}
}
\value{
The \code{\link{ggplot2}} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

A method that creates a plot for \code{\link{DataOrdinal}} object.
}
\note{
With more than 9 toxicity categories, toxicity symbols must be
specified manually.\cr With more than 5 toxicity categories, toxicity labels
must be specified manually.
}
