% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations-methods.R
\name{plot,GeneralSimulationsSummary,missing-method}
\alias{plot,GeneralSimulationsSummary,missing-method}
\alias{plot-GeneralSimulationsSummary-missing}
\title{Plot \code{GeneralSimulationsSummary}}
\usage{
\S4method{plot}{GeneralSimulationsSummary,missing}(
  x,
  y,
  type = c("nObs", "doseSelected", "propDLTs", "nAboveTarget"),
  axis_text_angle = 45,
  ...
)
}
\arguments{
\item{x}{(\code{GeneralSimulationsSummary})\cr the object we want to plot from.}

\item{y}{(\code{missing})\cr not used.}

\item{type}{(\code{character})\cr the types of plots you want to obtain, see details.}

\item{axis_text_angle}{(\code{number})\cr rotation angle for the MTD estimate
x-axis tick labels. Defaults to 45 degrees.}

\item{...}{not used.}
}
\value{
A single \code{ggplot} object if a single plot is
asked for, otherwise a \code{gtable} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Graphical display of the general simulation summary.

This plot method can be applied to \code{\link{GeneralSimulationsSummary}} objects in
order to summarize them graphically.
}
\details{
The following plot types are available:
\describe{
\item{\code{"nObs"}}{The distribution of the total number of patients in the
simulated trials. For trials with a placebo, only patients assigned to
an active dose are included.}
\item{\code{"doseSelected"}}{The distribution of the final selected dose (MTD)
across trials. A selected dose of zero indicates that the trial stopped
because all doses in the dose grid appeared too toxic.}
\item{\code{"propDLTs"}}{The distribution of the percentage of patients with
dose-limiting toxicities (DLTs). For trials with a placebo, this is the
percentage among patients assigned to an active dose.}
\item{\code{"nAboveTarget"}}{The distribution of the number of patients treated
at doses above the target toxicity interval, as determined by the \code{truth}
and \code{target} arguments supplied to
\code{\link{summary,GeneralSimulations-method}}.}
}
Any subset of these plot types can be requested with \code{type}.
}
