% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{plot.gtable}
\alias{plot.gtable}
\alias{print.gtable}
\title{Plot \code{gtable} Ojects}
\usage{
\method{plot}{gtable}(x, ...)

\method{print}{gtable}(x, ...)
}
\arguments{
\item{x}{(\code{gtable})\cr object to plot.}

\item{...}{additional parameters for \code{\link[grid:grid.draw]{grid::grid.draw()}}.}
}
\description{
This is needed because \code{crmPack} uses \code{\link[gridExtra:arrangeGrob]{gridExtra::arrangeGrob()}} to combine
\code{ggplot2} plots, and the resulting \code{gtable} object is not plotted otherwise
when implicitly printing it in the console, e.g.
}
