% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Samples-methods.R
\docType{methods}
\name{plot,Samples,ComboLogistic-method}
\alias{plot,Samples,ComboLogistic-method}
\title{Plotting dose-toxicity combo model fits}
\usage{
\S4method{plot}{Samples,ComboLogistic}(x, y, data, focus, resolution = 20,
  ...)
}
\arguments{
\item{x}{the \code{\linkS4class{Samples}} object}

\item{y}{the \code{\linkS4class{ComboLogistic}} object}

\item{data}{the \code{\linkS4class{DataCombo}} object}

\item{focus}{one or two drug names of this combo for which the model fit
plot should be produced}

\item{resolution}{number of points to be used in each dimension for 2D plots
(default: 20) - higher number is higher resolution}

\item{\dots}{passed to the single agent plotting method if \code{focus}
specifies only one drug name}
}
\value{
todo
}
\description{
Plotting dose-toxicity combo model fits
}

