% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Samples-methods.R
\name{plot,Samples,DALogisticLogNormal-method}
\alias{plot,Samples,DALogisticLogNormal-method}
\title{Plotting dose-toxicity model fits}
\usage{
\S4method{plot}{Samples,DALogisticLogNormal}(x, y, data, hazard = FALSE, ..., showLegend = TRUE)
}
\arguments{
\item{x}{the \code{\linkS4class{Samples}} object}

\item{y}{the \code{\linkS4class{DALogisticLogNormal}} object}

\item{data}{the \code{\linkS4class{DataDA}} object}

\item{hazard}{see \code{\link{fitPEM}} for the explanation}

\item{\dots}{not used}

\item{showLegend}{should the legend be shown? (default)}
}
\value{
This returns the \code{\link[ggplot2]{ggplot}}
object for the dose-toxicity model fit
}
\description{
Plotting dose-toxicity model fits
}
