% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Samples-methods.R
\docType{methods}
\name{plot,Samples,DualEndpoint-method}
\alias{plot,Samples,DualEndpoint-method}
\title{Plotting dose-toxicity and dose-biomarker model fits}
\usage{
\S4method{plot}{Samples,DualEndpoint}(x, y, data, extrapolate = TRUE,
  showLegend = FALSE, ...)
}
\arguments{
\item{x}{the \code{\linkS4class{Samples}} object}

\item{y}{the \code{\linkS4class{DualEndpoint}} object}

\item{data}{the \code{\linkS4class{DataDual}} object}

\item{extrapolate}{should the biomarker fit be extrapolated to the whole
dose grid? (default)}

\item{showLegend}{should the legend be shown? (not default)}

\item{\dots}{additional arguments for the parent method
\code{\link{plot,Samples,Model-method}}}
}
\value{
This returns the \code{\link[ggplot2]{ggplot}}
object with the dose-toxicity and dose-biomarker model fits
}
\description{
When we have the dual endpoint model,
also the dose-biomarker fit is shown in the plot
}

