% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Samples-methods.R
\docType{methods}
\name{plot,Samples,LogisticIndepBeta-method}
\alias{plot,Samples,LogisticIndepBeta-method}
\title{Plot dose-tox fit from 'LogisticIndepBeta' model class}
\usage{
\S4method{plot}{Samples,LogisticIndepBeta}(x, y, data, ...,
  xlab = "Dose level", ylab = "Probability of DLT [\%]",
  showLegend = TRUE)
}
\arguments{
\item{x}{the samples}

\item{y}{the LogisticIndepBeta object}
}
\description{
Plot dose-tox fit from 'LogisticIndepBeta' model class
}
\keyword{methods}

