% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Samples-methods.R
\name{plot,Samples,LogisticLogNormalCombo-method}
\alias{plot,Samples,LogisticLogNormalCombo-method}
\title{Plotting two-drug combination dose-toxicity model fits}
\usage{
\S4method{plot}{Samples,LogisticLogNormalCombo}(
  x,
  y,
  data,
  ...,
  xlab = NULL,
  ylab = NULL,
  fillLab = "Probability of DLT [\%]",
  showLegend = TRUE
)
}
\arguments{
\item{x}{the \code{\linkS4class{Samples}} object.}

\item{y}{the \code{\linkS4class{LogisticLogNormalCombo}} object.}

\item{data}{the \code{\linkS4class{DataCombo}} object.}

\item{\dots}{passed to \code{\link{fit}}.}

\item{xlab}{the x axis label. If \code{NULL}, the first drug name is used.}

\item{ylab}{the y axis label. If \code{NULL}, the second drug name is used.}

\item{fillLab}{the fill legend label.}

\item{showLegend}{should the legend be shown? (default)}
}
\value{
This returns the \code{\link[ggplot2]{ggplot}} object for the
two-drug combination dose-toxicity model fit.
}
\description{
Plotting two-drug combination dose-toxicity model fits
}
