% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Samples-methods.R
\docType{methods}
\name{plot,Samples,Model-method}
\alias{plot,Samples,Model-method}
\title{Plotting dose-toxicity model fits}
\usage{
\S4method{plot}{Samples,Model}(x, y, data, ..., xlab = "Dose level",
  ylab = "Probability of DLT [\%]", showLegend = TRUE)
}
\arguments{
\item{x}{the \code{\linkS4class{Samples}} object}

\item{y}{the \code{\linkS4class{Model}} object}

\item{data}{the \code{\linkS4class{Data}} object}

\item{xlab}{the x axis label}

\item{ylab}{the y axis label}

\item{showLegend}{should the legend be shown? (default)}

\item{\dots}{not used}
}
\value{
This returns the \code{\link[ggplot2]{ggplot}}
object for the dose-toxicity model fit
}
\description{
Plotting dose-toxicity model fits
}

