% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{plot.arrange}
\alias{plot.arrange}
\title{Plots arrange objects}
\usage{
\method{plot}{arrange}(x, ...)
}
\arguments{
\item{x}{the arrange object}

\item{\dots}{additional parameters for \code{\link[grid]{grid.draw}}}
}
\description{
Plots arrange objects
}

