% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Samples-methods.R
\docType{methods}
\name{plotDualResponses}
\alias{plotDualResponses}
\alias{plotDualResponses,ModelTox,ANY,ModelEff,ANY-method}
\alias{plotDualResponses,ModelTox,Samples,ModelEff,Samples-method}
\title{Plot of the DLE and efficacy curve side by side given a DLE pseudo model,
a DLE sample, an efficacy pseudo model and a given efficacy sample}
\usage{
plotDualResponses(DLEmodel, DLEsamples, Effmodel, Effsamples, data, ...)

\S4method{plotDualResponses}{ModelTox,Samples,ModelEff,Samples}(DLEmodel,
  DLEsamples, Effmodel, Effsamples, data, extrapolate = TRUE,
  showLegend = FALSE, ...)

\S4method{plotDualResponses}{ModelTox,ANY,ModelEff,ANY}(DLEmodel, DLEsamples,
  Effmodel, Effsamples, data, ...)
}
\arguments{
\item{DLEmodel}{the pseudo DLE model of \code{\linkS4class{ModelTox}} class object}

\item{DLEsamples}{the DLE samples of \code{\linkS4class{Samples}} clas object}

\item{Effmodel}{the pseudo efficacy model of \code{\linkS4class{ModelEff}} class object}

\item{Effsamples}{the Efficacy samples of \code{\linkS4class{Samples}} class object}

\item{data}{the data input of \code{\linkS4class{DataDual}} class object}
}
\description{
Plot of the DLE and efficacy curve side by side given a DLE pseudo model,
a DLE sample, an efficacy pseudo model and a given efficacy sample
}
\section{Methods (by class)}{
\itemize{
\item \code{DLEmodel = ModelTox,DLEsamples = Samples,Effmodel = ModelEff,Effsamples = Samples}: function todo

\item \code{DLEmodel = ModelTox,DLEsamples = ANY,Effmodel = ModelEff,Effsamples = ANY}: Plot the DLE and efficacy curve side by side given a DLE model, an Efiicacy model without any samples
}}
\keyword{methods}

