% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Samples-methods.R
\docType{methods}
\name{plotGain,ModelTox,ModelEff-method}
\alias{plotGain,ModelTox,ModelEff-method}
\title{describeIn plotGain Standard method}
\usage{
\S4method{plotGain}{ModelTox,ModelEff}(DLEmodel, Effmodel, data, ...)
}
\description{
describeIn plotGain Standard method
}

