% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Samples-methods.R
\name{plotGain}
\alias{plotGain}
\title{Plot the gain curve in addition with the DLE and Efficacy curve using a given DLE pseudo and
a given Efficacy Pseudo model}
\usage{
plotGain(DLEmodel, Effmodel, data, ...)
}
\arguments{
\item{DLEmodel}{a specified Dle model follow \code{\linkS4class{ModelTox}} class specification}

\item{Effmodel}{a specified Dle model follow \code{\linkS4class{ModelEff}} class specification}

\item{data}{the data input follow \code{\linkS4class{DataDual}} class specification}
}
\description{
Plot the gain curve in addition with the DLE and Efficacy curve using a given DLE pseudo and
a given Efficacy Pseudo model
}
\keyword{methods}

