% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\docType{class}
\name{positive_number}
\alias{positive_number}
\title{\code{positive_number}}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The \code{\link{positive_number}} class is a class to store not \code{NULL}, non \code{NA},
finite and strictly positive numerical value. It is mainly used to store
reference dose value in model classes.
}
