% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-methods.R
\name{prob,numeric,LogisticLogNormalOrdinal,Samples-method}
\alias{prob,numeric,LogisticLogNormalOrdinal,Samples-method}
\alias{prob-LogisticLogNormalOrdinal}
\title{Calculate a grade-specific probability of toxicity for a given dose.}
\usage{
\S4method{prob}{numeric,LogisticLogNormalOrdinal,Samples}(dose, model, samples, grade, cumulative = TRUE, ...)
}
\arguments{
\item{grade}{(\code{integer} or \code{integer_vector})\cr The toxicity grade for which probabilities are required}

\item{cumulative}{(\code{flag})\cr Should the returned probability be cumulative
(the default) or grade-specific?}
}
\description{
Calculate a grade-specific probability of toxicity for a given dose.
}
