\docType{methods}
\name{prob,numeric,Model,Samples-method}
\alias{prob,numeric,Model,Samples-method}
\title{Compute the probability for a given dose, given model and samples}
\usage{
\S4method{prob}{numeric,Model,Samples}(dose, model, samples, ...)
}
\arguments{
  \item{dose}{the dose}

  \item{model}{the \code{\linkS4class{Model}}}

  \item{samples}{the \code{\linkS4class{Samples}}}

  \item{...}{unused}
}
\description{
Compute the probability for a given dose, given model and
samples
}

