% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Model-methods.R
\docType{methods}
\name{prob,numeric,ModelTox,Samples-method}
\alias{prob,numeric,ModelTox,Samples-method}
\title{Compute the probability for a given dose, given Pseudo DLE model and samples}
\usage{
\S4method{prob}{numeric,ModelTox,Samples}(dose, model, samples, ...)
}
\arguments{
\item{dose}{the dose}

\item{model}{the \code{\linkS4class{ModelTox}} object}

\item{samples}{the \code{\linkS4class{Samples}}}

\item{\dots}{unused}
}
\value{
the vector (for \code{\linkS4class{ModelTox}} objects) of probability
samples.
}
\description{
Compute the probability for a given dose, given Pseudo DLE model and samples
}
\keyword{methods}

