% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-methods.R
\name{prob}
\alias{prob}
\alias{prob,numeric,LogisticLogNormalGrouped,Samples-method}
\alias{prob-LogisticLogNormalGrouped}
\alias{prob,numeric,LogisticLogNormalCombo,Samples-method}
\alias{prob-LogisticLogNormalCombo}
\alias{prob,matrix,LogisticLogNormalCombo,Samples-method}
\alias{prob-LogisticLogNormalCombo-matrix}
\alias{prob,numeric,LogisticIndepBeta,Samples-method}
\alias{prob-LogisticIndepBeta}
\alias{prob,numeric,LogisticIndepBeta,missing-method}
\alias{prob-LogisticIndepBeta-noSamples}
\title{Computing Toxicity Probabilities for a Given Dose, Model and Samples}
\usage{
prob(dose, model, samples, ...)

\S4method{prob}{numeric,LogisticLogNormalGrouped,Samples}(dose, model, samples, group, ...)

\S4method{prob}{numeric,LogisticLogNormalCombo,Samples}(dose, model, samples, ...)

\S4method{prob}{matrix,LogisticLogNormalCombo,Samples}(dose, model, samples, ...)

\S4method{prob}{numeric,LogisticIndepBeta,Samples}(dose, model, samples, ...)

\S4method{prob}{numeric,LogisticIndepBeta,missing}(dose, model, samples, ...)
}
\arguments{
\item{dose}{(\code{number} or \code{numeric})\cr the dose which is targeted.
The following recycling rule applies when \code{samples} is not missing: vectors
of size 1 will be recycled to the size of the sample
(i.e. \code{size(samples)}). Otherwise, \code{dose} must have the same
size as the sample.}

\item{model}{(\code{GeneralModel} or \code{ModelTox})\cr the model for single agent
dose escalation or pseudo DLE (dose-limiting events)/toxicity model.}

\item{samples}{(\code{Samples})\cr the samples of model's parameters that will be
used to compute toxicity probabilities. Can also be missing for some models.}

\item{...}{model specific parameters when \code{samples} are not used.}

\item{group}{(\code{character} or \code{factor})\cr for \code{\link{LogisticLogNormalGrouped}},
indicating whether to calculate the probability for the \code{mono} or for
the \code{combo} arm.}
}
\value{
A \code{proportion} or \code{numeric} vector with the toxicity probabilities,
or a numeric matrix for methods that evaluate multiple dose combinations at
once. If non-scalar \code{samples} were used, then every element in the returned
vector corresponds to one element of a sample. Hence, in this case, the
output vector is of the same length as the sample vector. If scalar
\code{samples} were used or no \code{samples} were used, e.g. for pseudo
DLE/toxicity \code{model}, then the output is of the same length as the length
of the \code{dose}. For matrix-valued dose inputs, the returned matrix contains
one column per dose combination. In the case of \code{LogisticLogNormalOrdinal},
the probabilities relate to toxicities of grade given by \code{grade}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

A function that computes the probability of the occurrence of a DLE at a
specified dose level, based on the model parameters (samples).
}
\details{
The \code{prob()} function computes the probability of toxicity for given
doses, using samples of the model parameter(s).
If you work with multivariate model parameters, then assume that your model
specific \code{prob()} method receives a samples matrix where the rows
correspond to the sampling index, i.e. the layout is then
\verb{nSamples x dimParameter}.
}
\section{Functions}{
\itemize{
\item \code{prob(dose = numeric, model = LogisticLogNormalGrouped, samples = Samples)}: method for \code{\link{LogisticLogNormalGrouped}} which needs \code{group}
argument in addition.

\item \code{prob(dose = numeric, model = LogisticLogNormalCombo, samples = Samples)}: method for \code{\link{LogisticLogNormalCombo}} for a single dose
combination provided as a named numeric vector.

\item \code{prob(dose = matrix, model = LogisticLogNormalCombo, samples = Samples)}: method for \code{\link{LogisticLogNormalCombo}} for one or more dose
combinations provided in the rows of a numeric matrix.

\item \code{prob(dose = numeric, model = LogisticIndepBeta, samples = Samples)}: compute toxicity probabilities of the occurrence of a DLE at
a specified dose level, based on the samples of \code{\link{LogisticIndepBeta}} model
parameters.

\item \code{prob(dose = numeric, model = LogisticIndepBeta, samples = missing)}: compute toxicity probabilities of the occurrence of a DLE at
a specified dose level, based on the \code{\link{LogisticIndepBeta}} model parameters.
All model parameters (except \code{dose}) should be present in the \code{model} object.

}}
\note{
The \code{\link[=prob]{prob()}} and \code{\link[=dose]{dose()}} functions are the inverse of
each other, for all \code{\link[=dose]{dose()}} methods for which its first argument, i.e. a
given independent variable that dose depends on, represents toxicity
probability.
}
\examples{
# Create some data.
my_data <- Data(
  x = c(0.1, 0.5, 1.5, 3, 6, 10, 10, 10),
  y = c(0, 0, 0, 0, 0, 0, 1, 0),
  cohort = c(0, 1, 2, 3, 4, 5, 5, 5),
  doseGrid = c(0.1, 0.5, 1.5, 3, 6, seq(from = 10, to = 80, by = 2))
)

# Initialize a model, e.g. 'LogisticLogNormal'.
my_model <- LogisticLogNormal(
  mean = c(-0.85, 1),
  cov = matrix(c(1, -0.5, -0.5, 1), nrow = 2),
  ref_dose = 56
)

# Get samples from posterior.
my_options <- McmcOptions(burnin = 100, step = 2, samples = 20)
my_samples <- mcmc(data = my_data, model = my_model, options = my_options)

# Posterior for Prob(DLT | dose = 50).
prob(dose = 50, model = my_model, samples = my_samples)

# Create data from the 'DataDual' class.
data_dual <- DataDual(
  x = c(25, 50, 25, 50, 75, 300, 250, 150),
  y = c(0, 0, 0, 0, 0, 1, 1, 0),
  w = c(0.31, 0.42, 0.59, 0.45, 0.6, 0.7, 0.6, 0.52),
  doseGrid = seq(from = 25, to = 300, by = 25)
)

# Initialize a toxicity model using 'LogisticIndepBeta' model.
dlt_model <- LogisticIndepBeta(
  binDLE = c(1.05, 1.8),
  DLEweights = c(3, 3),
  DLEdose = c(25, 300),
  data = data_dual
)

# Get samples from posterior.
dlt_sample <- mcmc(data = data_dual, model = dlt_model, options = my_options)

# Posterior for Prob(DLT | dose = 100).
prob(dose = 100, model = dlt_model, samples = dlt_sample)
prob(dose = c(50, 150), model = dlt_model)
}
\seealso{
\code{\link[=probFunction]{probFunction()}}, \code{\link[=dose]{dose()}}, \code{\link[=efficacy]{efficacy()}}.
}
