% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Model-class.R
\name{probComboLogistic}
\alias{probComboLogistic}
\title{Helper for the prob function of ComboLogistic}
\usage{
probComboLogistic(dose, alpha0, alpha1, eta, refDose)
}
\arguments{
\item{dose}{matrix, with each row corresponding to a
different combination (combis x drugs)}

\item{alpha0}{matrix of alpha0 samples (samples x drugs)}

\item{alpha1}{matrix of alpha1 samples (samples x drugs)}

\item{eta}{vector of eta samples (length samples)}

\item{refDose}{vector of reference doses (length drugs)}
}
\value{
matrix of sample probabilities
for each of the drug combinations (samples x combis)
}
\description{
Internal helper function. It will either use C++ or R code,
depending on "crmPackUsesCpp" option (determined in .onLoad hook)
}
\author{
Daniel Sabanes Bove \email{sabanesd@roche.com}
}
\keyword{internal}

