% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{probit}
\alias{probit}
\title{Shorthand for Probit Function}
\usage{
probit(x)
}
\arguments{
\item{x}{(\code{numeric})\cr the function argument.}
}
\value{
The probit of \code{x}, the quantile function of the standard normal distribution.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Computes the probit (inverse cumulative distribution function of standard normal) transformation.
}
\examples{
probit(0.5)
probit(c(0.025, 0.5, 0.975))
}
