% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-class.R
\name{ProbitLogNormal}
\alias{ProbitLogNormal}
\title{Initialization function for the "ProbitLogNormal" class}
\usage{
ProbitLogNormal(mu, Sigma)
}
\arguments{
\item{mu}{the prior mean vector}

\item{Sigma}{the prior covariance matrix}
}
\value{
the \code{\linkS4class{ProbitLogNormal}} object
}
\description{
Initialization function for the "ProbitLogNormal" class
}
\keyword{methods}

