% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations-class.R
\name{PseudoDualFlexiSimulations}
\alias{PseudoDualFlexiSimulations}
\title{Initialization function for 'PseudoDualFlexiSimulations' class}
\usage{
PseudoDualFlexiSimulations(sigma2betaWest, ...)
}
\arguments{
\item{sigma2betaWest}{please refer to \code{\linkS4class{PseudoDualFlexiSimulations}} class object}

\item{\dots}{additional parameters from \code{\linkS4class{PseudoDualSimulations}}}
}
\value{
the \code{\linkS4class{PseudoDualFlexiSimulations}} object
}
\description{
Initialization function for 'PseudoDualFlexiSimulations' class
}
