% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Simulations-class.R
\name{PseudoDualSimulations}
\alias{PseudoDualSimulations}
\title{Initialization function for 'DualPseudoSimulations' class}
\usage{
PseudoDualSimulations(fitEff, sigma2est, ...)
}
\arguments{
\item{fitEff}{please refer to \code{\linkS4class{PseudoDualSimulations}} class object}

\item{sigma2est}{please refer to \code{\linkS4class{PseudoDualSimulations}} class object}

\item{\dots}{additional parameters from \code{\linkS4class{PseudoSimulations}}}
}
\value{
the \code{\linkS4class{PseudoDualSimulations}} object
}
\description{
Initialization function for 'DualPseudoSimulations' class
}

