% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Simulations-class.R
\name{PseudoSimulations}
\alias{PseudoSimulations}
\title{Initialization function of the 'PseudoSimulations' class}
\usage{
PseudoSimulations(fit, stopReasons, ...)
}
\arguments{
\item{fit}{please refer to \code{\linkS4class{PseudoSimulations}} class object}

\item{stopReasons}{please refer to \code{\linkS4class{PseudoSimulations}} class object}

\item{\dots}{additional parameters from \code{\linkS4class{GeneralSimulations}}}
}
\value{
the \code{\linkS4class{PseudoSimulations}} object
}
\description{
Initialization function of the 'PseudoSimulations' class
}
\keyword{methods}

