% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{qinvGamma}
\alias{qinvGamma}
\title{Compute the quantile function of Inverse gamma distribution}
\usage{
qinvGamma(p, a, b, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{p}{vector of probabilities}

\item{a}{the shape parameter of the inverse gamma distribution}

\item{b}{the scale parameter of the inverse gamma distribution}

\item{lower.tail}{logical; if TRUE (default), probabilities are \code{P(X  > x)},
otherwise, \code{P(X <= x)}.}

\item{logical;}{FALSE (default) if TRUE, probabilities/densities p are
returned as \code{log(p)}}
}
\description{
Compute the quantile function of Inverse gamma distribution
}
\keyword{internal}
