% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Backfill-class.R
\docType{class}
\name{Recruitment-class}
\alias{Recruitment-class}
\alias{.Recruitment}
\alias{Recruitment}
\alias{.DefaultRecruitment}
\title{\code{Recruitment}}
\usage{
.DefaultRecruitment()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{\link{Recruitment}} is a virtual class for recruitment criteria, from which all
other specific recruitment criteria classes inherit.
The subclasses are used to specify the maximum number of backfill patients
that can be recruited relative to the main trial cohort size.
}
\note{
Typically, end users will not use the \code{.DefaultRecruitment()} function.
}
\seealso{
\code{\link{RecruitmentUnlimited}}, \code{\link{RecruitmentRatio}}.
}
