% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations-methods.R
\docType{class}
\name{Report}
\alias{Report}
\title{A Reference Class to represent sequentially updated reporting objects.}
\description{
A Reference Class to represent sequentially updated reporting objects.
}
\section{Fields}{

\describe{
\item{\code{object}}{The object from which to report}

\item{\code{df}}{the data frame to which columns are sequentially added}

\item{\code{dfNames}}{the names to which strings are sequentially added}
}}


