% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rinvGamma}
\alias{rinvGamma}
\title{Random Generation for the Inverse Gamma Distribution}
\usage{
rinvGamma(n, a, b)
}
\arguments{
\item{n}{(\code{count})\cr the number of observations.}

\item{a}{(\code{number})\cr the shape parameter of the inverse gamma distribution.}

\item{b}{(\code{number})\cr the scale parameter of the inverse gamma distribution.}
}
\value{
A vector of random values from the inverse gamma distribution.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\keyword{internal}
