% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design-class.R
\name{RuleDesign}
\alias{RuleDesign}
\title{Initialization function for "RuleDesign"}
\usage{
RuleDesign(nextBest, cohortSize, data, startingDose)
}
\arguments{
\item{nextBest}{see \code{\linkS4class{RuleDesign}}}

\item{cohortSize}{see \code{\linkS4class{RuleDesign}}}

\item{data}{see \code{\linkS4class{RuleDesign}}}

\item{startingDose}{see \code{\linkS4class{RuleDesign}}}
}
\value{
the \code{\linkS4class{RuleDesign}} object
}
\description{
Initialization function for "RuleDesign"
}
\keyword{methods}
