% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{safeInteger}
\alias{safeInteger}
\title{Safe conversion to integer vector}
\usage{
safeInteger(x)
}
\arguments{
\item{x}{the numeric vector}
}
\value{
the integer vector
}
\description{
Safe conversion to integer vector
}
\keyword{internal}
