% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{SafetyWindow-class}
\alias{SafetyWindow-class}
\alias{SafetyWindow}
\title{\code{SafetyWindow}}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{\link{SafetyWindow}} is a class for safety window.
}
\seealso{
\code{\link{SafetyWindowSize}}, \code{\link{SafetyWindowConst}}.
}
