% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\name{SafetyWindowConst}
\alias{SafetyWindowConst}
\title{Initialization function for \code{SafetyWindowConst}}
\usage{
SafetyWindowConst(gap, follow, follow_min)
}
\arguments{
\item{gap}{see \code{\linkS4class{SafetyWindowConst}}}

\item{follow}{see \code{\linkS4class{SafetyWindowConst}}}

\item{follow_min}{see \code{\linkS4class{SafetyWindowConst}}}
}
\value{
the \code{\linkS4class{SafetyWindowConst}} object
}
\description{
Initialization function for \code{SafetyWindowConst}
}
\keyword{methods}
