% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{SafetyWindowSize-class}
\alias{SafetyWindowSize-class}
\alias{.SafetyWindowSize}
\title{Safety window length based on cohort size.
This class is used to decide the rolling rule from the clinical perspective.}
\description{
\code{patientGap} is to be used as follows. If for example, the
cohort size is 4 and we want to specify three time intervals between these
four patients: The interval between the 1st and 2nd patient = 7 units of time
, the interval between the 2nd and 3rd patient = 5 units of time, the interval
between the 3rd and 4th patient = 3 units of time, then we specify
\code{patientGap} to be \code{c(7,5,3)}. Sometimes, we only think the interval
between the 1st and 2nd patient should be increased for the safety consideration
, and the rest time intervals can be the same, whatever the cohort size is, then
we specify \code{patientGap} to be \code{c(7,3)}. The package will automatically
repeat the last element of the vector for the rest time intervals.
}
\details{
Note that \code{sizeIntervals} is to be read as follows. For instance, When we
want to change the \code{patientGap} based on the cohort size, i.e. the time interval
between the 1st and 2nd patient = 9 units of time and the rest time intervals are
5 units of time when the cohort size is equal or larger than 4. And the time
interval between the 1st and 2nd patient = 7 units of time and the rest time
intervals are 3 units of time when the cohort size is smaller than 4, then we
specify \code{sizeIntervals} to be \code{c(0, 4)}. That means, the right
bound of the intervals are exclusive to the interval, and the last interval
goes from the last value until infinity.
}
\section{Slots}{

\describe{
\item{\code{patientGap}}{Observed period of the previous patient before the next patient
can be dosed}

\item{\code{sizeIntervals}}{An integer vector with the left bounds of the relevant
cohort size intervals}

\item{\code{patientFollow}}{The period of time that each patient in the cohort needs to be
followed before the next cohort open}

\item{\code{patientFollowMin}}{At least one patient in the cohort needs to be followed at
the minimal follow up time}
}}

\examples{
# As example, here is the rule for: 
#   having patientGap as (7,3,3,3,...) for cohort size <4
#   and having patientGap as (9,5,5,5...) for cohort size >=4

myWindowLength <- SafetyWindowSize(patientGap = list(c(7,3),c(9,5)),
                                   sizeIntervals = c(1,4),
                                   patientFollow = 7,
                                   patientFollowMin = 14)
}
\keyword{classes}
