% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\name{SafetyWindowSize}
\alias{SafetyWindowSize}
\title{Initialization function for \code{SafetyWindowSize}}
\usage{
SafetyWindowSize(patientGap, sizeIntervals, patientFollow, patientFollowMin)
}
\arguments{
\item{patientGap}{see \code{\linkS4class{SafetyWindowSize}}}

\item{sizeIntervals}{see \code{\linkS4class{SafetyWindowSize}}}

\item{patientFollow}{see \code{\linkS4class{SafetyWindowSize}}}

\item{patientFollowMin}{see \code{\linkS4class{SafetyWindowSize}}}
}
\value{
the \code{\linkS4class{SafetyWindowSize}} object
}
\description{
Initialization function for \code{SafetyWindowSize}
}
\keyword{methods}
