\docType{class}
\name{Samples-class}
\alias{Samples-class}
\title{Class for the MCMC output}
\description{
Class for the MCMC output
}
\section{Slots}{

\describe{
\item{\code{data}}{a list where each entry contains the samples of a (vector-valued)
parameter in a vector/matrix in the format (number of samples) x (dimension
of the parameter).}

\item{\code{options}}{the \code{\linkS4class{McmcOptions}} which have been used}
}}
\keyword{classes}

