% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Samples-class.R
\name{Samples}
\alias{Samples}
\title{Initialization function for "Samples"}
\usage{
Samples(data, options)
}
\arguments{
\item{data}{see \code{\linkS4class{Samples}}}

\item{options}{see \code{\linkS4class{Samples}}}
}
\value{
the \code{\linkS4class{Samples}} object
}
\description{
Initialization function for "Samples"
}
\keyword{methods}
