% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McmcOptions-methods.R, R/Samples-methods.R
\name{sampleSize}
\alias{sampleSize}
\alias{sampleSize,McmcOptions-method}
\alias{sampleSize-McmcOptions}
\alias{sampleSize,Samples-method}
\alias{sampleSize-Samples}
\title{Computing the Number of Samples}
\usage{
sampleSize(object, ...)

\S4method{sampleSize}{McmcOptions}(object, ...)

\S4method{sampleSize}{Samples}(object, ...)
}
\arguments{
\item{object}{(\code{McmcOptions} or \code{Samples})\cr object based on which the number
of samples is computed.}

\item{...}{not used.}
}
\value{
A number of samples.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

A method that computes the number of MCMC samples.
}
\section{Functions}{
\itemize{
\item \code{sampleSize(McmcOptions)}: compute the number of MCMC samples based on \code{McmcOptions}
object.

\item \code{sampleSize(Samples)}: get the number of MCMC samples from \code{Samples} object.

}}
\examples{
# Set up the MCMC option in order to have a burn-in of 10000 iterations and
# then take every other iteration up to a collection of 10000 samples.
my_options <- McmcOptions(burnin = 10000, step = 2, samples = 10000)

sampleSize(my_options)
# Set up the MCMC option in order to have a burn-in of 100 iterations and
# then take every other iteration up to a collection of 200 samples.
my_options <- McmcOptions(burnin = 100, step = 2, samples = 200)

my_samples <- Samples(
  data = list(alpha = rnorm(200), beta = rnorm(200)),
  options = my_options
)

sampleSize(my_samples)
}
