% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_design.R
\name{set_seed}
\alias{set_seed}
\title{Helper Function to Set and Save the RNG Seed}
\usage{
set_seed(seed = NULL)
}
\arguments{
\item{seed}{an object specifying if and how the random number generator
should be initialized ("seeded"). Either \code{NULL} (default) or an
integer that will be used in a call to \code{\link[=set.seed]{set.seed()}} before
simulating the response vectors. If set, the value is saved as the
\code{seed} slot of the returned object. The default, \code{NULL} will
not change the random generator state.}
}
\value{
The integer vector containing the random number generate state will
be returned, in order to call this function with this input to reproduce
the obtained simulation results.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This code is basically copied from \code{stats:::simulate.lm}.
}
