% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design-methods.R
\name{setSeed}
\alias{setSeed}
\title{Helper function to set and save the RNG seed}
\usage{
setSeed(seed = NULL)
}
\arguments{
\item{seed}{an object specifying if and how the random number generator
should be initialized (\dQuote{seeded}). Either \code{NULL} (default) or an
integer that will be used in a call to \code{\link{set.seed}} before
simulating the response vectors. If set, the value is saved as the
\code{seed} slot of the returned object. The default, \code{NULL} will
not change the random generator state.}
}
\value{
The RNGstate will be returned, in order to call this function
with this input to reproduce the obtained simulation results
}
\description{
This is basically copied from simulate.lm
}
\author{
Daniel Sabanes Bove \email{sabanesd@roche.com}
}
\keyword{programming}
