% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations-methods.R
\name{show,ComboSimulations-method}
\alias{show,ComboSimulations-method}
\alias{show-ComboSimulations}
\title{Show \code{ComboSimulations} Objects}
\usage{
\S4method{show}{ComboSimulations}(object)
}
\arguments{
\item{object}{(\code{ComboSimulations})\cr the object we want to print.}
}
\value{
Invisibly returns the object itself.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Display a brief representation of the \code{\link{ComboSimulations}} object.
}
