% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations-methods.R
\name{show,ComboSimulationsSummary-method}
\alias{show,ComboSimulationsSummary-method}
\alias{show-ComboSimulationsSummary}
\title{Show the Summary of Combination Simulations}
\usage{
\S4method{show}{ComboSimulationsSummary}(object)
}
\arguments{
\item{object}{(\code{ComboSimulationsSummary})
the object we want to print.}
}
\value{
Invisibly returns a data frame of the results with one row and
appropriate column names.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Display a summary of combination simulation results.
}
