% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Simulations-methods.R
\docType{methods}
\name{show,DualSimulationsSummary-method}
\alias{show,DualSimulationsSummary-method}
\title{Show the summary of the dual-endpoint simulations}
\usage{
\S4method{show}{DualSimulationsSummary}(object)
}
\arguments{
\item{object}{the \code{\linkS4class{DualSimulationsSummary}} object we want
to print}
}
\value{
invisibly returns a data frame of the results with one row and
appropriate column names
}
\description{
Show the summary of the dual-endpoint simulations
}
\keyword{methods}

