% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations-methods.R
\name{show,GeneralSimulations-method}
\alias{show,GeneralSimulations-method}
\alias{show-GeneralSimulations}
\title{Show \code{Simulations} Objects}
\usage{
\S4method{show}{GeneralSimulations}(object)
}
\arguments{
\item{object}{(\code{GeneralSimulations})\cr the object we want to print.}
}
\value{
Invisibly returns the object itself.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Display a brief representation of the \link{GeneralSimulations} object.
}
