\docType{methods}
\name{show,GeneralSimulations-summary-method}
\alias{show,GeneralSimulations-summary-method}
\title{Show the summary of the simulations}
\usage{
\S4method{show}{GeneralSimulations-summary}(object)
}
\arguments{
  \item{object}{the
  \code{\linkS4class{GeneralSimulations-summary}} object we
  want to print}
}
\value{
invisibly returns a data frame of the results with one row
and appropriate column names
}
\description{
Show the summary of the simulations
}
\keyword{methods}

