% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations-methods.R
\name{show,GeneralSimulationsSummary-method}
\alias{show,GeneralSimulationsSummary-method}
\alias{show-GeneralSimulationsSummary}
\title{Show the Summary of the Simulations}
\usage{
\S4method{show}{GeneralSimulationsSummary}(object)
}
\arguments{
\item{object}{(\code{GeneralSimulationsSummary})\cr the object we want to print.}
}
\value{
Invisibly returns a data frame of the results with one row and
appropriate column names.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Display a summary of general simulation results.
}
