% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-methods.R
\name{show,HierarchicalData-method}
\alias{show,HierarchicalData-method}
\alias{show-HierarchicalData}
\title{Show \code{HierarchicalData} Objects}
\usage{
\S4method{show}{HierarchicalData}(object)
}
\arguments{
\item{object}{(\code{HierarchicalData})\cr the object we want to print.}
}
\value{
Invisibly returns the object itself.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Display a brief representation of the \code{\link{HierarchicalData}} object.
}
