% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design-methods.R
\name{show,HierarchicalDesign-method}
\alias{show,HierarchicalDesign-method}
\alias{show-HierarchicalDesign}
\title{Show \code{HierarchicalDesign} Objects}
\usage{
\S4method{show}{HierarchicalDesign}(object)
}
\arguments{
\item{object}{(\code{HierarchicalDesign})\cr the object we want to print.}
}
\value{
Invisibly returns the object itself.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Display a brief representation of the \code{\link{HierarchicalDesign}} object.
}
