% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-methods.R
\name{show,HierarchicalModel-method}
\alias{show,HierarchicalModel-method}
\alias{show-HierarchicalModel}
\title{Show \code{HierarchicalModel} Objects}
\usage{
\S4method{show}{HierarchicalModel}(object)
}
\arguments{
\item{object}{(\code{HierarchicalModel})\cr the object we want to print.}
}
\value{
Invisibly returns the object itself.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Display a brief representation of the \code{\link{HierarchicalModel}} object.
}
