% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations-methods.R
\name{show,HierarchicalSimulations-method}
\alias{show,HierarchicalSimulations-method}
\alias{show-HierarchicalSimulations}
\title{Show \code{HierarchicalSimulations} Objects}
\usage{
\S4method{show}{HierarchicalSimulations}(object)
}
\arguments{
\item{object}{(\code{HierarchicalSimulations})\cr the object we want to print.}
}
\value{
Invisibly returns the object itself.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Display a brief representation of the \code{\link{HierarchicalSimulations}} object.
}
