% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations-methods.R
\name{show,HierarchicalSimulationsSummary-method}
\alias{show,HierarchicalSimulationsSummary-method}
\alias{show-HierarchicalSimulationsSummary}
\title{Show the Summary of Hierarchical Simulations}
\usage{
\S4method{show}{HierarchicalSimulationsSummary}(object)
}
\arguments{
\item{object}{(\code{HierarchicalSimulationsSummary})\cr the object we want to print.}
}
\value{
Invisibly returns a named list of arm-level summary data frames.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Display arm-specific hierarchical simulation summaries.
}
