% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Simulations-methods.R
\docType{methods}
\name{show,PseudoSimulationsSummary-method}
\alias{show,PseudoSimulationsSummary-method}
\title{Show the summary of the simulations}
\usage{
\S4method{show}{PseudoSimulationsSummary}(object)
}
\arguments{
\item{object}{the \code{\linkS4class{PseudoSimulationsSummary}} object we want
to print}
}
\value{
invisibly returns a data frame of the results with one row and
appropriate column names
}
\description{
Show the summary of the simulations
}
\examples{
##obtain the plot for the simulation results
##If only DLE responses are considered in the simulations
##Specified your simulations when no DLE samples are used
##(Please refer to desgin-method 'simulate TDDesign' examples for details)
mySim <- simulate(design,
                  args=NULL,
                  truth=myTruth,
                  nsim=10,
                  seed=819,
                  parallel=FALSE)
##If DLE samples are involved
##Please refer to design-method 'simulate TDsamplesDesign' examples for details
mySim <- simulate(design,
                  args=NULL,
                  truth=myTruth,
                  nsim=10,
                  seed=819,
                  mcmcOptions=options,
                  parallel=FALSE)
##Then produce a summary of your simulations
MYSUM <- summary(mySim,
         truth=myTruth)
##show the summary of the simulated results in a data frame
show(MYSUM)
}
\keyword{methods}

