\docType{methods}
\name{show,Simulations-summary-method}
\alias{show,Simulations-summary-method}
\title{Show the summary of the simulations}
\usage{
\S4method{show}{Simulations-summary}(object)
}
\arguments{
  \item{object}{the
  \code{\linkS4class{Simulations-summary}} object we want
  to print}
}
\value{
invisibly returns a data frame of the results with one row
and appropriate column names
}
\description{
Show the summary of the simulations
}
\keyword{methods}

