% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Simulations-methods.R
\docType{methods}
\name{show,SimulationsSummary-method}
\alias{show,SimulationsSummary-method}
\title{Show the summary of the simulations}
\usage{
\S4method{show}{SimulationsSummary}(object)
}
\arguments{
\item{object}{the \code{\linkS4class{SimulationsSummary}} object we want
to print}
}
\value{
invisibly returns a data frame of the results with one row and
appropriate column names
}
\description{
Show the summary of the simulations
}
\keyword{methods}

