% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design-methods.R
\name{simulate,DesignCombo-method}
\alias{simulate,DesignCombo-method}
\title{Simulate outcomes from a two-drug combination CRM design}
\usage{
\S4method{simulate}{DesignCombo}(
  object,
  nsim = 1L,
  seed = NULL,
  truth,
  args = NULL,
  firstSeparate = FALSE,
  mcmcOptions = McmcOptions(),
  parallel = FALSE,
  nCores = min(parallel::detectCores(), 5),
  derive = list(),
  ...
)
}
\arguments{
\item{object}{the \code{\link{DesignCombo}} object we want to simulate data from}

\item{nsim}{(\code{count})
the number of simulations (default: 1).}

\item{seed}{see \code{\link[=set_seed]{set_seed()}}}

\item{truth}{(\code{function})
a function that takes as input a dose combination (named numeric vector of
length 2) and returns the true DLT probability. Additional arguments can
be supplied in \code{args}.}

\item{args}{(\code{data.frame})
data frame with arguments for the \code{truth} function. The column names
correspond to the argument names and rows are recycled across simulations.}

\item{firstSeparate}{(\code{flag})
enroll the first patient separately from the rest of the cohort? If yes,
the cohort is closed when a DLT occurs in this first patient.}

\item{mcmcOptions}{(\code{\link{McmcOptions}})
MCMC options used for each trial evaluation.}

\item{parallel}{(\code{flag})
should simulation runs be parallelized?}

\item{nCores}{(\code{count})
number of cores used when \code{parallel = TRUE}.}

\item{derive}{(\code{list})
named list of functions deriving extra statistics from posterior toxicity
samples at the final recommended combination.}

\item{...}{not used.}
}
\value{
an object of class \code{\link{ComboSimulations}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\note{
Backfill cohorts are not yet implemented for \code{\link{DesignCombo}} simulations
and therefore lead to an error if used.
}
\examples{
# nolint start

design_combo <- .DefaultDesignCombo()

true_tox_combo <- function(dose) {
  plogis(-6 + 0.08 * dose[1] + 0.06 * dose[2] + 0.001 * dose[1] * dose[2])
}

options <- McmcOptions(
  burnin = 50,
  step = 1,
  samples = 50,
  rng_kind = "Mersenne-Twister",
  rng_seed = 1
)

my_sims_combo <- simulate(
  design_combo,
  truth = true_tox_combo,
  nsim = 1,
  seed = 819,
  mcmcOptions = options,
  parallel = FALSE
)

# nolint end
}
