% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Design-methods.R
\docType{methods}
\name{simulate,DualDesign-method}
\alias{simulate,DualDesign-method}
\title{Simulate outcomes from a dual-endpoint design}
\usage{
\S4method{simulate}{DualDesign}(object, nsim = 1L, seed = NULL, trueTox,
  trueBiomarker, args = NULL, sigma2W, rho = 0, firstSeparate = FALSE,
  mcmcOptions = McmcOptions(), parallel = FALSE, ...)
}
\arguments{
\item{object}{the \code{\linkS4class{DualDesign}} object we want to simulate
data from}

\item{nsim}{the number of simulations (default: 1)}

\item{seed}{see \code{\link{setSeed}}}

\item{trueTox}{a function which takes as input a dose (vector) and returns the
true probability (vector) for toxicity. Additional arguments can be supplied
in \code{args}.}

\item{trueBiomarker}{a function which takes as input a dose (vector) and
returns the true biomarker level (vector). Additional arguments can be
supplied in \code{args}.}

\item{args}{data frame with arguments for the \code{trueTox} and
\code{trueBiomarker} function. The column names correspond to the argument
names, the rows to the values of the arguments. The rows are appropriately
recycled in the \code{nsim} simulations.}

\item{sigma2W}{variance for the biomarker measurements}

\item{rho}{correlation between toxicity and biomarker measurements (default:
0)}

\item{firstSeparate}{enroll the first patient separately from the rest of
the cohort? (not default) If yes, the cohort will be closed if a DLT occurs
in this patient.}

\item{mcmcOptions}{object of class \code{\linkS4class{McmcOptions}},
giving the MCMC options for each evaluation in the trial. By default,
the standard options are used}

\item{parallel}{should the simulation runs be parallelized across the
clusters of the computer? (not default)}

\item{\dots}{not used}
}
\value{
an object of class \code{\linkS4class{DualSimulations}}
}
\description{
Simulate outcomes from a dual-endpoint design
}
\keyword{methods}

