% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design-methods.R
\name{simulate-HierarchicalDesign-method}
\alias{simulate,HierarchicalDesign-method}
\title{Simulate outcomes from a hierarchical CRM design}
\usage{
\S4method{simulate}{HierarchicalDesign}(
  object,
  nsim = 1L,
  seed = NULL,
  truth,
  truthResponse = plogis,
  args = NULL,
  firstSeparate = FALSE,
  mcmcOptions = McmcOptions(),
  parallel = FALSE,
  nCores = min(parallel::detectCores(), 5),
  derive = list(),
  ...
)
}
\arguments{
\item{object}{the \code{\link{HierarchicalDesign}} object we want to simulate data from.}

\item{nsim}{(\code{count})\cr the number of simulations.}

\item{seed}{see \code{\link{set_seed}}.}

\item{truth}{(\code{function} or named \code{list} of \code{function})\cr true DLT
probability function(s). If a list is supplied, names must match the
hierarchical arms.}

\item{truthResponse}{(\code{function} or named \code{list} of \code{function})\cr true
response probability function(s).}

\item{args}{(\code{data.frame})\cr arguments for the truth functions.}

\item{firstSeparate}{(\code{flag})\cr enroll the first patient of a cohort
separately?}

\item{mcmcOptions}{(\code{\link{McmcOptions}})\cr options for each MCMC fit.}

\item{parallel}{(\code{flag})\cr should simulation runs be parallelized?}

\item{nCores}{(\code{count})\cr number of cores used when \code{parallel = TRUE}.}

\item{derive}{(\code{list})\cr named list of additional statistics.}

\item{...}{not used.}
}
\value{
an object of class \code{\link{HierarchicalSimulations}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Simulate outcomes from a hierarchical CRM design.
}
