% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Design-methods.R
\docType{methods}
\name{simulate,TDDesign-method}
\alias{simulate,TDDesign-method}
\title{Methods to simulate using 'LogisticIndepBeta' model without samples}
\usage{
\S4method{simulate}{TDDesign}(object, nsim = 1L, seed = NULL, truth,
  args = NULL, firstSeparate = FALSE, parallel = FALSE, ...)
}
\description{
Methods to simulate using 'LogisticIndepBeta' model without samples
}
\keyword{methods}

