% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Design-methods.R
\docType{methods}
\name{simulate,TDsamplesDesign-method}
\alias{simulate,TDsamplesDesign-method}
\title{Methods to simulate using 'LogisticIndepBeta' model with samples}
\usage{
\S4method{simulate}{TDsamplesDesign}(object, nsim = 1L, seed = NULL, truth,
  args = NULL, firstSeparate = FALSE, mcmcOptions = McmcOptions(),
  parallel = FALSE, ...)
}
\description{
Methods to simulate using 'LogisticIndepBeta' model with samples
}
\keyword{methods}

