\docType{class}
\name{Simulations-summary-class}
\alias{Simulations-summary-class}
\title{Class for the summary of model-based simulations output}
\description{
Additional slots that require the model fits.
}
\section{Slots}{

\describe{
\item{\code{fitAtDoseMostSelected}}{fitted toxicity rate at dose most often selected}

\item{\code{meanFit}}{list with the average, lower (2.5%) and upper (97.5%)
quantiles of the mean fitted toxicity at each dose level}
}}
\keyword{classes}

