% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations-class.R
\name{Simulations}
\alias{Simulations}
\title{Initialization function for the "Simulations" class}
\usage{
Simulations(fit, stopReasons, ...)
}
\arguments{
\item{fit}{see \code{\linkS4class{Simulations}}}

\item{stopReasons}{see \code{\linkS4class{Simulations}}}

\item{\dots}{additional parameters from \code{\link{GeneralSimulations}}}
}
\value{
the \code{\linkS4class{Simulations}} object
}
\description{
Initialization function for the "Simulations" class
}
\keyword{methods}
