% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations-class.R
\docType{class}
\name{SimulationsSummary-class}
\alias{SimulationsSummary-class}
\alias{.SimulationsSummary}
\alias{SimulationsSummary}
\alias{.DefaultSimulationsSummary}
\title{\code{SimulationsSummary}}
\usage{
.DefaultSimulationsSummary()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

In addition to the slots in the parent class \code{\link{GeneralSimulationsSummary}},
it contains two slots with model fit information.
}
\section{Slots}{

\describe{
\item{\code{stop_report}}{(\code{matrix})\cr matrix of stopping rule outcomes}

\item{\code{fit_at_dose_most_selected}}{(\code{numeric})\cr fitted toxicity rate at dose most often selected}

\item{\code{additional_stats}}{(\code{list})\cr list of additional statistical summary}

\item{\code{overdose_prob}}{(\code{numeric})\cr posterior probabilities that toxicity at
the recommended dose exceeds the upper target bound, for simulations in
which a dose was recommended}

\item{\code{mean_fit}}{(\code{list})\cr list with the average, lower (2.5\%) and upper (97.5\%)
quantiles of the mean fitted toxicity at each dose level}
}}

\note{
Typically, end users will not use the \code{.DefaultSimulationsSummary()} function.
}
