% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-methods.R
\name{singleDrugData}
\alias{singleDrugData}
\alias{singleDrugData,DataCombo-method}
\alias{singleDrugData-DataCombo}
\title{Extracting Single-Drug Data from Combination Data}
\usage{
singleDrugData(object, drug, ...)

\S4method{singleDrugData}{DataCombo}(object, drug)
}
\arguments{
\item{object}{(\code{DataCombo})
object from which single-drug data are extracted.}

\item{drug}{(\code{string})
name of the drug to extract. Must be one of \code{object@drugNames}.}

\item{...}{not used.}
}
\value{
A \code{\link{Data}} object containing dose, dose grid and patient-level
outcomes corresponding to the selected drug.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

A method that extracts one drug-specific \code{\link{Data}} object from a
two-drug \code{\link{DataCombo}} object.
}
\examples{
# Example of extracting single-drug data from two-drug combination data.

my_combo_data <- DataCombo(
  x = cbind(
    drug1 = c(10, 10, 20, 20),
    drug2 = c(20, 20, 20, 40)
  ),
  y = c(0L, 1L, 0L, 0L),
  ID = 1L:4L,
  cohort = c(1L, 2L, 3L, 4L),
  doseGrid = list(
    drug1 = c(10, 20, 30),
    drug2 = c(20, 40)
  )
)

# Extract the data for one specific drug.
singleDrugData(my_combo_data, "drug1")
}
