\docType{methods}
\name{size,CohortSizeConst,ANY,Data-method}
\alias{size,CohortSizeConst,ANY,Data-method}
\title{Constant cohort size}
\usage{
\S4method{size}{CohortSizeConst,ANY,Data}(cohortSize, dose, data, ...)
}
\arguments{
  \item{cohortSize}{The rule, an object of class
  \code{\linkS4class{CohortSize}}}

  \item{dose}{the next dose}

  \item{data}{The data input, an object of class
  \code{\linkS4class{Data}}}

  \item{...}{additional arguments}
}
\description{
Constant cohort size
}

