\docType{methods}
\name{size,CohortSizeParts,ANY,DataParts-method}
\alias{size,CohortSizeParts,ANY,DataParts-method}
\title{Cohort size based on the parts}
\usage{
\S4method{size}{CohortSizeParts,ANY,DataParts}(cohortSize, dose, data, ...)
}
\arguments{
  \item{cohortSize}{The rule, an object of class
  \code{\linkS4class{CohortSize}}}

  \item{dose}{the next dose}

  \item{data}{The data input, an object of class
  \code{\linkS4class{Data}}}

  \item{...}{additional arguments}
}
\description{
Cohort size based on the parts
}

