% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rules-methods.R
\docType{methods}
\name{size}
\alias{size}
\alias{size,CohortSizeConst,ANY,Data-method}
\alias{size,CohortSizeDLT,ANY,Data-method}
\alias{size,CohortSizeMax,ANY,Data-method}
\alias{size,CohortSizeMin,ANY,Data-method}
\alias{size,CohortSizeParts,ANY,DataParts-method}
\alias{size,CohortSizeRange,ANY,Data-method}
\title{Determine the size of the next cohort}
\usage{
size(cohortSize, dose, data, ...)

\S4method{size}{CohortSizeRange,ANY,Data}(cohortSize, dose, data, ...)

\S4method{size}{CohortSizeDLT,ANY,Data}(cohortSize, dose, data, ...)

\S4method{size}{CohortSizeMax,ANY,Data}(cohortSize, dose, data, ...)

\S4method{size}{CohortSizeMin,ANY,Data}(cohortSize, dose, data, ...)

\S4method{size}{CohortSizeConst,ANY,Data}(cohortSize, dose, data, ...)

\S4method{size}{CohortSizeParts,ANY,DataParts}(cohortSize, dose, data, ...)
}
\arguments{
\item{cohortSize}{The rule, an object of class
\code{\linkS4class{CohortSize}}}

\item{dose}{the next dose}

\item{data}{The data input, an object of class \code{\linkS4class{Data}}}

\item{\dots}{additional arguments}
}
\value{
the size as integer value
}
\description{
This function determines the size of the next cohort.
}
\section{Methods (by class)}{
\itemize{
\item \code{cohortSize = CohortSizeRange,dose = ANY,data = Data}: Determine the cohort size based on the range into which the
next dose falls into

\item \code{cohortSize = CohortSizeDLT,dose = ANY,data = Data}: Determine the cohort size based on the number of DLTs so
far

\item \code{cohortSize = CohortSizeMax,dose = ANY,data = Data}: Size based on maximum of multiple cohort size rules

\item \code{cohortSize = CohortSizeMin,dose = ANY,data = Data}: Size based on minimum of multiple cohort size rules

\item \code{cohortSize = CohortSizeConst,dose = ANY,data = Data}: Constant cohort size

\item \code{cohortSize = CohortSizeParts,dose = ANY,data = DataParts}: Cohort size based on the parts
}}
\keyword{methods}

