% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_knitr_Design.R
\docType{class}
\name{StartingDose-class}
\alias{StartingDose-class}
\alias{.StartingDose}
\alias{StartingDose}
\alias{.DefaultStartingDose}
\title{\code{StartingDose}}
\usage{
StartingDose(starting_dose)

.DefaultStartingDose()
}
\arguments{
\item{starting_dose}{(\code{positive_number})\cr see slot definition.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{\link{StartingDose}} is a simple wrapper class for the \code{startingDose} slot of all
design classes. It is used internally by \code{knit_print} methods
}
\section{Slots}{

\describe{
\item{\code{starting_dose}}{(\code{numeric})\cr the starting dose}
}}

\note{
Typically, end users will not use the \code{.DefaultStartingDose()} function.
}
\keyword{internal}
