% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\name{StoppingAll}
\alias{StoppingAll}
\title{Initialization function for "StoppingAll"}
\usage{
StoppingAll(stopList)
}
\arguments{
\item{stopList}{see \code{\linkS4class{StoppingAll}}}
}
\value{
the \code{\linkS4class{StoppingAll}} object
}
\description{
Initialization function for "StoppingAll"
}
\keyword{methods}
