% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\name{StoppingAny}
\alias{StoppingAny}
\title{Initialization function for "StoppingAny"}
\usage{
StoppingAny(stopList)
}
\arguments{
\item{stopList}{see \code{\linkS4class{StoppingAny}}}
}
\value{
the \code{\linkS4class{StoppingAny}} object
}
\description{
Initialization function for "StoppingAny"
}
\keyword{methods}
