% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{StoppingCIRatio-class}
\alias{.StoppingCIRatio}
\alias{StoppingCIRatio-class}
\title{Stop based on a target ratio, the ratio of the upper to the lower
95\% credibility interval of the estimate}
\description{
Stop based on a target ratio, the ratio of the upper to the lower
95\% credibility interval of the estimate
}
\section{Slots}{

\describe{
\item{\code{targetRatio}}{the target ratio of the upper to the lower of the 95\% credibility interval of the
estimate that required to stop a trial}
}}
\examples{
##Define the target stopping ratio of 5
##This is a ratio of the upper to the lower 95\% credibility interval of the estimates
myStopping <- StoppingCIRatio(targetRatio=5)
}
\keyword{classes}

