% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rules-class.R
\name{StoppingCIRatio}
\alias{StoppingCIRatio}
\title{Initialization function for "StoppingCIRatio"}
\usage{
StoppingCIRatio(targetRatio)
}
\arguments{
\item{targetRatio}{please refer to \code{\linkS4class{StoppingCIRatio}} class object}
}
\value{
the \code{\linkS4class{StoppingCIRatio}} class object
}
\description{
Initialization function for "StoppingCIRatio"
}
\keyword{methods}

