% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\name{StoppingCohortsNearDose}
\alias{StoppingCohortsNearDose}
\title{Initialization function for "StoppingCohortsNearDose"}
\usage{
StoppingCohortsNearDose(nCohorts, percentage)
}
\arguments{
\item{nCohorts}{see \code{\linkS4class{StoppingCohortsNearDose}}}

\item{percentage}{see \code{\linkS4class{StoppingCohortsNearDose}}}
}
\value{
the \code{\linkS4class{StoppingCohortsNearDose}} object
}
\description{
Initialization function for "StoppingCohortsNearDose"
}
\keyword{methods}
