% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{StoppingExternal-class}
\alias{StoppingExternal-class}
\alias{.StoppingExternal}
\alias{StoppingExternal}
\alias{.DefaultStoppingExternal}
\title{\code{StoppingExternal}}
\usage{
StoppingExternal(report_label = NA_character_)

.DefaultStoppingExternal(report_label = NA_character_)
}
\arguments{
\item{report_label}{(\code{string} or \code{NA})\cr see slot definition.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{\link{StoppingExternal}} is the class for stopping based on an external flag.
}
\note{
Typically, end users will not use the \code{.DefaultStoppingExternal()} function.
}
\examples{
my_stopping <- StoppingExternal()
}
