% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{StoppingGstarCIRatio-class}
\alias{StoppingGstarCIRatio-class}
\alias{.StoppingGstarCIRatio}
\title{Stop based on a target ratio, the ratio of the upper to the lower
95\% credibility interval of the estimate of the minimum of the dose which gives the maximum gain (Gstar) and 
the TD end of trial, the dose with probability of DLE equals to the target 
probability of DLE used at the end of a trial.}
\description{
Stop based on a target ratio, the ratio of the upper to the lower
95\% credibility interval of the estimate of the minimum of the dose which gives the maximum gain (Gstar) and 
the TD end of trial, the dose with probability of DLE equals to the target 
probability of DLE used at the end of a trial.
}
\section{Slots}{

\describe{
\item{\code{targetRatio}}{the target ratio of the upper to the lower of the 95\% credibility interval of the 
estimate that required to stop a trial}

\item{\code{targetEndOfTrial}}{the target probability of DLE to be used at the end of a trial}
}}

\examples{
##Define the target stopping ratio of 5 and 
##the target probability of DLE to be used at the end of a trial
##This is a ratio of the upper to the lower 95\% credibility interval of the estimates
myStopping <- StoppingGstarCIRatio(targetRatio=5,
                                   targetEndOfTrial=0.3)
}
\keyword{classes}
