% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\name{StoppingGstarCIRatio}
\alias{StoppingGstarCIRatio}
\title{Initialization function for "StoppingGstarCIRatio"}
\usage{
StoppingGstarCIRatio(targetRatio, targetEndOfTrial)
}
\arguments{
\item{targetRatio}{please refer to \code{\linkS4class{StoppingGstarCIRatio}} class object}

\item{targetEndOfTrial}{please refer to \code{\linkS4class{StoppingGstarCIRatio}} class object}
}
\value{
the \code{\linkS4class{StoppingGstarCIRatio}} class object
}
\description{
Initialization function for "StoppingGstarCIRatio"
}
\keyword{methods}
