% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rules-class.R
\name{StoppingGstarsamplesCIRatio}
\alias{StoppingGstarsamplesCIRatio}
\title{Initialization function for "StoppingGstarsamplesCIRatio"}
\usage{
StoppingGstarsamplesCIRatio(targetRatio, targetEndOfTrial)
}
\arguments{
\item{targetRatio}{please refer to \code{\linkS4class{StoppingGstarsamplesCIRatio}} class object}

\item{targetEndOfTrial}{please refer to \code{\linkS4class{StoppingGstarsamplesCIRatio}} class object}
}
\value{
the \code{\linkS4class{StoppingGstarsamplesCIRatio}} class object
}
\description{
Initialization function for "StoppingGstarsamplesCIRatio"
}
\keyword{methods}

