% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\name{StoppingHighestDose}
\alias{StoppingHighestDose}
\title{Initialization function for "StoppingHighestDose"}
\usage{
StoppingHighestDose()
}
\value{
the \code{\linkS4class{StoppingHighestDose}} object
}
\description{
Initialization function for "StoppingHighestDose"
}
\keyword{methods}
