% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\name{StoppingList}
\alias{StoppingList}
\title{Initialization function for "StoppingList"}
\usage{
StoppingList(stopList, summary)
}
\arguments{
\item{stopList}{see \code{\linkS4class{StoppingList}}}

\item{summary}{see \code{\linkS4class{StoppingList}}}
}
\value{
the \code{\linkS4class{StoppingList}} object
}
\description{
Initialization function for "StoppingList"
}
\keyword{methods}
