% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{StoppingMaxGainCIRatio-class}
\alias{StoppingMaxGainCIRatio-class}
\alias{.StoppingMaxGainCIRatio}
\alias{StoppingMaxGainCIRatio}
\title{\code{StoppingMaxGainCIRatio}}
\usage{
StoppingMaxGainCIRatio(target_ratio, prob_target, report_label = character(0))
}
\arguments{
\item{target_ratio}{(\code{numeric})\cr see slot definition.}

\item{prob_target}{(\code{proportion})\cr see slot definition.}

\item{report_label}{(\code{character})\cr see slot definition.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{\link{StoppingMaxGainCIRatio}} is the class for testing a stopping rule that is based
on a target ratio of the 95\% credibility interval. Specifically, this is the
ratio of the upper to the lower bound of the 95\% credibility interval's
estimate of the:
(1) target dose (i.e. a dose that corresponds to a given target
probability of the occurrence of a DLT \code{prob_target}), or
(2) max gain dose (i.e. a dose which gives the maximum gain),
depending on which one out of these two is smaller.
}
\section{Slots}{

\describe{
\item{\code{target_ratio}}{(\code{numeric})\cr target for the ratio of the 95\% credibility
interval's estimate, that is required to stop a trial.}

\item{\code{prob_target}}{(\code{proportion})\cr the target probability of the occurrence
of a DLT.}
}}

\examples{

# Define the target stopping ratio (5) and the target probability of a DLT to
# be used (0.3).
my_stopping <- StoppingMaxGainCIRatio(target_ratio = 5, prob_target = 0.3)
}
