\docType{class}
\name{StoppingMaxPatients-class}
\alias{StoppingMaxPatients-class}
\title{Stop based on maximum number of patients}
\description{
Stop based on maximum number of patients
}
\section{Slots}{

\describe{
\item{\code{nPatients}}{maximum allowed number of patients}
}}
\keyword{classes}

