% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\name{StoppingMinCohorts}
\alias{StoppingMinCohorts}
\title{Initialization function for "StoppingMinCohorts"}
\usage{
StoppingMinCohorts(nCohorts)
}
\arguments{
\item{nCohorts}{see \code{\linkS4class{StoppingMinCohorts}}}
}
\value{
the \code{\linkS4class{StoppingMinCohorts}} object
}
\description{
Initialization function for "StoppingMinCohorts"
}
\keyword{methods}
