\docType{class}
\name{StoppingMinPatients-class}
\alias{StoppingMinPatients-class}
\title{Stop based on minimum number of patients}
\description{
Stop based on minimum number of patients
}
\section{Slots}{

\describe{
\item{\code{nPatients}}{minimum required number of patients}
}}
\keyword{classes}

