% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{StoppingMissingDose-class}
\alias{StoppingMissingDose-class}
\alias{.StoppingMissingDose}
\alias{StoppingMissingDose}
\alias{.DefaultStoppingMissingDose}
\title{\code{StoppingMissingDose}}
\usage{
StoppingMissingDose()

.DefaultStoppingMissingDose()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{\link{StoppingMissingDose}} is the class for stopping based on NA returned by
next best dose.
}
\note{
Typically, end users will not use the \code{.DefaultStoppingMissingDose()} function.
}
\examples{
# The rule for stopping the study if NA or Placebo is returned as
# next best dose.
my_stopping <- StoppingMissingDose()
}
