% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{StoppingMTDCV-class}
\alias{StoppingMTDCV-class}
\alias{.StoppingMTDCV}
\alias{StoppingMTDCV}
\title{Stop based on precision of MTD calculated as CV(MTD)}
\usage{
StoppingMTDCV(target, threshCV)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\section{Slots}{

\describe{
\item{\code{target}}{(\code{numeric})\cr toxicity target of MTD}

\item{\code{threshCV}}{(\code{numeric})\cr threshold for CV to be considered accurate enough
to stop the trial}
}}

\examples{

# As example, here is the rule for: 
#   stopping the study if the MTD estimation is precise enough, i.e. if robust
#   coefficient of variation of the MTD is below 40\%

myStopping <- StoppingMTDCV(target = 0.3,
                            threshCV = 40)
}
