% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\name{StoppingMTDdistribution}
\alias{StoppingMTDdistribution}
\title{Initialization function for "StoppingMTDdistribution"}
\usage{
StoppingMTDdistribution(target, thresh, prob)
}
\arguments{
\item{target}{see \code{\linkS4class{StoppingMTDdistribution}}}

\item{thresh}{see \code{\linkS4class{StoppingMTDdistribution}}}

\item{prob}{see \code{\linkS4class{StoppingMTDdistribution}}}
}
\value{
the \code{\linkS4class{StoppingMTDdistribution}} object
}
\description{
Initialization function for "StoppingMTDdistribution"
}
\keyword{methods}
