% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{StoppingOrdinal-class}
\alias{StoppingOrdinal-class}
\alias{.StoppingOrdinal}
\alias{StoppingOrdinal}
\alias{.DefaultStoppingOrdinal}
\title{\code{StoppingOrdinal}}
\usage{
StoppingOrdinal(grade, rule)

.DefaultStoppingOrdinal()
}
\arguments{
\item{grade}{(\code{integer})\cr see slot definition.}

\item{rule}{(\code{Stopping})\cr see slot definition.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{\link{StoppingOrdinal}} is the class for stopping based on a Stopping rule applied
to a specific toxicity grade in an ordinal CRM trial
}
\section{Slots}{

\describe{
\item{\code{grade}}{(\code{integer})\cr the grade to which the rule should be applied}

\item{\code{rule}}{(\code{Stopping})\cr the rule to apply}
}}

\note{
Typically, end users will not use the \code{.DefaultStoppingOrdinal()} function.
}
\examples{
StoppingOrdinal(
  1L,
  StoppingTargetProb(target = c(0.2, 0.35), prob = 0.6)
)
}
