% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{StoppingPatientsNearDose-class}
\alias{StoppingPatientsNearDose-class}
\alias{.StoppingPatientsNearDose}
\alias{StoppingPatientsNearDose}
\title{\code{StoppingPatientsNearDose}}
\usage{
StoppingPatientsNearDose(nPatients, percentage = 50)
}
\arguments{
\item{nPatients}{(\code{number})\cr see slot definition.}

\item{percentage}{(\code{number})\cr see slot definition.}
}
\description{
Stop based on number of patients near to next best dose
}
\section{Slots}{

\describe{
\item{\code{nPatients}}{number of required patients}

\item{\code{percentage}}{percentage (between 0 and 100) within the next best dose
the patients must lie}
}}

\examples{

# As example, here is the rule for: 
#   stopping the study if at least 9 patients were dosed at a dose within (1 +/- 0.2) 
#   of the next best dose

myStopping <- StoppingPatientsNearDose(nPatients = 9,
                                       percentage = 0.2)

}
\keyword{classes}
