% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{StoppingSpecificDose-class}
\alias{StoppingSpecificDose-class}
\alias{.StoppingSpecificDose}
\alias{StoppingSpecificDose}
\title{\code{StoppingSpecificDose}}
\usage{
StoppingSpecificDose(rule, dose)
}
\arguments{
\item{rule}{(\verb{stopping rule})\cr see slot definition.}

\item{dose}{(\code{number})\cr see slot definition.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{\link{StoppingSpecificDose}} is the class for testing a stopping rule at specific dose
of teh dose grid and not at the next best dose
}
\section{Slots}{

\describe{
\item{\code{rule}}{(\code{rule})\cr a stopping rule available in this package.}

\item{\code{dose}}{(\code{number})\cr a dose that is defined as part of the dose grid of the data.}
}}

