\docType{class}
\name{StoppingTargetBiomarker-class}
\alias{StoppingTargetBiomarker-class}
\title{Stop based on probability of target biomarker}
\description{
Stop based on probability of target biomarker
}
\section{Slots}{

\describe{
\item{\code{target}}{the biomarker level, relative to the maximum, that
needs to be reached}

\item{\code{prob}}{required target probability
for reaching sufficient precision}
}}
\keyword{classes}

