% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\name{StoppingTargetBiomarker}
\alias{StoppingTargetBiomarker}
\title{Initialization function for "StoppingTargetBiomarker"}
\usage{
StoppingTargetBiomarker(target, scale = c("relative", "absolute"), prob)
}
\arguments{
\item{target}{see \code{\linkS4class{StoppingTargetBiomarker}}}

\item{scale}{see \code{\linkS4class{StoppingTargetBiomarker}}}

\item{prob}{see \code{\linkS4class{StoppingTargetBiomarker}}}
}
\value{
the \code{\linkS4class{StoppingTargetBiomarker}} object
}
\description{
Initialization function for "StoppingTargetBiomarker"
}
\keyword{methods}
