% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{StoppingTargetProb-class}
\alias{.StoppingTargetProb}
\alias{StoppingTargetProb-class}
\title{Stop based on probability of target tox interval}
\description{
Stop based on probability of target tox interval
}
\section{Slots}{

\describe{
\item{\code{target}}{the target toxicity interval, e.g. \code{c(0.2, 0.35)}}

\item{\code{prob}}{required target toxicity probability (e.g. \code{0.4})
for reaching sufficient precision}
}}
\keyword{classes}

