% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\name{StoppingTargetProb}
\alias{StoppingTargetProb}
\title{Initialization function for "StoppingTargetProb"}
\usage{
StoppingTargetProb(target, prob)
}
\arguments{
\item{target}{see \code{\linkS4class{StoppingTargetProb}}}

\item{prob}{see \code{\linkS4class{StoppingTargetProb}}}
}
\value{
the \code{\linkS4class{StoppingTargetProb}} object
}
\description{
Initialization function for "StoppingTargetProb"
}
\keyword{methods}
