% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\name{StoppingTDCIRatio}
\alias{StoppingTDCIRatio}
\title{Initialization function for "StoppingTDCIRatio"}
\usage{
StoppingTDCIRatio(targetRatio, targetEndOfTrial)
}
\arguments{
\item{targetRatio}{please refer to \code{\linkS4class{StoppingTDCIRatio}} class object}

\item{targetEndOfTrial}{please refer to \code{\linkS4class{StoppingTDCIRatio}} class object}
}
\value{
the \code{\linkS4class{StoppingTDCIRatio}} class object
}
\description{
Initialization function for "StoppingTDCIRatio"
}
\keyword{methods}
