% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rules-class.R
\name{StoppingTDsamplesCIRatio}
\alias{StoppingTDsamplesCIRatio}
\title{Initialization function for "StoppingTDsamplesCIRatio"}
\usage{
StoppingTDsamplesCIRatio(targetRatio, targetEndOfTrial)
}
\arguments{
\item{targetRatio}{please refer to \code{\linkS4class{StoppingTDsamplesCIRatio}} class object}

\item{targetEndOfTrial}{please refer to \code{\linkS4class{StoppingTDsamplesCIRatio}} class object}
}
\value{
the \code{\linkS4class{StoppingTDsamplesCIRatio}} class object
}
\description{
Initialization function for "StoppingTDsamplesCIRatio"
}
\keyword{methods}

