% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-methods.R
\name{subset-Data}
\alias{subset-Data}
\alias{[,Data,numeric,missing,missing-method}
\alias{[,Data,logical,missing,missing-method}
\title{Subsetting Operator for the Data Class}
\usage{
\S4method{[}{Data,numeric,missing,missing}(x, i)

\S4method{[}{Data,logical,missing,missing}(x, i)
}
\arguments{
\item{x}{(\code{Data})\cr what to subset.}

\item{i}{(\code{integer} or \code{logical})\cr indices or logical vector
for subsetting observations.}

\item{j}{not used.}

\item{drop}{not used.}
}
\value{
A \code{\link{Data}} object with the selected observations.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Subset observations (patients) from a \code{\link{Data}} object
using numeric or logical indexing.
}
