% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations-methods.R
\name{summary,ComboSimulations-method}
\alias{summary,ComboSimulations-method}
\alias{summary-ComboSimulations}
\title{Summarize \code{ComboSimulations}}
\usage{
\S4method{summary}{ComboSimulations}(object, truth, target = c(0.2, 0.35), ...)
}
\arguments{
\item{object}{(\code{ComboSimulations})
the object we want to summarize.}

\item{truth}{(\code{function})
optional function mapping a dose combination to a toxicity probability.
It can accept one argument (length-2 numeric vector or one-row matrix) or
two numeric arguments (\code{drug1}, \code{drug2}).}

\item{target}{(\code{numeric})
optional target toxicity interval used only when \code{truth} is supplied.}

\item{...}{additional arguments can be supplied here for \code{truth}.}
}
\value{
An object of class \code{\link{ComboSimulationsSummary}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Summarize two-drug combination simulations.
}
