% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Simulations-methods.R
\docType{methods}
\name{summary,DualSimulations-method}
\alias{summary,DualSimulations-method}
\title{Summarize the dual-endpoint design simulations, relative to given true
dose-toxicity and dose-biomarker curves}
\usage{
\S4method{summary}{DualSimulations}(object, trueTox, trueBiomarker,
  target = c(0.2, 0.35), ...)
}
\arguments{
\item{object}{the \code{\linkS4class{DualSimulations}} object we want to
summarize}

\item{trueTox}{a function which takes as input a dose (vector) and returns the
true probability (vector) for toxicity.}

\item{trueBiomarker}{a function which takes as input a dose (vector) and
returns the true biomarker level (vector).}

\item{target}{the target toxicity interval (default: 20-35\%) used for the
computations}

\item{\dots}{Additional arguments can be supplied here for \code{trueTox}
and \code{trueBiomarker}}
}
\value{
an object of class \code{\linkS4class{DualSimulationsSummary}}
}
\description{
Summarize the dual-endpoint design simulations, relative to given true
dose-toxicity and dose-biomarker curves
}
\keyword{methods}

