% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations-methods.R
\name{summary,GeneralSimulations-method}
\alias{summary,GeneralSimulations-method}
\alias{summary-GeneralSimulations}
\title{Summarize the \code{GeneralSimulations}, Relative to a Given Truth}
\usage{
\S4method{summary}{GeneralSimulations}(object, truth, target = c(0.2, 0.35), ...)
}
\arguments{
\item{object}{(\code{GeneralSimulations})\cr the object we want to summarize.}

\item{truth}{(\code{function})\cr a function which takes as input a dose (vector)
and returns the true probability (vector) for toxicity.}

\item{target}{(\code{numeric})\cr the target toxicity interval (default: 20-35\%)
used for the computations.}

\item{...}{additional arguments can be supplied here for \code{truth}.}
}
\value{
An object of class \code{\link{GeneralSimulationsSummary}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Summarize simulations relative to a given true dose-toxicity curve.
}
