% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations-methods.R
\name{summary,HierarchicalSimulations-method}
\alias{summary,HierarchicalSimulations-method}
\alias{summary-HierarchicalSimulations}
\title{Summarize Hierarchical Design Simulations}
\usage{
\S4method{summary}{HierarchicalSimulations}(object, truth, target = c(0.2, 0.35), ...)
}
\arguments{
\item{object}{(\code{HierarchicalSimulations})\cr the object we want to summarize.}

\item{truth}{(\code{function} or named \code{list} of \code{function})\cr true DLT
probability function(s). If a list is supplied, names must match the
hierarchical arms.}

\item{target}{(\code{numeric} or named \code{list} of \code{numeric})\cr target toxicity
interval(s). If a list is supplied, names must match the hierarchical arms.}

\item{...}{additional arguments can be supplied here for \code{truth}.}
}
\value{
An object of class \code{\link{HierarchicalSimulationsSummary}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Summarize hierarchical simulations by delegating to the corresponding
arm-level simulation summary methods.
}
