% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Simulations-methods.R
\docType{methods}
\name{summary,PseudoDualFlexiSimulations-method}
\alias{summary,PseudoDualFlexiSimulations-method}
\title{Summary for Pseudo Dual responses simulations given a pseudo DLE model and the Flexible efficacy model.}
\usage{
\S4method{summary}{PseudoDualFlexiSimulations}(object, trueDLE, trueEff,
  targetEndOfTrial = 0.3, targetDuringTrial = 0.35, ...)
}
\arguments{
\item{object}{the \code{\linkS4class{PseudoDualFlexiSimulations}} object we want to summarize}

\item{trueDLE}{a function which takes as input a dose (vector) and returns the true probability of DLE (vector)}

\item{trueEff}{a vector which takes as input the true mean efficacy values at all dose levels (in order)}

\item{targetEndOfTrial}{the target probability of DLE that are used at the end of a trial. Default at 0.3.}

\item{targetDuringTrial}{the target probability of DLE that are used during the trial. Default at 0.35.}

\item{\dots}{Additional arguments can be supplied here for \code{trueDLE} and \code{trueEff}}
}
\value{
an object of class \code{\linkS4class{PseudoDualSimulationsSummary}}
}
\description{
Summary for Pseudo Dual responses simulations given a pseudo DLE model and the Flexible efficacy model.
}
\examples{
##obtain the plot for the simulation results
##If DLE and efficacy responses are considered in the simulations and the 'EffFlexi' class is used 
## for the efficacy model
Effmodel<- EffFlexi(Eff=c(1.223, 2.513),Effdose=c(25,300),
                    sigma2=c(a=0.1,b=0.1),sigma2betaW=c(a=20,b=50),smooth="RW2",data=data)
## todo: how does the Effmodel come into the below code?
##Specify the simulations
##Please refer to design-method 'simulate DualResponsesSamplesFlexiDesign' examples for details
mySim<-simulate(object=design,
                args=NULL,
                trueDLE=myTruthDLE,
                trueEff=myTruthEff,
                trueSigma2=0.025,
                trueSigma2betaW=1,
                nsim=1,
                seed=819,
                parallel=FALSE)
##summarize the simulation results
summary(mySim,
        trueDLE=myTruthDLE,
        trueEff=myTruthEff)
}
\keyword{methods}

