% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Simulations-methods.R
\docType{methods}
\name{summary,PseudoSimulations-method}
\alias{summary,PseudoSimulations-method}
\title{Summarize the simulations, relative to a given truth}
\usage{
\S4method{summary}{PseudoSimulations}(object, truth, targetEndOfTrial = 0.3,
  targetDuringTrial = 0.35, ...)
}
\arguments{
\item{object}{the \code{\linkS4class{PseudoSimulations}} object we want to
summarize}

\item{truth}{a function which takes as input a dose (vector) and returns the
true probability (vector) for toxicity}

\item{targetEndOfTrial}{the target probability of DLE wanted to achieve at the end of a trial}

\item{targetDuringTrial}{the target probability of DLE wanted to achieve during a trial}

\item{\dots}{Additional arguments can be supplied here for \code{truth}}
}
\value{
an object of class \code{\linkS4class{PseudoSimulationsSummary}}
}
\description{
Summarize the simulations, relative to a given truth
}
\examples{
##obtain the plot for the simulation results
##If only DLE responses are considered in the simulations
##Specified your simulations when no DLE samples are used
##(Please refer to desgin-method 'simulate TDDesign' examples for details)
mySim <- simulate(design,
                  args=NULL,
                  truth=myTruth,
                  nsim=10,
                  seed=819,
                  parallel=FALSE)
##If DLE samples are involved
##Please refer to design-method 'simulate TDsamplesDesign' examples for details
mySim <- simulate(design,
                  args=NULL,
                  truth=myTruth,
                  nsim=10,
                  seed=819,
                  mcmcOptions=options,
                  parallel=FALSE)
##Then produce a summary of your simulations
summary(mySim,
        truth=myTruth)
}
\keyword{methods}

