\docType{methods}
\name{summary,Simulations-method}
\alias{summary,Simulations-method}
\title{Summarize the model-based design simulations, relative to a given truth}
\usage{
\S4method{summary}{Simulations}(object, truth, target = c(0.2, 0.35), ...)
}
\arguments{
  \item{object}{an object for which a summary is desired.}

  \item{...}{additional arguments affecting the summary
  produced.}
}
\value{
an object of class \code{\linkS4class{Simulations-summary}}
}
\description{
Summarize the model-based design simulations, relative to a
given truth
}
\keyword{methods}

