% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Design-class.R
\docType{class}
\name{TDsamplesDesign-class}
\alias{.TDsamplesDesign}
\alias{TDsamplesDesign-class}
\title{Design class using DLE responses only with samples}
\description{
Design class using DLE responses only with samples
}
\keyword{class}

