% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design-class.R
\name{ThreePlusThreeDesign}
\alias{ThreePlusThreeDesign}
\title{Creates a new 3+3 design object from a dose grid}
\usage{
ThreePlusThreeDesign(doseGrid)
}
\arguments{
\item{doseGrid}{the dose grid to be used}
}
\value{
the object of class \code{\linkS4class{RuleDesign}} with the
3+3 design
}
\description{
Creates a new 3+3 design object from a dose grid
}
\examples{

# inizialing a 3+3 design 
myDesign <- ThreePlusThreeDesign(doseGrid=c(5, 10, 15, 25, 35, 50, 80))
  
  
}
\author{
Daniel Sabanes Bove \email{sabanesd@roche.com}
}
\keyword{programming}
