% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-methods.R
\name{tidy,DataMixture-method}
\alias{tidy,DataMixture-method}
\alias{tidy-DataMixture}
\title{Tidy Method for the \code{\link{DataMixture}} Class}
\usage{
\S4method{tidy}{DataMixture}(x, ...)
}
\value{
The \code{\link{tibble}} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

A method that tidies a \code{\link{DataMixture}} object.
}
\section{Usage Notes}{

The prior observations are indicated by a \code{Cohort} value of \code{0} in the returned
\code{tibble}.
}

\examples{
.DefaultData() \%>\%tidy()
.DefaultDataOrdinal() \%>\% tidy()
.DefaultDataGrouped() \%>\% tidy()
.DefaultDAData() \%>\% tidy()
}
