% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-methods.R
\name{tidy,DataParts-method}
\alias{tidy,DataParts-method}
\alias{tidy-DataParts}
\title{Tidy Method for the \code{\link{DataParts}} Class}
\usage{
\S4method{tidy}{DataParts}(x, ...)
}
\value{
The \code{\link{tibble}} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

A method that tidies a \code{\link{DataParts}} object.
}
\examples{
.DefaultData() \%>\%tidy()
.DefaultDataOrdinal() \%>\% tidy()
.DefaultDataGrouped() \%>\% tidy()
.DefaultDAData() \%>\% tidy()
}
