% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-methods.R
\name{tidy,GeneralData-method}
\alias{tidy,GeneralData-method}
\alias{tidy-GeneralData}
\alias{tidy,DataGrouped-method}
\alias{tidy-DataGrouped}
\title{Tidy Method for the \code{\link{GeneralData}} Class}
\usage{
\S4method{tidy}{GeneralData}(x, ...)

\S4method{tidy}{DataGrouped}(x, ...)
}
\value{
The \code{\link{tibble}} object.

The \code{\link{tibble}} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

A method that tidies a \code{\link{GeneralData}} object.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

A method that tidies a \code{\link{DataGrouped}} object.
}
\examples{
.DefaultData() \%>\%tidy()
.DefaultDataOrdinal() \%>\% tidy()
.DefaultDataGrouped() \%>\% tidy()
.DefaultDAData() \%>\% tidy()
.DefaultData() \%>\%tidy()
.DefaultDataOrdinal() \%>\% tidy()
.DefaultDataGrouped() \%>\% tidy()
.DefaultDAData() \%>\% tidy()
}
