% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-class.R
\name{TITELogisticLogNormal}
\alias{TITELogisticLogNormal}
\title{Initialization function for the "TITELogisticLogNormal" class}
\usage{
TITELogisticLogNormal(weightMethod = c("linear", "adaptive"), ...)
}
\arguments{
\item{weightMethod}{see \code{\linkS4class{TITELogisticLogNormal}}}

\item{\dots}{see \code{\linkS4class{LogisticLogNormal}}}
}
\value{
the \code{\linkS4class{TITELogisticLogNormal}} object
}
\description{
Initialization function for the "TITELogisticLogNormal" class
}
\keyword{methods}
