% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-methods.R
\name{update,DataCombo-method}
\alias{update,DataCombo-method}
\alias{update-DataCombo}
\title{Updating \code{DataCombo} Objects}
\usage{
\S4method{update}{DataCombo}(
  object,
  x,
  y,
  response = rep(NA_integer_, length(y)),
  ID = length(object@ID) + seq_along(y),
  new_cohort = TRUE,
  check = TRUE,
  backfill = FALSE,
  cohort = NULL,
  ...
)
}
\arguments{
\item{object}{(\code{DataCombo})}

\item{x}{(\code{numeric})
named dose combination with one entry for each drug.
Names must match \code{object@drugNames}.}

\item{y}{(\code{integer})
the DLT vector (0/1 vector) for all patients in this cohort.}

\item{response}{(\code{integer})
the efficacy response vector (0/1 vector). May contain \code{NA}.}

\item{ID}{(\code{integer})
the patient IDs.}

\item{new_cohort}{(\code{flag})
if \code{TRUE} (default) the new data are assigned to a new cohort.}

\item{check}{(\code{flag})
whether the validation of the updated object should be conducted.}

\item{backfill}{(\code{flag})
whether the new patients being added are from a backfill cohort.}

\item{cohort}{(\code{int})
if provided, the new patients will be assigned to this cohort index.
If \code{NULL} (default), the cohort index will be determined based on the
\code{new_cohort} parameter.}

\item{...}{not used.}
}
\value{
The new, updated \code{\link{DataCombo}} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

A method that updates existing \code{\link{DataCombo}} object with new data.
}
\examples{
# Create some data of class 'DataCombo'.
my_data <- DataCombo(
  x = cbind(
    drug1 = c(10, 10, 10, 20, 20, 20),
    drug2 = c(20, 20, 20, 20, 20, 20)
  ),
  y = c(0, 0, 1, 0, 0, 0),
  doseGrid = list(drug1 = c(10, 20, 30), drug2 = c(20, 40))
)

# Update the data with a new cohort.
my_data1 <- update(
  my_data,
  x = c(drug1 = 30, drug2 = 40),
  y = c(0L, 1L, 0L)
)
my_data1
}
