% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Model-methods.R
\docType{methods}
\name{update,Effloglog-method}
\alias{update,Effloglog-method}
\title{Update method for the 'Effloglog'Model class}
\usage{
\S4method{update}{Effloglog}(object, data, ...)
}
\arguments{
\item{object}{is the model which follow \code{\linkS4Class{Effloglog}} class object}
}
\description{
Update method for the 'Effloglog'Model class
}
\keyword{methods}

