% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-methods.R
\name{update-HierarchicalData-method}
\alias{update,HierarchicalData-method}
\alias{update-HierarchicalData}
\title{Updating \code{HierarchicalData} Objects}
\usage{
\S4method{update}{HierarchicalData}(object, arm, ..., check = TRUE)
}
\arguments{
\item{object}{(\code{HierarchicalData})\cr object you want to update.}

\item{arm}{(\code{string})\cr name of the arm to update.}

\item{...}{arguments passed to the selected arm's \code{update} method.}

\item{check}{(\code{flag})\cr whether the validation of the updated object should
be conducted.}
}
\value{
The new, updated \code{\link{HierarchicalData}} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

A method that updates one arm in a \code{\link{HierarchicalData}} object with new data.
}
