% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Model-methods.R
\docType{methods}
\name{update,LogisticIndepBeta-method}
\alias{update,LogisticIndepBeta-method}
\title{Update method for the 'LogisticIndepBeta'Model class}
\usage{
\S4method{update}{LogisticIndepBeta}(object, data, ...)
}
\arguments{
\item{object}{is the model which follow \code{\linkS4class{LogisticIndepBeta}} class object}
}
\description{
Update method for the 'LogisticIndepBeta'Model class
}
\keyword{method}

