% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchicalDesign-validity.R
\name{v_arm_condition}
\alias{v_arm_condition}
\alias{v_arm_finished_condition}
\alias{v_arm_min_dose_condition}
\title{Internal Helper Functions for Validation of \code{\link{ArmCondition}} Objects}
\usage{
v_arm_finished_condition(object)

v_arm_min_dose_condition(object)
}
\arguments{
\item{object}{(\code{ArmCondition})\cr object to validate.}
}
\value{
A \code{character} vector with the validation failure messages,
or \code{TRUE} in case validation passes.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

These functions are only used internally to validate the format of an input
\code{\link{ArmCondition}} or inherited classes and therefore not exported.
}
\section{Functions}{
\itemize{
\item \code{v_arm_finished_condition()}: validates that the \code{\link{ArmFinishedCondition}}
object contains a valid \code{arm_name} slot.

\item \code{v_arm_min_dose_condition()}: validates that the \code{\link{ArmMinDoseCondition}}
object contains valid \code{arm_name} and \code{min_dose} slots.

}}
