% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Backfill-validity.R
\name{v_backfill}
\alias{v_backfill}
\title{Internal Helper Functions for Validation of \code{\link{Backfill}} Objects}
\usage{
v_backfill(object)
}
\arguments{
\item{object}{(\code{Backfill})\cr object to validate.}
}
\value{
A \code{character} vector with the validation failure messages,
or \code{TRUE} in case validation passes.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

These functions are only used internally to validate the format of an input
\code{\link{Backfill}} object and therefore not exported.
}
\section{Functions}{
\itemize{
\item \code{v_backfill()}: validates that the \code{\link{Backfill}} object
contains valid slots.

}}
