% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchicalDesign-validity.R
\name{v_hierarchical_design}
\alias{v_hierarchical_design}
\alias{v_design_arm}
\title{Internal Helper Functions for Validation of \code{\link{HierarchicalDesign}} Objects}
\usage{
v_design_arm(object)

v_hierarchical_design(object)
}
\arguments{
\item{object}{(\code{HierarchicalDesign})\cr object to validate.}
}
\value{
A \code{character} vector with the validation failure messages,
or \code{TRUE} in case validation passes.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

These functions are only used internally to validate the format of an input
\code{\link{HierarchicalDesign}} or related classes and therefore not exported.
}
\section{Functions}{
\itemize{
\item \code{v_design_arm()}: validates that the \code{\link{DesignArm}} object contains valid slots.

\item \code{v_hierarchical_design()}: validates that the \code{\link{HierarchicalDesign}} object
contains valid hierarchical arm metadata.

}}
