% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-validity.R
\name{v_model_objects}
\alias{v_model_objects}
\alias{v_general_model}
\alias{v_model}
\alias{v_model_log_normal}
\alias{v_model_logistic_kadane}
\alias{v_model_logistic_normal_mix}
\alias{v_model_logistic_normal_fixed_mix}
\alias{v_model_logistic_log_normal_mix}
\alias{v_model_dual_endpoint}
\alias{v_model_dual_endpoint_RW}
\title{Internal Helper Functions for Validation of \code{\link{AllModels}} Objects}
\usage{
v_general_model(object)

v_model(object)

v_model_log_normal(object)

v_model_logistic_kadane(object)

v_model_logistic_normal_mix(object)

v_model_logistic_normal_fixed_mix(object)

v_model_logistic_log_normal_mix(object)

v_model_dual_endpoint(object)

v_model_dual_endpoint_RW(object)
}
\arguments{
\item{object}{(\code{AllModels})\cr object to validate.}
}
\value{
A \code{character} vector with the validation failure messages,
or \code{TRUE} in case validation passes.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

These functions are only used internally to validate the format of an input
\code{\link{AllModels}} or inherited classes and therefore not exported.
}
\section{Functions}{
\itemize{
\item \code{v_general_model}: validates that the names of the
arguments in \code{init} function are included in \code{datanames} slot.

\item \code{v_model}: validates that the names of the
arguments in \code{dose} and \code{prob} functions contains \code{prob} and \code{dose}
respectively, as well as they match \code{sample} slot of the \code{object}.

\item \code{v_model_log_normal}: validates that the \code{ref_dose} is a positive scalar.

\item \code{v_model_logistic_kadane}: validates that the logistic Kadane model
parameters are valid.

\item \code{v_model_logistic_normal_mix}: validates that \code{weightpar} and \code{ref_dose} are valid.

\item \code{v_model_logistic_normal_fixed_mix}: validates that \code{component} is a list with
valid \code{ModelParamsNormal} objects as well as \code{weights} and \code{ref_dose} are
correct.

\item \code{v_model_logistic_log_normal_mix}: validates that \code{ref_dose} is valid and
\code{share_weight} represents probability.

\item \code{v_model_dual_endpoint}: validates that \code{\link{DualEndpoint}} class slots are valid.

\item \code{v_model_dual_endpoint_RW}: validates that \code{\link{DualEndpointRW}} class slots are valid.
}}

