% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelParams-validity.R
\name{v_model_params}
\alias{v_model_params}
\alias{v_model_params_normal}
\title{Internal Helper Functions for Validation of Model Parameters Objects}
\usage{
v_model_params_normal(object)
}
\arguments{
\item{object}{(\code{ModelParamsNormal})\cr multivariate normal parameters object
to validate.}
}
\value{
A \code{character} vector with the validation failure messages,
or \code{TRUE} in case validation passes.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

These functions are only used internally to validate the format of an object
with model parameters or inherited classes and therefore not exported.
}
\section{Functions}{
\itemize{
\item \code{v_model_params_normal()}: a helper function that validates multivariate normal
parameters.

}}
