% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-validity.R
\name{v_NextBestNCRMLoss}
\alias{v_NextBestNCRMLoss}
\title{These functions are only used internally to validate the format of an input
\code{\link{NextBestNCRMLoss}} or inherited classes and therefore not exported.}
\usage{
v_NextBestNCRMLoss(object)
}
\arguments{
\item{object}{(\code{NextBestNCRMLoss})\cr object to validate.}
}
\value{
A \code{character} vector with the validation failure messages,
or \code{TRUE} in case validation passes.
}
\description{
These functions are only used internally to validate the format of an input
\code{\link{NextBestNCRMLoss}} or inherited classes and therefore not exported.
}
\section{Functions}{
\itemize{
\item \code{v_NextBestNCRMLoss}: validates the \code{\link{.NextBestNCRMLoss}} object
contains valid probability target and percentage threshold.
}}

