% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations-validity.R, R/Simulation-validity.R
\name{v_pseudo_simulations}
\alias{v_pseudo_simulations}
\title{Internal Helper Functions for Validation of \code{\link{PseudoSimulations}} Objects}
\usage{
v_pseudo_simulations(object)
}
\arguments{
\item{object}{(\code{PseudoSimulations})\cr object to validate.}
}
\value{
A \code{character} vector with the validation failure messages,
or \code{TRUE} in case validation passes.

A \code{character} vector with the validation failure messages,
or \code{TRUE} in case validation passes.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

These functions are only used internally to validate the format of an input
\code{\link{PseudoSimulations}} or inherited classes and therefore not exported.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

These functions are only used internally to validate the format of an input
\code{\link{PseudoSimulations}} or inherited classes and therefore not exported.
}
\section{Functions}{
\itemize{
\item \code{v_pseudo_simulations()}: validates that the \code{\link{PseudoSimulations}} object
contains valid \code{fit}, \code{FinalTDtargetEndOfTrialEstimates} , \code{FinalTDtargetDuringTrialAtDoseGrid},
\code{FinalTDtargetEndOfTrialAtDoseGrid} , \code{FinalTDEOTCIs}, \code{FinalTDEOTRatios}, \code{FinalCIs}, \code{FinalRatios},
object and valid \code{stopReasons} simulations.

}}
