% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Backfill-validity.R
\name{v_recruitment}
\alias{v_recruitment}
\alias{v_recruitment_ratio}
\title{Internal Helper Functions for Validation of \code{\link{Recruitment}} Objects}
\usage{
v_recruitment_ratio(object)
}
\arguments{
\item{object}{(\code{Recruitment})
object to validate.}
}
\value{
A \code{character} vector with the validation failure messages,
or \code{TRUE} in case validation passes.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

These functions are only used internally to validate the format of an input
\code{\link{Recruitment}} or inherited classes and therefore not exported.
}
\section{Functions}{
\itemize{
\item \code{v_recruitment_ratio()}: validates that the \code{\link{RecruitmentRatio}} object
contains valid \code{ratio} slot.

}}
