% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-validity.R
\name{v_safety_window}
\alias{v_safety_window}
\alias{v_safety_window_size}
\alias{v_safety_window_const}
\title{Internal Helper Functions for Validation of \code{\link{SafetyWindow}} Objects}
\usage{
v_safety_window_size(object)

v_safety_window_const(object)
}
\arguments{
\item{object}{(\code{SafetyWindow})\cr object to validate.}
}
\value{
A \code{character} vector with the validation failure messages,
or \code{TRUE} in case validation passes.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

These functions are only used internally to validate the format of an input
\code{\link{SafetyWindow}} or inherited classes and therefore not exported.
}
\section{Functions}{
\itemize{
\item \code{v_safety_window_size()}: validates that the \code{\link{SafetyWindowSize}} object
contains valid slots.

\item \code{v_safety_window_const()}: validates that the \code{\link{SafetyWindowConst}} object
contains valid slots.

}}
