% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Samples-validity.R
\name{v_samples_objects}
\alias{v_samples_objects}
\alias{v_samples}
\title{Internal Helper Functions for Validation of \code{\link{Samples}} Objects}
\usage{
v_samples(object)
}
\arguments{
\item{object}{(\code{Samples})\cr object to validate.}
}
\value{
A \code{character} vector with the validation failure messages, or \code{TRUE}
in case validation passes.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

These functions are only used internally to validate the format of an input
\code{\link{Samples}} or inherited classes and therefore not exported.
}
\section{Functions}{
\itemize{
\item \code{v_samples()}: validates that the \code{\link{Samples}} object contains
valid \code{data} slot.

}}
