% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_knitr_Design.R
\name{v_starting_dose}
\alias{v_starting_dose}
\title{Internal Helper Functions for Validation of \code{\link{StartingDose}} Objects}
\arguments{
\item{object}{(\code{StartingDose})\cr object to validate.}
}
\value{
A \code{character} vector with the validation failure messages,
or \code{TRUE} in case validation passes.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

These functions are only used internally to validate the format of an input
\code{\link{RuleDesign}} or inherited classes and therefore not exported.
}
