% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-validity.R
\name{validate_data_objects}
\alias{validate_data_objects}
\alias{validate_subjects}
\alias{h_doses_unique_per_cohort}
\alias{validate_data}
\alias{validate_data_dual}
\alias{validate_data_parts}
\title{Internal Helper Functions for Validation of \code{\link{GeneralData}} Objects}
\usage{
validate_subjects(object)

h_doses_unique_per_cohort(dose, cohort)

validate_data(object)

validate_data_dual(object)

validate_data_parts(object)
}
\arguments{
\item{object}{(\code{GeneralData})\cr object to validate.}

\item{dose}{(\code{numeric})\cr dose values.}

\item{cohort}{(\code{integer})\cr cohort indices parallel to \code{doses}.}
}
\value{
A character vector with the validation failure messages,
or \code{TRUE} in case validation passes.

\code{TRUE} if \code{dose} is unique per \code{cohort}, otherwise \code{FALSE}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

These functions are only used internally to validate the format of an input
\code{\link{GeneralData}} or inherited classes and therefore not exported.
}
\section{Functions}{
\itemize{
\item \code{validate_subjects}: validates that the \code{\link{GeneralData}}
object contains unique \code{ID}, non-negative \code{cohort} indices and
\code{ID} and \code{cohort} vectors which are of the same length \code{nObs}.

\item \code{h_doses_unique_per_cohort}: helper function which verifies whether
the \code{dose} values are unique in each and every different \code{cohort}.

\item \code{validate_data}: validates that the \code{\link{Data}} object contains
valid elements with respect to their types, dependency and length.

\item \code{validate_data_dual}: validates that the \code{\link{DataDual}} object
contains valid biomarker vector with respect to its type and the length.

\item \code{validate_data_parts}: validates that the \code{\link{DataParts}} object
contains valid elements with respect to their types, dependency and length.
}}

