% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McmcOptions-validity.R
\name{validate_mcmcoptions_objects}
\alias{validate_mcmcoptions_objects}
\alias{validate_mcmc_options}
\title{Internal Helper Functions for Validation of \code{\link{McmcOptions}} Objects}
\usage{
validate_mcmc_options(object)
}
\arguments{
\item{object}{(\code{McmcOptions})\cr object to validate.}
}
\value{
A \code{character} vector with the validation failure messages,
or \code{TRUE} in case validation passes.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

These functions are only used internally to validate the format of an input
\code{\link{McmcOptions}} or inherited classes and therefore not exported.
}
\section{Functions}{
\itemize{
\item \code{validate_mcmc_options}: validates that the \code{\link{McmcOptions}}
object contains valid integer scalars \code{iterations}, \code{burnin} and \code{step}
as well as proper parameters for Random Number Generator.
}}

