% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-validity.R
\name{validate_increments_hsr_beta}
\alias{validate_increments_hsr_beta}
\alias{validate_stopping_lowest_dose_hsr_beta}
\alias{validate_stopping}
\alias{validate_stopping_mtd_cv}
\title{Internal Helper Functions for Validation of \code{\link{Stopping}} Objects}
\usage{
validate_increments_hsr_beta(object)

validate_stopping_lowest_dose_hsr_beta(object)

validate_stopping_mtd_cv(object)
}
\arguments{
\item{object}{(\code{Stopping})\cr object to validate.}
}
\value{
A \code{character} vector with the validation failure messages,
or \code{TRUE} in case validation passes.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

These functions are only used internally to validate the format of an input
\code{\link{Stopping}} or inherited classes and therefore not exported.
}
\section{Functions}{
\itemize{
\item \code{validate_increments_hsr_beta}: validates that the \code{\link{IncrementsHSRBeta}}
object contains valid probability target, threshold and shape parameters.

\item \code{validate_stopping_lowest_dose_hsr_beta}: validates that the \code{\link{StoppingLowestDoseHSRBeta}}
object contains valid probability target, threshold and shape parameters.

\item \code{validate_stopping_mtd_cv}: validates that the \code{\link{StoppingMTDCV}} object
contains valid probability target and percentage threshold.
}}

