% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Model-class.R
\docType{class}
\name{AllModels-class}
\alias{.AllModels}
\alias{AllModels-class}
\title{Class for All models
This is a class where all other model inherit}
\description{
Class for All models
This is a class where all other model inherit
}
\keyword{classes}

