% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-class.R
\docType{class}
\name{AllModels-class}
\alias{AllModels-class}
\alias{.AllModels}
\alias{AllModels}
\title{\code{AllModels}}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{\link{AllModels}} is a parent class for all models.
}
