% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-class.R
\docType{class}
\name{AllModels-class}
\alias{AllModels-class}
\alias{.AllModels}
\title{Class for All models
This is a class where all models inherit.}
\description{
Class for All models
This is a class where all models inherit.
}
\section{Slots}{

\describe{
\item{\code{datanames}}{The names of all data slots that are used in all models. 
In particularly, those are also used in the \code{datamodel} and/or
\code{priormodel} definition for \code{\linkS4class{GeneralModel}}.}
}}

\seealso{
\code{\linkS4class{GeneralModel}}, \code{\linkS4class{ModelPseudo}}
}
\keyword{classes}
