\docType{methods}
\name{&,StoppingAll,Stopping-method}
\alias{&,StoppingAll,Stopping-method}
\title{The method combining a stopping list and an atomic}
\usage{
\S4method{&}{StoppingAll,Stopping}(e1, e2)
}
\arguments{
  \item{e1}{\code{\linkS4class{StoppingAll}} object}

  \item{e2}{\code{\linkS4class{Stopping}} object}
}
\value{
The modified \code{\linkS4class{StoppingAll}} object
}
\description{
The method combining a stopping list and an atomic
}
\keyword{methods}

