\name{approximate}
\alias{approximate}
\title{Approximate posterior with (log) normal distribution}
\usage{
approximate(object, ...)
}
\arguments{
  \item{object}{the object}

  \item{\dots}{unused}
}
\value{
the approximation model
}
\description{
Approximate posterior with (log) normal distribution
}
\keyword{methods}

