% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Data-methods.R
\docType{methods}
\name{as.list,Data-method}
\alias{as.list,Data-method}
\title{as.list method for the "Data" class}
\usage{
\S4method{as.list}{Data}(x, ...)
}
\arguments{
\item{x}{the \code{\linkS4class{Data}} object we want to convert}

\item{\dots}{unused}
}
\value{
a list of all slots in \code{x}
}
\description{
as.list method for the "Data" class
}
\keyword{methods}

