% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-methods.R
\name{as.list,GeneralData-method}
\alias{as.list,GeneralData-method}
\title{as.list method for the "GeneralData" class}
\usage{
\S4method{as.list}{GeneralData}(x, ...)
}
\arguments{
\item{x}{the \code{\linkS4class{GeneralData}} object we want to convert}

\item{\dots}{unused}
}
\value{
a list of all slots in \code{x}
}
\description{
as.list method for the "GeneralData" class
}
\examples{

# Create some data of class 'Data'
myData <- Data(x=c(0.1,0.5,1.5,3,6,10,10,10),
               y=c(0,0,0,0,0,0,1,0),
               doseGrid=c(0.1,0.5,1.5,3,6,
               seq(from=10,to=80,by=2)))

# Converting Data object to list
as.list(myData)
}
\keyword{methods}
