% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmate.R
\name{check_length}
\alias{check_length}
\alias{assert_length}
\alias{test_length}
\alias{expect_length}
\title{Check if an argument is one of the given lengths}
\usage{
check_length(x, len)

assert_length(x, len, .var.name = checkmate::vname(x), add = NULL)

test_length(x, len)

expect_length(x, len, info = NULL, label = vname(x))
}
\arguments{
\item{x}{(\code{any})\cr any object for which \code{\link[base:length]{base::length()}} function is defined.}

\item{len}{(\code{integer})\cr a vector of allowable lengths.}
}
\value{
\code{TRUE} if successful, otherwise a string with the error message.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Check that \code{x} is one of the lengths specified by the \code{len} parameter.
}
\examples{
x <- 1:5
check_length(x, len = 5)
check_length(x, len = c(1, 5))
check_length(x, len = c(2, 10))
}
\seealso{
\code{\link{assertions}} for more details.
}
