% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{CohortSize-class}
\alias{CohortSize-class}
\alias{CohortSize}
\title{\code{CohortSize}}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{\link{CohortSize}} is a class for for cohort sizes.
}
\seealso{
\code{\link{CohortSizeRange}}, \code{\link{CohortSizeDLT}}, \code{\link{CohortSizeConst}},
\code{\link{CohortSizeParts}}, \code{\link{CohortSizeMin}}, \code{\link{CohortSizeMin}}.
}
