% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{CohortSize-class}
\alias{CohortSize-class}
\title{The virtual class for cohort sizes}
\description{
The virtual class for cohort sizes
}
\seealso{
\code{\linkS4class{CohortSizeMax}},
\code{\linkS4class{CohortSizeMin}},
\code{\linkS4class{CohortSizeRange}},
\code{\linkS4class{CohortSizeDLT}},
\code{\linkS4class{CohortSizeConst}},
\code{\linkS4class{CohortSizeParts}}
}
\keyword{classes}
