% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{CohortSizeConst-class}
\alias{.CohortSizeConst}
\alias{CohortSizeConst-class}
\title{Constant cohort size}
\description{
This class is used when the cohort size should be kept constant.
}
\section{Slots}{

\describe{
\item{\code{size}}{the constant integer size}
}}
\examples{

# This is to have along the study a constant cohort size of 3
mySize <- CohortSizeConst(size=3)

}
\keyword{classes}

