% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{CohortSizeConst-class}
\alias{CohortSizeConst-class}
\alias{.CohortSizeConst}
\alias{CohortSizeConst}
\alias{.DefaultCohortSizeConst}
\title{\code{CohortSizeConst}}
\usage{
CohortSizeConst(size)

.DefaultCohortSizeConst()
}
\arguments{
\item{size}{(\code{number})\cr see slot definition.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{\link{CohortSizeConst}} is the class for fixed and constant size of cohort.
}
\section{Slots}{

\describe{
\item{\code{size}}{(\code{integer})\cr cohort size.}
}}

\note{
Typically, end users will not use the \code{.DefaultCohortSizeConst()} function.
}
\examples{

# Cohort of size 3, constant along the study.
my_size <- CohortSizeConst(size = 3)
}
