% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{CohortSizeDLT-class}
\alias{.CohortSizeDLT}
\alias{CohortSizeDLT-class}
\title{Cohort size based on number of DLTs}
\description{
Cohort size based on number of DLTs
}
\section{Slots}{

\describe{
\item{\code{DLTintervals}}{an integer vector with the left bounds of the relevant
DLT intervals}

\item{\code{cohortSize}}{an integer vector of the same length with the cohort
sizes in the \code{DLTintervals}}
}}
\keyword{classes}

