% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{CohortSizeRange-class}
\alias{.CohortSizeRange}
\alias{CohortSizeRange-class}
\title{Cohort size based on dose range}
\description{
Cohort size based on dose range
}
\section{Slots}{

\describe{
\item{\code{intervals}}{a vector with the left bounds of the relevant dose intervals}

\item{\code{cohortSize}}{an integer vector of the same length with the cohort
sizes in the \code{intervals}}
}}
\keyword{classes}

