% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{CohortSizeRange-class}
\alias{CohortSizeRange-class}
\alias{.CohortSizeRange}
\title{Cohort size based on dose range}
\description{
Cohort size based on dose range
}
\section{Slots}{

\describe{
\item{\code{intervals}}{a vector with the left bounds of the relevant dose intervals}

\item{\code{cohortSize}}{an integer vector of the same length with the cohort
sizes in the \code{intervals}}
}}

\examples{

# As example, here is the rule for: 
#   having cohort of size 1 for doses <30
#   and having cohort of size 3 for doses >=30

mySize <- CohortSizeRange(intervals=c(0, 30),
                          cohortSize=c(1, 3))

}
\keyword{classes}
