% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crmPack-package.R
\docType{package}
\name{crmPack-package}
\alias{crmPack-package}
\alias{crmPack}
\title{Object-oriented implementation of CRM designs}
\description{
Object-oriented implementation of CRM designs
}
\author{
Daniel Sabanes Bove \email{sabanesd@roche.com}
}
\keyword{package}
