% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{crmPackExample}
\alias{crmPackExample}
\title{Open the Example PDF for crmPack}
\usage{
crmPackExample()
}
\value{
Called for side effects.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Calling this helper function opens the \code{example.pdf} document,
residing in the \code{doc} subfolder of the package installation directory.
}
