% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{crmPackHelp}
\alias{crmPackHelp}
\title{Open the Browser with Help Pages for crmPack}
\usage{
crmPackHelp()
}
\value{
Called for side effects.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This convenience function opens your browser with the help pages for crmPack.
}
