% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{crmPackHelp}
\alias{crmPackHelp}
\title{Open the browser with help pages for crmPack}
\usage{
crmPackHelp()
}
\value{
nothing
}
\description{
This convenience function opens your browser with the help pages for
crmPack.
}
\author{
Daniel Sabanes Bove \email{sabanesd@roche.com}
}
\keyword{documentation}
