% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-class.R
\docType{class}
\name{DALogisticLogNormal-class}
\alias{DALogisticLogNormal-class}
\alias{.DALogisticLogNormal}
\title{Logistic model with bivariate (log) normal prior and data augmentation}
\description{
This is a modified data augmented CRM with logistic regression model using
a bivariate normal prior on the intercept and log slope parameters.
This class inherits from the normal logistic model class.
}
\details{
todo: need to include here formula for the lambda prior.
}
\section{Slots}{

\describe{
\item{\code{npiece}}{the number of pieces in the \code{PEM}}

\item{\code{l}}{a vector used in the lambda prior}

\item{\code{C_par}}{a parameter used in the lambda prior,
according to Liu's paper, \code{C_par=2} is recommended.}

\item{\code{conditionalPEM}}{is a conditional piecewise-exponential model used?
(default) Otherwise an unconditional model is used.}
}}

\examples{
npiece <- 10
Tmax <- 60

lambda_prior <- function(k)
{
  npiece / (Tmax * (npiece - k + 0.5))
}

model <- DALogisticLogNormal(mean=c(-0.85,1),
                           cov=matrix(c(1, -0.5, -0.5, 1),
                                      nrow=2),
                           refDose=56,
                           npiece=npiece,
                           l=as.numeric(t(apply(as.matrix(c(1:npiece), 1, npiece),
                                                2,
                                                lambda_prior))),
                           C_par=2)
}
\keyword{classes}
