% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations-class.R
\docType{class}
\name{DASimulations-class}
\alias{DASimulations-class}
\alias{.DASimulations}
\alias{.DefaultDASimulations}
\title{Class for the simulations output from DA based designs}
\usage{
.DefaultDASimulations()
}
\description{
This class captures the trial simulations from DA based
designs. In comparison to the parent class \code{\linkS4class{Simulations}},
it contains additional slots to capture the time to DLT fits, additional
parameters and the trial duration.
}
\section{Slots}{

\describe{
\item{\code{trialduration}}{the vector of trial duration values for all simulations.}
}}

\note{
Typically, end users will not use the \code{.DASimulations()} function.  This
function has a noticeable execution time.
}
\keyword{classes}
