% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-class.R
\docType{class}
\name{Data-class}
\alias{Data-class}
\alias{.Data}
\title{\code{Data}}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

The \code{\linkS4class{Data}} is a class for the data input.
It inherits from \code{\linkS4class{GeneralData}}.
}
\section{Slots}{

\describe{
\item{\code{x}}{(\code{numeric})\cr the doses for the patients.}

\item{\code{y}}{(\code{integer})\cr the vector of toxicity events (0 or 1 integers).}

\item{\code{doseGrid}}{(\code{numeric})\cr the vector of all possible doses (sorted),
i.e. the dose grid.}

\item{\code{nGrid}}{(\code{integer})\cr number of gridpoints.}

\item{\code{xLevel}}{(\code{integer})\cr the levels for the doses the patients have been given,
w.r.t \code{doseGrid}.}

\item{\code{placebo}}{(\code{logical})\cr if \code{TRUE} the first dose level
in the \code{doseGrid}is considered as PLACEBO.}
}}

\examples{
# create some data from the class 'Data'
myData <- Data(x=c(0.1,0.5,1.5,3,6,10,10,10),
               y=c(0,0,0,0,0,0,1,0),
               doseGrid=c(0.1,0.5,1.5,3,6,
                          seq(from=10,to=80,by=2)))
}
\keyword{classes}
