% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Data-class.R
\docType{class}
\name{Data-class}
\alias{.Data}
\alias{Data-class}
\title{Class for the data input}
\description{
Class for the data input
}
\section{Slots}{

\describe{
\item{\code{x}}{the doses for the patients}

\item{\code{y}}{the vector of toxicity events (0 or 1 integers)}

\item{\code{ID}}{unique patient IDs (integer vector)}

\item{\code{cohort}}{the cohort indices (sorted values from 0, 1, 2, ...)}

\item{\code{doseGrid}}{the vector of all possible doses (sorted), i.e. the dose
grid}

\item{\code{nObs}}{number of observations}

\item{\code{nGrid}}{number of gridpoints}

\item{\code{xLevel}}{the levels for the doses the patients have been given}
}}
\keyword{classes}

