% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Data-class.R
\docType{class}
\name{DataCombo-class}
\alias{.DataCombo}
\alias{DataCombo-class}
\title{Class for the data input in combo trials}
\description{
This class inherits from \code{\linkS4class{GeneralData}}.
}
\section{Slots}{

\describe{
\item{\code{x}}{a matrix with the doses of all \code{nDrugs} drugs (columns) for the
\code{nObs} patients (rows). The column names are the \code{drugNames}.}

\item{\code{y}}{the vector of toxicity events (0 or 1 integers)}

\item{\code{doseGrid}}{a list containing a vector of all possible doses (sorted) for
each of the \code{nDrugs} drugs (named with \code{drugNames}).}

\item{\code{nDrugs}}{number of drugs}

\item{\code{drugNames}}{character vector with the drug names}

\item{\code{nGrid}}{vector with the number of gridpoints for each of the drugs}

\item{\code{xLevel}}{an integer matrix with the levels for the doses the patients
have been given, same dimensions as \code{x}.}
}}
\keyword{classes}

