% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-class.R
\docType{class}
\name{DataDA-class}
\alias{DataDA-class}
\alias{.DataDA}
\title{Class for the time-to-DLT augmented data input}
\description{
This is a subclass of \code{\linkS4class{Data}}, so contains all
slots from \code{\linkS4class{Data}}, and in addition DLT free survival
values. (Note that \dQuote{survival} values here refers to the time that 
the subject did not experience any DLT yet, and is not referring to 
deaths.)
}
\section{Slots}{

\describe{
\item{\code{u}}{the continuous vector of DLT free survival values}

\item{\code{Tmax}}{the DLT observation period}

\item{\code{t0}}{time of initial dosing for each patient}
}}

\examples{
data <- DataDA(x=c(0.1, 0.5, 1.5, 3, 6, 10, 10, 10),
              y=c(0, 0, 1, 1, 0, 0, 1, 0),
              doseGrid=
                c(0.1, 0.5, 1.5, 3, 6,
                  seq(from=10, to=80, by=2)),
              u=c(42,30,15,5,20,25,30,60),
              t0=c(0,-15,-30,-40,-55,-70,-75,-85),
              Tmax=60)

#set up an empty data set;

emptydata <- DataDA(doseGrid=c(0.1, 0.5,1, 1.5, 3, 6,
                               seq(from=10, to=80, by=2)),Tmax=60)
}
\keyword{classes}
