% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Data-class.R
\docType{class}
\name{DataDual-class}
\alias{.DataDual}
\alias{DataDual-class}
\title{Class for the dual endpoint data input}
\description{
This is a subclass of \code{\linkS4class{Data}}, so contains all
slots from \code{\linkS4class{Data}}, and in addition biomarker
values.
}
\section{Slots}{

\describe{
\item{\code{w}}{the continuous vector of biomarker values}
}}
\keyword{classes}

