% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-class.R
\docType{class}
\name{DataDual-class}
\alias{DataDual-class}
\alias{.DataDual}
\title{Class for the dual endpoint data input}
\description{
This is a subclass of \code{\linkS4class{Data}}, so contains all
slots from \code{\linkS4class{Data}}, and in addition biomarker
values.
}
\section{Slots}{

\describe{
\item{\code{w}}{the continuous vector of biomarker values}
}}

\examples{
# Create some data from the class 'DataDual'
myData <- DataDual(x=c(0.1,0.5,1.5,3,6,10,10,10),
                   y=c(0,0,0,0,0,0,1,0),
                   w=rnorm(8),
                   doseGrid=c(0.1,0.5,1.5,3,6,
                              seq(from=10,to=80,by=2)))
}
\keyword{classes}
