% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design-class.R
\name{Design}
\alias{Design}
\title{Initialization function for "Design"}
\usage{
Design(model, stopping, increments, PLcohortSize = CohortSizeConst(0L),
  ...)
}
\arguments{
\item{model}{see \code{\linkS4class{Design}}}

\item{stopping}{see \code{\linkS4class{Design}}}

\item{increments}{see \code{\linkS4class{Design}}}

\item{PLcohortSize}{see \code{\linkS4class{Design}}}

\item{\dots}{additional arguments for \code{\link{RuleDesign}}}
}
\value{
the \code{\linkS4class{Design}} object
}
\description{
Initialization function for "Design"
}
\keyword{methods}
