% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{dinvGamma}
\alias{dinvGamma}
\title{Compute the density of Inverse gamma distribution}
\usage{
dinvGamma(x, a, b, log = FALSE, normalize = TRUE)
}
\arguments{
\item{x}{vector of quantiles}

\item{a}{the shape parameter of the inverse gamma distribution}

\item{b}{the scale parameter of the inverse gamm distribution}

\item{log}{logical; if TRUE, probabilities p are given as log(p)}

\item{normalize}{logical; if TRUE, the output will be normalized}
}
\description{
Compute the density of Inverse gamma distribution
}
\keyword{internal}
