% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Model-methods.R
\docType{methods}
\name{dose}
\alias{dose}
\alias{dose,numeric,Model,Samples-method}
\title{Compute the doses for a given probability, given model and samples}
\usage{
dose(prob, model, samples, ...)

\S4method{dose}{numeric,Model,Samples}(prob, model, samples, ...)
}
\arguments{
\item{prob}{the probability}

\item{model}{the \code{\linkS4class{Model}}}

\item{samples}{the \code{\linkS4class{Samples}}}

\item{\dots}{unused}
}
\description{
Compute the doses for a given probability, given model and samples
}
\section{Methods (by class)}{
\itemize{
\item \code{prob = numeric,model = Model,samples = Samples}: 
}}
\keyword{methods}

