% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-methods.R
\name{doseFunction}
\alias{doseFunction}
\alias{doseFunction,GeneralModel-method}
\alias{doseFunction-GeneralModel}
\alias{doseFunction,ModelTox-method}
\alias{doseFunction-ModelTox}
\title{Getting the Dose Function for a Given Model Type}
\usage{
doseFunction(model, ...)

\S4method{doseFunction}{GeneralModel}(model, ...)

\S4method{doseFunction}{ModelTox}(model, ...)
}
\arguments{
\item{model}{(\code{GeneralModel} or \code{ModelTox})\cr the model.}

\item{...}{model specific parameters.}
}
\value{
A \code{function} that computes doses for a given toxicity probability
and the model.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

A function that returns a function that computes the dose reaching a
specific target probability, based on the model specific parameters.
}
\examples{
my_model <- LogisticLogNormal(
  mean = c(-0.85, 1),
  cov = matrix(c(1, -0.5, -0.5, 1), nrow = 2),
  ref_dose = 50
)

dose_fun <- doseFunction(my_model, alpha0 = 2, alpha1 = 3)
dose_fun(0.6)
}
\seealso{
\code{\link{dose}}, \code{\link{probFunction}}.
}
