% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Design-class.R
\docType{class}
\name{DualDesign-class}
\alias{.DualDesign}
\alias{DualDesign-class}
\title{Class for the dual-endpoint CRM design}
\description{
This class has special requirements for the \code{model} and \code{data}
slots in comparison to the parent class \code{\linkS4class{Design}}:
}
\section{Slots}{

\describe{
\item{\code{model}}{the model to be used, an object of class
\code{\linkS4class{DualEndpoint}}}

\item{\code{data}}{what is the dose grid, any previous data, etc., contained
in an object of class \code{\linkS4class{DataDual}}

Note that the \code{NextBest} slot can be of any class, this allows for easy
comparison with recommendation methods that don't use the
biomarker information.}
}}
\keyword{classes}

