% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-class.R
\name{DualEndpoint}
\alias{DualEndpoint}
\title{Initialization function for the "DualEndpoint" class}
\usage{
DualEndpoint(mu, Sigma, refDose = 1, useLogDose = FALSE, sigma2W, rho)
}
\arguments{
\item{mu}{see \code{\linkS4class{DualEndpoint}}}

\item{Sigma}{see \code{\linkS4class{DualEndpoint}}}

\item{refDose}{see \code{\linkS4class{DualEndpoint}} (default: 1)}

\item{useLogDose}{see \code{\linkS4class{DualEndpoint}} 
(default: \code{FALSE})}

\item{sigma2W}{see \code{\linkS4class{DualEndpoint}}}

\item{rho}{see \code{\linkS4class{DualEndpoint}}}
}
\value{
the \code{\linkS4class{DualEndpoint}} object
}
\description{
Initialization function for the "DualEndpoint" class
}
\keyword{methods}
