% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-class.R
\name{DualEndpointBeta}
\alias{DualEndpointBeta}
\title{Initialization function for the "DualEndpointBeta" class}
\usage{
DualEndpointBeta(E0, Emax, delta1, mode, ref_dose_beta, ...)
}
\arguments{
\item{E0}{see \code{\linkS4class{DualEndpointBeta}}}

\item{Emax}{see \code{\linkS4class{DualEndpointBeta}}}

\item{delta1}{see \code{\linkS4class{DualEndpointBeta}}}

\item{mode}{see \code{\linkS4class{DualEndpointBeta}}}

\item{ref_dose_beta}{see \code{\linkS4class{DualEndpointBeta}}}

\item{\dots}{additional parameters, see \code{\linkS4class{DualEndpoint}}}
}
\value{
the \code{\linkS4class{DualEndpointBeta}} object
}
\description{
Initialization function for the "DualEndpointBeta" class
}
\keyword{methods}
