% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-class.R
\name{DualEndpointEmax}
\alias{DualEndpointEmax}
\title{Initialization function for the "DualEndpointEmax" class}
\usage{
DualEndpointEmax(E0, Emax, ED50, refDoseEmax, ...)
}
\arguments{
\item{E0}{see \code{\linkS4class{DualEndpointEmax}}}

\item{Emax}{see \code{\linkS4class{DualEndpointEmax}}}

\item{ED50}{see \code{\linkS4class{DualEndpointEmax}}}

\item{refDoseEmax}{see \code{\linkS4class{DualEndpointEmax}}}

\item{\dots}{additional parameters, see \code{\linkS4class{DualEndpoint}}}
}
\value{
the \code{\linkS4class{DualEndpointEmax}} object
}
\description{
Initialization function for the "DualEndpointEmax" class
}
\keyword{methods}
