% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Design-class.R
\docType{class}
\name{DualResponsesDesign-class}
\alias{.DualResponsesDesign}
\alias{DualResponsesDesign-class}
\title{Class of design which based on one pseudo DLE and one pseudo efficacy model without samples}
\description{
Class of design which based on one pseudo DLE and one pseudo efficacy model without samples
}
\keyword{class}

