% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design-class.R
\name{DualResponsesDesign}
\alias{DualResponsesDesign}
\title{Initialization function for 'DualResponsesDesign"}
\usage{
DualResponsesDesign(Effmodel, data, ...)
}
\arguments{
\item{Effmodel}{please refer to \code{\linkS4class{DualResponsesDesign}} class object}

\item{data}{please refer to \code{\linkS4class{DualResponsesDesign}} class object}

\item{\dots}{additional arguments for \code{\link{TDDesign}}}
}
\value{
the \code{\linkS4class{DualResponsesDesign}} class object
}
\description{
Initialization function for 'DualResponsesDesign"
}
\keyword{methods}
