% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations-class.R
\docType{class}
\name{DualSimulations-class}
\alias{DualSimulations-class}
\alias{.DualSimulations}
\title{Class for the simulations output from dual-endpoint model based designs}
\description{
This class captures the trial simulations from dual-endpoint model based
designs. In comparison to the parent class \code{\linkS4class{Simulations}},
it contains additional slots to capture the dose-biomarker fits, and the
sigma2W and rho estimates.
}
\section{Slots}{

\describe{
\item{\code{rhoEst}}{the vector of final posterior median rho estimates}

\item{\code{sigma2West}}{the vector of final posterior median sigma2W estimates}

\item{\code{fitBiomarker}}{list with the final dose-biomarker curve fits}
}}

\keyword{classes}
