% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations-class.R
\docType{class}
\name{DualSimulationsSummary-class}
\alias{DualSimulationsSummary-class}
\alias{.DualSimulationsSummary}
\alias{DualSimulationsSummary}
\alias{.DefaultDualSimulationsSummary}
\title{\code{DualSimulationsSummary}}
\usage{
.DefaultDualSimulationsSummary()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
This class captures the summary of dual-endpoint simulations output.
In comparison to its parent class \code{\link{SimulationsSummary}}, it has additional slots.
}
\section{Slots}{

\describe{
\item{\code{biomarker_fit_at_dose_most_selected}}{(\code{numeric})\cr fitted biomarker level at most often selected dose.}

\item{\code{mean_biomarker_fit}}{(\code{list})\cr list with average, lower (2.5\%) and upper (97.5\%) quantiles of
mean fitted biomarker level at each dose}
}}

\note{
Typically, end users will not use the \code{.DefaultDualSimulationsSummary()} function.
}
