% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations-class.R
\docType{class}
\name{DualSimulationsSummary-class}
\alias{DualSimulationsSummary-class}
\alias{.DualSimulationsSummary}
\title{Class for the summary of dual-endpoint simulations output}
\description{
In addition to the slots in the parent class
\code{\linkS4class{SimulationsSummary}}, it contains two slots for the
biomarker model fit information.
}
\details{
Note that objects should not be created by users, therefore no
initialization function is provided for this class.
}
\section{Slots}{

\describe{
\item{\code{biomarkerFitAtDoseMostSelected}}{fitted biomarker level at dose most often selected}

\item{\code{meanBiomarkerFit}}{list with the average, lower (2.5%) and upper (97.5%)
quantiles of the mean fitted biomarker level at each dose level}
}}

\keyword{classes}
