% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Model-class.R
\docType{class}
\name{EffFlexi-class}
\alias{.EffFlexi}
\alias{EffFlexi-class}
\title{Class for the efficacy model in non-parameteric or flexible form using pseudo data prior
This is a class where no parameter model and the mean efficacy responses are estimated at
each dose levels aiming to capture different shapes for the efficacy curve.
The first order or the second order random walk model can be used foe smoothing data}
\description{
Class for the efficacy model in non-parameteric or flexible form using pseudo data prior
This is a class where no parameter model and the mean efficacy responses are estimated at
each dose levels aiming to capture different shapes for the efficacy curve.
The first order or the second order random walk model can be used foe smoothing data
}
\keyword{class}

