% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-class.R
\name{EffFlexi}
\alias{EffFlexi}
\title{Initialization function for the "EffFlexi" class}
\usage{
EffFlexi(Eff, Effdose, sigma2, sigma2betaW, smooth = c("RW1", "RW2"),
  data)
}
\arguments{
\item{Eff}{the pseudo efficacy responses}

\item{Effdose}{the corresponding dose levels for the pseudo efficacy responses}

\item{sigma2}{the prior variance of the efficacy responses which can be specified 
with a single positive scalar or with two positive scalar values fot the shape \code{a} and 
the rate \code{b} parameters of the inverse gamma distribtuion.}

\item{sigma2betaW}{the prior variance of the random walk model used for smoothing which can be
specified with a single postive scalar or with two positive scalars representing the shape \code{a}
and the rate \code{b} parameter of the inverse gamma distribution.}

\item{smooth}{used for smoothing data for this efficacy model. That is either the "RW1", the 
first-order random walk model or "RW2", the second-order random walk model is used of the mean
efficacy responses.}

\item{data}{the input data to update estimates of model parameters and 
follow the \code{\linkS4class{DataDual}} object class specification}
}
\value{
the \code{\linkS4class{EffFlexi}} class object
}
\description{
Initialization function for the "EffFlexi" class
}
\keyword{methods}
