% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Design-methods.R
\docType{methods}
\name{examine}
\alias{examine}
\alias{examine,Design-method}
\alias{examine,RuleDesign-method}
\title{Obtain hypothetical trial course table for a design}
\usage{
examine(object, ...)

\S4method{examine}{Design}(object, mcmcOptions = McmcOptions(), ...)

\S4method{examine}{RuleDesign}(object, ...)
}
\arguments{
\item{object}{the design (\code{\linkS4class{Design}} or
\code{\linkS4class{RuleDesign}} object) we want to examine}

\item{mcmcOptions}{object of class \code{\linkS4class{McmcOptions}},
giving the MCMC options for each evaluation in the trial. By default,
the standard options are used}

\item{\dots}{additional arguments (see methods)}
}
\value{
The data frame
}
\description{
This generic function takes a design and generates a dataframe
showing the beginning of several hypothetical trial courses under
the design. This means, from the generated dataframe one can read off:
- how many cohorts are required in the optimal case (no DLTs observed) in
  order to reach the highest dose of the specified dose grid
- assuming no DLTs are observed until a certain dose level, what the next
  recommended dose is for all possible number of DLTs observed
- the actual relative increments that will be used in these cases
- whether the trial would stop at a certain cohort
Examining the "single trial" behavior of a dose escalation design is
the first important step in evaluating a design, and cannot be replaced by
studying solely the operating characteristics in "many trials". The cohort
sizes are also taken from the design, assuming no DLTs occur until the dose
listed.
}
\section{Methods (by class)}{
\itemize{
\item \code{Design}: Examine a model-based CRM

\item \code{RuleDesign}: Examine a rule-based design
}}
\keyword{methods}
\keyword{regression}

