% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-class.R
\docType{class}
\name{GeneralData-class}
\alias{GeneralData-class}
\alias{.GeneralData}
\title{Class for general data input}
\description{
Class for general data input
}
\section{Slots}{

\describe{
\item{\code{ID}}{unique patient IDs (integer vector)}

\item{\code{cohort}}{the cohort indices (sorted values from 0, 1, 2, ...)}

\item{\code{nObs}}{number of observations}
}}

\keyword{classes}
