% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-class.R
\docType{class}
\name{GeneralModel-class}
\alias{GeneralModel-class}
\alias{.GeneralModel}
\title{No Intitialization function for this
General class for model input}
\description{
This is the general model class, from which all other specific models for 
involving BUGS (the software) for computing result. 
It inherits all slots from \code{\linkS4class{AllModels}}
}
\details{
The \code{datamodel} must obey the convention that the data input is
called exactly as in the corresponding data class.
All prior distributions for parameters should be contained in the
model function \code{priormodel}. The background is that this can
be used to simulate from the prior distribution, before obtaining any
data.
}
\section{Slots}{

\describe{
\item{\code{datamodel}}{a function representing the BUGS data model specification
(see the details above)}

\item{\code{priormodel}}{a function representing the BUGS prior specification
(see the details above)}

\item{\code{modelspecs}}{a function computing the list of the data model and prior
model specifications that are required for fully specifying them (e.g. prior
parameters, reference dose, etc.), based on the data
slots that are then required as arguments of this function. This will then
be passed to BUGS for the computations.}

\item{\code{init}}{a function computing the list of starting values for parameters
required to be initialized in the MCMC sampler, based on the
data slots that are then required as arguments of this
function}

\item{\code{sample}}{names of all parameters from which you would like to save the
MCMC samples.}
}}

\seealso{
\code{\linkS4class{Model}}
}
\keyword{classes}
