% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations-class.R
\docType{class}
\name{GeneralSimulationsSummary-class}
\alias{GeneralSimulationsSummary-class}
\alias{.GeneralSimulationsSummary}
\alias{GeneralSimulationsSummary}
\alias{.DefaultGeneralSimulationsSummary}
\alias{.DefaultPseudoSimulationsSummary}
\title{\code{GeneralSimulationsSummary}}
\usage{
.DefaultGeneralSimulationsSummary()

.DefaultPseudoSimulationsSummary()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This class captures the summary of general simulations output. Note that
objects should not be created by users. Therefore, no initialization
function is provided for this class.
}
\section{Slots}{

\describe{
\item{\code{target}}{(\code{numeric})\cr target toxicity interval}

\item{\code{targetDoseInterval}}{(\code{numeric})\cr corresponding target dose interval}

\item{\code{nsim}}{(\code{integer})\cr number of simulations}

\item{\code{propDLTs}}{(\code{ANY})\cr proportions of DLTs in the trials}

\item{\code{meanToxRisk}}{(\code{numeric})\cr mean toxicity risks for the patients}

\item{\code{doseSelected}}{(\code{numeric})\cr doses selected as MTD}

\item{\code{toxAtDosesSelected}}{(\code{numeric})\cr true toxicity at doses selected}

\item{\code{propAtTarget}}{(\code{numeric})\cr Proportion of trials selecting target MTD}

\item{\code{doseMostSelected}}{(\code{numeric})\cr dose most often selected as MTD}

\item{\code{obsToxRateAtDoseMostSelected}}{(\code{numeric})\cr observed toxicity rate at dose most often selected}

\item{\code{nObs}}{(\code{ANY})\cr number of patients overall}

\item{\code{nAboveTarget}}{(\code{integer})\cr number of patients treated above target tox interval}

\item{\code{doseGrid}}{(\code{numeric})\cr the dose grid that has been used}

\item{\code{placebo}}{(\code{logical})\cr set to TRUE (default is FALSE) for a design with placebo}
}}

\note{
Typically, end users will not use the \code{.DefaultGeneralSimulationsSummary()} function.

Typically, end users will not use the \code{.DefaultPseudoSimulationsSummary()} function.
}
