% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design-methods.R
\name{get_result_list}
\alias{get_result_list}
\title{Helper function to obtain simulation results list}
\usage{
get_result_list(fun, nsim, vars, parallel, n_cores)
}
\arguments{
\item{fun}{the simulation function for a single iteration, which takes as
single parameter the iteration index.}

\item{nsim}{number of simulations to be conducted.}

\item{vars}{names of the variables.}

\item{parallel}{should the simulation runs be parallelized across the
clusters of the computer?}

\item{n_cores}{how many cores should be used for parallel computing?}
}
\value{
The list with all simulation results (one iteration corresponds
to one list element).
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

The function \code{fun} can use variables that are visible to itself.
The names of these variables have to given in the vector \code{vars}.
}
\keyword{internal}
\keyword{programming}
