% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{h_calc_report_label_percentage}
\alias{h_calc_report_label_percentage}
\title{calculate percentage of true stopping rules for report label output}
\usage{
h_calc_report_label_percentage(stop_report)
}
\arguments{
\item{stop_report}{object from summary method}
}
\description{
calculates true column means and converts output into percentages
before combining the output with the report label; output is passed
to \code{\link[=show]{show()}} and output with cat to console
}
