% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{h_default_if_empty}
\alias{h_default_if_empty}
\title{Getting the default label for empty labels}
\usage{
h_default_if_empty(label, default)
}
\arguments{
\item{label}{contains the current label value and checks if the label is
an empty string, i.e. \code{""}. If so, this function replaces it with the default
label as specified by \code{default}. If not the label is returned as it is.}

\item{default}{contains the default label the replaces the empty string in label}
}
\value{
a label containing the updated (or remaining) label string.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

A simple helper function that replaces empty labels by the default label.
}
\examples{
h_default_if_empty(label = "", default = "default label")
h_default_if_empty(label = "custom label", default = "default label")
}
