% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_knitr_GeneralData.R
\name{h_get_formatted_dosegrid}
\alias{h_get_formatted_dosegrid}
\title{Format a \code{doseGrid} for Printing}
\usage{
h_get_formatted_dosegrid(grid, units = NA)
}
\arguments{
\item{grid}{(\code{double})\cr the dose grid}

\item{units}{(\code{character})\cr The units in which the values in \code{doseGrid} are
measured.  Appended to each value in \code{doseGrid} when \code{knit_print}ed.  The
default, \code{NA}, omits the units.}
}
\value{
A character string containing the formatted dose grid.  If the grid
is \code{c(1, 2, 3)} and \code{units} is \code{"mg"}, the returned value is \code{"1 mg, 2 mg and 3 mg"}.
}
\description{
Format a \code{doseGrid} for Printing
}
\keyword{internal}
