% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-class.R
\name{h_hierarchical_compile_modelspecs}
\alias{h_hierarchical_compile_modelspecs}
\title{Compile the Hierarchical Model-Specification Function}
\usage{
h_hierarchical_compile_modelspecs(models_to_arms, parameter_pools)
}
\arguments{
\item{models_to_arms}{(\code{list})\cr named arm-specific models.}

\item{parameter_pools}{(\code{list})\cr exchangeable parameter specification from
\code{\link[=HierarchicalModel]{HierarchicalModel()}}.}
}
\value{
A function returning the JAGS data list.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Creates the \code{modelspecs} function used by \code{\link[=mcmc]{mcmc()}} to prepare the JAGS data
needed by the dynamically compiled hierarchical model.
}
\keyword{internal}
