% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_rules.R
\name{h_next_best_marker}
\alias{h_next_best_marker}
\title{Add the Next-Best Dose Marker}
\usage{
h_next_best_marker(plot, dose, y, dose_grid, dose_scale)
}
\arguments{
\item{plot}{(\code{ggplot}) plot to decorate.}

\item{dose}{(\code{number}) next best dose.}

\item{y}{(\code{number}) vertical marker position.}

\item{dose_grid}{(\code{numeric}) ordered dose grid.}

\item{dose_scale}{(\code{string}) dose-axis scale.}
}
\value{
The decorated \code{ggplot2} object.
}
\description{
Add the Next-Best Dose Marker
}
\keyword{internal}
