% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_rules.R
\name{h_next_best_tdsamples_plot}
\alias{h_next_best_tdsamples_plot}
\title{Building the Plot for \code{nextBest-NextBestTDsamples} Method.}
\usage{
h_next_best_tdsamples_plot(
  target_in_trial_samples,
  target_trial_end_samples,
  target_in_trial_est,
  target_trial_end_est,
  dose_grid_range,
  nextBest,
  doselimit,
  next_best_dose
)
}
\arguments{
\item{target_in_trial_samples}{(\code{numeric})\cr vector of in-trial samples.}

\item{target_trial_end_samples}{(\code{numeric})\cr vector of end-of-trial samples.}

\item{target_in_trial_est}{(\code{number})\cr target in-trial estimate.}

\item{target_trial_end_est}{(\code{number})\cr target end-of-trial estimate.}

\item{dose_grid_range}{(\code{numeric})\cr range of dose grid.}

\item{nextBest}{(\code{NextBestTDsamples})\cr the rule for the next best dose.}

\item{doselimit}{(\code{number})\cr the maximum allowed next dose.}

\item{next_best_dose}{(\code{number})\cr next best dose.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Helper function which creates the plot for \code{\link[=nextBest-NextBestTDsamples]{nextBest-NextBestTDsamples()}}
method.
}
