% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{h_null_if_na}
\alias{h_null_if_na}
\title{Getting \code{NULL} for \code{NA}}
\usage{
h_null_if_na(x)
}
\arguments{
\item{x}{any atomic object of length \code{1}. For the definition of "atomic",
see \code{\link[=is.atomic]{is.atomic()}}.}
}
\value{
\code{NULL} if \code{x} is \code{NA}, otherwise, \code{x}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

A simpler helper function that replaces \code{NA} object by \code{NULL} object.
}
\examples{
h_null_if_na(NA)
}
