% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-methods.R
\name{h_plot_data_dataordinal}
\alias{h_plot_data_dataordinal}
\alias{plot-Data}
\title{Helper Function for the Plot Method of the Data and DataOrdinal Classes}
\usage{
h_plot_data_dataordinal(x, blind = FALSE, legend = TRUE, ...)
}
\arguments{
\item{x}{(\code{Data})\cr object we want to plot.}

\item{blind}{(\code{flag})\cr indicates whether to blind the data.
If \code{TRUE}, then placebo subjects are reported at the same level
as the active dose level in the corresponding cohort,
and DLTs are always assigned to the first subjects in a cohort.}

\item{legend}{(\code{flag})\cr whether the legend should be added.}

\item{...}{not used.}

\item{y}{(\code{missing})\cr missing object, for compatibility with the generic
function.}
}
\value{
The \code{\link{ggplot2}} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

A method that creates a plot for \code{\link{Data}}  and \code{\link{DataOrdinal}}objects.
}
