% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{h_plot_data_df}
\alias{h_plot_data_df}
\title{Helper Function for the Plot Method of subclasses of DataGeneral}
\usage{
h_plot_data_df(data, ...)
}
\arguments{
\item{data}{(\code{DataGeneral})\cr object from which data is extracted and converted
into a data frame.}

\item{...}{further arguments passed to class-specific methods.}
}
\value{
\code{data.frame} containing columns for patient, cohort, dose and toxicity grade
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

A method that transforms \code{\link{DataGeneral}}  objects into a \code{tibble} suitable for
plotting with \code{ggplot2} methods
}
