% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{h_plot_data_df}
\alias{h_plot_data_df}
\title{Prepare Data for Plotting}
\usage{
h_plot_data_df(data, blind = FALSE)
}
\arguments{
\item{data}{(\code{Data})\cr object from which data is extracted and converted
into a data frame.}

\item{blind}{(\code{flag})\cr should data be blinded?
If \code{TRUE}, then for each cohort, all DLTs are assigned to the first subjects
in the cohort. In addition, the placebo (if any) is set to the active dose
level for that cohort.}
}
\value{
A \verb{data frame} object with values to plot.

This is to blind the data.
For each cohort, all DLTs are assigned to the first subjects in the cohort.
In addition, the placebo (if any) is set to the active dose level for that cohort.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This helper function prepares a data.frame object based on \code{Data} class object.
The resulting data frame is used by the plot function for \code{Data} class objects.
}
