% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations-methods.R
\name{h_plot_doses_tried}
\alias{h_plot_doses_tried}
\title{Helper Function to Create Doses Tried Plot}
\usage{
h_plot_doses_tried(sim_doses, dose_grid)
}
\arguments{
\item{sim_doses}{(\code{list})\cr list of simulated doses per trial.}

\item{dose_grid}{(\code{numeric})\cr dose grid.}
}
\value{
A \code{ggplot} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Creates a bar plot showing average proportions of doses tested.
}
\keyword{internal}
