% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations-methods.R
\name{h_plot_simulation_trajectory}
\alias{h_plot_simulation_trajectory}
\title{Helper Function to Create Trajectory Plot}
\usage{
h_plot_simulation_trajectory(sim_doses, max_patients, has_placebo)
}
\arguments{
\item{sim_doses}{(\code{list})\cr list of simulated doses per trial.}

\item{max_patients}{(\code{integer})\cr maximum number of patients.}

\item{has_placebo}{(\code{flag})\cr whether the design includes placebo.}
}
\value{
A \code{ggplot} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Creates a trajectory plot showing dose level statistics across patients.
}
\keyword{internal}
