% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_simulations.R
\name{h_summarize_add_stats}
\alias{h_summarize_add_stats}
\title{Helper function to calculate average across iterations for each additional
reporting parameter
extracts parameter names as specified by user and averaged the values
for each specified parameter to \code{\link[=show]{show()}} and output with cat to console}
\usage{
h_summarize_add_stats(stats_list)
}
\arguments{
\item{stats_list}{object from simulation with nested parameter values
(sublist for each parameter)}
}
\value{
list of parameter names and averaged values for console output
}
\description{
Helper function to calculate average across iterations for each additional
reporting parameter
extracts parameter names as specified by user and averaged the values
for each specified parameter to \code{\link[=show]{show()}} and output with cat to console
}
