% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_design.R
\name{h_unpack_stopit}
\alias{h_unpack_stopit}
\title{Helper function to recursively unpack stopping rules and return lists with
logical value and label given}
\usage{
h_unpack_stopit(stopit_tree)
}
\arguments{
\item{stopit_tree}{object from simulate method}
}
\value{
named list
}
\description{
Helper function to recursively unpack stopping rules and return lists with
logical value and label given
}
