% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{Increments-class}
\alias{Increments-class}
\alias{Increments}
\title{\code{Increments}}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{\link{Increments}} is a virtual class for controlling increments, from which all
other specific increments classes inherit.
}
\seealso{
\code{\link{IncrementsRelative}}, \code{\link{IncrementsRelativeDLT}},
\code{\link{IncrementsDoseLevels}}, \code{\link{IncrementsHSRBeta}}, \code{\link{IncrementsMin}}.
}
