% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{Increments-class}
\alias{Increments-class}
\alias{Increments}
\title{The virtual class for controlling increments}
\description{
The virtual class for controlling increments
}
\seealso{
\code{\linkS4class{IncrementsRelative}},
\code{\linkS4class{IncrementsRelativeDLT}},
\code{\linkS4class{IncrementsRelativeParts}},
\code{\link{IncrementsNumDoseLevels}}
\code{\link{IncrementsHSRBeta}}
}
\keyword{classes}
