% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{IncrementsNumDoseLevels-class}
\alias{IncrementsNumDoseLevels-class}
\alias{.IncrementsNumDoseLevels}
\title{Increments control based on number of dose levels}
\description{
Increments control based on number of dose levels
}
\section{Slots}{

\describe{
\item{\code{maxLevels}}{scalar positive integer for the number of maximum 
dose levels to increment for the next dose. It defaults to 1, 
which means that no dose skipping is allowed - the next dose 
can be maximum one level higher than the current dose.}
}}

\examples{

# As example, here is the rule for: 
# maximum skip one dose level, that is 2 dose levels higher is maximum
# increment
myIncrements <- IncrementsNumDoseLevels(maxLevels=2)

}
\keyword{classes}
