\docType{class}
\name{IncrementsRelative-class}
\alias{IncrementsRelative-class}
\title{Increments control based on relative differences in intervals}
\description{
Increments control based on relative differences in
intervals
}
\section{Slots}{

\describe{
\item{\code{intervals}}{a vector with the bounds of the relevant intervals of length
\code{n}}

\item{\code{increments}}{a vector of length \code{n-1} with the maximum allowable
relative increments in the \code{intervals}}
}}
\keyword{classes}

