% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{joinBodies}
\alias{joinBodies}
\title{Helper function to join two function bodies}
\usage{
joinBodies(body1, body2)
}
\arguments{
\item{body1}{first body}

\item{body2}{second body}
}
\value{
joined body
}
\description{
Helper function to join two function bodies
}
\keyword{internal}
\keyword{programming}
