% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{joinModels}
\alias{joinModels}
\title{Helper function to join two BUGS models}
\usage{
joinModels(model1, model2)
}
\arguments{
\item{model1}{first model}

\item{model2}{second model}
}
\value{
joined model
}
\description{
Helper function to join two BUGS models
}
\keyword{internal}
\keyword{programming}
