% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_knitr_CohortSize.R,
%   R/helpers_kintr_Increments.R, R/helpers_knitr_GeneralData.R,
%   R/helpers_knitr_GeneralModel.R, R/helpers_knitr_Stopping.R
\name{knit_print}
\alias{knit_print}
\alias{knit_print.CohortSizeConst}
\alias{knit_print.CohortSizeRange}
\alias{knit_print.CohortSizeDLT}
\alias{knit_print.CohortSizeParts}
\alias{knit_print.CohortSizeMax}
\alias{knit_print.CohortSizeMin}
\alias{knit_print.CohortSizeOrdinal}
\alias{knit_print.IncrementsRelative}
\alias{knit_print.IncrementsRelativeDLT}
\alias{knit_print.IncrementsDoseLevels}
\alias{knit_print.IncrementsHSRBeta}
\alias{knit_print.IncrementsMin}
\alias{knit_print.IncrementsOrdinal}
\alias{knit_print.IncrementsRelativeParts}
\alias{knit_print.IncrementsRelativeDLTCurrent}
\alias{knit_print.GeneralData}
\alias{knit_print.DualEndpoint}
\alias{knit_print.ModelParamsNormal}
\alias{knit_print.GeneralModel}
\alias{knit_print.LogisticKadane}
\alias{knit_print.LogisticKadaneBetaGamma}
\alias{knit_print.LogisticLogNormal}
\alias{knit_print.LogisticLogNormalMixture}
\alias{knit_print.LogisticLogNormalSub}
\alias{knit_print.LogisticNormalMixture}
\alias{knit_print.LogisticNormalFixedMixture}
\alias{knit_print.OneParLogNormalPrior}
\alias{knit_print.OneParExpPrior}
\alias{knit_print.LogisticLogNormalGrouped}
\alias{knit_print.LogisticLogNormalOrdinal}
\alias{knit_print.StoppingMaxGainCIRatio}
\alias{knit_print.StoppingTDCIRatio}
\alias{knit_print.StoppingTargetBiomarker}
\alias{knit_print.StoppingLowestDoseHSRBeta}
\alias{knit_print.StoppingMTDCV}
\alias{knit_print.StoppingMTDdistribution}
\alias{knit_print.StoppingHighestDose}
\alias{knit_print.StoppingSpecificDose}
\alias{knit_print.StoppingTargetProb}
\alias{knit_print.StoppingMinCohorts}
\alias{knit_print.StoppingMinPatients}
\alias{knit_print.StoppingPatientsNearDose}
\alias{knit_print.StoppingCohortsNearDose}
\alias{knit_print.StoppingMissingDose}
\title{Render a \code{CohortSizeConst} Object}
\usage{
\method{knit_print}{CohortSizeConst}(x, ..., asis = TRUE, label = c("participant", "participants"))

\method{knit_print}{CohortSizeRange}(x, ..., asis = TRUE)

\method{knit_print}{CohortSizeDLT}(x, ..., asis = TRUE)

\method{knit_print}{CohortSizeParts}(x, ..., asis = TRUE, label = c("participant", "participants"))

\method{knit_print}{CohortSizeMax}(x, ..., asis = TRUE)

\method{knit_print}{CohortSizeMin}(x, ..., asis = TRUE)

\method{knit_print}{CohortSizeOrdinal}(x, ..., asis = TRUE)

\method{knit_print}{IncrementsRelative}(x, ..., asis = TRUE)

\method{knit_print}{IncrementsRelativeDLT}(x, ..., asis = TRUE)

\method{knit_print}{IncrementsDoseLevels}(x, ..., asis = TRUE)

\method{knit_print}{IncrementsHSRBeta}(x, ..., asis = TRUE)

\method{knit_print}{IncrementsMin}(x, ..., asis = TRUE)

\method{knit_print}{IncrementsOrdinal}(x, ..., asis = TRUE)

\method{knit_print}{IncrementsRelativeParts}(x, ..., asis = TRUE, labels = c("toxicity", "toxicities"))

\method{knit_print}{IncrementsRelativeDLTCurrent}(x, ..., asis = TRUE, labels = c("DLT", "DLTs"))

\method{knit_print}{GeneralData}(
  x,
  ...,
  asis = TRUE,
  labels = c("participant", "participants"),
  full_grid = FALSE,
  summarise = c("none", "dose", "cohort"),
  summarize = summarise,
  units = NA,
  format_func = function(x) x
)

\method{knit_print}{DualEndpoint}(
  x,
  ...,
  asis = TRUE,
  use_values = TRUE,
  fmt = "\%5.2f",
  units = NA,
  biomarker_name = "a PD biomarker"
)

\method{knit_print}{ModelParamsNormal}(
  x,
  use_values = TRUE,
  fmt = "\%5.2f",
  params = c("alpha", "beta"),
  preamble = "The prior for &theta; is given by\\\\n",
  asis = TRUE,
  theta = "\\\\theta",
  ...
)

\method{knit_print}{GeneralModel}(
  x,
  ...,
  params = c("alpha", "beta"),
  asis = TRUE,
  use_values = TRUE,
  fmt = "\%5.2f",
  units = NA
)

\method{knit_print}{LogisticKadane}(x, ..., asis = TRUE, use_values = TRUE, fmt = "\%5.2f", units = NA)

\method{knit_print}{LogisticKadaneBetaGamma}(x, ..., asis = TRUE, use_values = TRUE, fmt = "\%5.2f", units = NA)

\method{knit_print}{LogisticLogNormal}(
  x,
  ...,
  use_values = TRUE,
  fmt = "\%5.2f",
  params = c(`\\\\alpha` = "alpha", `log(\\\\beta)` = "beta"),
  preamble = "The prior for &theta; is given by\\\\n",
  asis = TRUE
)

\method{knit_print}{LogisticLogNormalMixture}(x, ..., asis = TRUE, use_values = TRUE, fmt = "\%5.2f", units = NA)

\method{knit_print}{LogisticLogNormalSub}(
  x,
  ...,
  use_values = TRUE,
  fmt = "\%5.2f",
  params = c(`\\\\alpha` = "alpha", `log(\\\\beta)` = "beta"),
  preamble = "The prior for &theta; is given by\\\\n",
  asis = TRUE
)

\method{knit_print}{LogisticNormalMixture}(x, ..., asis = TRUE, use_values = TRUE, fmt = "\%5.2f", units = NA)

\method{knit_print}{LogisticNormalFixedMixture}(x, ..., asis = TRUE, use_values = TRUE, fmt = "\%5.2f", units = NA)

\method{knit_print}{OneParLogNormalPrior}(x, ..., asis = TRUE, use_values = TRUE, fmt = "\%5.2f")

\method{knit_print}{OneParExpPrior}(x, ..., asis = TRUE)

\method{knit_print}{LogisticLogNormalGrouped}(
  x,
  ...,
  use_values = TRUE,
  fmt = "\%5.2f",
  params = c(`\\\\alpha` = "alpha", `\\\\beta` = "beta", `log(\\\\delta_0)` = "delta_0",
    `log(\\\\delta_1)` = "delta_1"),
  preamble = "The prior for &theta; is given by\\\\n",
  asis = TRUE
)

\method{knit_print}{LogisticLogNormalOrdinal}(
  x,
  ...,
  use_values = TRUE,
  fmt = "\%5.2f",
  params = NA,
  preamble = "The prior for &theta; is given by\\\\n",
  asis = TRUE
)

\method{knit_print}{StoppingMaxGainCIRatio}(
  x,
  ...,
  dose_label = "the next best dose",
  tox_label = "toxicity",
  fmt_string =
    paste0("\%sIf the ratio of the upper and lower limits of the posterior 95\%\% ",
    "credible interval for the probability of \%s at the smaller of the doses ",
    "given by the end-of-trial TD estimate and the G* estimate is les than or ",
    "equal to "),
  asis = TRUE
)

\method{knit_print}{StoppingTDCIRatio}(
  x,
  ...,
  dose_label = "the next best dose",
  tox_label = "toxicity",
  fmt_string =
    paste0("\%sIf, at \%s, the ratio of the upper and lower limits of the posterior ",
    "95\%\% credible interval for the probability of \%s is less than or equal to "),
  asis = TRUE
)

\method{knit_print}{StoppingTargetBiomarker}(
  x,
  ...,
  dose_label = "the next best dose",
  biomarker_label = "the target biomarker",
  fmt_string =
    paste0("\%sIf, at \%s, the posterior probability that \%s is in the range (\%.2f, \%.2f)\%s is \%.0f\%\% or more."),
  asis = TRUE
)

\method{knit_print}{StoppingLowestDoseHSRBeta}(
  x,
  ...,
  tox_label = "toxicity",
  fmt_string =
    paste0("\%sIf, using a Hard Stopping Rule with a prior of Beta(\%.0f, \%.0f), the ",
    "lowest dose in the dose grid has a posterior probability of \%s of ",
    "\%.0f\%\% or more."),
  asis = TRUE
)

\method{knit_print}{StoppingMTDCV}(
  x,
  ...,
  fmt_string =
    paste0("\%sIf the posterior estimate of the robust coefficient of variation of the MTD, ",
    "expressed as a percentage, is both greater than zero and less than or ",
    "equal to \%.0f\%\%."),
  asis = TRUE
)

\method{knit_print}{StoppingMTDdistribution}(
  x,
  ...,
  fmt_string =
    "\%sIf the mean posterior probability of \%s at \%.0f\%\% of \%s is at least \%4.2f.",
  dose_label = "the next best dose",
  tox_label = "toxicity",
  asis = TRUE
)

\method{knit_print}{StoppingHighestDose}(
  x,
  ...,
  dose_label = "the highest dose in the dose grid",
  asis = TRUE
)

\method{knit_print}{StoppingSpecificDose}(x, ..., dose_label = as.character(x@dose), asis = TRUE)

\method{knit_print}{StoppingTargetProb}(
  x,
  ...,
  fmt_string =
    "\%sIf the probability of \%s at \%s is in the range [\%4.2f, \%4.2f] is at least \%4.2f.",
  dose_label = "the next best dose",
  tox_label = "toxicity",
  asis = TRUE
)

\method{knit_print}{StoppingMinCohorts}(x, ..., asis = TRUE)

\method{knit_print}{StoppingMinPatients}(x, ..., label = "participants", asis = TRUE)

\method{knit_print}{StoppingPatientsNearDose}(
  x,
  ...,
  dose_label = "the next best dose",
  label = "participants",
  asis = TRUE
)

\method{knit_print}{StoppingCohortsNearDose}(x, ..., dose_label = "the next best dose", asis = TRUE)

\method{knit_print}{StoppingMissingDose}(x, ..., asis = TRUE)
}
\arguments{
\item{x}{(\code{ModelParamsNormal})\cr the object to be rendered}

\item{...}{Not used at present}

\item{asis}{(\code{flag})\cr Not used at present}

\item{label}{(\code{character})\cr the term used to label participants}

\item{labels}{(\code{character})\cr How to describe the participants in the trial.
See Usage Notes below.}

\item{full_grid}{(\code{flag})\cr Should the full dose grid appear in the output table
or simply those doses for whom at least one evaluable participant is available?
Ignored unless \code{summarise == "dose"}.}

\item{summarise}{(\code{character})\cr How to summarise the observed data.  The default,
\code{"none"}, lists observed data at the participant level.  \code{"dose"} presents
participant counts by dose and \code{"cohort"} by cohort.}

\item{summarize}{(\code{character})\cr Synonym for \code{summarise}}

\item{units}{(\code{character})\cr The units in which the values in \code{doseGrid} are
measured.  Appended to each value in \code{doseGrid} when \code{knit_print}ed.  The
default, \code{NA}, omits the units.}

\item{format_func}{(\code{function})\cr The function used to format the participant table.
The default applies no formatting.  Obvious alternatives include \code{kableExtra::kable_styling}.}

\item{use_values}{(\code{flag})\cr print the values associated with hyperparameters,
or the symbols used to define the hyper-parameters.  That is, for example, mu or 1.}

\item{fmt}{(\code{character})\cr the \code{sprintf} format string used to render
numerical values.  Ignored if \code{use_values} is \code{FALSE}.}

\item{biomarker_name}{(\code{character})\cr A description of the biomarker}

\item{params}{(\code{character})\cr The names of the model parameters.  See Usage
Notes below.}

\item{preamble}{(\code{character})\cr The text used to introduce the LaTeX representation
of the model}

\item{theta}{(\code{character})\cr the LaTeX representation of the theta vector}

\item{dose_label}{(\code{character})\cr the term used to describe the target dose}

\item{fmt_string}{(\code{character})\cr the character string that defines the format
of the output}

\item{biomarker_label}{(\code{character})\cr the term used to describe the biomarker}
}
\value{
a character string that represents the object in markdown.

The markdown representation of the object, as a character string

A character string containing a LaTeX rendition of the object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

We provide additional utility functions to allow human-friendly rendition of
crmPack objects in Markdown and Quarto files

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\section{Usage Notes}{

\code{label} describes the trial's participants.

It should be a character vector of length 1 or 2.  If of length 2, the first
element describes a \code{cohort_size} of 1 and the second describes all other
\code{cohort_size}s.  If of length 1, the character \code{s} is appended to the value
when \code{cohort_size} is not 1.


The default value of \code{col.names} is \code{c("Lower", "Upper", "Cohort size")} and
that of \code{caption} is \code{"Defined by the dose to be used in the next cohort"}.
These values can be overridden by passing \code{col.names} and \code{caption} in the
function call.


The by default, the columns are labelled \code{Lower}, \code{Upper} and \verb{Cohort size}.  The table's caption is
\verb{Defined by the number of DLTs so far observed}.  These values can be overridden by passing
\code{col.names} and \code{caption} in the function call.


The default value of \code{col.names} is \code{c("Min", "Max", "Increment")} and that
of \code{caption} is \code{"Defined by highest dose administered so far"}.  These
values can be overridden by passing \code{col.names} and \code{caption} in the function
call.


The default value of \code{col.names} is \code{c("Min", "Max", "Increment")} and that
of \code{caption} is \code{"Defined by number of DLTs reported so far"}. These values
can be overridden by passing \code{col.names} and \code{caption} in the function call.


\code{labels} defines how toxicities are described.

It should be a character vector of length 1 or 2.  If of length 2, the first
element describes a single toxicity and the second describes all other
toxicity counts.  If of length 1, the character \code{s} is appended to the value
describing a single toxicity.


The default value of \code{col.names} is \code{c("Min", "Max", "Increment")} and that
of \code{caption} is \code{"Defined by number of DLTs in the current cohort"}. These values
can be overridden by passing \code{col.names} and \code{caption} in the function call.

\code{labels} defines how toxicities are described.

It should be a character vector of length 1 or 2.  If of length 2, the first
element describes a single toxicity and the second describes all other
toxicity counts.  If of length 1, the character \code{s} is appended to the value
describing a single toxicity.


\code{labels} describes the trial's participants.

It should be a character vector of length 1 or 2.  If of length 2, the first
element describes a single participant and the second describes all other
situations.  If of length 1, the character \code{s} is appended to the value
when the number of participants is not 1.
The default values of \code{col.names} and \code{caption} vary depending on the summary
requested. The default values can be overridden by passing \code{col.names} and
\code{caption} in the function call.


\code{params} must be a character vector of length equal to that of \code{x@mean} (and
\code{x@cov}).  Its values represent the parameters of the model as entries in the
vector \code{theta}, on the left-hand side of "~" in the definition of the prior.
If named, names should be valid LaTeX, escaped as usual for R character variables.
For example, \code{"\\\\alpha"} or \code{"\\\\beta_0"}.  If unnamed, names are constructed by
pre-pending an escaped backslash to each value provided.


\code{label} describes the trial's participants.

It should be a character vector of length 1 or 2.  If of length 2, the first
element describes a \code{cohort_size} of 1 and the second describes all other
\code{cohort_size}s.  If of length 1, the character \code{s} is appended to the value
when \code{cohort_size} is not 1.


The default value of \code{col.names} is \code{c("Lower", "Upper", "Cohort size")} and
that of \code{caption} is \code{"Defined by the dose to be used in the next cohort"}.
These values can be overridden by passing \code{col.names} and \code{caption} in the
function call.


The by default, the columns are labelled \code{Lower}, \code{Upper} and \verb{Cohort size}.  The table's caption is
\verb{Defined by the number of DLTs so far observed}.  These values can be overridden by passing
\code{col.names} and \code{caption} in the function call.


The default value of \code{col.names} is \code{c("Min", "Max", "Increment")} and that
of \code{caption} is \code{"Defined by highest dose administered so far"}.  These
values can be overridden by passing \code{col.names} and \code{caption} in the function
call.


The default value of \code{col.names} is \code{c("Min", "Max", "Increment")} and that
of \code{caption} is \code{"Defined by number of DLTs reported so far"}. These values
can be overridden by passing \code{col.names} and \code{caption} in the function call.


\code{labels} defines how toxicities are described.

It should be a character vector of length 1 or 2.  If of length 2, the first
element describes a single toxicity and the second describes all other
toxicity counts.  If of length 1, the character \code{s} is appended to the value
describing a single toxicity.


The default value of \code{col.names} is \code{c("Min", "Max", "Increment")} and that
of \code{caption} is \code{"Defined by number of DLTs in the current cohort"}. These values
can be overridden by passing \code{col.names} and \code{caption} in the function call.

\code{labels} defines how toxicities are described.

It should be a character vector of length 1 or 2.  If of length 2, the first
element describes a single toxicity and the second describes all other
toxicity counts.  If of length 1, the character \code{s} is appended to the value
describing a single toxicity.


\code{labels} describes the trial's participants.

It should be a character vector of length 1 or 2.  If of length 2, the first
element describes a single participant and the second describes all other
situations.  If of length 1, the character \code{s} is appended to the value
when the number of participants is not 1.
The default values of \code{col.names} and \code{caption} vary depending on the summary
requested. The default values can be overridden by passing \code{col.names} and
\code{caption} in the function call.


\code{params} must be a character vector of length equal to that of \code{x@mean} (and
\code{x@cov}).  Its values represent the parameters of the model as entries in the
vector \code{theta}, on the left-hand side of "~" in the definition of the prior.
If named, names should be valid LaTeX, escaped as usual for R character variables.
For example, \code{"\\\\alpha"} or \code{"\\\\beta_0"}.  If unnamed, names are constructed by
pre-pending an escaped backslash to each value provided.
}

\seealso{
\code{\link{knit_print}} for more details.
}
