% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_kintr_Increments.R,
%   R/helpers_knitr_CohortSize.R
\name{knit_print.IncrementsRelative}
\alias{knit_print.IncrementsRelative}
\alias{knit_print.IncrementsRelativeDLT}
\alias{knit_print.IncrementsDoseLevels}
\alias{knit_print.IncrementsHSRBeta}
\alias{knit_print.IncrementsMin}
\alias{knit_print.IncrementsOrdinal}
\alias{knit_print.IncrementsRelativeParts}
\alias{knit_print.IncrementsRelativeDLTCurrent}
\alias{knit_print}
\alias{knit_print.CohortSizeConst}
\alias{knit_print.CohortSizeRange}
\alias{knit_print.CohortSizeDLT}
\alias{knit_print.CohortSizeParts}
\alias{knit_print.CohortSizeMax}
\alias{knit_print.CohortSizeMin}
\alias{knit_print.CohortSizeOrdinal}
\title{Render a IncrementsRelative Object}
\usage{
\method{knit_print}{IncrementsRelative}(x, ..., asis = TRUE)

\method{knit_print}{IncrementsRelativeDLT}(x, ..., asis = TRUE)

\method{knit_print}{IncrementsDoseLevels}(object, ..., asis = TRUE)

\method{knit_print}{IncrementsHSRBeta}(x, ..., asis = TRUE)

\method{knit_print}{IncrementsMin}(x, ..., asis = TRUE)

\method{knit_print}{IncrementsOrdinal}(x, ..., asis = TRUE)

\method{knit_print}{IncrementsRelativeParts}(x, ..., asis = TRUE, labels = c("toxicity", "toxicities"))

\method{knit_print}{IncrementsRelativeDLTCurrent}(x, ..., asis = TRUE, labels = c("DLT", "DLTs"))

\method{knit_print}{CohortSizeConst}(x, ..., asis = TRUE, label = c("participant", "participants"))

\method{knit_print}{CohortSizeRange}(x, ..., asis = TRUE)

\method{knit_print}{CohortSizeDLT}(x, ..., asis = TRUE)

\method{knit_print}{CohortSizeParts}(x, ..., asis = TRUE, label = c("participant", "participants"))

\method{knit_print}{CohortSizeMax}(x, ..., asis = TRUE)

\method{knit_print}{CohortSizeMin}(x, ..., asis = TRUE)

\method{knit_print}{CohortSizeOrdinal}(x, ..., asis = TRUE)
}
\arguments{
\item{x}{(\code{CohortSize})\cr The object to knit_print.}

\item{...}{passed through to the \code{knit_print} method of the standard rule}

\item{asis}{(\code{flag})\cr Should the return value be wrapped in a call to \code{knitr::asis_output}?}

\item{labels}{(\code{character})\cr The word used to describe toxicities.  See
Usage Notes below.}

\item{label}{(\code{character})\cr The word used to label the participants.  See Usage Notes below.}
}
\value{
a character string that represents the object in markdown.

The markdown representation of the object, as a character string
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

We provide additional utility functions to allow human-friendly rendition of
crmPack objects in Markdown and Quarto files

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\section{Usage Notes}{

The default value of \code{col.names} is \code{c("Min", "Max", "Increment")} and that
of \code{caption} is \code{"Defined by highest dose administered so far"}.  These
values can be overridden by passing \code{col.names} and \code{caption} in the function
call.


The default value of \code{col.names} is \code{c("Min", "Max", "Increment")} and that
of \code{caption} is \code{"Defined by number of DLTs reported so far"}. These values
can be overridden by passing \code{col.names} and \code{caption} in the function call.


\code{labels} defines how toxicities are described.

It should be a character vector of length 1 or 2.  If of length 2, the first
element describes a single toxicity and the second describes all other
toxicity counts.  If of length 1, the character \code{s} is appended to the value
describing a single toxicity.


The default value of \code{col.names} is \code{c("Min", "Max", "Increment")} and that
of \code{caption} is \code{"Defined by number of DLTs in the current cohort"}. These values
can be overridden by passing \code{col.names} and \code{caption} in the function call.

\code{labels} defines how toxicities are described.

It should be a character vector of length 1 or 2.  If of length 2, the first
element describes a single toxicity and the second describes all other
toxicity counts.  If of length 1, the character \code{s} is appended to the value
describing a single toxicity.


\code{label} describes the trial's participants.

It should be a character vector of length 1 or 2.  If of length 2, the first
element describes a \code{cohort_size} of 1 and the second describes all other
\code{cohort_size}s.  If of length 1, the character \code{s} is appended to the value
when \code{cohort_size} is not 1.


The by default, the columns are labelled \code{Lower}, \code{Upper} and \verb{Cohort size}.  The table's caption is
\verb{Defined by the number of DLTs so far observed}.  These values can be overridden by passing
\code{col.names} and \code{caption} in the function call.


The default value of \code{col.names} is \code{c("Min", "Max", "Increment")} and that
of \code{caption} is \code{"Defined by highest dose administered so far"}.  These
values can be overridden by passing \code{col.names} and \code{caption} in the function
call.


The default value of \code{col.names} is \code{c("Min", "Max", "Increment")} and that
of \code{caption} is \code{"Defined by number of DLTs reported so far"}. These values
can be overridden by passing \code{col.names} and \code{caption} in the function call.


\code{labels} defines how toxicities are described.

It should be a character vector of length 1 or 2.  If of length 2, the first
element describes a single toxicity and the second describes all other
toxicity counts.  If of length 1, the character \code{s} is appended to the value
describing a single toxicity.


The default value of \code{col.names} is \code{c("Min", "Max", "Increment")} and that
of \code{caption} is \code{"Defined by number of DLTs in the current cohort"}. These values
can be overridden by passing \code{col.names} and \code{caption} in the function call.

\code{labels} defines how toxicities are described.

It should be a character vector of length 1 or 2.  If of length 2, the first
element describes a single toxicity and the second describes all other
toxicity counts.  If of length 1, the character \code{s} is appended to the value
describing a single toxicity.


\code{label} describes the trial's participants.

It should be a character vector of length 1 or 2.  If of length 2, the first
element describes a \code{cohort_size} of 1 and the second describes all other
\code{cohort_size}s.  If of length 1, the character \code{s} is appended to the value
when \code{cohort_size} is not 1.


The by default, the columns are labelled \code{Lower}, \code{Upper} and \verb{Cohort size}.  The table's caption is
\verb{Defined by the number of DLTs so far observed}.  These values can be overridden by passing
\code{col.names} and \code{caption} in the function call.
}

\seealso{
\code{\link{knit_print}} for more details.
}
