% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-class.R
\name{LogisticIndepBeta}
\alias{LogisticIndepBeta}
\title{Intialization function for "LogisticIndepBeta" class}
\usage{
LogisticIndepBeta(binDLE, DLEdose, DLEweights, data)
}
\arguments{
\item{binDLE}{the number of subjects observed with a DLE, the pseudo DLE responses}

\item{DLEdose}{the corresponding dose levels for the pseudo DLE responses, pseudo dose levels}

\item{DLEweights}{the total number of subjects treated at each of the dose levels, pseudo weights}

\item{data}{the input data to update estimates of model parameters and 
follow the \code{\linkS4class{Data}} object class specification}
}
\value{
the \code{\linkS4class{LogisticIndepBeta}}
}
\description{
Intialization function for "LogisticIndepBeta" class
}
\keyword{methods}
