% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-class.R
\docType{class}
\name{LogisticNormal-class}
\alias{LogisticNormal-class}
\alias{.LogisticNormal}
\title{Standard logistic model with bivariate normal prior}
\description{
This is the usual logistic regression model with a bivariate normal prior on
the intercept and slope.
}
\details{
The covariate is the natural logarithm of the dose \eqn{x} divided by
the reference dose \eqn{x^{*}}:

\deqn{logit[p(x)] = \alpha + \beta \cdot \log(x/x^{*})}
where \eqn{p(x)} is the probability of observing a DLT for a given dose
\eqn{x}.

The prior is
\deqn{(\alpha, \beta) \sim Normal(\mu, \Sigma)}

The slots of this class contain the mean vector, the covariance and
precision matrices of the bivariate normal distribution, as well as the
reference dose.
}
\section{Slots}{

\describe{
\item{\code{mean}}{the prior mean vector \eqn{\mu}}

\item{\code{cov}}{the prior covariance matrix \eqn{\Sigma}}

\item{\code{prec}}{the prior precision matrix \eqn{\Sigma^{-1}}}

\item{\code{refDose}}{the reference dose \eqn{x^{*}}}
}}

\examples{
# Define the dose-grid
emptydata <- Data(doseGrid = c(1, 3, 5, 10, 15, 20, 25, 40, 50, 80, 100))


model <- LogisticNormal(mean = c(-0.85, 1),
                        cov = matrix(c(1, -0.5, -0.5, 1), nrow = 2),
                        refDose = 50)

options <- McmcOptions(burnin=100,
                       step=2,
                       samples=1000)

options(error=recover)
mcmc(emptydata, model, options)
}
\keyword{classes}
