% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{logit}
\alias{logit}
\title{Shorthand for logit function}
\usage{
logit(x)
}
\arguments{
\item{x}{the function argument}
}
\value{
the logit(x)
}
\description{
Shorthand for logit function
}
\keyword{programming}
