% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rules-methods.R
\docType{methods}
\name{maxSize}
\alias{maxSize}
\alias{maxSize,CohortSize-method}
\title{"MAX" combination of cohort size rules}
\usage{
maxSize(...)

\S4method{maxSize}{CohortSize}(...)
}
\arguments{
\item{\dots}{Objects of class \code{\linkS4class{CohortSize}}}
}
\value{
the combination as an object of class
\code{\linkS4class{CohortSizeMax}}
}
\description{
This function combines cohort size rules by taking
the maximum of all sizes.
}
\section{Methods (by class)}{
\itemize{
\item \code{CohortSize}: The method combining cohort size rules by taking maximum
}}
\seealso{
\code{\link{minSize}}
}
\keyword{methods}

