% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mcmc.R
\docType{methods}
\name{mcmc}
\alias{mcmc}
\alias{mcmc,Data,LogisticNormal,McmcOptions-method}
\alias{mcmc,Data,Model,McmcOptions-method}
\title{Obtain posterior samples for all model parameters}
\usage{
mcmc(data, model, options, ...)

\S4method{mcmc}{Data,Model,McmcOptions}(data, model, options,
  program = c("JAGS", "OpenBUGS", "WinBUGS"), verbose = FALSE, ...)

\S4method{mcmc}{Data,LogisticNormal,McmcOptions}(data, model, options,
  verbose = FALSE, ...)
}
\arguments{
\item{data}{The data input, an object of class \code{\linkS4class{Data}}}

\item{model}{The model input, an object of class \code{\linkS4class{Model}}}

\item{options}{MCMC options, an object of class
\code{\linkS4class{McmcOptions}}}

\item{program}{the program which shall be used: either \dQuote{JAGS} (default),
\dQuote{OpenBUGS} or \dQuote{WinBUGS}}

\item{verbose}{shall progress bar and messages be printed? (not default)}

\item{\dots}{unused}
}
\value{
The posterior samples, an object of class
\code{\linkS4class{Samples}}.
}
\description{
Obtain posterior samples for all model parameters
}
\details{
This is the function to actually run the MCMC machinery to produce posterior
samples from all model parameters and required derived values. It is a
generic function, so that customized versions may be conveniently defined
for specific subclasses of Data, Model, and McmcOptions input.
}
\section{Methods (by class)}{
\itemize{
\item \code{data = Data,model = Model,options = McmcOptions}: Standard method which uses JAGS/BUGS

\item \code{data = Data,model = LogisticNormal,options = McmcOptions}: The fast method for the LogisticNormal class
}}
\keyword{methods}
\keyword{regression}

