% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mcmc.R
\docType{methods}
\name{mcmc}
\alias{mcmc}
\alias{mcmc,Data,LogisticNormal,McmcOptions-method}
\alias{mcmc,GeneralData,GeneralModel,McmcOptions-method}
\title{Obtain posterior samples for all model parameters}
\usage{
mcmc(data, model, options, ...)

\S4method{mcmc}{GeneralData,GeneralModel,McmcOptions}(data, model, options,
  program = c("JAGS", "OpenBUGS", "WinBUGS"), verbose = FALSE, ...)

\S4method{mcmc}{Data,LogisticNormal,McmcOptions}(data, model, options,
  verbose = FALSE, ...)
}
\arguments{
\item{data}{The data input, an object of class \code{\linkS4class{GeneralData}}}

\item{model}{The model input, an object of class \code{\linkS4class{GeneralModel}}}

\item{options}{MCMC options, an object of class
\code{\linkS4class{McmcOptions}}}

\item{program}{the program which shall be used: either \dQuote{JAGS} (default),
\dQuote{OpenBUGS} or \dQuote{WinBUGS}}

\item{verbose}{shall progress bar and messages be printed? (not default)}

\item{\dots}{unused}
}
\value{
The posterior samples, an object of class
\code{\linkS4class{Samples}}.
}
\description{
This is the function to actually run the MCMC machinery to produce posterior
samples from all model parameters and required derived values. It is a
generic function, so that customized versions may be conveniently defined
for specific subclasses of GeneralData, GeneralModel, and McmcOptions input.
}
\details{
Reproducible samples can be obtained by setting the seed via
\code{\link{set.seed}} before in the user code as usual. However, note that
because the RNG sampler used is external to R, running this MCMC function
will not change the seed position -- that is, the repeated call to this
function will then result in exactly the same output.
}
\section{Methods (by class)}{
\itemize{
\item \code{data = GeneralData,model = GeneralModel,options = McmcOptions}: Standard method which uses JAGS/BUGS

\item \code{data = Data,model = LogisticNormal,options = McmcOptions}: The fast method for the LogisticNormal class
}}
\keyword{methods}
\keyword{regression}

