% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McmcOptions-class.R
\docType{class}
\name{McmcOptions-class}
\alias{McmcOptions-class}
\alias{.McmcOptions}
\title{Class for the three canonical MCMC options}
\description{
Class for the three canonical MCMC options
}
\section{Slots}{

\describe{
\item{\code{iterations}}{number of MCMC iterations}

\item{\code{burnin}}{number of burn-in iterations which are not saved}

\item{\code{step}}{only every step-th iteration is saved after the burn-in}
}}

\examples{

# Set up MCMC option in order to have a burn-in of 10000 iterations and
# then take every other iteration up to a collection of 10000 samples
options <- McmcOptions(burnin=10000,
                       step=2,
                       samples=10000)

}
\keyword{classes}
