% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rules-methods.R
\docType{methods}
\name{minSize}
\alias{minSize}
\alias{minSize,CohortSize-method}
\title{"MIN" combination of cohort size rules}
\usage{
minSize(...)

\S4method{minSize}{CohortSize}(...)
}
\arguments{
\item{\dots}{Objects of class \code{\linkS4class{CohortSize}}}
}
\value{
the combination as an object of class
\code{\linkS4class{CohortSizeMin}}
}
\description{
This function combines cohort size rules by taking
the minimum of all sizes.
}
\section{Methods (by class)}{
\itemize{
\item \code{CohortSize}: The method combining cohort size rules by taking minimum
}}
\seealso{
\code{\link{maxSize}}
}
\keyword{methods}

