% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-class.R
\docType{class}
\name{ModelNormal-class}
\alias{ModelNormal-class}
\alias{.ModelNormal}
\alias{ModelNormal}
\title{\code{ModelNormal}}
\usage{
ModelNormal(mean, cov, ref_dose = 0)
}
\arguments{
\item{mean}{(\code{numeric})\cr the prior mean vector.}

\item{cov}{(\code{matrix})\cr the prior covariance matrix.}

\item{ref_dose}{(\code{number})\cr the reference dose.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{\link{ModelNormal}} is the class for a model with a reference dose and bivariate
normal prior on the model parameters \code{alpha0} and \code{alpha1} (or its functions),
i.e.: \deqn{(f1(alpha0, f2(alpha1)) ~ Normal(mean, cov),} where \code{f1}, \code{f2}
are some functions as specified in \code{priormodel} slot. By default, \code{f1} is
just identity and \code{f2} is natural logarithm, which means that \code{alpha1} has
log-normal distribution in such case. The parameter \code{alpha1} has a log-normal
distribution by default to ensure positivity of \code{alpha1} which ensures
\code{exp(alpha1) > 1}.
The slots of this class contain the mean vector, the covariance and
precision matrices of the bivariate normal distribution, as well as the
reference dose. Note that the precision matrix is an inverse of the
covariance matrix in the \code{JAGS}.
All ("normal") model specific classes inherit from this class.
}
\section{Slots}{

\describe{
\item{\code{mean}}{(\code{numeric})\cr the prior mean vector.}

\item{\code{cov}}{(\code{matrix})\cr the prior covariance matrix.}

\item{\code{prec}}{the prior precision matrix, which is an inverse matrix of the \code{cov}.}

\item{\code{ref_dose}}{(\code{number})\cr the reference dose.}
}}

\seealso{
\code{\link{ModelNormal}}, \code{\link{LogisticNormal}}, \code{\link{LogisticLogNormal}},
\code{\link{LogisticLogNormalSub}}, \code{\link{ProbitLogNormal}}, \code{\link{ProbitLogNormalLogDose}}.
}
