% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-class.R
\docType{class}
\name{ModelPseudo-class}
\alias{ModelPseudo-class}
\alias{.ModelPseudo}
\alias{ModelPseudo}
\title{\code{ModelPseudo}}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{\link{ModelPseudo}} is the parent class for models that express their prior in
the form of pseudo data (as if there is some data before the trial starts).
}
\seealso{
\code{\link{GeneralModel}}.
}
