% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Model-class.R
\docType{class}
\name{ModelPseudo-class}
\alias{.ModelPseudo}
\alias{ModelPseudo-class}
\title{Class of models using prior in form of Pseudo data}
\description{
Class of models using prior in form of Pseudo data
}
\keyword{classes}

