% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-class.R
\docType{class}
\name{ModelPseudo-class}
\alias{ModelPseudo-class}
\alias{.ModelPseudo}
\title{Class of models using expressing their prior in form of Pseudo data}
\description{
This is the Pseudo model class, from which all models where their prior 
are expressed in form of pseudo data (as if some data are 
available before the trial starts) inherit. It also inherits all slots
from \code{\linkS4class{AllModels}}.No slots for this class
}
\seealso{
\code{\linkS4class{LogisticIndepBeta}},
\code{\linkS4class{Effloglog}},
\code{\linkS4class{EffFlexi}}
}
\keyword{classes}
