% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{myBayesLogit}
\alias{myBayesLogit}
\title{Do MCMC sampling for Bayesian logistic regression model}
\usage{
myBayesLogit(y, X, m0, P0, options)
}
\arguments{
\item{y}{0/1 vector of responses}

\item{X}{design matrix}

\item{m0}{prior mean vector}

\item{P0}{precision matrix}

\item{options}{McmcOptions object}
}
\value{
the matrix of samples (samples x parameters)
}
\description{
Do MCMC sampling for Bayesian logistic regression model
}
\keyword{internal}
