% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{NextBest-class}
\alias{NextBest-class}
\title{The virtual class for finding next best dose}
\description{
The virtual class for finding next best dose
}
\seealso{
\code{\linkS4class{NextBestMTD}},
\code{\linkS4class{NextBestNCRM}},
\code{\linkS4class{NextBestDualEndpoint}},
\code{\linkS4class{NextBestThreePlusThree}}
}
\keyword{classes}
