% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rules-methods.R
\docType{methods}
\name{nextBest}
\alias{nextBest}
\alias{nextBest,NextBestDualEndpoint,numeric,Samples,DualEndpoint,Data-method}
\alias{nextBest,NextBestMTD,numeric,Samples,Model,Data-method}
\alias{nextBest,NextBestNCRM,numeric,Samples,Model,Data-method}
\alias{nextBest,NextBestNCRM,numeric,Samples,Model,DataParts-method}
\alias{nextBest,NextBestThreePlusThree,missing,missing,missing,Data-method}
\title{Find the next best dose}
\usage{
nextBest(nextBest, doselimit, samples, model, data, ...)

\S4method{nextBest}{NextBestMTD,numeric,Samples,Model,Data}(nextBest, doselimit,
  samples, model, data, ...)

\S4method{nextBest}{NextBestNCRM,numeric,Samples,Model,Data}(nextBest,
  doselimit, samples, model, data, ...)

\S4method{nextBest}{NextBestNCRM,numeric,Samples,Model,DataParts}(nextBest,
  doselimit, samples, model, data, ...)


  \S4method{nextBest}{NextBestThreePlusThree,missing,missing,missing,Data}(nextBest,
  doselimit, samples, model, data, ...)


  \S4method{nextBest}{NextBestDualEndpoint,numeric,Samples,DualEndpoint,Data}(nextBest,
  doselimit, samples, model, data, ...)
}
\arguments{
\item{nextBest}{The rule, an object of class \code{\linkS4class{NextBest}}}

\item{doselimit}{The maximum allowed next dose. If this is an empty (length
0) vector, then no dose limit will be applied in the course of dose
recommendation calculation, and a corresponding warning is given.}

\item{samples}{the \code{\linkS4class{Samples}} object}

\item{model}{The model input, an object of class \code{\linkS4class{Model}}}

\item{data}{The data input, an object of class \code{\linkS4class{Data}}}

\item{\dots}{possible additional arguments without method dispatch}
}
\value{
a list with the next best dose (element \code{value})
on the grid defined in \code{data}, and a plot depicting this recommendation
(element \code{plot})
}
\description{
Compute the recommended next best dose.
}
\details{
This function outputs the next best dose recommendation based on the
corresponding rule \code{nextBest}, the posterior \code{samples} from the
\code{model} and the underlying \code{data}.
}
\section{Methods (by class)}{
\itemize{
\item \code{nextBest = NextBestMTD,doselimit = numeric,samples = Samples,model = Model,data = Data}: Find the next best dose based on the MTD rule

\item \code{nextBest = NextBestNCRM,doselimit = numeric,samples = Samples,model = Model,data = Data}: Find the next best dose based on the NCRM method

\item \code{nextBest = NextBestNCRM,doselimit = numeric,samples = Samples,model = Model,data = DataParts}: Find the next best dose based on the NCRM method when
two parts trial is used

\item \code{nextBest = NextBestThreePlusThree,doselimit = missing,samples = missing,model = missing,data = Data}: Find the next best dose based on the 3+3 method

\item \code{nextBest = NextBestDualEndpoint,doselimit = numeric,samples = Samples,model = DualEndpoint,data = Data}: Find the next best dose based on the dual endpoint
model
}}
\keyword{methods}

