% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{NextBestDualEndpoint-class}
\alias{.NextBestDualEndpoint}
\alias{NextBestDualEndpoint-class}
\title{The class with the input for finding the next dose
based on the dual endpoint model}
\description{
This rule first excludes all doses that exceed the probability
\code{maxOverdoseProb} of having an overdose toxicity, as specified by the
overdose interval \code{overdose}. Then, it picks under the remaining
admissible doses the one that maximizes the probability to have at least
\code{target} biomarker level, relative to the maximum biomarker level
across the dose grid.
}
\section{Slots}{

\describe{
\item{\code{target}}{the biomarker level, relative to the maximum, that
needs to be reached. For example, 0.9 means that a dose with 90%
of the maximum biomarker level is considered as having reached
sufficient biomarker level.}

\item{\code{overdose}}{the overdose toxicity interval}

\item{\code{maxOverdoseProb}}{maximum overdose probability that is allowed}
}}
\keyword{classes}

