% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{NextBestInfTheory-class}
\alias{NextBestInfTheory-class}
\alias{.NextBestInfTheory}
\alias{NextBestInfTheory}
\title{\code{NextBestInfTheory}}
\usage{
NextBestInfTheory(target, asymmetry)
}
\arguments{
\item{target}{(\code{proportion})\cr see slot definition.}

\item{asymmetry}{(\code{number})\cr see slot definition.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{\link{NextBestInfTheory}} is the class for next best dose that is based on
information theory as proposed in https://doi.org/10.1002/sim.8450.
}
\section{Slots}{

\describe{
\item{\code{target}}{(\code{proportion})\cr target toxicity probability.}

\item{\code{asymmetry}}{(\code{number})\cr value of the asymmetry exponent in the
divergence function that describes the rate of penalization for overly
toxic does. It must be a value from (0, 2) interval.}
}}

