% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{NextBestMaxGain-class}
\alias{.NextBestMaxGain}
\alias{NextBestMaxGain-class}
\title{Class for next best with maximum gain based on DLE and Efficacy responses}
\description{
Class for next best with maximum gain based on DLE and Efficacy responses
}
\keyword{class}

