% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\name{NextBestMaxGain}
\alias{NextBestMaxGain}
\title{Initialization function for the class 'NextBestMaxGain'}
\usage{
NextBestMaxGain(DLEDuringTrialtarget, DLEEndOfTrialtarget)
}
\arguments{
\item{DLEDuringTrialtarget}{please refer to \code{\linkS4class{NextBestMaxGain}} class object}

\item{DLEEndOfTrialtarget}{please refer to \code{\linkS4class{NextBestMaxGain}} class object}
}
\value{
the \code{\linkS4class{NextBestMaxGain}} class object
}
\description{
Initialization function for the class 'NextBestMaxGain'
}
\keyword{methods}
