% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{NextBestMinDist-class}
\alias{NextBestMinDist-class}
\alias{.NextBestMinDist}
\alias{NextBestMinDist}
\alias{.DefaultNextBestMinDist}
\title{\code{NextBestMinDist}}
\usage{
NextBestMinDist(target)

.DefaultNextBestMinDist()
}
\arguments{
\item{target}{(\code{proportion})\cr see slot definition.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{\link{NextBestMinDist}} is the class for next best dose that is based on minimum
distance to target probability.

Typically, end users will not use the \code{.DefaultNextBestMinDist()} function.
}
\section{Slots}{

\describe{
\item{\code{target}}{(\code{proportion})\cr single target toxicity probability, except
0 or 1.}
}}

\examples{
# In the example below, the MTD is defined as the dose with the toxicity rate
# with minimal distance to the target of 30\%.
next_best_min_dist <- NextBestMinDist(target = 0.3)
.DefaultNextBestMinDist()
}
