% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{NextBestNCRMLoss-class}
\alias{NextBestNCRMLoss-class}
\alias{.NextBestNCRMLoss}
\title{Next best dose based on NCRM rule and loss function}
\description{
This class is similar to NCRM class, the only difference is the addition of
loss function. As in NCRM rule, first admissible doses are found, which are those
with probability to fall in overdose category being below
\code{maxOverdoseProb}. Next, within the admissible doses, the loss function is
calculated, i.e. \code{losses} \%*\% \code{target}. Finally, the corresponding
dose with lowest loss function (Bayes risk) is recommended for the next dose.
}
\section{Slots}{

\describe{
\item{\code{target}}{A numeric vector specifying the target toxicity interval
(limits included)}

\item{\code{overdose}}{A numeric vector specifying the overdose toxicity interval
(lower limit excluded, upper limit included).}

\item{\code{unacceptable}}{An optional numeric vector specifying the unacceptable toxicity interval
(lower limit excluded, upper limit included).}

\item{\code{maxOverdoseProb}}{A numeric value specifying the maximum overdose
(overdose or excessive + unacceptable) probability that is allowed.}

\item{\code{losses}}{A numeric vector specifying the loss function. If the \code{unacceptable}
is provided, the vector length must be 4, otherwise is 3.}
}}

\note{
The loss function should be a vector of either 3 or 4 values.
This is because the loss function values must be specified for each interval,
that is under-dosing, target toxicity, and overdosing toxicity or
under-dosing, target toxicity, overdosing (excessive) toxicity, and unacceptable
toxicity intervals.
}
