% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\name{NextBestNCRMLoss}
\alias{NextBestNCRMLoss}
\title{Initialization function for "NextBestNCRMLoss"}
\usage{
NextBestNCRMLoss(
  target,
  overdose,
  unacceptable = c(1, 1),
  max_overdose_prob,
  losses
)
}
\arguments{
\item{target}{see \linkS4class{NextBestNCRMLoss}}

\item{overdose}{see \linkS4class{NextBestNCRMLoss}}

\item{unacceptable}{see \linkS4class{NextBestNCRMLoss}}

\item{losses}{see \linkS4class{NextBestNCRMLoss}}

\item{maxOverdoseProb}{see \linkS4class{NextBestNCRMLoss}}
}
\value{
the \linkS4class{NextBestNCRMLoss} object
}
\description{
Initialization function for "NextBestNCRMLoss"
}
