% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{NextBestTD-class}
\alias{NextBestTD-class}
\alias{.NextBestTD}
\title{Next best dose based on Pseudo DLE model without sample}
\description{
The class is to find the next best dose for allocation and the dose for final recommendation 
at the end of a trial without involving any samples. This is a class for which only
 DLE response will be incorporated for the dose-allocation.
This is only based on the probabilities of
the occurrence of a DLE obtained by using the modal estimates of the model paramters.
There are two inputs inputs which are the two target 
probabilities of the occurrence of a DLE used during trial
and used at the end of trial, for finding the next best dose for allocation and the dose 
for recommendation at the end of the trial.
It is only suitable to use with the model specified in \code{ModelTox} class.
}
\section{Slots}{

\describe{
\item{\code{targetDuringTrial}}{the target probability of the occurrrence of a DLE to be used
during the trial}

\item{\code{targetEndOfTrial}}{the target probability of the occurrence of a DLE to be used at the end 
of the trial. This target is particularly used to recommend the dose for which its posterior 
probability of the occurrence of a DLE is equal to this target}
}}

\examples{
##define the NextBestTD class (no samples and based a pseudo DLE model)
##specified the target probability of the occurrence of a DLE during the trial be 0.35
##specified the target probability of the occurrence of a DLE at the end of trial be 0.3

myNextBest <-NextBestTD(targetDuringTrial=0.35,
                        targetEndOfTrial=0.3)
}
\keyword{class}
