% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\name{NextBestTDsamples}
\alias{NextBestTDsamples}
\title{Initialization function for class "NextBestTDsamples"}
\usage{
NextBestTDsamples(targetDuringTrial, targetEndOfTrial, derive)
}
\arguments{
\item{targetDuringTrial}{please refer to \code{\linkS4class{NextBestTDsamples}} class object}

\item{targetEndOfTrial}{please refer to \code{\linkS4class{NextBestTDsamples}} class object}

\item{derive}{please refer to \code{\linkS4class{NextBestTDsamples}} class object}
}
\value{
the \code{\linkS4class{NextBestTDsamples}} class object
}
\description{
Initialization function for class "NextBestTDsamples"
}
\keyword{methods}
