% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{NextBestThreePlusThree-class}
\alias{NextBestThreePlusThree-class}
\alias{.NextBestThreePlusThree}
\alias{NextBestThreePlusThree}
\title{\code{NextBestThreePlusThree}}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{\link{NextBestThreePlusThree}} is the class for next best dose that
implements the classical 3+3 dose recommendation. No input is required,
hence this class has no slots.
}
\examples{
# Next best dose class object using the classical 3+3 design.
my_next_best <- NextBestThreePlusThree()
}
