% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{NextBestThreePlusThree-class}
\alias{NextBestThreePlusThree-class}
\alias{.NextBestThreePlusThree}
\title{The class with the input for finding the next dose in target interval}
\description{
Implements the classical 3+3 dose recommendation.
No input is required, hence this class has no slots.
}
\examples{

## Choose the next best dose using the  classical 3+3 design.
## No input is required, hence this function has no parameters.
myNextBest <- NextBestThreePlusThree()


}
\keyword{classes}
