% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Backfill-methods.R
\name{openCohort}
\alias{openCohort}
\title{Open / recruit backfill patients into a cohort?}
\usage{
openCohort(opening, cohort, data, dose, ...)
}
\arguments{
\item{opening}{(\code{Opening})\cr opening rule to be applied.}

\item{cohort}{(\code{int})\cr backfill cohort index.}

\item{data}{(\code{Data})\cr current trial data.}

\item{dose}{(\code{numeric})\cr the recommended next best dose.}

\item{...}{further arguments (not used).}
}
\value{
\code{TRUE} if this backfill cohort can be opened / recruited into,
\code{FALSE} otherwise.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
