\docType{methods}
\name{or-Stopping-Stopping}
\alias{or-Stopping-Stopping}
\title{The method combining two atomic stopping rules}
\usage{
\S4method{|}{Stopping,Stopping}(e1, e2)
}
\arguments{
  \item{e1}{First \code{\linkS4class{Stopping}} object}

  \item{e2}{Second \code{\linkS4class{Stopping}} object}
}
\value{
The \code{\linkS4class{StoppingAny}} object
}
\description{
The method combining two atomic stopping rules
}
\keyword{methods}

