\docType{methods}
\name{or-Stopping-StoppingAny}
\alias{or-Stopping-StoppingAny}
\title{The method combining a stopping list and an atomic}
\usage{
\S4method{|}{StoppingAny,Stopping}(e1, e2)
}
\arguments{
  \item{e1}{\code{\linkS4class{StoppingAny}} object}

  \item{e2}{\code{\linkS4class{Stopping}} object}
}
\value{
The modified \code{\linkS4class{StoppingAny}} object
}
\description{
The method combining a stopping list and an atomic
}
\keyword{methods}

