% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rules-methods.R
\docType{methods}
\name{or-StoppingAny-Stopping}
\alias{or-StoppingAny-Stopping}
\alias{|,Stopping,StoppingAny-method}
\title{The method combining an atomic and a stopping list}
\usage{
\S4method{|}{Stopping,StoppingAny}(e1, e2)
}
\arguments{
\item{e1}{\code{\linkS4class{Stopping}} object}

\item{e2}{\code{\linkS4class{StoppingAny}} object}
}
\value{
The modified \code{\linkS4class{StoppingAny}} object
}
\description{
The method combining an atomic and a stopping list
}
\keyword{methods}

