% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Data-methods.R
\docType{methods}
\name{plot,Data,missing-method}
\alias{plot,Data,missing-method}
\title{Plot method for the "Data" class}
\usage{
\S4method{plot}{Data,missing}(x, y, ...)
}
\arguments{
\item{x}{the \code{\linkS4class{Data}} object we want to plot}

\item{y}{missing}

\item{\dots}{not used}
}
\value{
the \code{\link[ggplot2]{ggplot}} object
}
\description{
Plot method for the "Data" class
}
\keyword{methods}

