% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Data-methods.R
\docType{methods}
\name{plot,DataCombo,missing-method}
\alias{plot,DataCombo,missing-method}
\title{Plot method for the "DataCombo" class}
\usage{
\S4method{plot}{DataCombo,missing}(x, y, select = head(x@drugNames, 2L),
  shorten = 0.03, ...)
}
\arguments{
\item{x}{the \code{\linkS4class{DataDual}} object we want to plot}

\item{y}{missing}

\item{select}{two drug names that we want to use for plotting. Defaults
to the first two drug names.}

\item{shorten}{relative amount of shortening the arrows. default: 0.05}

\item{\dots}{not used}
}
\value{
the \code{\link[ggplot2]{ggplot}} object
}
\description{
Plot method for the "DataCombo" class
}
\keyword{methods}

