% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-methods.R
\name{plot,DataDA,missing-method}
\alias{plot,DataDA,missing-method}
\title{Plot method for the "DataDA" class
todo: insert example code in below file (empty still)}
\usage{
\S4method{plot}{DataDA,missing}(x, y, blind = FALSE, ...)
}
\arguments{
\item{x}{the \code{\linkS4class{DataDA}} object we want to plot}

\item{y}{missing}

\item{blind}{Logical (default FALSE) if to blind the data}

\item{\dots}{not used}
}
\value{
the \code{\link[ggplot2]{ggplot}} object
}
\description{
Plot method for the "DataDA" class
todo: insert example code in below file (empty still)
}
\examples{

}
\keyword{methods}
