% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Data-methods.R
\docType{methods}
\name{plot,DataDual,ModelEff-method}
\alias{plot,DataDual,ModelEff-method}
\title{Plot the fitted dose-efficacy curve given a pseudo Efficacy model and data
without samples}
\usage{
\S4method{plot}{DataDual,ModelEff}(x, y, ..., xlab = "Dose level",
  ylab = "Expected Efficacy", showLegend = TRUE)
}
\arguments{
\item{x}{the data input from the \code{\linkS4class{DataDual}} class object}

\item{y}{the model from the \code{\linkS4class{ModelEff}} class object}
}
\description{
Plot the fitted dose-efficacy curve given a pseudo Efficacy model and data
without samples
}

