% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Simulations-methods.R
\docType{methods}
\name{plot,DualSimulations,missing-method}
\alias{plot,DualSimulations,missing-method}
\title{Plot dual-endpoint simulations}
\usage{
\S4method{plot}{DualSimulations,missing}(x, y, type = c("trajectory",
  "dosesTried", "sigma2W", "rho"), ...)
}
\arguments{
\item{x}{the \code{\linkS4class{DualSimulations}} object we want
to plot from}

\item{y}{missing}

\item{type}{the type of plots you want to obtain.}

\item{\dots}{not used}
}
\value{
A single \code{\link[ggplot2]{ggplot2}} object if a single plot is
asked for, otherwise a \code{\link{gridExtra}{gTree}} object.
}
\description{
This plot method can be applied to \code{\linkS4class{DualSimulations}}
objects in order to summarize them graphically. In addition to the standard
plot types, there is
\describe{
\item{sigma2W}{Plot a boxplot of the final biomarker variance estimates in
the simulated trials}
\item{rho}{Plot a boxplot of the final correlation estimates in
the simulated trials}
}
}
\keyword{methods}

