% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Simulations-methods.R
\docType{methods}
\name{plot,GeneralSimulations,missing-method}
\alias{plot,GeneralSimulations,missing-method}
\title{Plot simulations}
\usage{
\S4method{plot}{GeneralSimulations,missing}(x, y, type = c("trajectory",
  "dosesTried"), ...)
}
\arguments{
\item{x}{the \code{\linkS4class{GeneralSimulations}} object we want
to plot from}

\item{y}{missing}

\item{type}{the type of plots you want to obtain.}

\item{\dots}{not used}
}
\value{
A single \code{\link[ggplot2]{ggplot2}} object if a single plot is
asked for, otherwise a \code{\link{gridExtra}{gTree}} object.
}
\description{
Summarize the simulations with plots
}
\details{
This plot method can be applied to \code{\linkS4class{GeneralSimulations}}
objects in order to summarize them graphically. Possible \code{type}s of
plots at the moment are: \describe{ \item{trajectory}{Summary of the
trajectory of the simulated trials} \item{dosesTried}{Average proportions of
the doses tested in patients} } You can specify one or both of these in the
\code{type} argument.
}
\examples{
##obtain the plot for the simulation results
##If only DLE responses are considered in the simulations
##Specified your simulations when no DLE samples are used
##(Please refer to desgin-method 'simulate TDDesign' examples for details)
mySim <- simulate(design,
                  args=NULL,
                  truth=myTruth,
                  nsim=10,
                  seed=819,
                  parallel=FALSE)
##If DLE samples are involved
##Please refer to design-method 'simulate TDsamplesDesign' examples for details
mySim <- simulate(design,
                  args=NULL,
                  truth=myTruth,
                  nsim=10,
                  seed=819,
                  mcmcOptions=options,
                  parallel=FALSE)
##Then produce plot of you simulted results
print(plot(mySim))
}
\keyword{methods}

