% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations-methods.R
\name{plot,GeneralSimulationsSummary,missing-method}
\alias{plot,GeneralSimulationsSummary,missing-method}
\alias{plot-GeneralSimulationsSummary-missing}
\title{Plot \code{GeneralSimulationsSummary}}
\usage{
\S4method{plot}{GeneralSimulationsSummary,missing}(
  x,
  y,
  type = c("nObs", "doseSelected", "propDLTs", "nAboveTarget"),
  axis_text_angle = 45,
  ...
)
}
\arguments{
\item{x}{(\code{GeneralSimulationsSummary})\cr the object we want to plot from.}

\item{y}{(\code{missing})\cr not used.}

\item{type}{(\code{character})\cr the types of plots you want to obtain.}

\item{axis_text_angle}{(\code{number})\cr rotation angle for the MTD estimate
x-axis tick labels. Defaults to 45 degrees.}

\item{...}{not used.}
}
\value{
A single \code{ggplot} object if a single plot is
asked for, otherwise a \code{gtable} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Graphical display of the general simulation summary.

This plot method can be applied to \code{\link{GeneralSimulationsSummary}} objects in
order to summarize them graphically. Possible \code{type}s of plots at the moment
are:
\describe{
\item{nObs}{Distribution of the number of patients in the simulated trials}
\item{doseSelected}{Distribution of the final selected doses in the trials.
Note that this can include zero entries, meaning that the trial was
stopped because all doses in the dose grid appeared too toxic.}
\item{propDLTs}{Distribution of the proportion of patients with DLTs in the
trials}
\item{nAboveTarget}{Distribution of the number of patients treated at doses
which are above the target toxicity interval (as specified by the
\code{truth} and \code{target} arguments to \code{\link{summary,GeneralSimulations-method}})}
}
You can specify any subset of these in the \code{type} argument.
}
