% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations-methods.R
\docType{methods}
\name{plot,GeneralSimulationsSummary,missing-method}
\alias{plot,GeneralSimulationsSummary,missing-method}
\title{Graphical display of the general simulation summary}
\usage{
\S4method{plot}{GeneralSimulationsSummary,missing}(x, y, type = c("nObs",
  "doseSelected", "propDLTs", "nAboveTarget"), ...)
}
\arguments{
\item{x}{the \code{\linkS4class{GeneralSimulationsSummary}} object we want
to plot from}

\item{y}{missing}

\item{type}{the types of plots you want to obtain.}

\item{\dots}{not used}
}
\value{
A single \code{\link[ggplot2]{ggplot}} object if a single plot is
asked for, otherwise a \code{\link{gridExtra}{gTree}} object.
}
\description{
This plot method can be applied to
\code{\linkS4class{GeneralSimulationsSummary}} objects in order to
summarize them graphically. Possible \code{type}s of plots at the moment
are:
}
\details{
\describe{
\item{nObs}{Distribution of the number of patients in the simulated trials}
\item{doseSelected}{Distribution of the final selected doses in the trials.
Note that this can include zero entries, meaning that the trial was stopped
because all doses in the dose grid appeared too toxic.}
\item{propDLTs}{Distribution of the proportion of patients with DLTs in the
trials}
\item{nAboveTarget}{Distribution of the number of patients treated at doses
which are above the target toxicity interval (as specified by the
\code{truth} and \code{target} arguments to
\code{\link{summary,GeneralSimulations-method}})}
}
You can specify any subset of these in the \code{type} argument.
}
\keyword{methods}
