% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Simulations-methods.R
\docType{methods}
\name{plot,PseudoDualFlexiSimulations,missing-method}
\alias{plot,PseudoDualFlexiSimulations,missing-method}
\title{This plot method can be applied to \code{\linkS4class{PseudoDualFlexiSimulations}}
objects in order to summarize them graphically. Possible \code{type}s of
plots at the moment are: \describe{ \item{trajectory}{Summary of the
trajectory of the simulated trials} \item{dosesTried}{Average proportions of
the doses tested in patients} \item{sigma2}{The variance of the efficacy responses}
\item{sigma2betaW}{The variance of the random walk model}}
You can specify one or both of these in the
\code{type} argument.}
\usage{
\S4method{plot}{PseudoDualFlexiSimulations,missing}(x, y,
  type = c("trajectory", "dosesTried", "sigma2", "sigma2betaW"), ...)
}
\arguments{
\item{x}{the \code{\linkS4class{PseudoDualFlexiSimulations}} object we want
to plot from}

\item{y}{missing}

\item{type}{the type of plots you want to obtain.}

\item{\dots}{not used}
}
\value{
A single \code{\link[ggplot2]{ggplot2}} object if a single plot is
asked for, otherwise a \code{\link{gridExtra}{gTree}} object.
}
\description{
This plot method can be applied to \code{\linkS4class{PseudoDualFlexiSimulations}}
objects in order to summarize them graphically. Possible \code{type}s of
plots at the moment are: \describe{ \item{trajectory}{Summary of the
trajectory of the simulated trials} \item{dosesTried}{Average proportions of
the doses tested in patients} \item{sigma2}{The variance of the efficacy responses}
\item{sigma2betaW}{The variance of the random walk model}}
You can specify one or both of these in the
\code{type} argument.
}
\examples{
##obtain the plot for the simulation results
##If DLE and efficacy responses are considered in the simulations
##The efficacy model must be of 'EffFlexi' class
Effmodel<- EffFlexi(Eff=c(1.223, 2.513),Effdose=c(25,300),
                    sigma2=c(a=0.1,b=0.1),sigma2betaW=c(a=20,b=50),smooth="RW2",data=data)
## todo: how does the Effmodel come into the below code?
##Specify your simulations. Please refers to the example in design-method 'simulate' 
## DualResponsesSamplesFlexiDesign for details
mySim<-simulate(object=design,
                args=NULL,
                trueDLE=myTruthDLE,
                trueEff=myTruthEff,
                trueSigma2=0.025,
                trueSigma2betaW=1,
                nsim=1,
                seed=819,
                parallel=FALSE)
##plot this simulated results
print(plot(mySim))
}
\keyword{methods}

