% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Simulations-methods.R
\docType{methods}
\name{plot,PseudoDualSimulationsSummary,missing-method}
\alias{plot,PseudoDualSimulationsSummary,missing-method}
\title{Plot the summary of Pseudo Dual Simulations summary}
\usage{
\S4method{plot}{PseudoDualSimulationsSummary,missing}(x, y, type = c("nObs",
  "doseSelected", "propDLE", "nAboveTargetEndOfTrial", "meanFit", "meanEffFit"),
  ...)
}
\arguments{
\item{x}{the \code{\linkS4class{PseudoDualSimulationsSummary}} object we want to plot from}

\item{y}{missing}

\item{type}{the types of plots you want to obtain.}

\item{\dots}{not used}
}
\value{
A single \code{\link[ggplot2]{ggplot2}} object if a single plot is
asked for, otherwise a \code{\link{gridExtra}{gTree}} object.
}
\description{
This plot method can be applied to \code{\linkS4class{PseudoDualSimulationsSummary}} objects in order
to summarize them graphically. Possible \code{type} of plots at the moment are those listed in
\code{\link{plot,PseudoSimulationsSummary,missing-method}} plus:
\describe{\item{meanEffFit}{Plot showing the fitted dose-efficacy curve. If no samples are involved, only the
average fitted dose-efficacy curve across the trials will be ploted. If samples (DLE and efficacy) are involved,
the average fitted dose-efficacy curve across the trials, together with the 95\% credibility interval; and comparison
with the assumed truth (as specified by the \code{trueEff} argument to
\code{\link{summary,PseudoDualSimulations-method}})}}
You can specify any subset of these in the \code{type} argument.
}
\examples{
##obtain the plot of the summary for the simulation results
##If DLE and efficacy responses are considered in the simulations
##Specified your simulations when no samples are used
##(Please refer to desgin-method 'simulate DualResponsesDesign' examples for details)
mySim <-simulate(object=design,
                 args=NULL,
                 trueDLE=myTruthDLE,
                 trueEff=myTruthEff,
                 trueNu=1/0.025,
                 nsim=10,
                 seed=819,
                 parallel=FALSE)
##If DLE and efficacy samples are involved
##Please refer to design-method 'simulate DualResponsesSamplesDesign' examples for details
mySim<-simulate(design,
                args=NULL,
                trueDLE=myTruthDLE,
                trueEff=myTruthEff,
                trueNu=1/0.025,
                nsim=10,
                mcmcOptions=options,
                seed=819,
                parallel=FALSE)
##OR if the 'EffFlexi' class is used 
## for the efficacy model
Effmodel<- EffFlexi(Eff=c(1.223, 2.513),Effdose=c(25,300),
                    sigma2=c(a=0.1,b=0.1),sigma2betaW=c(a=20,b=50),smooth="RW2",data=data)
##Specify the simulations
##Please refer to design-method 'simulate DualResponsesSamplesFlexiDesign' examples for details
mySim<-simulate(object=design,
                args=NULL,
                trueDLE=myTruthDLE,
                trueEff=myTruthEff,
                trueSigma2=0.025,
                trueSigma2betaW=1,
                nsim=1,
                seed=819,
                parallel=FALSE)
##Then produce a summary of your simulations
MYSUM <- summary(mySim,
                 trueDLE=myTruthDLE,
                  trueEff=myTruthEff)

##Then plot the summary of the simulations
print(plot(MYSUM))
}
\keyword{methods}

