% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Simulations-methods.R
\docType{methods}
\name{plot,PseudoSimulationsSummary,missing-method}
\alias{plot,PseudoSimulationsSummary,missing-method}
\title{Plot summaries of the pseudo simulations}
\usage{
\S4method{plot}{PseudoSimulationsSummary,missing}(x, y, type = c("nObs",
  "doseSelected", "propDLE", "nAboveTargetEndOfTrial", "meanFit"), ...)
}
\arguments{
\item{x}{the \code{\linkS4class{PseudoSimulationsSummary}} object we want
to plot from}

\item{y}{missing}

\item{type}{the types of plots you want to obtain.}

\item{\dots}{not used}
}
\value{
A single \code{\link[ggplot2]{ggplot2}} object if a single plot is
asked for, otherwise a \code{\link{gridExtra}{gTree}} object.
}
\description{
Graphical display of the simulation summary
}
\details{
This plot method can be applied to \code{\linkS4class{PseudoSimulationsSummary}}
objects in order to summarize them graphically. This can be used when only DLE responses are involved
in the simulations. This also applied to results with or without samples generated during the simulations
}
\examples{
##obtain the plot for the simulation results
##If only DLE responses are considered in the simulations
##Specified your simulations when no DLE samples are used
##(Please refer to desgin-method 'simulate TDDesign' examples for details)
mySim <- simulate(design,
                  args=NULL,
                  truth=myTruth,
                  nsim=10,
                  seed=819,
                  parallel=FALSE)
##If DLE samples are involved
##Please refer to design-method 'simulate TDsamplesDesign' examples for details
mySim <- simulate(design,
                  args=NULL,
                  truth=myTruth,
                  nsim=10,
                  seed=819,
                  mcmcOptions=options,
                  parallel=FALSE)
##Then produce a summary of your simulations
MYSUM <- summary(mySim,
                 truth=myTruth)
##plot the summary of the simulations
print(plot(MYSUM))
}
\keyword{methods}

