% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Samples-methods.R
\docType{methods}
\name{plot,Samples,ModelEff-method}
\alias{plot,Samples,ModelEff-method}
\title{Plot of the Efficacy pseudo Model with samples}
\usage{
\S4method{plot}{Samples,ModelEff}(x, y, data, ..., xlab = "Dose level",
  ylab = "Expected Efficacy", showLegend = TRUE)
}
\arguments{
\item{x}{bla}

\item{y}{bli todo}
}
\description{
Plot of the Efficacy pseudo Model with samples
}
\keyword{methods}

