% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Simulations-methods.R
\docType{methods}
\name{plot,SimulationsSummary,missing-method}
\alias{plot,SimulationsSummary,missing-method}
\title{Plot summaries of the model-based design simulations}
\usage{
\S4method{plot}{SimulationsSummary,missing}(x, y, type = c("nObs",
  "doseSelected", "propDLTs", "nAboveTarget", "meanFit"), ...)
}
\arguments{
\item{x}{the \code{\linkS4class{SimulationsSummary}} object we want
to plot from}

\item{y}{missing}

\item{type}{the types of plots you want to obtain.}

\item{\dots}{not used}
}
\value{
A single \code{\link[ggplot2]{ggplot2}} object if a single plot is
asked for, otherwise a \code{\link{gridExtra}{gTree}} object.
}
\description{
Graphical display of the simulation summary
}
\details{
This plot method can be applied to \code{\linkS4class{SimulationsSummary}}
objects in order to summarize them graphically. Possible \code{type} of
plots at the moment are those listed in
\code{\link{plot,GeneralSimulationsSummary,missing-method}} plus:
\describe{
\item{meanFit}{Plot showing the average fitted dose-toxicity curve across
the trials, together with 95\% credible intervals, and comparison with the
assumed truth (as specified by the \code{truth} argument to
\code{\link{summary,Simulations-method}})}
}
You can specify any subset of these in the \code{type} argument.
}
\keyword{methods}

