% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{plot.gtable}
\alias{plot.gtable}
\title{Plots gtable objects}
\usage{
\method{plot}{gtable}(x, ...)
}
\arguments{
\item{x}{the gtable object}

\item{\dots}{additional parameters for \code{\link[grid]{grid.draw}}}
}
\description{
Plots gtable objects
}
