% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Model-methods.R
\docType{methods}
\name{prob}
\alias{prob}
\alias{prob,matrix,ComboLogistic,Samples-method}
\alias{prob,numeric,Model,Samples-method}
\title{Compute the probability for a given dose, given model and samples}
\usage{
prob(dose, model, samples, ...)

\S4method{prob}{numeric,Model,Samples}(dose, model, samples, ...)

\S4method{prob}{matrix,ComboLogistic,Samples}(dose, model, samples, ...)
}
\arguments{
\item{dose}{the dose}

\item{model}{the \code{\linkS4class{Model}} or
\code{\linkS4class{ComboLogistic}} object}

\item{samples}{the \code{\linkS4class{Samples}}}

\item{\dots}{unused}
}
\value{
either the vector (for \code{\linkS4class{Model}} objects) or the
matrix (for \code{\linkS4class{ComboLogistic}} objects) of probability
samples.
}
\description{
Compute the probability for a given dose, given model and samples
}
\section{Methods (by class)}{
\itemize{
\item \code{dose = numeric,model = Model,samples = Samples}: 

\item \code{dose = matrix,model = ComboLogistic,samples = Samples}: 
}}
\keyword{methods}

