% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Model-methods.R
\docType{methods}
\name{prob}
\alias{prob}
\alias{prob,numeric,ModelOrComboLogistic,Samples-method}
\title{Compute the probability for a given dose, given model and samples}
\usage{
prob(dose, model, samples, ...)

\S4method{prob}{numeric,ModelOrComboLogistic,Samples}(dose, model, samples, ...)
}
\arguments{
\item{dose}{the dose}

\item{model}{the \code{\linkS4class{Model}} or
\code{\linkS4class{ComboLogistic}} object}

\item{samples}{the \code{\linkS4class{Samples}}}

\item{\dots}{unused}
}
\description{
Compute the probability for a given dose, given model and samples
}
\section{Methods (by class)}{
\itemize{
\item \code{dose = numeric,model = ModelOrComboLogistic,samples = Samples}: 
}}
\keyword{methods}

