\name{prob}
\alias{prob}
\title{Compute the probability for a given dose, given model and samples}
\usage{
prob(dose, model, samples, ...)
}
\arguments{
  \item{dose}{the dose}

  \item{model}{the \code{\linkS4class{Model}}}

  \item{samples}{the \code{\linkS4class{Samples}}}

  \item{\dots}{unused}
}
\description{
Compute the probability for a given dose, given model and
samples
}
\keyword{methods}

