% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-methods.R
\name{probFunction}
\alias{probFunction}
\alias{probFunction,GeneralModel-method}
\alias{probFunction-GeneralModel}
\alias{probFunction,ModelTox-method}
\alias{probFunction-ModelTox}
\title{Getting the Prob Function for a Given Model Type}
\usage{
probFunction(model, ...)

\S4method{probFunction}{GeneralModel}(model, ...)

\S4method{probFunction}{ModelTox}(model, ...)
}
\arguments{
\item{model}{(\code{GeneralModel} or \code{ModelTox})\cr the model.}

\item{...}{model specific parameters.}
}
\value{
A \code{function} that computes toxicity probabilities for a given dose
and the model.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

A function that returns a function that computes the toxicity probabilities
for a given dose, model and the model parameters.
}
\examples{
my_model <- LogisticLogNormal(
  mean = c(-0.85, 1),
  cov = matrix(c(1, -0.5, -0.5, 1), nrow = 2),
  ref_dose = 50
)

prob_fun <- probFunction(my_model, alpha0 = 2, alpha1 = 3)
prob_fun(30)
}
\seealso{
\code{\link{prob}}, \code{\link{doseFunction}}.
}
