% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{probit}
\alias{probit}
\title{Shorthand for probit function}
\usage{
probit(x)
}
\arguments{
\item{x}{the function argument}
}
\value{
the probit(x)
}
\description{
Shorthand for probit function
}
\keyword{programming}
