% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations-class.R
\docType{class}
\name{PseudoDualFlexiSimulations-class}
\alias{PseudoDualFlexiSimulations-class}
\alias{.PseudoDualFlexiSimulations}
\alias{PseudoDualFlexiSimulations}
\alias{.DefaultPseudoDualFlexiSimulations}
\title{\code{PseudoDualFlexiSimulations}}
\usage{
PseudoDualFlexiSimulations(sigma2_beta_west, ...)

.DefaultPseudoDualFlexiSimulations()
}
\arguments{
\item{sigma2_beta_west}{(\code{numeric})\cr the vector of the final posterior mean sigma2betaW estimates}

\item{\dots}{additional parameters from \code{\link{PseudoDualSimulations}}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
This class captures the trial simulations design using both the DLE and
efficacy responses using \code{\link{EffFlexi}} efficacy model.
It extends \code{\link{PseudoDualSimulations}} by adding the capability to capture the sigma2betaW estimates.
}
\section{Slots}{

\describe{
\item{\code{sigma2_beta_west}}{(\code{numeric})\cr the vector of the final posterior mean sigma2betaW estimates}
}}

\note{
Typically, end users will not use the \code{.DefaultPseudoFlexiSimulations()} function.
}
