% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Simulations-class.R
\docType{class}
\name{PseudoDualSimulations-class}
\alias{.PseudoDualSimulations}
\alias{PseudoDualSimulations-class}
\title{Class for Pseudo Simulation based on DLE and efficacy pseudo model}
\description{
Class for Pseudo Simulation based on DLE and efficacy pseudo model
}
\keyword{class}

