% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Simulations-class.R
\docType{class}
\name{PseudoDualSimulations-class}
\alias{.PseudoDualSimulations}
\alias{PseudoDualSimulations-class}
\title{This is a class which captures the trial simulations design using both the
DLE and efficacy responses. The design of model from \code{\linkS4class{ModelTox}}
class and the efficacy model from \code{\linkS4class{ModelEff}} class
(except \code{\linkS4class{EffFlexi}} class). It contains all slots from
\code{\linkS4class{GeneralSimulations}} and \code{\linkS4class{PseudoSimulations}} object.
In comparison to the parent class \code{\linkS4class{PseudoSimulations}},
it contains additional slots to
capture the dose-efficacy curve and the sigma2 estimates.}
\description{
This is a class which captures the trial simulations design using both the
DLE and efficacy responses. The design of model from \code{\linkS4class{ModelTox}}
class and the efficacy model from \code{\linkS4class{ModelEff}} class
(except \code{\linkS4class{EffFlexi}} class). It contains all slots from
\code{\linkS4class{GeneralSimulations}} and \code{\linkS4class{PseudoSimulations}} object.
In comparison to the parent class \code{\linkS4class{PseudoSimulations}},
it contains additional slots to
capture the dose-efficacy curve and the sigma2 estimates.
}
\section{Slots}{

\describe{
\item{\code{fitEff}}{list of the final values. If DLE and efficacy samples are gerneated, it contains the
final fitted values. If no DLE and efficacy samples are used, it contains the modal estimates of the
parameters in the two models and the posterior estimates of the probabilities of the occurrence of a
DLE and the expected efficacy responses.}

\item{\code{sigma2est}}{the vector of the final posterior mean sigma2 estimates}
}}
\keyword{class}

