% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Simulations-class.R
\docType{class}
\name{PseudoDualSimulationsSummary-class}
\alias{.PseudoDualSimulationsSummary}
\alias{PseudoDualSimulationsSummary-class}
\title{Class for the summary of the dual responses simulations using pseudo models}
\description{
It contains all slots from \code{\linkS4class{PseudoSimulationsSummary}} object. In addition to
the slots in the parent class \code{\linkS4class{PseudoSimulationsSummary}}, it contains two
more slots for the efficacy model fit information.
}
\details{
Note that objects should not be created by users, therefore no initialization function
is provided for this class.
}
\section{Slots}{

\describe{
\item{\code{EffFitAtDoseMostSelected}}{fitted expected mean efficacy value at dose most often
selected}

\item{\code{meanEffFit}}{list with mean, lower (2.5%) and upper (97.5%) quantiles of the fitted expected
efficacy value at each dose level.}
}}
\keyword{class}

