% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations-class.R
\docType{class}
\name{PseudoDualSimulationsSummary-class}
\alias{PseudoDualSimulationsSummary-class}
\alias{.PseudoDualSimulationsSummary}
\title{Class for the summary of the dual responses simulations using pseudo models}
\description{
It contains all slots from \code{\linkS4class{PseudoSimulationsSummary}} object. In addition to 
the slots in the parent class \code{\linkS4class{PseudoSimulationsSummary}}, it contains four 
more slots for the efficacy model fit information.
}
\details{
Note that objects should not be created by users, therefore no initialization function
is provided for this class.
}
\section{Slots}{

\describe{
\item{\code{targetGstar}}{the target dose level such that its gain value is at maximum}

\item{\code{targetGstarAtDoseGrid}}{the dose level at dose Grid closest and below Gstar}

\item{\code{GstarSummary}}{the six-number table summary (lowest, 25th, 50th (median), 75th percentile, mean 
and highest value) of the final Gstar values obtained across all simulations}

\item{\code{ratioGstarSummary}}{the six-number summary table of the ratios of the upper to the lower 95\%
credibility intervals of the final Gstar across all simulations}

\item{\code{EffFitAtDoseMostSelected}}{fitted expected mean efficacy value at dose most often
selected}

\item{\code{meanEffFit}}{list with mean, lower (2.5%) and upper (97.5%) quantiles of the fitted expected 
efficacy value at each dose level.}
}}

\keyword{class}
