% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Simulations-class.R
\docType{class}
\name{PseudoSimulations-class}
\alias{.PseudoSimulations}
\alias{PseudoSimulations-class}
\title{Class for Pseudo simulation}
\description{
Class for Pseudo simulation
}
\keyword{class}

