% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations-class.R
\name{PseudoSimulations}
\alias{PseudoSimulations}
\title{Initialization function of the 'PseudoSimulations' class}
\usage{
PseudoSimulations(
  fit,
  FinalTDtargetDuringTrialEstimates,
  FinalTDtargetEndOfTrialEstimates,
  FinalTDtargetDuringTrialAtDoseGrid,
  FinalTDtargetEndOfTrialAtDoseGrid,
  FinalTDEOTCIs,
  FinalTDEOTRatios,
  FinalCIs,
  FinalRatios,
  stopReasons,
  ...
)
}
\arguments{
\item{fit}{please refer to \code{\linkS4class{PseudoSimulations}} class object}

\item{FinalTDtargetDuringTrialEstimates}{please refer to \code{\linkS4class{PseudoSimulations}} class object}

\item{FinalTDtargetEndOfTrialEstimates}{please refer to \code{\linkS4class{PseudoSimulations}} class object}

\item{FinalTDtargetDuringTrialAtDoseGrid}{please refer to \code{\linkS4class{PseudoSimulations}} class object}

\item{FinalTDtargetEndOfTrialAtDoseGrid}{please refer to \code{\linkS4class{PseudoSimulations}} class object}

\item{FinalTDEOTCIs}{please refer to \code{\linkS4class{PseudoSimulations}} class object}

\item{FinalTDEOTRatios}{please refer to \code{\linkS4class{PseudoSimulations}} class object}

\item{FinalCIs}{please refer to \code{\linkS4class{PseudoSimulations}} class object}

\item{FinalRatios}{please refer to \code{\linkS4class{PseudoSimulations}} class object}

\item{stopReasons}{please refer to \code{\linkS4class{PseudoSimulations}} class object}

\item{\dots}{additional parameters from \code{\linkS4class{GeneralSimulations}}}
}
\value{
the \code{\linkS4class{PseudoSimulations}} object
}
\description{
Initialization function of the 'PseudoSimulations' class
}
\keyword{methods}
