% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rinvGamma}
\alias{rinvGamma}
\title{The random generation of the Inverse gamma distribution}
\usage{
rinvGamma(n, a, b)
}
\arguments{
\item{n}{the number of observations}

\item{a}{the shape parameter of the inverse gamma distribution}

\item{b}{the scale parameter of the inverse gamm distribution}
}
\description{
The random generation of the Inverse gamma distribution
}
\keyword{internal}
