% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{SafetyWindow-class}
\alias{SafetyWindow-class}
\title{The virtual class for safety window}
\description{
The virtual class for safety window
}
\seealso{
\code{\linkS4class{SafetyWindowSize}},
\code{\linkS4class{SafetyWindowConst}},
}
\keyword{classes}
