% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{SafetyWindowConst-class}
\alias{SafetyWindowConst-class}
\alias{.SafetyWindowConst}
\title{Constant safety window length}
\description{
This class is used when the \code{gap} should be kept constant.
}
\section{Slots}{

\describe{
\item{\code{gap}}{the constant gap between patients.}

\item{\code{follow}}{how long to follow each patient.}

\item{\code{follow_min}}{minimum follow up.}
}}

\examples{
# This is to have along the study constant parameters settings of safety window
# length, regardless of the cohort size.
my_win_len <- SafetyWindowConst(
  gap = c(7, 5, 3),
  follow = 7,
  follow_min = 14
)
}
\keyword{classes}
