% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{SafetyWindowConst-class}
\alias{SafetyWindowConst-class}
\alias{.SafetyWindowConst}
\title{Constant safety window length}
\description{
This class is used when the \code{patientGap} should be kept constant.
}
\section{Slots}{

\describe{
\item{\code{patientGap}}{the constant gap between patients.}

\item{\code{patientFollow}}{how long to follow each patient.}

\item{\code{patientFollowMin}}{minimum follow up.}
}}

\examples{
# This is to have along the study constant parameters settings of safety window length, whatever the cohort size is
myWindowLength <- SafetyWindowConst(patientGap = c(7,5,3),
                                    patientFollow = 7,
                                    patientFollowMin = 14)
}
\keyword{classes}
