% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\name{SafetyWindowConst}
\alias{SafetyWindowConst}
\title{Initialization function for \code{SafetyWindowConst}}
\usage{
SafetyWindowConst(patientGap, patientFollow, patientFollowMin)
}
\arguments{
\item{patientGap}{see \code{\linkS4class{SafetyWindowConst}}}

\item{patientFollow}{see \code{\linkS4class{SafetyWindowConst}}}

\item{patientFollowMin}{see \code{\linkS4class{SafetyWindowConst}}}
}
\value{
the \code{\linkS4class{SafetyWindowConst}} object
}
\description{
Initialization function for \code{SafetyWindowConst}
}
\keyword{methods}
