\name{sampleSize}
\alias{sampleSize}
\title{Compute the number of samples for a given MCMC options triple}
\usage{
sampleSize(mcmcOptions)
}
\arguments{
  \item{mcmcOptions}{the \code{\linkS4class{McmcOptions}}
  object}
}
\value{
the resulting sample size
}
\description{
Compute the number of samples for a given MCMC options
triple
}
\keyword{programming}

