% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McmcOptions-methods.R
\name{sampleSize}
\alias{sampleSize}
\alias{sampleSize,McmcOptions-method}
\alias{sampleSize-McmcOptions-method}
\title{Computing the Number of Samples}
\usage{
sampleSize(object, ...)

\S4method{sampleSize}{McmcOptions}(object, ...)
}
\arguments{
\item{object}{(\code{McmcOptions})\cr object based on which the number of samples
is computed.}

\item{...}{not used.}
}
\value{
A number of samples for a given MCMC options.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

A method that computes the number of samples for a given MCMC options triple.
}
\examples{
# Set up the MCMC option in order to have a burn-in of 10000 iterations and
# then take every other iteration up to a collection of 10000 samples.
my_options <- McmcOptions(burnin = 10000, step = 2, samples = 10000)

sampleSize(my_options)
}
