% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McmcOptions-methods.R
\name{sampleSize}
\alias{sampleSize}
\title{Compute the number of samples for a given MCMC options triple}
\usage{
sampleSize(mcmcOptions)
}
\arguments{
\item{mcmcOptions}{the \code{\linkS4class{McmcOptions}} object}
}
\value{
the resulting sample size
}
\description{
Compute the number of samples for a given MCMC options triple
}
\examples{

# Set up MCMC option in order to have a burn-in of 10000 iterations and
# then take every other iteration up to a collection of 10000 samples
options <- McmcOptions(burnin=10000,
                       step=2,
                       samples=10000)

sampleSize(options)

}
\keyword{programming}
