% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/McmcOptions-methods.R
\name{sampleSize}
\alias{sampleSize}
\title{Compute the number of samples for a given MCMC options triple}
\usage{
sampleSize(mcmcOptions)
}
\arguments{
\item{mcmcOptions}{the \code{\linkS4class{McmcOptions}} object}
}
\value{
the resulting sample size
}
\description{
Compute the number of samples for a given MCMC options triple
}
\keyword{programming}

