% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations-methods.R
\docType{methods}
\name{show,GeneralSimulationsSummary-method}
\alias{show,GeneralSimulationsSummary-method}
\title{Show the summary of the simulations}
\usage{
\S4method{show}{GeneralSimulationsSummary}(object)
}
\arguments{
\item{object}{the \code{\linkS4class{GeneralSimulationsSummary}} object we want
to print}
}
\value{
invisibly returns a data frame of the results with one row and
appropriate column names
}
\description{
Show the summary of the simulations
}
\keyword{methods}
