% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Design-methods.R
\docType{methods}
\name{simulate,DualResponsesDesign-method}
\alias{simulate,DualResponsesDesign-method}
\title{Methods to simulate without any samples based on one Pseudo DLE and one Pseudo Efficacy model}
\usage{
\S4method{simulate}{DualResponsesDesign}(object, nsim = 1L, seed = NULL,
  trueDLE, trueEff, trueNu, args = NULL, firstSeparate = FALSE,
  parallel = FALSE, ...)
}
\description{
Methods to simulate without any samples based on one Pseudo DLE and one Pseudo Efficacy model
}
\keyword{method}

