% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierarchicalDesign-methods.R
\name{simulate,HierarchicalDesign-method}
\alias{simulate,HierarchicalDesign-method}
\title{Simulate outcomes from a hierarchical CRM design}
\usage{
\S4method{simulate}{HierarchicalDesign}(
  object,
  nsim = 1L,
  seed = NULL,
  truth,
  truthResponse = plogis,
  args = NULL,
  firstSeparate = FALSE,
  mcmcOptions = McmcOptions(),
  parallel = FALSE,
  nCores = min(parallel::detectCores(), 5),
  derive = list(),
  ...
)
}
\arguments{
\item{object}{the \code{\link{HierarchicalDesign}} object we want to simulate data from.}

\item{nsim}{(\code{count})\cr the number of simulations.}

\item{seed}{see \code{\link[=set_seed]{set_seed()}}.}

\item{truth}{(\code{function} or named \code{list} of \code{function})\cr true DLT
probability function(s). If a list is supplied, names must match the
hierarchical arms.}

\item{truthResponse}{(\code{function} or named \code{list} of \code{function})\cr true
response probability function(s).}

\item{args}{(\code{data.frame})\cr arguments for the truth functions.}

\item{firstSeparate}{(\code{flag})\cr enroll the first patient separately from the rest of
the cohort? (not default) If yes, the cohort will be closed if a DLT occurs
in this patient.}

\item{mcmcOptions}{(\link{McmcOptions})\cr object of class \code{\link{McmcOptions}},
giving the MCMC options for each evaluation in the trial. By default,
the standard options are used}

\item{parallel}{(\code{flag})\cr should the simulation runs be parallelized across the
clusters of the computer? (not default)}

\item{nCores}{(\code{count})\cr how many cores should be used for parallel computing?
Defaults to the number of cores on the machine, maximum 5.}

\item{derive}{(\code{list})\cr a named list of functions which derives statistics, based on the
vector of posterior MTD samples. Each list element must therefore accept
one and only one argument, which is a numeric vector, and return a number.}

\item{...}{not used}
}
\value{
an object of class \code{\link{HierarchicalSimulations}}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
