% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Design-methods.R
\docType{methods}
\name{simulate,RuleDesign-method}
\alias{simulate,RuleDesign-method}
\title{Simulate outcomes from a rule-based design}
\usage{
\S4method{simulate}{RuleDesign}(object, nsim = 1L, seed = NULL, truth,
  args = NULL, parallel = FALSE, ...)
}
\arguments{
\item{object}{the \code{\linkS4class{RuleDesign}} object we want to simulate
data from}

\item{nsim}{the number of simulations (default: 1)}

\item{seed}{see \code{\link{setSeed}}}

\item{truth}{a function which takes as input a dose (vector) and returns the
true probability (vector) for toxicity. Additional arguments can be supplied
in \code{args}.}

\item{args}{data frame with arguments for the \code{truth} function. The
column names correspond to the argument names, the rows to the values of the
arguments. The rows are appropriately recycled in the \code{nsim}
simulations.}

\item{parallel}{should the simulation runs be parallelized across the
clusters of the computer? (not default)}

\item{\dots}{not used}
}
\value{
an object of class \code{\linkS4class{GeneralSimulations}}
}
\description{
Simulate outcomes from a rule-based design
}
\keyword{methods}

