\docType{class}
\name{Simulations-class}
\alias{Simulations-class}
\title{Class for the simulations output from model based designs}
\description{
This class captures the trial simulations from model based
designs. Additional slots fit and stopReasons compared to
the general class.
}
\section{Slots}{

\describe{
\item{\code{fit}}{list with the final fits}

\item{\code{stopReasons}}{list of stopping reasons for each simulation run}
}}
\keyword{classes}

