\name{size}
\alias{size}
\title{Determine the size of the next cohort}
\usage{
size(cohortSize, dose, data, ...)
}
\arguments{
  \item{cohortSize}{The rule, an object of class
  \code{\linkS4class{CohortSize}}}

  \item{dose}{the next dose}

  \item{data}{The data input, an object of class
  \code{\linkS4class{Data}}}

  \item{\dots}{additional arguments}
}
\value{
the size as integer value
}
\description{
This function determines the size of the next cohort.
}
\keyword{methods}

