\docType{class}
\name{Stopping-class}
\alias{Stopping-class}
\title{The virtual class for stopping rules}
\description{
The virtual class for stopping rules
}
\seealso{
\code{\linkS4class{StoppingList}},
\code{\linkS4class{StoppingMaxPatients}},
\code{\linkS4class{StoppingCohortsNearDose}},
\code{\linkS4class{StoppingPatientsNearDose}},
\code{\linkS4class{StoppingMinCohorts}},
\code{\linkS4class{StoppingTargetProb}}
\code{\linkS4class{StoppingMTDdistribution}},
\code{\linkS4class{StoppingTargetBiomarker}}
}
\keyword{classes}

