% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{Stopping-class}
\alias{Stopping-class}
\alias{Stopping}
\title{\code{Stopping}}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{\link{Stopping}} is a class for for stopping rules.
}
\seealso{
\code{\link{StoppingList}}, \code{\link{StoppingCohortsNearDose}}, \code{\link{StoppingPatientsNearDose}},
\code{\link{StoppingMinCohorts}}, \code{\link{StoppingMinPatients}}, \code{\link{StoppingTargetProb}},
\code{\link{StoppingMTDdistribution}}, \code{\link{StoppingTargetBiomarker}}, \code{\link{StoppingHighestDose}}
\code{\link{StoppingMTDCV}}, \code{\link{StoppingLowestDoseHSRBeta}}.
}
