% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{StoppingAny-class}
\alias{.StoppingAny}
\alias{StoppingAny-class}
\title{Stop based on fullfillment of any stopping rule}
\description{
This class can be used to combine multiple stopping rules with an OR
operator.
}
\details{
\code{stopList} contains all stopping rules, which are again objects of
class \code{\linkS4class{Stopping}}. Any of these rules must be fulfilled in
order that the result of this rule is to stop.
}
\section{Slots}{

\describe{
\item{\code{stopList}}{list of stopping rules}
}}
\keyword{classes}

