% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{StoppingCohortsNearDose-class}
\alias{.StoppingCohortsNearDose}
\alias{StoppingCohortsNearDose-class}
\title{Stop based on number of cohorts near to next best dose}
\description{
Stop based on number of cohorts near to next best dose
}
\section{Slots}{

\describe{
\item{\code{nCohorts}}{number of required cohorts}

\item{\code{percentage}}{percentage (between 0 and 100) within the next best dose
the cohorts must lie}
}}
\keyword{classes}

