% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{StoppingCohortsNearDose-class}
\alias{StoppingCohortsNearDose-class}
\alias{.StoppingCohortsNearDose}
\alias{StoppingCohortsNearDose}
\title{\code{StoppingCohortsNearDose}}
\usage{
StoppingCohortsNearDose(nCohorts = 2L, percentage = 50)
}
\arguments{
\item{nCohorts}{(\code{number})\cr see slot definition.}

\item{percentage}{(\code{number})\cr see slot definition.}
}
\description{
Stop based on number of cohorts near to next best dose
}
\section{Slots}{

\describe{
\item{\code{nCohorts}}{number of required cohorts}

\item{\code{percentage}}{(\code{number})\cr percentage (between and including 0 and 100)
within the next best dose the cohorts must lie.}
}}

\examples{

# As example, here is the rule for: 
#   stopping the study if at least 3 cohorts were dosed at a dose within (1 +/- 0.2) 
#   of the next best dose

myStopping <- StoppingCohortsNearDose(nCohorts = 3,
                                      percentage = 0.2)

}
\keyword{classes}
