% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{StoppingDoseStabilized-class}
\alias{StoppingDoseStabilized-class}
\alias{.StoppingDoseStabilized}
\alias{StoppingDoseStabilized}
\alias{.DefaultStoppingDoseStabilized}
\title{\code{StoppingDoseStabilized}}
\usage{
StoppingDoseStabilized(nCohorts = 1L, report_label = NA_character_)

.DefaultStoppingDoseStabilized()
}
\arguments{
\item{nCohorts}{(\code{number})\cr see slot definition.}

\item{report_label}{(\code{string} or \code{NA})\cr see slot definition.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{\link{StoppingDoseStabilized}} is the class for stopping when the next best dose
has stabilized. Stabilization occurs when the current next best dose is the
same as the dose administered to each of the most recent consecutive
\code{nCohorts} cohorts.

For combination trials, both dose components must be the same.
}
\section{Slots}{

\describe{
\item{\code{nCohorts}}{(\code{number})\cr number of consecutive cohorts required to have
received the current next best dose.}
}}

\note{
Typically, end users will not use the
\code{.DefaultStoppingDoseStabilized()} function.
}
\examples{
# Stop when the next best dose has remained unchanged for three consecutive
# cohorts.
my_stopping <- StoppingDoseStabilized(nCohorts = 3)
}
