% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{StoppingMinCohorts-class}
\alias{StoppingMinCohorts-class}
\alias{.StoppingMinCohorts}
\alias{StoppingMinCohorts}
\alias{.DefaultStoppingMinCohorts}
\title{\code{StoppingMinCohorts}}
\usage{
StoppingMinCohorts(nCohorts)

.DefaultStoppingMinCohorts()
}
\arguments{
\item{nCohorts}{(\code{number})\cr see slot definition.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{\link{StoppingMinCohorts}} is the class for stopping based on minimum number of
cohorts.

Typically, end users will not use the \code{.DefaultStoppingMinCohorts()} function.
}
\section{Slots}{

\describe{
\item{\code{nCohorts}}{(\code{number})\cr minimum required number of cohorts.}
}}

\examples{

# As example, here is the rule for stopping the study if at least 6 cohorts
# were already dosed.
my_stopping <- StoppingMinCohorts(nCohorts = 6)
.DefaultStoppingMinCohorts()
}
