% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{StoppingMinCohorts-class}
\alias{.StoppingMinCohorts}
\alias{StoppingMinCohorts-class}
\title{Stop based on minimum number of cohorts}
\description{
Stop based on minimum number of cohorts
}
\section{Slots}{

\describe{
\item{\code{nCohorts}}{minimum required number of cohorts}
}}
\examples{

# As example, here is the rule for: 
#   stopping the study if at least 6 cohorts were already dosed

myStopping <- StoppingMinCohorts(nCohorts = 6)

}
\keyword{classes}

