% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{StoppingMinPatients-class}
\alias{.StoppingMinPatients}
\alias{StoppingMinPatients-class}
\title{Stop based on minimum number of patients}
\description{
Stop based on minimum number of patients
}
\section{Slots}{

\describe{
\item{\code{nPatients}}{minimum allowed number of patients}
}}
\keyword{classes}

