% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{StoppingMinPatients-class}
\alias{StoppingMinPatients-class}
\alias{.StoppingMinPatients}
\alias{StoppingMinPatients}
\alias{.DefaultStoppingMinPatients}
\title{\code{StoppingMinPatients}}
\usage{
StoppingMinPatients(nPatients = 20L, report_label = NA_character_)

.DefaultStoppingMinPatients()
}
\arguments{
\item{nPatients}{(\code{number})\cr see slot definition.}

\item{report_label}{(\code{string} or \code{NA})\cr see slot definition.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{\link{StoppingMinPatients}} is the class for stopping based on minimum number of
patients
}
\section{Slots}{

\describe{
\item{\code{nPatients}}{(\code{number})\cr minimum allowed number of patients.}
}}

\note{
Typically, end users will not use the \code{.DefaultStoppingMinPatients()} function.
}
\examples{

# As example, here is the rule for stopping the study if at least 20 patients
# were already dosed
my_stopping <- StoppingMinPatients(nPatients = 20)
}
