% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{StoppingMinPatients-class}
\alias{StoppingMinPatients-class}
\alias{.StoppingMinPatients}
\title{Stop based on minimum number of patients}
\description{
Stop based on minimum number of patients
}
\section{Slots}{

\describe{
\item{\code{nPatients}}{minimum allowed number of patients}
}}

\examples{

# As example, here is the rule for: 
#   stopping the study if at least 20 patients were already dosed

myStopping <- StoppingMinPatients(nPatients = 20)

}
\keyword{classes}
