% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\name{StoppingMinPatients}
\alias{StoppingMinPatients}
\title{Initialization function for "StoppingMinPatients"}
\usage{
StoppingMinPatients(nPatients)
}
\arguments{
\item{nPatients}{see \code{\linkS4class{StoppingMinPatients}}}
}
\value{
the \code{\linkS4class{StoppingMinPatients}} object
}
\description{
Initialization function for "StoppingMinPatients"
}
\keyword{methods}
