% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{StoppingMTDCV-class}
\alias{StoppingMTDCV-class}
\alias{.StoppingMTDCV}
\alias{StoppingMTDCV}
\title{\code{StoppingMTDCV}}
\usage{
StoppingMTDCV(target = 0.3, thresh_cv = 40)
}
\arguments{
\item{target}{(\code{proportion})\cr see slot definition.}

\item{thresh_cv}{(\code{number})\cr see slot definition.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\code{\link{StoppingMTDCV}} is a class for stopping rule based on precision of MTD
which is calculated as the coefficient of variation (CV) of the MTD.
}
\section{Slots}{

\describe{
\item{\code{target}}{(\code{proportion})\cr toxicity target of MTD.}

\item{\code{thresh_cv}}{(\code{number})\cr threshold for CV to be considered accurate enough
to stop the trial.}
}}

\examples{
# Stopping the study if the MTD estimation is precise enough, i.e. if robust
# coefficient of variation of the MTD is below 40\%.
my_stopping <- StoppingMTDCV(target = 0.3, thresh_cv = 40)
}
