% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{StoppingMTDdistribution-class}
\alias{StoppingMTDdistribution-class}
\alias{.StoppingMTDdistribution}
\alias{StoppingMTDdistribution}
\title{\code{StoppingMTDdistribution}}
\usage{
StoppingMTDdistribution(target, thresh, prob)
}
\arguments{
\item{target}{(\code{proportion})\cr see slot definition in \code{\link{StoppingMTDdistribution}}.}

\item{thresh}{(\code{proportion})\cr see slot definition in \code{\link{StoppingMTDdistribution}}.}

\item{prob}{(\code{proportion})\cr see slot definition in \code{\link{StoppingMTDdistribution}}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{\link{StoppingMTDdistribution}} is the class for stopping based on MTD distribution.
It is used for the cases where the stopping occurs when there exists
\code{prob} probability above a threshold of \code{tresh} of the current MTD been
reached. Here, the MTD is defined as the dose for which \code{prob(DLT) = target}.
}
\section{Slots}{

\describe{
\item{\code{target}}{(\code{proportion})\cr the target toxicity probability defining the MTD.}

\item{\code{thresh}}{(\code{proportion})\cr the threshold relative to the MTD.}

\item{\code{prob}}{(\code{proportion})\cr required probability.}
}}

\examples{

# As example, here is the rule for stopping the study if there is at least 0.9
# probability that MTD > 0.5 * next_dose. Here MTD is defined as the dose for
# which prob(DLT) = 0.33
my_stopping <- StoppingMTDdistribution(
  target = 0.33,
  thresh = 0.5,
  prob = 0.9
)
}
