\docType{class}
\name{StoppingPatientsNearDose-class}
\alias{StoppingPatientsNearDose-class}
\title{Stop based on number of patients near to next best dose}
\description{
Stop based on number of patients near to next best dose
}
\section{Slots}{

\describe{
\item{\code{nPatients}}{number of required patients}

\item{\code{percentage}}{percentage (between 0 and 100) within the next best dose
the patients must lie}
}}
\keyword{classes}

