% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Rules-class.R
\docType{class}
\name{StoppingTargetBiomarker-class}
\alias{.StoppingTargetBiomarker}
\alias{StoppingTargetBiomarker-class}
\title{Stop based on probability of target biomarker}
\description{
Stop based on probability of target biomarker
}
\section{Slots}{

\describe{
\item{\code{target}}{the biomarker level, relative to the maximum, that needs to be
reached. So this must be a probability (1 is allowed here)}

\item{\code{prob}}{required target probability for reaching sufficient precision}
}}
\keyword{classes}

