\docType{class}
\name{StoppingTargetProb-class}
\alias{StoppingTargetProb-class}
\title{Stop based on probability of target tox interval}
\description{
Stop based on probability of target tox interval
}
\section{Slots}{

\describe{
\item{\code{target}}{the target toxicity interval}

\item{\code{prob}}{required target toxicity probability
for reaching sufficient precision}
}}
\keyword{classes}

