% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rules-methods.R
\docType{methods}
\name{stopTrial}
\alias{stopTrial}
\alias{stopTrial,StoppingAll,ANY,ANY,ANY,ANY-method}
\alias{stopTrial,StoppingAny,ANY,ANY,ANY,ANY-method}
\alias{stopTrial,StoppingCohortsNearDose,numeric,ANY,ANY,Data-method}
\alias{stopTrial,StoppingList,ANY,ANY,ANY,ANY-method}
\alias{stopTrial,StoppingMTDdistribution,numeric,Samples,Model,ANY-method}
\alias{stopTrial,StoppingMinCohorts,ANY,ANY,ANY,Data-method}
\alias{stopTrial,StoppingMinPatients,ANY,ANY,ANY,Data-method}
\alias{stopTrial,StoppingPatientsNearDose,numeric,ANY,ANY,Data-method}
\alias{stopTrial,StoppingTargetBiomarker,numeric,Samples,DualEndpoint,ANY-method}
\alias{stopTrial,StoppingTargetProb,numeric,Samples,Model,ANY-method}
\title{Stop the trial?}
\usage{
stopTrial(stopping, dose, samples, model, data, ...)

\S4method{stopTrial}{StoppingList,ANY,ANY,ANY,ANY}(stopping, dose, samples,
  model, data, ...)

\S4method{stopTrial}{StoppingAll,ANY,ANY,ANY,ANY}(stopping, dose, samples,
  model, data, ...)

\S4method{stopTrial}{StoppingAny,ANY,ANY,ANY,ANY}(stopping, dose, samples,
  model, data, ...)

\S4method{stopTrial}{StoppingCohortsNearDose,numeric,ANY,ANY,Data}(stopping,
  dose, samples, model, data, ...)

\S4method{stopTrial}{StoppingPatientsNearDose,numeric,ANY,ANY,Data}(stopping,
  dose, samples, model, data, ...)

\S4method{stopTrial}{StoppingMinCohorts,ANY,ANY,ANY,Data}(stopping, dose,
  samples, model, data, ...)

\S4method{stopTrial}{StoppingMinPatients,ANY,ANY,ANY,Data}(stopping, dose,
  samples, model, data, ...)

\S4method{stopTrial}{StoppingTargetProb,numeric,Samples,Model,ANY}(stopping,
  dose, samples, model, data, ...)


  \S4method{stopTrial}{StoppingMTDdistribution,numeric,Samples,Model,ANY}(stopping,
  dose, samples, model, data, ...)


  \S4method{stopTrial}{StoppingTargetBiomarker,numeric,Samples,DualEndpoint,ANY}(stopping,
  dose, samples, model, data, ...)
}
\arguments{
\item{stopping}{The rule, an object of class
\code{\linkS4class{Stopping}}}

\item{dose}{the recommended next best dose}

\item{samples}{the \code{\linkS4class{Samples}} object}

\item{model}{The model input, an object of class \code{\linkS4class{Model}}}

\item{data}{The data input, an object of class \code{\linkS4class{Data}}}

\item{\dots}{additional arguments}
}
\value{
logical value: \code{TRUE} if the trial can be stopped, \code{FALSE}
otherwise. It should have an attribute \code{message} which gives the reason
for the decision.
}
\description{
This function returns whether to stop the trial.
}
\section{Methods (by class)}{
\itemize{
\item \code{stopping = StoppingList,dose = ANY,samples = ANY,model = ANY,data = ANY}: Stop based on multiple stopping rules

\item \code{stopping = StoppingAll,dose = ANY,samples = ANY,model = ANY,data = ANY}: Stop based on fulfillment of all multiple stopping
rules

\item \code{stopping = StoppingAny,dose = ANY,samples = ANY,model = ANY,data = ANY}: Stop based on fulfillment of any stopping rule

\item \code{stopping = StoppingCohortsNearDose,dose = numeric,samples = ANY,model = ANY,data = Data}: Stop based on number of cohorts near to next best dose

\item \code{stopping = StoppingPatientsNearDose,dose = numeric,samples = ANY,model = ANY,data = Data}: Stop based on number of patients near to next best
dose

\item \code{stopping = StoppingMinCohorts,dose = ANY,samples = ANY,model = ANY,data = Data}: Stop based on minimum number of cohorts

\item \code{stopping = StoppingMinPatients,dose = ANY,samples = ANY,model = ANY,data = Data}: Stop based on minimum number of patients

\item \code{stopping = StoppingTargetProb,dose = numeric,samples = Samples,model = Model,data = ANY}: Stop based on probability of target tox interval

\item \code{stopping = StoppingMTDdistribution,dose = numeric,samples = Samples,model = Model,data = ANY}: Stop based on MTD distribution

\item \code{stopping = StoppingTargetBiomarker,dose = numeric,samples = Samples,model = DualEndpoint,data = ANY}: Stop based on probability of targeting biomarker
}}
\keyword{methods}

