% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Simulations-methods.R
\docType{methods}
\name{summary,Simulations-method}
\alias{summary,Simulations-method}
\title{Summarize the model-based design simulations, relative to a given truth}
\usage{
\S4method{summary}{Simulations}(object, truth, target = c(0.2, 0.35), ...)
}
\arguments{
\item{object}{the \code{\linkS4class{Simulations}} object we want to
summarize}

\item{truth}{a function which takes as input a dose (vector) and returns the
true probability (vector) for toxicity}

\item{target}{the target toxicity interval (default: 20-35\%) used for the
computations}

\item{\dots}{Additional arguments can be supplied here for \code{truth}}
}
\value{
an object of class \code{\linkS4class{SimulationsSummary}}
}
\description{
Summarize the model-based design simulations, relative to a given truth
}
\keyword{methods}

