% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Design-class.R
\docType{class}
\name{TDDesign-class}
\alias{.TDDesign}
\alias{TDDesign-class}
\title{Design class using DLE responses only without samples}
\description{
Design class using DLE responses only without samples
}
\keyword{class}

