% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design-class.R
\name{TDDesign}
\alias{TDDesign}
\title{Initialization function for 'TDDesign' class}
\usage{
TDDesign(model, stopping, increments, PLcohortSize = CohortSizeConst(0L), ...)
}
\arguments{
\item{model}{please refer to \code{\linkS4class{TDDesign}} class object}

\item{stopping}{please refer to \code{\linkS4class{TDDesign}} class object}

\item{increments}{please refer to \code{\linkS4class{TDDesign}} class object}

\item{PLcohortSize}{see \code{\linkS4class{TDDesign}}}

\item{\dots}{additional arguments for \code{\linkS4class{RuleDesign}}}
}
\value{
the \code{\linkS4class{TDDesign}} class object
}
\description{
Initialization function for 'TDDesign' class
}
\keyword{methods}
