% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design-class.R
\docType{class}
\name{TDsamplesDesign-class}
\alias{TDsamplesDesign-class}
\alias{.TDsamplesDesign}
\alias{TDsamplesDesign}
\title{\code{TDsamplesDesign}}
\usage{
TDsamplesDesign(
  model,
  stopping,
  increments,
  pl_cohort_size = CohortSizeConst(0L),
  ...
)
}
\arguments{
\item{model}{(\code{ModelTox})\cr see slot definition.}

\item{stopping}{(\code{Stopping})\cr see slot definition.}

\item{increments}{(\code{Increments})\cr see slot definition.}

\item{pl_cohort_size}{(\code{CohortSize})\cr see slot definition.}

\item{...}{
  Arguments passed on to \code{\link[=RuleDesign]{RuleDesign}}
  \describe{
    \item{\code{nextBest}}{(\code{NextBest})\cr see slot definition.}
    \item{\code{cohortSize}}{(\code{CohortSize})\cr see slot definition.}
    \item{\code{data}}{(\code{Data})\cr see slot definition.}
    \item{\code{startingDose}}{(\code{number})\cr see slot definition.}
  }}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\code{\link{TDsamplesDesign}} is the class of design based only on DLT responses using
\code{\link{ModelTox}} class model (i.e. \code{\link{LogisticIndepBeta}}) as well as MCMC samples
obtained for this model.
}
\section{Slots}{

\describe{
\item{\code{model}}{(\code{ModelTox})\cr the pseudo DLT model to be used.}

\item{\code{stopping}}{(\code{Stopping})\cr stopping rule(s) for the trial.}

\item{\code{increments}}{(\code{Increments})\cr how to control increments between dose levels.}

\item{\code{pl_cohort_size}}{(\code{CohortSize})\cr rules for the cohort sizes for placebo,
if any planned (defaults to constant 0 placebo patients).}
}}

