% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrmPackClass-methods.R, R/Rules-methods.R,
%   R/Samples-methods.R
\name{tidy}
\alias{tidy}
\alias{tidy,CrmPackClass-method}
\alias{tidy-CrmPackClass}
\alias{tidy,IncrementsRelative-method}
\alias{tidy-IncrementsRelative}
\alias{tidy,CohortSizeDLT-method}
\alias{tidy-CohortSizeDLT}
\alias{tidy,CohortSizeRange-method}
\alias{tidy-CohortSizeRange}
\alias{tidy,CohortSizeParts-method}
\alias{tidy-CohortSizeParts}
\alias{tidy,NextBestNCRM-method}
\alias{tidy-NextBestNCRM}
\alias{tidy,Samples-method}
\alias{tidy-Samples}
\title{Tidying \code{CrmPackClass} objects}
\usage{
tidy(x, ...)

\S4method{tidy}{CrmPackClass}(x, ...)

\S4method{tidy}{IncrementsRelative}(x, ...)

\S4method{tidy}{CohortSizeDLT}(x, ...)

\S4method{tidy}{CohortSizeRange}(x, ...)

\S4method{tidy}{CohortSizeParts}(x, ...)

\S4method{tidy}{IncrementsRelative}(x, ...)

\S4method{tidy}{NextBestNCRM}(x, ...)

\S4method{tidy}{Samples}(x, ...)
}
\arguments{
\item{x}{(\code{CrmPackClass})\cr The object to be tidied}

\item{...}{\cr Not used at present}
}
\value{
A (list of) tibble(s) represnting the object in tidy form.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

In the spirit of the \code{broom} package, provide a method to convert a
\code{CrmPackClass} object to a (list of) tibble9s).

Following the pronciples of the \code{broom} package, convert a \code{CrmPackClass}
object to a (list of) tibbles.  This is a basic, default representation
}
\examples{
CohortSizeConst(3) |> tidy()
CohortSizeRange(intervals=c(0, 20), cohort_size =c(1,3)) |> tidy()
CohortSizeParts(cohort_sizes = c(1, 3)) |> tidy()
CohortSizeRange(intervals=c(0, 20), cohort_size =c(1,3)) |> tidy()
NextBestNCRM(
  target = c(0.2, 0.35),
  overdose = c(0.35, 1),
  max_overdose_prob = 0.25
) |> tidy()
}
\keyword{methods}
