% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model-class.R
\docType{class}
\name{TITELogisticLogNormal-class}
\alias{TITELogisticLogNormal-class}
\alias{.TITELogisticLogNormal}
\title{Standard logistic model with bivariate (log) normal prior (for TITE-CRM use)}
\description{
This is a TITE-CRM based on a logistic regression model using
a bivariate normal prior on the intercept and log slope parameters.
This class inherits from the normal logistic model class.
}
\section{Slots}{

\describe{
\item{\code{weightMethod}}{weight function method: linear or addaptive
(this was used in Liu, Yin and Yuan's paper)}
}}

\keyword{classes}
