% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-methods.R
\name{update,Data-method}
\alias{update,Data-method}
\alias{update-Data-method}
\title{Update method for the \code{\link{Data}} class.}
\usage{
\S4method{update}{Data}(
  object,
  x,
  y,
  ID = length(object@ID) + seq_along(y),
  new_cohort = TRUE,
  ...
)
}
\arguments{
\item{object}{(\code{Data})\cr object we want to update.}

\item{x}{(\code{numeric})\cr the dose level (one level only!).}

\item{y}{(\code{integer})\cr the DLT vector (0/1 vector) for all patients in this cohort.}

\item{ID}{the patient IDs.}

\item{new_cohort}{(\code{flag})\cr if \code{TRUE} (default) the new data are assigned to a new cohort.}

\item{\dots}{not used.}
}
\value{
The new, updated \code{\link{Date}} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

A method that updates existing \code{\link{Data}} object with new data.
}
\examples{
# Create some data of class 'Data'.
my_data <- Data(
  x = c(0.1, 0.5, 1.5, 3, 6, 10, 10, 10),
  y = c(0, 0, 0, 0, 0, 0, 1, 0),
  doseGrid = c(0.1, 0.5, 1.5, 3, 6, seq(from = 10, to = 80, by = 2))
)

# Update the data with a new cohort.
my_data1 <- update(my_data, x = 20, y = c(0L, 1L, 1L))
}
