% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-methods.R
\name{update,DataDA-method}
\alias{update,DataDA-method}
\title{Update method for the "DataDA" class}
\usage{
\S4method{update}{DataDA}(
  object,
  factDLTs,
  factSurv,
  factT0,
  thisDose,
  trialtime,
  ID = NULL,
  newCohort = TRUE,
  ...
)
}
\arguments{
\item{object}{the old \code{\linkS4class{DataDA}} object}

\item{factDLTs}{the simulated DLT outcome for all patients}

\item{factSurv}{the simulated DLT time for all patients}

\item{factT0}{the time that patients start DLT observation window}

\item{thisDose}{the current dose of the cohort}

\item{ID}{the patient IDs}

\item{newCohort}{logical: if TRUE (default) the new data are assigned
to a new cohort}

\item{\dots}{not used}
}
\value{
the new \code{\linkS4class{DataDA}} object
}
\description{
Update observations in the \code{\linkS4class{DataDA}} object
}
\examples{
thisData <- DataDA(x=c(0.1, 0.5, 1.5, 3, 6, 10, 10, 10),
                   y=c(0, 0, 1, 1, 0, 0, 1, 0),
                   doseGrid=
                     c(0.1, 0.5, 1.5, 3, 6,
                       seq(from=10, to=80, by=2)),
                   u=c(42,30,15,5,20,25,30,60),
                   t0=c(0,-15,-30,-40,-55,-70,-75,-85),
                   Tmax=60)


## todo: update this example with new function call structure 
factDLTs <- thisData@y
factSurv <- thisData@u
factT0   <- thisData@t0
thisT0   <- 5
thisDose <- 20
thisDLTs <- 0
thisSurv <- 20

tempData <- update(object=thisData,
                   factDLTs= c(factDLTs,thisDLTs),  ####the y will be updated according to u 
                   factSurv= c(factSurv,thisSurv),
                   factT0= c(factT0,thisT0),
                   thisDose= thisDose,
                   trialtime= 8)  ####this is the globle timeline for a trial 
}
\keyword{methods}
