% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Data-methods.R
\docType{methods}
\name{update,DataDual-method}
\alias{update,DataDual-method}
\title{Update method for the "DataDual" class}
\usage{
\S4method{update}{DataDual}(object, x, y, w, ID = (if (length(object@ID))
  max(object@ID) else 0L) + seq_along(y), ...)
}
\arguments{
\item{object}{the old \code{\linkS4class{DataDual}} object}

\item{x}{the dose level (one level only!)}

\item{y}{the DLT vector (0/1 vector), for all patients in this cohort}

\item{w}{the biomarker vector, for all patients in this cohort}

\item{ID}{the patient IDs}

\item{\dots}{not used}
}
\value{
the new \code{\linkS4class{DataDual}} object
}
\description{
Add new data to the \code{\linkS4class{DataDual}} object
}
\keyword{methods}

