% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rules-validity.R
\name{v_next_best}
\alias{v_next_best}
\alias{v_next_best_mtd}
\title{Internal Helper Functions for Validation of \code{\link{NextBest}} Objects}
\usage{
v_next_best_mtd(object)
}
\arguments{
\item{object}{(\code{NextBest})\cr object to validate.}
}
\value{
A \code{character} vector with the validation failure messages,
or \code{TRUE} in case validation passes.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

These functions are only used internally to validate the format of an input
\code{\link{NextBest}} or inherited classes and therefore not exported.
}
\section{Functions}{
\itemize{
\item \code{v_next_best_mtd}: validates that the \code{\link{NextBestMTD}} object
contains valid \code{target} probability and \code{derive} function.
}}

