% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulations-validity.R, R/Simulation-validity.R
\name{v_pseudo_simulations}
\alias{v_pseudo_simulations}
\alias{v_pseudo_dual_simulations}
\alias{v_pseudo_dual_flex_simulations}
\alias{v_da_simulations}
\title{Internal Helper Functions for Validation of \code{\link{PseudoSimulations}} Objects}
\usage{
v_pseudo_simulations(object)

v_pseudo_dual_simulations(object)

v_pseudo_dual_flex_simulations(object)

v_pseudo_simulations(object)

v_pseudo_dual_simulations(object)

v_pseudo_dual_flex_simulations(object)

v_da_simulations(object)
}
\arguments{
\item{object}{(\code{PseudoSimulations})\cr object to validate.}
}
\value{
A \code{character} vector with the validation failure messages,
or \code{TRUE} in case validation passes.

A \code{character} vector with the validation failure messages,
or \code{TRUE} in case validation passes.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

These functions are only used internally to validate the format of an input
\code{\link{PseudoSimulations}} or inherited classes and therefore not exported.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

These functions are only used internally to validate the format of an input
\code{\link{PseudoSimulations}} or inherited classes and therefore not exported.
}
\section{Functions}{
\itemize{
\item \code{v_pseudo_simulations()}: validates that the \code{\link{PseudoSimulations}} object
contains valid \code{fit}, \code{FinalTDtargetEndOfTrialEstimates} , \code{FinalTDtargetDuringTrialAtDoseGrid},
\code{FinalTDtargetEndOfTrialAtDoseGrid} , \code{FinalTDEOTCIs}, \code{FinalTDEOTRatios}, \code{FinalCIs}, \code{FinalRatios},
object and valid \code{stopReasons} simulations.

\item \code{v_pseudo_dual_simulations()}: validates that the \code{\link{PseudoDualSimulations}} object
contains valid \code{fitEff}, \code{FinalGstarEstimates} , \code{FinalGstarAtDoseGrid},
\code{FinalGstarCIs} , \code{FinalGstarRatios}, \code{FinalOptimalDose}, \code{FinalOptimalDoseAtDoseGrid}
object and valid \code{sigma2est} simulations.

\item \code{v_pseudo_dual_flex_simulations()}: validates that the \code{\link{PseudoDualFlexiSimulations}}
object contains valid \code{sigma2betaWest} vector of the final posterior mean
sigma2betaW estimates.\code{FinalGstarEstimates} , \code{FinalGstarAtDoseGrid},

\item \code{v_pseudo_simulations()}: validates that the \code{\link{PseudoSimulations}} object
contains valid \code{fit}, \code{FinalTDtargetEndOfTrialEstimates} , \code{FinalTDtargetDuringTrialAtDoseGrid},
\code{FinalTDtargetEndOfTrialAtDoseGrid} , \code{FinalTDEOTCIs}, \code{FinalTDEOTRatios}, \code{FinalCIs}, \code{FinalRatios},
object and valid \code{stopReasons} simulations.

\item \code{v_pseudo_dual_simulations()}: validates that the \code{\link{PseudoDualSimulations}} object
contains valid \code{fitEff}, \code{FinalGstarEstimates} , \code{FinalGstarAtDoseGrid},
\code{FinalGstarCIs} , \code{FinalGstarRatios}, \code{FinalOptimalDose}, \code{FinalOptimalDoseAtDoseGrid}
object and valid \code{sigma2est} simulations.

\item \code{v_pseudo_dual_flex_simulations()}: validates that the \code{\link{PseudoDualFlexiSimulations}}
object contains valid \code{sigma2betaWest} vector of the final posterior mean sigma2betaW
estimates.\code{FinalGstarEstimates} , \code{FinalGstarAtDoseGrid},

\item \code{v_da_simulations()}: validates that the \code{\link{DASimulations}} object
contains valid \code{trialduration} the vector of trial duration values for all simulations.

}}
