% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\docType{class}
\name{Validate}
\alias{Validate}
\title{\code{Validate}}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

The \code{\link{Validate}} class is a Reference Class
to help programming validation for new S4 classes.
}
\details{
Starting from an empty \code{msg} vector, with each check
that is returning \code{FALSE} the vector gets a new element - the string
explaining the failure of the validation.
}
\section{Fields}{

\describe{
\item{\code{msg}}{the message character vector}
}}


