% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-validity.R
\name{validate_data}
\alias{validate_data}
\alias{validate_subjects}
\alias{h_is_dose_unique}
\title{Internal Helper Functions for Validation of \code{GeneralData} Objects.}
\usage{
validate_subjects(object)

h_is_dose_unique(dose, cohort)

validate_data(object)
}
\arguments{
\item{object}{(\code{GeneralData})\cr object to validate.}

\item{dose}{(\code{numeric})\cr vector of doses.}

\item{cohort}{(\code{integer})\cr vector of cohorts corresponding to \code{doses}.}
}
\value{
A character vector with the validation failure messages,
or \code{TRUE} in case validation passes.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

These functions are only used internally to validate the format of an input
\code{\linkS4class{GeneralData}} and \code{\linkS4class{Data}} object and therefore not exported.
}
\section{Functions}{
\itemize{
\item \code{validate_subjects}: validates that the \code{\linkS4class{GeneralData}} object contains
unique \code{ID}, non-negative \code{cohort} indices and
\code{ID} and \code{cohort} vectors are of the same length equal to \code{nObs}.

\item \code{h_is_dose_unique}: helper function.
It returns \code{TRUE} if \code{dose} is unique per \code{cohort}.

\item \code{validate_data}: validates that the \code{\linkS4class{Data}} object contains
valid elements with respect to their types, dependency and length.
}}

